Update golden_test.sh to be aware of RELEASE_AIDL_USE_UNFROZEN
The script now reads the current value of the build flag from source.
For "check" operation, it uses that value to determine which
golden_output files to check against.
For "update" operation, it uses the value to determin which
golden_output files to update, then it modifies the value of the
RELEASE_AIDL_USE_UNFROZEN flag and reruns it. It will kick off the build
again to build the interfaces with this new value. This can take
multiple minutes because it needs to regenrate the build files. It then
reset the flag value to the previous value, which causes the next build
to regenerate the build files again.

Test: m aidl_unittests
Test: golden_test.sh check
Test: golden_test.sh update
Test: modify some golden files and rerun both check/update
Bug: 299364399

Change-Id: Ieb2bbbe21b1489f5094d1c1c14a68e08266a49c7