linaro_swr: A generic build target using software rendering
linaro_swr is an attempt to prepare a generic arm64 build
target booting with software rendering support SwANGLE
i.e. GLES implementation (ANGLE) on top of SwiftShader’s
vulkan (Pastel) implementation.

Though it is developed and tested for Qcom target devices
for now, but the plan is to make it generic enough to boot
on any generic arm64 device.

Change-Id: I135b4d2148330368ff298cb5d4bd290a67a94422
Signed-off-by: Amit Pundir <amit.pundir@linaro.org>