Update build to pick up sources from `metalava` dir to match upstream
The sources in `metalava` directory are built into the `metalava-main`
library (instead of `metalava`) because `metalava` has to be built in
`/tools/metalava` because some Bazel rules hard code its path.

Due to the issues described in b/300447782 some builds were left in a
state that broke incremental builds so it was necessary to add a
`CleanSpec.mk` file to clean those files up.

Bug: 300224510
Bug: 300447782
Test: m checkapi
Test: m metalava && out/host/linux-x86/bin/metalava --version
Test: b run //tools/metalava/metalava:metalava --version
Change-Id: I570940bf58ac6f9ea90f03d8d7a2bc3e3a4d3d61