gfxstream: meson qnx cross build
Pkgconfig wasn't able to resolve the dependencies for qnx build, so the
libraries and headers are found and added manually. Hopefully we can get
pkgconfig working soon to clean this up.

A `qnx` directory is added to the project root for qnx build scripts and
cross files. A script also handles the aemu dependency. Run 'make' to
build cross-compiled aemu deps, and then gfxstream.  Cmake toolchain
file is added for building the aemu dependancy.

BUG=287082350
TEST=qnx cross compile: cd qnx || make

Change-Id: I1793356c42bfbc65c285fc7da5b7f13461f9dd4e