Inline single use help constants
Previously, because different tests expected slightly different help
output the common parts were extracted into separate constants that
could be reused to reduce the cost of changing the help. However, the
preceding changes have simplified the help output and made it more
consistent so the individual constants are only used in a single place.
This change inlines and removes those constants.

Bug: 286023667
Test: ./gradlew
Change-Id: I396bc43777665f8d597e6714e746a74b1e7efc6f