Remove unused code from ResourceType
The file was originally copied from `layoutlib` and so had lots of code
that was not actually used by metalava. It was kept to keep consistency
with the original but now that it has been converted to Kotlin there is
little point in keeping it so this change removes it.

Bug: 301160262
Test: ./gradlew
Change-Id: I926e1e7ef91c8f736851c5d67b5b63b32b250cae