Parameterize the CompatibilityCheckAndroidApisTest
This change is very careful to keep the diff size to a minimum. e.g. it
adds an unnecessary `let` block to avoid changing the indentation of
the main test code which would require reformatting and increase the
size of the change diff. That will be removed in a follow up
reformatting only change.

Bug: 301282006
Test: ./gradlew
Change-Id: Id199f7569efe2c7bfa80947a6797dc42b77d553c