Combine different data sources into one list
Previously, the list of `ApiLevelCheck` was created dynamically from
two separate maps. This change combines them into a single fixed list
of `ApiLevelCheck` instances.

The set of issues to suppress are converted from a comma separate
string into arrays of issues. That makes it easier to read and see
when new issues are added.

Instead of `disabled` defaulting to `true` for 27 and `false` for the
other checks each check explicitly sets `disabled = true` except for
27. That will make it easier for individual checks to be enabled in
following changes.

Bug: 301282006
Test: ./gradlew
Change-Id: I05d52c0e8357ef1afd8dc3f8f3a7337ab9ce1414