riscv64: Fix crash in ExecuteNterpWithClinitImpl
ArtMethod::declaring_class_ is a GcRoot, which is only 32 bits wide.

With the 64-bit load, if the adjacent ArtMethod::access_flags_ field is
nonzero, the pointer in t0 will be invalid, and dereferencing it will
cause a segmentation fault. This was observed as system_server crashing
during the second boot of the device.

Test: Boot AOSP on HiFive Unmatched twice
Change-Id: Iaee482f6fcf963dc6d330650aa4be58311a807ad