Convert SignatureFileCache to a class
The `SignatureFileCache` maintains a cache from `File` to
`TextCodebase` to allow them to be reused. Previously, when it was an
`object` that cache was kept across tests, effectively being a memory
leak. This change converts it to a class and passes an instance of it
to where it is needed. So instead of there being one global instance
there is a separate instance per test which will allow the memory to
be garbage collected.

Bug: 301405075
Test: ./gradlew
Change-Id: I692b45d27efb95dc74c59fa2afdb4c2f323c61b2