Add ActionContext to encapsulate data common to processFlags actions
Previously, if contextual information needed to be passed to some of
the methods called by `processFlags` that perform the different actions
it supports then they would need to be added to the method parameters
because they do not have a common class. Over time the number of those
parameters have built up and a pattern has emerged of lots of methods
taking the same parameters so it makes sense to try and dedup those
parameters and make it easier to add new ones.

One way to do that would be to add them to a separate class but the
problem with that is that the methods are also called from separate
commands. Eventually, they will be organized better but for now we
need an approach that reduces maintenance costs while maintaining the
flexibility that we need due to this intermediate stage of development.

So, this change adds an `ActionContext` to encapsulate the contextual
information and then takes advantage of Kotlin's extension function
mechanism to modify the methods, adding this as a receiver and removing
the now unnecessary parameters.

Bug: 301405075
Test: ./gradlew
Change-Id: Ibae8b9b487f8b80405d8d32b0680dc5f65dce080