Delete the MRD check in PacketRetransmittingState.
Consider that MRD constant defined in RenewState and RebindState is a
bit redundant with known T2 or valid lifetime expiry, actually both MRD
and T2 or valid lifetime expiry should be triggered simultaneously, to
simplify the implementation and given Confirm message isn't supported
yet, just delete the MRD check in PacketRetransmittingState, and rely on
the rebindTimeout and expirationTimeout on behalf of MRD in RenewState
and RebindState, which have been scheduled at BoundState#enter.

Then we can also delete the onMessageExchangeFailed callback from
RenewState and RebindState, which should be never called (both MRC and
MRD is 0 equivalently for RenewState and RebindState).

Bug: 260934173
Test: atest NetworkStackRootTests
Change-Id: Ic608217dd980d1855ced9a0cbe81929c5f827c5e