Move SignatureFileLoader cache to SignatureFileCache
Previously, the `SignatureFileLoader` provided two functions. The first
was to load signature files and handle `ApiParseException`s. The second
was to cache some loads so that the `TextCodebase` can be reused. This
change splits the caching functionality out into a separate
`SignatureFileCache` as some callers want caching and others do not.

The separate commands generally do not want caching, because:
1. They are simple enough that if they want to reuse the same
   `TextCodebase` they can just pass it around. Part of the reason it
   is used in `Driver.kt` is because there is so much going on with
   so many different signature files that it's impossible to keep
   track.
2. In some cases, such as `ConvertJarsToSignatureFiles` which can load
   multiple large signature files one at a time, use them and then move
   on to the next, caching would just increase memory usage for no
   performance benefit.

The calls from `processFlags` generally do want caching.

Bug: 301405075
Test: ./gradlew
Change-Id: Ifafaa4f578133b77f0380acddb0c16dba40a7e1c