Reuse ClassResolver once created
Previously, each action that required its own `ClassResolver` would
create one. That was because creating a `ClassResolver` instance can be
expensive and so it should only be created when necessary. However,
once it is created it should be reused for the same reason.

This change adds the `ClassResolverProvider` to do just that.

Bug: 301405075
Test: ./gradlew
Change-Id: I04c3c60eb0e9754410473f4f58fca9151852a4c6