Fix the compatibility check fast path
Previously, it read the signature file as UTF-8 which caused problems
if it was actually a jar file. This change makes sure that it is a
`.txt` file and then optimizes the file check by comparing length and
then comparing the files in chunks.

That allows the previous disabled test that was throwing OOM to be
re-enabled.

Bug: 301282006
Test: ./gradlew
Change-Id: Ifdfd78238cb6d05596b8cdaf8c4a01321e39ff8c