Add test for the fast path compatibility check
The fast path through the compatibility check is a performance
optimization that has no discernible effect from outside, apart from
making the code run faster, something that cannot easily be relied upon
while testing. As the check for the fast path is buried deep within the
command line processing it is not easy to test in isolation, which
means that it was not previously tested. That was an issue because
there are some cases where it currently fails, causing an OOM error.

This change uses a global var to tunnel the result of the fast path
check out from the compatibility code and surface it to the test. It is
nasty but it seems like the best way to avoid the catch-22 of having to
refactor the code to be able to test it but needing a test to be able
to refactor the code safely.

It adds tests for when the fast path is taken and when it is not. It
also adds a currently ignored test that throws an OOM error.

Bug: 301282006
Test: ./gradlew
Change-Id: I4112e04e084af4e1107f1241898b0b305ec52247