Remove --omit-locations option
This test only command line option does not appear to have any impact
on the only test that uses it so this change removes it. If a similar
capabililty is required in future then it can be added by simply
allowing the test to provide additional mappings from test specific
files to a fixed label that can be used by `cleanupMessage`. There is
no need to have a test only command line option to do that.

Bug: 301282006
Test: ./gradlew
Change-Id: Ib65c0dd37cb99b12f33e1a10d81b80f40840570d