Use longer names for temporary vars in OR_RETURN/OR_FATAL.
Before this change, the macros declare a variable called `tmp`, which
may shadow a variable above that has the same name and cause a problem
when `expr` expends to something that contains `tmp`.

For example:

Result<int> tmp(1);
int a = OR_RETURN(std::move(tmp));
                            ^^^
          The `tmp` in the macro shadows it.

Without -Wall, this can lead to undefined behavior. Fortunately, the
Android build system uses -Wall so that it can catch this issue and
throw a compilation error, so undefined behavior will not happen.
However, it's still not great that the macros don't allow the operand to
expend to something that contains `tmp`.

This change mitigates the issue by using a longer name with a leading
"__" for the temporary variable.

Test: Presubmit
Change-Id: I5035570a21b6f0e1baf60a1d089a4b5eabd299f9