Autoupdate of 2 submodules
*build/soong updated to 370e34172f28f7bd8bdfa7b3b8a846523f2eba33
Merge "Generate a ndk_sysroot target in bp2build" into main
*system/tools/aidl updated to 2136a11a9a45615c51d40e3283f71a12045eb2fc
Pass deps to SetStubsForDynamicDeps

This function now requires an additional arg for `deps`. This will be
used to add `ndk_sysroot` to deps of sdk variants of cc libraries.

Update the usage of this function in system/tools/aidl

Test: TH
Bug: 300504837
Change-Id: Id6ad0c254f65403ee273b4fbc56e20521160f886