tp: add annotation support to protozero_to_json
This CL adds support for emitting the __anntotations key at the root of
metrics protos emitted as JSON. This used to work in the past when we
used proto_to_json but when we moved to protozero_to_json, that broke
this support. This is necessary because Chrome's TBMv3 depends on this.

While we're here, remove proto_to_json from the codebase as nothing uses
this anymore.

Change-Id: Ieffe76fdb674a04f43ac621e132a59a0676c3f8f