Allow BaseOptionGroupTest based tests to check more than options
Previously, the `runTest` method would only pass the `options` to the
supplied `test` lambda which only allowed callers to check the state of
the options. This change refactors that to switch to calling the `test`
lambda on a `Result` object which allows for additional information to
be passed in. That necessitated the use of `it` in the lambdas to be
replaced with `options`.

Bug: 286023667
Test: ./gradlew
Change-Id: Ic327007ef7e3aa586efe87371284fe72721536cb