Verify an error reported by ReporterOptions exits with non-zero
There is some logic in the `MainCommand` which checks to see if any
errors were reported to any of the `DefaultReporter` instances. If they
were and `--pass-baseline-updates` was not specified then it will
throw an exception which will cause metalava to exit with a non-zero
exit code. This change adds a test to verify that behavior works when
the error is reported during parsing of `ReporterOptions`. That is to
ensure that a follow up change which refactors this does not break
that behavior.

Bug: 295013652
Test: ./gradlew
Change-Id: I7db5862bf7a29829ada51217f9c673bbaae27831