Use `ReporterOptions.reporter` for bootstrapping
Previously, the `ReporterOptions.reporter` was used both for handling
issues that were reported by `ReporterOptions` and also for general
reporter issues. This change uses that just for bootstrapping and
creates another `DefaultReporter` in `Options`, alongside the
`reporterApiLint` and `reporterCompatibilityReleased`.

A follow up change will rename `ReporterOptions` as it is no longer
configuring a `Reporter`.

Bug: 295013652
Test: ./gradlew
Change-Id: I9ef2b2c3c7b73f55c8c7bb801a5ec175c542b69a