Create Manifest lazily when requested
Previously, the `Manifest` would be created during parsing of the Clikt
options, which is done before `Options.parse()` is called. This change
splits the `manifest` property into 2, an option that produces an
optional `File` and a lazily created property that constructs the
`Manifest` when requested.

This is needed because a follow up change needs to initialize the
`reporter` option in `parse()` alongside `reporterApiLint` and as that
is called after the Clikt options are parsed that will break the
`manifest` option.

Bug: 295013652
Test: ./gradlew
Change-Id: Ic96093694f5e69bc1fa2e782c96739c1a0af82b9