Separate --version handling
Previously, the `--version` option was supported using the built in
Clikt `versionOption` function but that had a couple of problems.
1. It would compute the version eagerly, even if it was not actually
   needed, which was unnecessary. Computing the version involved
   reading a resource file so while not a lot of work was worth
   avoiding. This was hacked around in https://r.android.com/2701733.
2. It threw a `PrintMessage` error which is used for a number of
   different reasons in Clikt so it would be difficult (hacky) to
   differentiate between them. That is necessary for the follow up
   change which adds support for the `--print-stack-trace` option.

This change addresses the issues by using the `eagerOption` function
(which is what `versionOption` uses under the covers) directly and
simply throwing a special `PrintVersionException` which is then caught
and it is at that point that the version is computed and output.

Bug: 301969474
Test: ./gradlew
Change-Id: I130a17a3116cf1d59ef4e8e74a00bfc596d171f9