Correct handling ArtMethod array during class linking
In case of userfaultfd, we were setting declaring_class_ of all
ArtMethods to nullptr. This could cause problem if a GC is taking place
simultaneously and it is visiting the ArtMethod array at the same time.

Instead, we should convert the array to a 'no gc-root array' so that
compaction code skips it. The marking-phase will simply skips it as it
will no longer be pointed by the class.

Bug: 301880194
Test: test infra
Change-Id: I72007dd94ad71aa9ccc6005eb6d99b5a8c1ad7ee