[scudo] Update header without read-modify-write operation (#66955)
We used to update the deallocated block with
atomic_compare_exchange_strong to ensure the concurrent double-free will
be detected. However, this operation incurs huge performance overhead
which takes over 50% execution time in deallocate(). Given that we
already have the checksum to guard the most double-free cases and other
block verifications in the primary allocator, use atomic-store instead.

GitOrigin-RevId: 54ddd0762bd0dac31c9e2732bf807cea1785f091
Change-Id: I8914b848b60f657249563722c944920ed9ae39a8