Autoupdate of 2 submodules
*build/soong updated to dec9ce6693cd21c4fb765206c2bd1d72c6b736f3
Merge "support aidl bp2build changes" into main
*system/tools/aidl updated to 6186c842efc229c55fc30f91f6d92141541d1bbb
Make aidl phony mod responsible for bp2build

This effective no-op refactoring prepares for allowlist v2. Allowlist v2
will require that, to be bp2build-on, a module's dependencies must be
bp2build-on. Previously, the phony module created by aidl_inteface was
not the same module that generated its bp2build definition, so any rdeps
of that phony module could not possibly be bp2build-on. This change
remedies this.

Fixes: 301676937
Test: m nothing
Test: Presubmits
Change-Id: Id2676856dbe967a629a0af10b62c3c0f62a65a06