Push use of Psi model classes into single method
This does not quite remove all references to Psi elements but it does
push the references down into a single method which will make it easier
to remove in a follow up CL.

One slight issue is that for some reason implicit constructors are
ignored when adding api level information but it is not clear why that
should be. It may just be unintentional behavior/limitation arising
from use of the Psi module. Added a TODO to follow up later.

Bug: 302290849
Test: ./gradlew
Change-Id: I03fecaaf2a0d58d04e96e1eddf869f8f95596a28