Autoupdate of 10 submodules
*external/rust/crates/const-oid updated to 8392c57c0e2db7dc8d1f19e827bce621c8af3825
Add baremetal _nostd variant

Note that product_available and vendor_available are dropped from the
existing target because they are irrelevant for a _host target (as per
the change in aosp/2763825).

Add a visibility restriction to the new variant to preserve the
limitation that this crate should not be used in general Android code.

Bug: 303192701
Test: build both variants
Change-Id: Ifd819cfab949aed0f20c4da90af618411216a6cb

*external/rust/crates/der updated to 7944c99fec8ce2247e2e8119c2a375b9a665732a
Add baremetal _nostd variant

Note that product_available and vendor_available are dropped from the
existing target because they are irrelevant for a _host target (as per
the change in aosp/2763825).

Add a visibility restriction to the new variant to preserve the
limitation that this crate should not be used in general Android code.

Bug: 303192701
Test: build both variants
Change-Id: Ib9f3a7b59d8c3549f23c9590a324240dbbb3860a

*external/rust/crates/flagset updated to 9e90425e467a13d0c057c4906675eebf39aba403
Add baremetal _nostd variant

Note that product_available and vendor_available are dropped from the
existing target because they are irrelevant for a _host target (as per
the change in aosp/2763825).

Add a visibility restriction to the new variant to preserve the
limitation that this crate should not be used in general Android code.

Bug: 303192701
Test: build both variants
Change-Id: I7bceaf8dbb593b1cb9a3ae6071d46870b28f351a

*external/rust/crates/hex updated to 7b1a8296425d7dce6a31e46cc0ca45b985d67b56
Add baremetal _nostd variant

Bug: 303192701
Test: build both variants
Change-Id: Ia9a78aed110f58f67b03a9325748081391d0189c

*external/rust/crates/pkcs1 updated to 9581b421ccd6c22f93dc133bda2c275a61cbe225
Add baremetal _nostd variant

Note that product_available and vendor_available are dropped from the
existing target because they are irrelevant for a _host target (as per
the change in aosp/2763825).

Add a visibility restriction to the new variant to preserve the
limitation that this crate should not be used in general Android code.

Bug: 303192701
Test: build both variants
Change-Id: I158dab8eda7562ba634c2a9f72dfbe490609ae50

*external/rust/crates/pkcs8 updated to 3ce2d523649d3cb4224e6202cb712667be0f1b92
Add baremetal _nostd variant

Note that product_available and vendor_available are dropped from the
existing target because they are irrelevant for a _host target (as per
the change in aosp/2763825).

Add a visibility restriction to the new variant to preserve the
limitation that this crate should not be used in general Android code.

Bug: 303192701
Test: build both variants
Change-Id: Id40ee8bf67004e6e380ff9fd361938c5799331e1

*external/rust/crates/sec1 updated to e785f6b306b1d277d68509cb2284b8810d3f852e
Add baremetal _nostd variant

Note that product_available and vendor_available are dropped from the
existing target because they are irrelevant for a _host target (as per
the change in aosp/2763825).

Add a visibility restriction to the new variant to preserve the
limitation that this crate should not be used in general Android code.

Bug: 303192701
Test: build both variants
Change-Id: I36d5dc3d6a22a750963b0ffb592f65880ae60ec0

*external/rust/crates/spki updated to f931d5da30e4a7cf5e5f6f9c25519d00a87880bf
Add baremetal _nostd variant

Note that product_available and vendor_available are dropped from the
existing target because they are irrelevant for a _host target (as per
the change in aosp/2763825).

Add a visibility restriction to the new variant to preserve the
limitation that this crate should not be used in general Android code.

Bug: 303192701
Test: build both variants
Change-Id: I6126d328055a1a70934e43ab504a35dc8bef4577

*external/rust/crates/x509-cert updated to 17e0d735ebf09a7663026055a6505ae1731719dc
Add baremetal _nostd variant

Note that product_available and vendor_available are dropped from the
existing target because they are irrelevant for a _host target (as per
the change in aosp/2763825).

Add a visibility restriction to the new variant to preserve the
limitation that this crate should not be used in general Android code.

Bug: 303192701
Test: build both variants
Change-Id: Ic00325e867e90f418a1f852410f0fd678033730e

*external/rust/crates/zeroize updated to a71b01282be197b1b7541718562e8186b7e3bdce
Add alloc to _nostd variant

Bug: 303192701
Test: build both variants
Change-Id: I724d92d37fa553c84ef65c0e0ebd80672e7c829e