Cleanup model.text java API
Previously, the text model used `@Throws(ApiParseException::class)` in
lots of different places. They were introduced during the conversion
from Java to Kotlin. However, most of them, in fact all but one of them
are unnecessary because they are no longer (if they ever were) called
from Java code.

So, this removes the unnecessary ones, leaving only the one of the
method that is called from Java. It also adds a missing `@MetalavaApi`
to that method.

Bug: 302654813
Test: ./gradlew
Change-Id: I800d1a95efeb067e1f10ae5a6c0569cbf5c62f2f