Work around psi annotations applying to both type and item
Per https://docs.oracle.com/javase/specs/jls/se11/html/jls-9.html#jls-9.7.4,
when an annotation is only TYPE_USE, it should only apply to a
type. If it is both TYPE_USE and e.g. METHOD, it should apply to
both a method's return type and the method itself (similar for
parameters and fields).

For now, this leaves type-use nullness annotations on types, to
be changed once type modifiers handle nullability.

Bug: 300081840
Test: Updates existing tests

Change-Id: I87c61f64834159d53d0daa92058603f5640b42d1