Add suite tests for toErasedTypeString and findMethod
This ensures consistent behavior of the `TypeItem.toErasedTypeString()`
and the `findMethod(String, String)` method that uses it across the
different model providers.

The main reason tests for `findMethod(String, String)` were added is
because some of the changes to `toErasedTypeString` will cause some
existing uses of `findMethod` in other tests to fail and need fixing.
Adding comprehensive testing here ensures that changing those other
tests does not reduce test coverage.

In addition to adding tests for the current behavior this documents the
expected behavior of the two methods. There are some discrepancies
between the current model provider implementations and that documented
behavior which will be fixed in follow up changes.

Bug: 303839034
Test: ./gradlew
Change-Id: I4a2efd555c78a8d70acf5b3f403eeddae450ba06