Replace "..." with "[]" in TextTypeItem.toErasedTypeString
Previously, varargs parameters were represented using the "..." suffix
instead of the "[]" suffix in the value returned from
`TestTypeItem.toErasedTypeString()`. This change switches it to use
"[]" instead.

That also fixes an issue with the handling of generic types with
explicit bounds when used as a varargs parameter. Previously, it would
not replace it with the bounds, e.g. the `T` in `T...` would not be
replaced with `Number`, if `T` was defined as `<T extends Number>`.
However, the `T` in `T[]` would be replaced with `Number` given the
same definition for `T`. That is because `substituteTypeParameters()`
did not handle `...` correctly. As this change always replaces `...`
with `[]` before `substituteTypeParameters()` is called that is no
longer an issue.

It was necessary to update a couple of calls to `findMethod()` in
another test as that relied on the previous behavior of
`TextTypeItem.toErasedTypeString()`.

Bug: 303839034
Test: ./gradlew
Change-Id: I42e40a143ff339d889cbfd9af8ce308693cb7ae1