Replace "..." suffix with "[]" during Psi type erasure
Previously, when erasing types a type that was `T...` would be erased
to `Object[]`, replacing the `...` suffix with `[]`. That was
inconsistent with how a non-generic varargs type was handled. It was
caused by a problem with `TypeConversionUtil.erasure` which this change
works around by handling `PsiEllipsisType` specially.

The new behavior is to always replace `...` with `[]` when erasing
types. This breaks a couple more of the tests on the text model but
they will be fixed in follow up changes.

Bug: 303839034
Test: ./gradlew
Change-Id: Ic103880ac3dd4561b404b1407215ebffdcd333b1