[scudo] Improve the message of region exhaustion (#68444)
In this CL, we move the printing of allocator stats from primary.h to
combined.h. This will also dump the secondary stats and reduce the log
spam when an OOM happens

Also change the symbol `F` to `E` to indicate region pages exhausted. It
means the region can't map more pages for blocks but it may still have
free blocks to allocate. `F` may hint the failure of fatel error in the
region. Also update the related comments.

GitOrigin-RevId: b53ff43d36d8fea451d34dfea14df37ca3d2cead
Change-Id: I1a3d720da63aa5803e17700c57d2d088acab68b6