Demonstrate error in generating from text stubs with classpath
This change adds a test to demonstrate the unexpected behavior in
generating from-text stubs with `--classpath` argument. When a parent
class does not possess a default constructor, metalava selects an
optimal constructor and calls it in the generated stub of the class
constructo to avoid compile error. However, from-text stub currently
does not insert the `super(args...)` signature if the parent class is
defined in the jar file passed via `--classpath` argument.

Test: ./gradlew
Bug: 303687754
Change-Id: Id066c0e855ae06301a4ff6e965bdf7b128a33c7f