Add constructors to all classes in class hierarchy
Previously, the `ApiAnalyzer.addConstructors(Predicate)` method only
added constructors to classes which were in the `ApiAnalyzer.codebase`
`Codebase`. That meant when classes were loaded from another
`Codebase`, e.g. such as when the `Codebase` is loaded from a signature
file with option `--api-class-resolution api:classpath`, classes from
the other `Codebase` were assumed to have a default no-args
constructor.

The reason that is true is because constructors were only added to
`ClassItem`s whose `Item.tag` field was set to `false`. However, only
classes in the `ApiAnalyzer.codebase` had their `tag` cleared by the
`codebase.clearTags()` call.

This change fixes the issue by simply using a `MutableSet<ClassItem>`
to track which `ClassItem`s have been visited. It uses the object's
identity as the key so should not adversely affect the performance.

Bug: 303687754
Test: ./gradlew
Change-Id: Id375e8aeacb41d5bd18dd5ce8793fb544038f465