Pull setting of cls.stubConstructor from nested blocks
Previously, the `cls.stubConstructor` property was set in two separat
blocks inside different conditionals which made it difficult to work
out when it was set to a non-null value and when wit was set to a null
value. This change pulls the assignment out of those conditionals,
merges the conditionals and makes `null` an explicit value.

Bug: 304226685
Test: ./gradlew
Change-Id: I0228e4426720601b263afcb89f22e700d44b103e