Simplify setting of `superConstructor`
When the `ApiAnalyzer` code that set `superConstructor` was first
written it had to deal with the `superConstructor` being set from the
sources. However, that is no longer true and this is the only place
where `superConstructor` is set and it is always `null` before the
`addConstructors` method is called. This change simplifies the setting
and ensures that all constructors use the same value for
`superConstructor`.

Bug: 304226685
Test: ./gradlew
Change-Id: I76a638dd049c744eeffa4cd53b40df76dd7e24f3