Add stub constructors even when --api-class-resolution=api:classpath
Previously, the constructors needed when generating stubs were not
added to the model when `--api-class-resolution=api:classpath`. The
reason for that was because the `addMissingConcreteMethods` which was
also called from (`addMissingItemsRequiredForGeneratingStubs`) would
fail if passed a codebase that mixed `TextClassItem` and `PsiClassItem`
in the same class hierarchy.

This change just moves the logic for handling constructors out of the
`if (options.apiClassResolution == ApiClassResolution.API) {...}`
block, leaving the call to `addMissingConcreteMethods()` inside.

Bug: 303687754
Test: ./gradlew
Change-Id: Id29e22cc46f0305317161772d46d295d3719a592