Autoupdate of 2 submodules
*art updated to 543be883e3b20604b940d7ad9ef5cadcd08c8976
Add an artd method to copy the embedded profile from an apk file.

Context: go/platform-support-for-baseline-profiles

Bug: 257532944
Test: atest art_standalone_artd_tests
Change-Id: I16d638aa3f281fa0bdb437bf6a27826795394a86
Merged-In: I16d638aa3f281fa0bdb437bf6a27826795394a86

*frameworks/proto_logging updated to b298c526d52db76952cd7909b069e34b8ff9f5d7
Add fields to count packages in BackgroundDexoptJobEnded atom

This new fields should help to differentiate the "full"
background DexOpt runs (where almost every package needs to
be re-optimized, usually following a mainline or OTA update)
and the "incremental" runs (where the number of package to
re-optimize should be way lower) - at the moment the two
cases are treated in the same way when it comes to metrics,
with the possible effect that the average duration of the
background DexOpt job is not accurate because the two extreme
cases are mixed together.

Bug: 290340704
Test: atest art_standalone_artd_tests
Test: atest ArtServiceTests
Change-Id: Iebb845e395b52e51f415feb171a88bbbd23e2610
Merged-In: Iebb845e395b52e51f415feb171a88bbbd23e2610