Autoupdate of 3 submodules
*build/soong updated to c5be9361b67ed84ccb8d765151c340a92cc7fff9
Merge "Retain concrete overrides of abstract methods by default" into main
*external/apache-http updated to dde1797b4add8591f327259e5e4f5bee061d7639
Merge "Prepare for keeping concrete overrides of abstract methods" into main
*external/icu updated to e7647727041f7257c8eb901f6b64744e09ab87f4
Prepare for keeping concrete overrides of abstract methods

Currently, metalava will discard concrete overrides of abstract methods
from the API signature files which causes problems when attempting to
generate stubs from those files and also discards information that may
be important in API reviews. This change is the first step in the
process of changing that behavior.

This change hard codes the current behavior for discarding the concrete
overrides (`add-additional-overrides=no`) into those signature files
that will (or are likely) to be affected by the change in behavior.
That allows the default behavior to be changed for all the other files
without affecting these.

Follow up changes will remove the hard coding from the signature files
and will update the contents which will include the concrete overrides
of the abstract methods.

More information and the background for this change is available at
go/add-overriding-one-pager.

Test: m checkapi
Bug: 299366704
Change-Id: I237ce1ccf203508a2dcbce95dba3da84ca9cee63