Autoupdate of 9 submodules
*frameworks/base updated to fc9164906edfe1cd7b56b27592ebfc28658f0a70
Merge "Prepare for keeping concrete overrides of abstract methods" into main
*packages/modules/Bluetooth updated to 7953818dc29bcf407b05f664e6e4397017f8555a
Merge "Prepare for keeping concrete overrides of abstract methods" into main
*packages/modules/Connectivity updated to 415773ed06dc8d516eacc1d2514bd697d866c422
Merge "Prepare for keeping concrete overrides of abstract methods" into main
*packages/modules/Permission updated to 9c9709abdeccc109988106729eb29af39d266a3e
Prepare for keeping concrete overrides of abstract methods

Currently, metalava will discard concrete overrides of abstract methods
from the API signature files which causes problems when attempting to
generate stubs from those files and also discards information that may
be important in API reviews. This change is the first step in the
process of changing that behavior.

This change hard codes the current behavior for discarding the concrete
overrides (`add-additional-overrides=no`) into those signature files
that will (or are likely) to be affected by the change in behavior.
That allows the default behavior to be changed for all the other files
without affecting these.

Follow up changes will remove the hard coding from the signature files
and will update the contents which will include the concrete overrides
of the abstract methods.

More information and the background for this change is available at
go/add-overriding-one-pager.

Test: m checkapi
Bug: 299366704
Change-Id: Ia1562532938bc078b739dfe8b3910b6c741cde72

*packages/modules/Scheduling updated to a1ad79b15017899c7d420f048ee7dd3478f1d0d8
Prepare for keeping concrete overrides of abstract methods

Currently, metalava will discard concrete overrides of abstract methods
from the API signature files which causes problems when attempting to
generate stubs from those files and also discards information that may
be important in API reviews. This change is the first step in the
process of changing that behavior.

This change hard codes the current behavior for discarding the concrete
overrides (`add-additional-overrides=no`) into those signature files
that will (or are likely) to be affected by the change in behavior.
That allows the default behavior to be changed for all the other files
without affecting these.

Follow up changes will remove the hard coding from the signature files
and will update the contents which will include the concrete overrides
of the abstract methods.

More information and the background for this change is available at
go/add-overriding-one-pager.

Test: m checkapi
Bug: 299366704
Change-Id: I16cef4319b4a6331196a225c547f65e9baff185e

*packages/modules/StatsD updated to e920e2c0d1056cb2cb27417e643aed510d489ce9
Merge "Prepare for keeping concrete overrides of abstract methods" into main
*packages/modules/Virtualization updated to 189606e08d77229a61e70b1087a79ffa6deea0a8
Merge "Prepare for keeping concrete overrides of abstract methods" into main
*packages/modules/Wifi updated to 5d3552b1b6644803f954a2310faf9b4da1635025
Prepare for keeping concrete overrides of abstract methods

Currently, metalava will discard concrete overrides of abstract methods
from the API signature files which causes problems when attempting to
generate stubs from those files and also discards information that may
be important in API reviews. This change is the first step in the
process of changing that behavior.

This change hard codes the current behavior for discarding the concrete
overrides (`add-additional-overrides=no`) into those signature files
that will (or are likely) to be affected by the change in behavior.
That allows the default behavior to be changed for all the other files
without affecting these.

Follow up changes will remove the hard coding from the signature files
and will update the contents which will include the concrete overrides
of the abstract methods.

More information and the background for this change is available at
go/add-overriding-one-pager.

Test: m checkapi
Bug: 299366704
Change-Id: Iedd8b6c3ad2d69cbdd83559b446d0ea66fe2e151

*packages/services/Car updated to 02e7f40db182455005a58e48e35d45b2e582533a
Prepare for keeping concrete overrides of abstract methods

Currently, metalava will discard concrete overrides of abstract methods
from the API signature files which causes problems when attempting to
generate stubs from those files and also discards information that may
be important in API reviews. This change is the first step in the
process of changing that behavior.

This change hard codes the current behavior for discarding the concrete
overrides (`add-additional-overrides=no`) into those signature files
that will (or are likely) to be affected by the change in behavior.
That allows the default behavior to be changed for all the other files
without affecting these.

Follow up changes will remove the hard coding from the signature files
and will update the contents which will include the concrete overrides
of the abstract methods.

More information and the background for this change is available at
go/add-overriding-one-pager.

Test: m checkapi
Bug: 299366704
Change-Id: Iaa665f1a678c37123e0fe1bda1a675282e733e05