Fix up the custom config used for Scudo.
Add a first version of the svelte config and the appropriate changes
to build using that version.

Update the config to include a copy of the size class data, so that all
Android config data is contained in one place and not depedendant on
anything from upstream.

Add a typedef for DefaultConfig and fix the tests so they use the actual
custom config.

Turn off the MTE possible unless compiled for aarch64.

Add a library and test target that verifies that the right configs are
being used from the custom config. This will verify that the custom
config is used properly.

Test: Unit tests pass.
Change-Id: Ie72c0a93f6264a1a83341333aa9adde308d56b35