Autoupdate of 4 submodules
*packages/modules/Permission updated to c5d238019c80109962235fa4359f21235ea7c916
Add additional overriding methods to signature files

This change adds overriding methods that are essential to compile the
stubs generated from the signature files.

More information and the background for this change is available at go/add-overriding-one-pager.

Test: m checkbuild --build-from-text-stub
Bug: 299366704
Change-Id: Ibccbf010cada6d0e72514f17c4744b785f57989b

*packages/modules/Scheduling updated to 6dd0b6252fda87d59ffb772e921276b45a896670
Add additional overriding methods to signature files

This change adds overriding methods that are essential to compile the
stubs generated from the signature files.

More information and the background for this change is available at go/add-overriding-one-pager.

Test: m checkbuild --build-from-text-stub
Bug: 299366704
Change-Id: Ic63569612c348126478bebdbab673f903ed4dd89

*packages/modules/StatsD updated to ad8f5291df6434a17c41253fbd63c4ab0119c44e
Add additional overriding methods to signature files

This change adds overriding methods that are essential to compile the
stubs generated from the signature files.

More information and the background for this change is available at go/add-overriding-one-pager.

Test: m checkbuild --build-from-text-stub
Bug: 299366704
Change-Id: I356923a0d1f9cc4480b636c6036103896ace8e9f

*packages/modules/Virtualization updated to 2e774ff0b63d351d50b24fbc3fec2d40b8d00964
Merge "Add additional overriding methods to signature files" into main