Always require path to newFolder(...)
Calling `TemporaryFolder.newFolder()` with no parameters will create a
folder with a random name. That makes errors reported differ from one
run to the next making it difficult to understand what is happening and
impossible to test errors.

This change adds a required `path` parameter and fixes up any existing
calls. This did cause a `SignatureToJDiffCommandTest` to fail as it
was basically performing two separate, distinct tests, each of which
tried to create the same directory, which is not allowed. That was
fixed by split the method into two.

Bug: 305396465
Test: ./gradlew
Change-Id: I01897786a93f92f077dcd0ac037f8679b12bf0c8