Fix toybox-gtests and re-enable them for postsubmit
A last minute change to toybox-gtests wasn't tested in presubmit
because the CL to enable the test in presubmit had to be separated
out into a later CL.  /bin/bash doesn't exist on the device (only
/bin/sh).  Failures when execing the test subprocesses would cause
RunTest to return and then continue out of the fork conditional,
causing the subproces to continue into the test running loop and
start trying to run more tests as if it were the parent process.

Fix the path to /system/bin/sh (via _PATH_BSHELL) on the device,
rename RunTest to ExecTest to make it clear it shouln't return on
success, and _exit with an error when ExecTest returns.

Bug: 305475166
Test: atest toybox-gtests
Change-Id: I9ee3be597f75a9de3b5c8ebb72298d24834a6c09