Autoupdate of 3 submodules
*art updated to c8c2953cffe8ff9d2204f4efdff9100eb9fa6df0
Use product vndk version to check if product vndk is deprecated

There are some logic in libnativeloader to check if product is
treblelized by checking ro.product.vndk.version. However, this property
is no longer valid to check if product is treblelized, because of vndk
deprecation. This makes libnativeloader to consider as product is not
treblelized when the property is empty from the VNDK deprecation. This
causes unexpected error from libnativeloader which allows product apps
to use system / system_ext libraries when the product partition is VNDK
deprecated. Product is force-treblelized from R, so it is safe to assume
that product partition is always treblelized as long as the device is
treblelized. This change removes existing check of
ro.product.vndk.version, and use this check only to confirm if product
VNDK is deprecated.

Bug: 290159430
Test: Cuttlefish build and run succeeded
Test: libnativeloader_e2e_tests passed
Change-Id: Ieee3ff2dde4244e7c46420ad8dde5f8e5cf249e4

*frameworks/base updated to 8fa8c07660f0e48e24fd68c91bde8654e61ef7ce
Merge "No need to check product vndk version" into main
*system/linkerconfig updated to 22f223e241b18c3fef1fe50f12fdad5a7ff363ee
Merge "Product namespace is required despite of VNDK deprecation" into main