Fix atrace event format
To comply with Perfetto's expected atrace format, change the way we are
writing traces, for exampel by substituting the second parameter
from tid to pid (apparently ftrace already provides the tid, but it's
not easy to figure out the pid from there).
In addition, create a separate Tracer for every instruction, as in the
case of multiple threads writing to the same fd, the trace gets
corrupted. This is probably an overkill as there should be more
instructions than threads, but so far is the simplest solution to
implement, and can be changed if this is seen as a performance
regression.

Test: ctest
Bug: 305023314
Change-Id: Icaebf9e2f4752062972d1b8a82b8c4f951de6bf7
Signed-off-by: Alessio Balsini <balsini@google.com>