Add the public API to the conscrypt platform API
When generating the core platform stubs from sources metalava
automatically adds in the public API (i.e. the classes and members that
are not annotated). However, when generating the stubs from text those
classes and members are not added causing build failures.

This is not an issue for other modules like art as they do not have a
separate `core-platform` stubs, instead it has been migrated to the
more standard `public` extended by `system` extended by `module-lib`
structure. The build automatically handles that my flattening the
hierarchy of signature files into one from which it generates the
stubs. e.g. the system stubs are generated from `public` + `system`.

Bug b/305928481 has been raised to switch conscrypt over to that
structure but in the meantime this change explicitly adds the public
API to the `api/platform/current.txt` by adding the public API classes
to the sources and adding `--show-unannotated` to the command line.

As a side effect that adds `OpenSSLSocketImpl.getHandshakeSession()` to
the API too. That is not actually a change in the API, but rather is
making the file a more accurate representation of the API as That
method was already generated in the stubs when building from sources.
That was necessary because implements a method from the `SSLSocket`
interface which is part of `libcore's` public API.

Test: m droid
Bug: 299366704
Change-Id: Id59112e685426dca4147e3fe381a9779803bc354