libavb_rs: rename verify.rs -> ops.rs
I originally thought we would put both the callback ops and the
verification API in a single `verify.rs` file, but the two pieces are
large and disjoint enough that it will help readability to split them.

This CL moves the callback ops to `ops.rs`, and a future CL will add the
verification APIs to `verify.rs`.

Bug: b/290110273
Test: atest
Change-Id: I43b597ef10d3a1a5cc90e362439ae35b7576dc9e