Don't get/set VSOCK "port" with grpc_sockaddr_*_port.
The VSOCK "port" has nothing to do with IP ports, so it's not
appropriate to get/set it from these functions. The documentation
for grpc_sockaddr_set_port and grpc_sockaddr_get_port makes it
clear that these functions are intended to refer to IP ports.
This brings the patch more in line with upstream support introduced
in 1.57.0.

Bug: 279548314
Test: Presubmit, cuttlefish_wmediumd_control_test
https://android-build.corp.google.com/abtd/run/L07100000963584554/

Change-Id: I599a50b4b4c61025a23973d959195fa897322ab7