Extract comparison of types into sameType
Previously, the `MethodItem.sameSignature` just used `t1 != t2` to
check whether two return or parameter types are the same for the
purposes of eliding the method. However, that does not work because
it does not handle bounds properly. This change extracts the check
into `sameType` method that documents the expected behavior and will
allow the behavior to be changed for both return and parameter types
in the same place.

Bug: 306333334
Test: ./gradlew
Change-Id: Ic920f6054bdfd046ad34bd271244ba30ee3282e4