Prevent CovariantReturnType annotations from being output
Previously, the `dalvik.annotation.codegen.CovariantReturnType` was
being written into stubs. It could also have been written into API
signature files but was not simply because the methods to which it
had been applied were themselves not output to the signature file
because they had (by definition) the same signature as the super
method.

However, the annotation is intended for implementation use only so it
should not appear in the stubs or the signature so this change removes
it altogether.

Bug: 306245305
Test: ./gradlew
Change-Id: I0d3b3862d8f5e9164551e27e47413ab9e40d9866