Move call to reallyHideFlaggedSystemApis before ApiAnalyzer.computeApi()
Previously, `reallyHideFlaggedSystemApis` was run after
`loadFromSources`, after `ApiAnalyzer.computeApi()` and
`ApiLint.check()` so any changes it made would not affect the api or
linter as expected.

This change moves the call before then so that it will affect those
properly, yet after `ApiAnalyzer.mergeExternalInclusionAnnotations()`
so it can affect annotations merged in from other sources.

This is not the correct solution but it is better than what was before
and does fix some issues in `FlaggedApiTest`. It does require an
extra `UnflaggedApi` issue to be reported in one of the still incorrect
tests but that is correct behavior given the way the test is broken.

Bug: 306242007
Test: ./gradlew
Change-Id: Ia27d8e9a7db70949a1c4bbcbf72a48ae076ec423