Add ShowabilityTest to cover some combinations of show/hide annotations
Change https://r.android.com/2716076 had to be reverted because it
caused some failures when attempting to build Android as it had some
combinations of show/hide annotations that were not being tested. This
change improves the test coverage.

Two tests are added with the show/hide annotations in different order.
That is because the annotations are processed in order and so that does
affect how they are processed. The tests ensure that does not affect
the generated output.

Bug: 292484829
Test: ./gradlew
      # Cherry picked https://r.android.com/2716076 and verified that
      # the new tests failed.
Change-Id: If6e1cbd8a874083d8b238c5513d9912bcd26a26c