Prevent private classes from being added to the stubs
The stubs for the intra core and core platform APIs include not just
the API items that are annotated with `@IntraCoreApi` and
`@CorePlatformApi` respectively but also any API items that are
unannotated. Unfortunately, these unannotated API items are not tracked
in any API signature file so they can easily leak out.

This change hides a few classes that are in the stubs which should not
be. It also adds the files used to generate the intra core and core
platform APIs so if they do have any unannotated classes that would be
included in their stubs unintentionally then they would be output to
the `public/current.txt` file and so easily detected.

Bug: 274805756
Test: m update-api
      # Make sure that no API .txt files were changed.
      m legacy.i18n.module.platform.api.stubs stable.i18n.module.platform.api.stubs
      dexdump out/target/product/vsoc_x86_64/obj/JAVA_LIBRARIES/legacy.i18n.module.platform.api.stubs_intermediates/javalib.jar | grep -E "ByteBufferIterator|CharsetICU|Icu4cMetadata"
      dexdump out/target/product/vsoc_x86_64/obj/JAVA_LIBRARIES/stable.i18n.module.platform.api.stubs_intermediates/javalib.jar | grep -E "ByteBufferIterator|CharsetICU|Icu4cMetadata"
      # The previous two commands should output nothing, before this change they
      # output dexdump entries for the affected classes, afterwards they output
      # nothing.
Change-Id: I43d814aa4f762db42dc6c3f91f4c51a5f2ee6913