Tradefed: Don't remove communal profile
Bedstead automatically removes all users (except for the System and Main
users) when a CTS test is run. The assumption is that those are the only
pre-existing users on the device. However, this assumption is false in
the case of a Communal Profile, which may pre-exist on the device. We
must ensure that we don't delete this user when initiating the CTS test
(if the user already exists).

Bug: 297103765
Test: atest android.multiuser.cts.UserManagerTest
Test: tapas tradefed-all && m && tools/tradefederation/core/tradefed.sh run singleCommand host -n --class com.android.tradefed.targetprep.RunOnSecondaryUserTargetPreparerTest
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:75b37cb4793445db28ccde3251bb7157c7393c14)
Merged-In: Ifa8ecb0f7a5fa2179bbdeb559bdac3abd5d90463
Change-Id: Ifa8ecb0f7a5fa2179bbdeb559bdac3abd5d90463