cgroup: Handle trailing new line in cgroup.controllers
The last item in cgroup.controllers (misc or rdma in my case)
contained a new line character which caused the controller search to
fail.

This commit avoids including the newline character inside the name
comparison.

The search failure caused the "cgroup_regression_test.sh" test to fail
with a confusing error when it tries to mount a V1 subsys thus
removing the V2 and causing the available set of V2s to change between
scans.

According to the V2 docs subsys names can only include lowercase
characters and '_'. So we strictly look for those characters.

The newline (and delimiting space) is just what the kernel currently
prints. IDK if it is specified anywhere, but if it changes then the
error should be obvious.

Fixes: 310da3784 ("Add new CGroups APIs")
Change-Id: Iad436c8bc7a854553da962f90d8623359a7654fb
Signed-off-by: Richard Palethorpe <rpalethorpe@suse.com>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Reviewed-by: Marius Kittler <mkittler@suse.de>
(cherry picked from commit 5292c46e5032011337d73119f27d365f29140004)
Test: N/A
Signed-off-by: Edward Liaw <edliaw@google.com>