Compute Showability for an Item by combining individual properties
Previously, `Showability` objects were treated as a set of fixed
values, like an `enum`. However, that led to problems where some
`Showability` objects could not be combined because there was no
suitable `Showability` to use, resulting in errors being reported.
This change combines the `Showability` instances by combining the
individual properties so there is no combination that cannot be handled
which avoids errors.

Bug: 292484829
Test: ./gradlew
Change-Id: Ib8d154aded572c6559255774e61bce2a1a943fbd