Updates in Android.bp to build libdav1d.a
Dav1d requires few files to be compiled twice, once with BITDEPTH
macro set to 8 and once with that set to 16. Current Android.bp was
building all the files twice instead of doing this for a select few
files. This is now fixed by compiling only the required files twice.

Also with this change, instead of depending on two libraries, the
consumers of this library now need to depend on one library "libdav1d".

Bug: 307665541
Test: dav1d

Change-Id: I30d90d2d35bce5e60d3ddee083deae4e07928159