Save time in remounting rw.
Saves about 20 secs.
Stop waiting for the device to become fully booted (sys.boot_completed) when all we are doing
is making the drives rw.  We are about to push files and reboot again.
This only affects the one extra reboot we do when we detect read-only
drives (after a user-debug flash).

Test: atest
Test: adevice update  # after flash
Change-Id: Ic3ab04a041a982965a32226fd0779ad110bee59a