Separate show/recursive from forStubsOnly in Showability
Previously, when an item was annotated with a
`--show-for-stub-purposes-annotation` it would set showability to
`SHOW_FOR_STUBS` which set `show`, `recursive` and `forStubsOnly` all
to `true`. This change separates them so that `SHOW_FOR_STUBS` only
sets `forStubsOnly` to `true`. That means that the `show` and
`recursive` properties are only set to `true` for annotations that add
an item to the current API, not the API being extended.

This is in preparation for the follow up change which improves the
process for computing the `Showability` for an `Item` from the
`Showability` from `AnnotationItem`s by doing the merge on an
individual property level. Separating the properties makes that a
little simpler.

The behavior is preserved by modifying the `show()` and
`showNonRecursive()` methods to return `true` if `forStubsOnly` is `true`.

Bug: 292484829
Test: ./gradlew
Change-Id: I27d33ff21787e4847846d1ba9c8904b03d45b286