Improve storage/GPT buffer management
Store all information fields in the Gpt struct in the provided buffer so
that the buffer is self contained and can fully reconstruct the Gpt.
This makes it easier for user code to move/manage buffer without needing
to resync GPT.

Also added a few helper classes in EFI application that internally
manage buffers required by libstorage API to simplify usage.

Change-Id: I52f8ef1f45f6fad379de5c2fad65e6f29793393d