Combine hasShowAnnotationInherited/onlyShowForStubPurposesInherited
Previously, when determining whether an API was only included for use
in stubs the `ApiPredicate.includeOnlyForStubPurposes()` method would
first check to see whether an item had a show annotation (using
`Item.hasShowAnnotationInherited()` and then check to see whether the
show annotation was for stubs only using
`Item.onlyShowForStubPurposesInherited()`. That was inefficient as each
of those would require checking the following, in order, until it found
a show annotation (or none):
1. The `Item`'s annotations.
2. If the item was a `MethodItem` then its super methods (recursively
   back of the class hierarchy).
3. Then its parent item, e.g. containing class or package.

This change combines the two methods into a single extension function
`Item.apiMembership()` that computes whether the `Item` belongs to the
`base` or `current` API surface, or no API surface. It is an extension
function (rather than a method on `Item`) that is private to
`ApiPredicate` as it is only used by that class.

Bug: 292484829
Test: ./gradlew
Change-Id: Iba8fd4d669b8f7b6f3ceeda6675ebdea08627186