Support hiding unstable APIs
Previously, it was not possible to hide an `Item` that had been
annotated with a show annotation because show annotations override hide
annotations. This change adds support for a special annotation for
hiding unstable APIs that can override the show annotation.

Currently, the hiding of unstable APIs is triggered by passing
`--hide-annotation android.annotation.FlaggedApi` but future changes
will add support for using other annotations.

Bug: 306242007
Test: ./gradlew
Change-Id: I94d4c81a64347f0a14d6f2fe612c1be8e6014fde