Enable the std feature of coset
By default coset doesn't depend on std. But we have separate build
targets for std and nostd clients, and there is no reason not to
enable coset's std feature for the former.

(The benefit is that CosetError then implements Error, which avoids
some small but tedious boilerplate in client code.)

Test: atest coset_test_src_lib
Change-Id: I743bd2f3e37c82d72e333c83534f31f53b9ce3ed