Fix repeated dexopt when an app has a DM file with no profile.
This bug is triggered when the following conditions are all met:

- The app has a DM file that contains no profile.
- The app has an empty reference profile or no reference profile.
- The current profiles don't have enough entries to trigger a profile
  merge.

In such a case, ART Service tries to use a profile from the DM file,
and profman unexpectedly returns 0 when processing the DM file, which
indicates that the DM file contains a non-empty profile, while it does
not. Then ART Service thinks that dexopt is needed because the app
should be "speed-profile", but there is actually no profile, so
dex2oat falls back to "verify". This is repeated on every bg-dexopt run.

The bug was introduced by r.android.com/2760747.

Bug: 308174155
Test: atest ArtGtestsTargetChroot:ProfileCompilationInfoTest

Change-Id: Ib91529760f4ea12729b3b8132a63b7e2da7dcbfa