Realign logic for has_sufficient_devices
The existing logic in has_sufficient_devices only verifies the device
number when the user applies a serial number. However, it returns the status regardless of whether the user sets a serial number or not. This change aligns the checking logic and the return value of the method, and also updates the method's name to has_set_sufficient_devices.


Bug: 309041856
Test: m atest ; atest-dev CtsWifiAwareTestCases
Test: m atest ; atest-dev CtsWifiAwareTestCases --serial $ANDROID_SERIAL
Change-Id: If5f77747d4b5ec464364da5d82af19dbb6bd8ad9