Assign default bug component to targets in this directory.
This CL is being created to improve the test attribution in android. It has been 
found that the tests owned by this OWNERS are not not being attributed to a
buganiser component.
This is part of a bigger effort to attribute all of the aosp code. 

If you think that the buganiser component is not correct please update
this CL with correct component and drop a +2.
We will use the updated component and merge the CL

Bug: 309090038
Change-Id: I3f2838347b5500392e6b5abaf33c36cce074e0a4
Test: N/A