Autoupdate of 2 submodules
*external/ltp updated to a67850ef348d5c9e974a4c7254b4a238083b88d5
Sandbox LTP genrules

gen_ltp_config used to read files from the source tree,
but this cl refactors it to package all the files it
needs into the tool itself, so it's more hermetic and
all the genrules don't need to explicitly list the files
the tool needs.

go/roboleaf-busy-beavers-sandboxing

Bug: 307824623
Test: diff'd the files produced by the ltp_config_* genrules before/after this change
Change-Id: Ica09f157980d89bc190ca44c75d64b9d7dec17b3

*test/vts-testcase/kernel updated to d52bf60eb8fbe12be4095ad3f32f4b9302392a4c
Sandbox LTP genrules

gen_ltp_config used to read files from the source tree,
but this cl refactors it to package all the files it
needs into the tool itself, so it's more hermetic and
all the genrules don't need to explicitly list the files
the tool needs.

go/roboleaf-busy-beavers-sandboxing

Bug: 307824623
Test: diff'd the files produced by the ltp_config_* genrules before/after this change
Change-Id: Ic31928653fdec492208b9dd1c6a337de7b108ba1