Add extra error msg when tracked module causes ninja errors.
We had a report of a user tracking a module and then removing it from
the repo.
The error message we output wasn't clear.
Now it suggests untracking the module if ninja error mentions the module
name.

Also fixed fakes.rs error from last commit that is only caught with
`cargo test`. (android.bp is missing whole_program_tests)

Test: atest
Test: adevice update
Test: adevice track BadModule && adevice update
Change-Id: I077fde1fd07fdc6fb1181abc1813786f292a6725