Fix ahat accounting for cleaned native registrations
If a registered native allocation is cleaned explicitly via the
thunk provided by RegisterNativeAllocation, but that thunk hasn't
yet been GC'ed, the associated heap dump currently reports that the
native allocation is live. Fix this by checking the cleaner's liveliness
in the associated linked list.

This aligns with Perfetto behavior for native registration allocation
tracking.

Test: atest ahat-tests --host
Bug: 235395968
Change-Id: Icc275de7b45e176846cf5932895692ae6eb8e166