Make it fatal if we cannot initialize private region for jit code cache
We should be able to usually initialize a private region. The rest of
the code assumes that private region has been initialized and that the
code and data mspace have been initialized. If private region couldn't
be initialized we cannot JIT at all. We should be able to initialize
private region always. If we couldn't then it means something else is
not working and we need to fix it. So instead of logging a warning make
it fatal to catch any issues earlier.

Bug: 307441080
Test: art/test.py
Change-Id: I6a090164ffa4b45378400d930a470bbaa56d3535