Autoupdate of 2 submodules
*build/soong updated to a19c9141aa31b68ca54d4af6a48a26973dd606bf
Merge changes from topic "apex-elf-checker" into main

* changes:
  Add check for unwanted_transitive_deps
  Add JoinWithPrefixSuffixAndSeperator() utility

*system/apex updated to b778223625d2bb955c5a9f519e211cf9b97588da
Add apex_elf_checker tool

A new tool checks ELF files in APEX. For now, --unwanted <names> is the
only check. It fails when any of unwanted names is found in NEEDED of
any ELF files in the APEX.

This is to support the apex.unwanted_transitive_deps property of the
build script.

Bug: 295593640
Test: m com.android.virt
Change-Id: If5adf536ccf573173f26f9f73cff0ff89bda0fa4