Correct parsing issue with type params
Previously, when a top-level `>` was reached in a type param string,
it was assumed to be the end of the param, but that isn't always
the case, like when the next characters are array brackets.

If the param does end after the `>`, the next character will either
be `,`, and the param will be added to the list, or the closing `>`,
and the param will be added to the list. So there is no need to add
the param when the `>` is reached.

Bug: 309149849
Test: Added test for previously incorrect case, existing tests pass

Change-Id: Iae8b8baa6dcc50945cde1773e2ee37581629270e