Dex verifier: Rename 'begin_' to 'offset_base_address_'
The name 'begin_' will be ambiguous in the presence of dex container,
since we have the beginning of dex file and of the dex container.
Rename the variable and other uses to be more semantically clear.

Similarly, cache the "data" section, since dex container deprecates
those header fields, so we will need more control over the variable.

Test: fuzz locally
Change-Id: I9eb70dbafe8add08e62578628cb1680dcb13221a