Use embedded_launcher: true in python tools
These scripts generate files, and they include the command
used to generate them in the file itself.

When using a non-embedded-launcher python binary, the binary
is a zip file that extracts itself to a temporary folder
before running. The temporary folder is non-deterministic,
and that non-deterministic path will end up in the file since
they write the sys.argv to the file.

embedded_launcher binaries do not extract the zip, and thus
don't have any non-deterministic paths.

Bug: 307824623
Test: Presubmits
Change-Id: Ia1d2ab97d587cf2f7a2d7a524b29ad875985577c