Ignore swap areas during GetProcMountsEntriesForPath.
GetProcMountsEntriesForPath assumes that every entry in `/proc/mounts`
has a `mount_point` (a.k.a, `fs_file`) field that is a non-empty
absolute path. This is not true for swap areas. According to fstab(5):

> For swap area, this field should be specified as `none'.

There are also some implementations on the Internet that uses an
arbitrary string for this field for swap areas.

After this change, GetProcMountsEntriesForPath ignores swap areas.

This change also improves the logging of the CHECK.

Bug: 310519261
Change-Id: I0c3945e3a80fc9db9b1c6cf9449d0239243279c3
Test: Presubmit