Add KUnit test module runner
This new binary test type expects the KUnit test modules (*.ko file) to
be pushed over to the device. It mounts `debugfs` if it's not already
mounted, loads the modules via `insmod`, and locates and parses the KTAP
file located in `debugfs` to translate the results suitable for the TF
framework.

If `debugfs` is already mounted, care is taken to not unmount it during
cleanup as to not disrupt potential code coverage that may be happening.

This binary target test is different then the others as it's a one
binary to many test results, instead of one to one. Therefore, it
handles its own testStarted/Failed/Ended calls instead of relying on
the base class functionality.

Collect only mode, dry run, is not currently implemented and will throw
a UnsupportedOperationException exception.

The mount/unmount/check debugfs calls have been moved out of the
GcovKernelCodeCoverageCollector class into a shared location inside
INativeDevice/NativeDevice.

Additional verification logic was added to the KTapResultParser to catch
incomplete KTAP files and malformed files.

On device testing:
 * Build kunit modules:
   * `$ tools/bazel build //common:kunit_x86_64`
 * Launch cvd with built kernel
 * Create test config similar:
```
<configuration description="kunit">
    <target_preparer class="com.android.tradefed.targetprep.RootTargetPreparer"/>

    <target_preparer class="com.android.tradefed.targetprep.PushFilePreparer">
        <option name="cleanup" value="true" />
	<option name="push-file" key="/usr/local/google/home/joefradley/kunit_modules" value="/data/local/tmp/kunit_modules" />
    </target_preparer>

    <result_reporter class="com.android.tradefed.result.ConsoleResultReporter" />

    <test class="com.android.tradefed.testtype.binary.KUnitModuleTest" >
	<option name='binary' key='dev_addr_lists_test' value='/data/local/tmp/kunit_modules/dev_addr_lists_test.ko' />
  </test>
</configuration>
```
 * Run config
   * `$ tradefed.sh run tmp.xml`

Bug: 309621188
Test: KUnitModuleTestTest, KTapResultParserTest, ExecutableTargetTestTest, ExecutableHostTestTest

Change-Id: I36f4919eb741e015bc4d3be4e2fec60dcd5f5596
Signed-off-by: Joe Fradley <joefradley@google.com>