Add includeDocumentationInStubs to StubWriterConfig
Avoids having to access `options.includeDocumentationInStubs` directly.
Previously, `options.includeDocumentationInStubs` and `docstubs` were
accessed separately and every use OR'd them together so that if
`docStubs` was `true` then it was as if
`options.includeDocumentationInStubs` was `true`. This change avoids
passing in `docStubs` and instead just sets
`config.includeDocumentationInStubs` to `true` if `docStubs` is `true`.

Bug: 310644782
Test: ./gradlew
Change-Id: I70a3664ad4d18c9de545d46031092d76cc4662ae