Add tests for broken --exclude-documentation-from-stubs
The documentation for `--exclude-documentation-from-stubs` says that it
has no effect on `--doc-stubs`. However, if it is specified then it
will prevent imports being added to doc stubs when necessary.

Conversely, if it is not specified then it should write out all the
documentation but it does not output `package-info.java`.

This change adds tests to show the broken behavior. A follow up change
will fix them.

Bug: 310644782
Test: ./gradlew
Change-Id: I86ae3b633e8d3d46c5457e5e6fb77ed8a73e0eaa