ReadFully: don't use stale errno on EOF
Since EOF is not an error condition for read(2), we weren't setting
errno despite returning false. This wasn't explicitly wrong, because
we didn't mention errno in the docs for ReadFully, but some clients
were logging it.

There's no good errno value for "short read" - we picked ENODATA as
the least worst option.

Bug: 302723053
Test: mma
Test: binderRpcTest on Linux Binder
Change-Id: I439a3a49a96bcb49baff0799fbfc4ac0d09ca66e