RESTRICT AUTOMERGE: Throw an error when enforcing implicit reads for a ReadRequestWithFilters and aggregations.
Note that this is a behaviour change for developers who have not been
checking for this exception. However, the APK already returns the same
error, and so it is likely that developers already check for this,
making this change acceptable.

Fix: 313791275
Fix: 313791841
Test: atest CtsHealthConnectHostSideDeviceTestCases
Change-Id: Ie29c02e0192ea861107d64397e3133af7c18a4b9
(cherry picked from commit 1c4b363c3573335f4e5ed7f89399cb4dfda4b5f2)