Change MemoryBuffer to use a vector.
In addition, add an offset value so that this can be used for the
compressed section data.

The modified MemoryBuffer now has a max size to prevent corrupted
elf data from causing huge allocations in the vector.

This class is only used for global data and a compressed .debug_frame in
the library code. The limit of 10MB is way over what any valid existing
globals data section is expected to be. A 50MB shared library only contains
a .debug_frame that is < 100KB in size. Therefore, 10MB should be able to
handle any valid large shared library with a valid large .debug_frame.

MemoryBuffer is used for some tests too, but those are all using
relatively small values.

Bug: 309857311

Test: All unit tests pass.
Change-Id: I184efea87d7c1f74202071e4943c2a762435ac11