Fix the ClassLoaderContext mismatch error message.
Before this change, ClassLoaderContext encoding relied on
`info.opened_dex_files`. This doesn't work for the error message
presentation because we no longer open dex files for ClassLoaderContext
verification since https://r.android.com/1531339. This problem leads to
empty dex paths in the error message and causes confusion. Note that it
does not affect the correctness of ClassLoaderContext verification.

This change fixes the presentation problem.

Bug: 311126962
Test: m test-art-host-gtest-art_runtime_tests

Change-Id: I83972aba7d06a086b7684e2c7ac7640e135e0ee0