Move PackageDocs into model.psi
Previously, there was one use of `PackageDocs` outside the `model.psi`
package, in `DocAnalyzer`. However, that only checked to make sure that
`getPackageDocs()` did not return `null`. However, as `DocAnalyzer` is
only called when generating doc stubs (or normal stubs with enhanced
documentation) it made no sense for it to be called on something other
than sources and in that case `getPackageDocs()` always returned a
non-null value.

This change removed that use, along with `Codebase.getPackageDocs()`
and moved `PackageDocs` into `model.psi` where it is used.

Bug: 312028307
Test: ./gradlew
Change-Id: I8c7bf781a90775e63d3cdc3845166622983e25a0