Fix AdbSshConnectionTest snapshot tests to take multiple arguments
Split the command for snapshot/restore being passed to Ssh as opposed to
passsing it as 1 long stream of text.
The tests were incorrect in nature as they should accept split
arguments.
Lastly, the binary location in Oxygen was returned with a newline which
would cause arguments passed to run as a separate command, as such
.trim() is necessary to get rid of newlines.

Bug: 310266582
Test: m tradefed-all && tools/tradefederation/core/tradefed.sh run singleCommand \
host -n --console-result-reporter:suppress-passed-tests --class \
com.android.tradefed.device.connection.AdbSshConnectionTest

Change-Id: Id6163e87625447f4dda13e98aca03e36aa5a903c