More 004-NativeAllocations failure output
Produce more debugging output for a failing test.

As part of that, we allocate additional finalizable objects, and
print how many of them were finalized. This should allow us to
tell if there were young GCs, but the PhantomReference was
promoted. Also print whether the reference was enqueued after
explicitly triggering a GC.

Mark various allocation functions noinline to make sure that
we don't accidentally keep the allocated objects live. I don't really
expect this to be the issue, but ...

Increase the delay after we fail until we recheck the reference.

Bug: 123921229
Test: Inspect message when PhantomReference is kept live.
Change-Id: Ic5cc9aba3ed383def3174ef0fc1c6b4ed61678a1