Autoupdate of 2 submodules
*cts updated to 0fd1a8f20f68c6e015164132be1a65be0c023829
Merge "Add dependency of CtsDeviceInfo.apk to csuite." into main
*test/app_compat/csuite updated to a320c1fbd64880726a719bdb3148e15f650a7e25
Create a new test plan ats-app-crawl to run csuite crawl tests on ATS
with AnTS relay.

Need to collect device info with DeviceInfoCollector to upload the test
results to AnTS. However, there is an option name conflict and we will
have to add an --alt-package option as the alternative of --package option.

Test: run on ATS.
Change-Id: If26162c0b35a54d5136c9067924d9c49595740a5