Accept multi-prefixes on receiving DHCPv6 packet.
It's valid for the server to not include a prefix in an IA_PD option
in certain scenarios (specifically in a reply to Renew / Rebind, which
means do not extend the prefix. e.g. the list of IA_Prefix option is
empty), only ignore the packet iff the prefix(es) do exist and all of
prefix(es) are invalid.

Bug: 260934173
Test: atest NetworkStackIntegrationtTests
Test: manually connecting to OpenWrt AP to test
Change-Id: I629de0d9d151bdecbb10284c2c9b3e506e48540b