Support multiple signature file inputs in signature-to-dex
Previously, it would accept two arguments an API file and an output
file into which dex signatures were written. This change adds support
for multiple files by changing `apiFile` to `apiFiles` and the
`dexFile` argument with a required `--out` option.

The existing uses in Android will be broken by this and will need
migrating at the same time this change is merged into it.

Bug: 307737035
Test: ./gradlew
Change-Id: I9262496a26d12c9946d6fd127ef1727fa031caf3