Make BadFuture check more resilient to follow up changes
Previously, the check was disabled when `superClass` was null which
meant that the follow up change to set `superClass` to null for
interfaces would prevent the check from being performed.

Also, when selecting whether to use "extend" or "implement" as the verb
in the report it would assume that it should use "implement" unless the
future interface was in the "extends". That caused two issues:
1. It uses "implement" instead of "extend" when the bad future type is
   not the first interface in an interface's extends list.
2. The follow up change would always use "implement" when an interface
   class was extending a bad future interface as the bad future would
   never be set as the `superClass`.

This change:
1. Runs the BadFuture check even when `superClass` is null.
2. Updates the test to make the bad future interface second in the
   extends list. That fails without the following change.
3. Updates the logic for determining which verb to use to fix the
   issue.

Bug: 312558482
Test: ./gradlew
Change-Id: I4ea3f449f3b44fd2e62691562a17e5e93cf9a144