Simplify SignatureWriter in preparation for follow up changes
The `SignatureWriter` handling of interfaces and super classes is quite
complex due to the first interface being used as the super class with
some parts of the interface list written out by
`writeSuperClassStatement` and the remainder by `writeInterfaceList`.

This change separates those concerns and only the super class is
written out by the former and only the interface list is written out
by the latter. The `writeInterfaceList` is also simplified by
separating the reconstruction and sorting of the interface list from
the writing which should make the follow up change which does not
affect the writing simpler.

Bug: 312558482
Test: ./gradlew
Change-Id: Ia7aa0b6a422f35891b09fc5c31d52cc548f5dbe4