Add sort-whole-extends-list FileFormat property
The legacy behavior of always outputting the first type in an
interface's extends list if it is visible first and only sorting the
remaining types in the list is confusing and can lead to instability
in the API as a change to the source to reorder the interfaces could
result in API changes even though there was no actual change in the
API. This change adds the `sort-whole-extends-list` file format
property to allow that behavior to be configured on a per file basis
to allow the API signature files to be migrated.

Setting `sort-whole-extends-list` will sort the whole list using a
total ordering to ensure the result is independent of the source
order.

Bug: 307926200
Test: ./gradlew
Change-Id: Id279ad2ad2a471c342518676dfc40c6d49a19796