Add total order comparators for TypeItem and ClassItem
Previously, the `TypeItem.comparator` only defined a partial order over
`TypeItem`s using the `fullName` which meant that if two items matched
on `fullName` then source order (or at least, input to the sort order)
would be used. That makes the API signature less stable as changes to
the sources which do not actually change the API would change the API
signature files, and other generated files like stubs and jdiff files.

This change adds a total ordering comparator to `TypeItem` and uses it
in `JDiffXmlWriter` and renames the `TypeItem.comparator` to
`partialOrderComparator` and deprecates it. It also cleans up the
comparators in [ClassItem].

It also adds a test to show the result of partial ordering.

Follow up changes will fix the other uses of the comparator.

Bug: 307926200
Test: ./gradlew
Change-Id: Ia9131a81590877574eb18f2a0914272332fedc5c