Reduce complexity in the test runner mechanism
The `test_runner_handler` module is a collection of static functions
that instantiate `TestRunner` instances every time we need to call one
of the interface methods. This makes it difficult to store and reuse
state in runner implementations. This also is error prone since
different call sites create the runners with different constructor
parameters.

This change modifies the runner mechanism to use class instances and
consolidates the instantiation point to a single location.

Test: atest-dev hello_world_test
Test: atest-dev hello_world_test --host
Test: atest-dev --test-mapping tools/tradefederation/core
Test: atest-dev --test-mapping tools/tradefederation/core --device-only
Test: atest-dev --test-mapping tools/tradefederation/core --host
Test: atest-dev nearby_share_stress_test hello_world_test
Change-Id: Ia3a2422d0de653e61a6e561e13e5debcf6ef7033