Autoupdate of 2 submodules
*device/google/cuttlefish updated to 6105c00e9600fb6b871dba44e1c0bc15d07c6308
Merge "keymint: add in sha-256 impl" into main
*system/keymint updated to 8339014fc42130ea56b03bf56911d814e3b0bf45
Use hash of vbkey in externally visible places

Add a new crypto trait for performing SHA-256, and add a BoringSSL based
implementation of that trait.

Add a helper method that applies this trait to the verified boot key in
the boot info, but only if it is bigger than 32 bytes.  This allows for
systems where the bootloader already reports a hashed boot key.

Use this helper in two places that expose the verified boot public key:
- generation of attestation extension
- transfer of root-of-trust.

Bug: 305103215
Test: VtsAidlKeyMintTargetTest
Change-Id: I3b4fee005f08ca2e7703a3d40fa5997c01145a31