Switch test suite to inject parameters into fields
Previously, the test suite passed parameters into the constructors.
That required each test class to pass the `TestParameters` from its
constructor to te `BaseModelTest` class. That was even more complex
when parameterizing on a second dimension. This change switches to
injecting the parameters into fields. It makes the `BaseModelTest`
a little more complicated as it cannot rely on the field being set
before initialization but does simplify the rest of the classes.

Bug: 302290849
Test: ./gradlew
Change-Id: I80642aaa0ffead7ab649875906f45ac3d989bf5a