Simplify tests that are parameterized in two dimensions
Tests like `CommonErasedTypeStringTest` are parameterized in two
dimensions, the first is by runner as returned by
`BaseModelTest.testParameters` and the second is by its own collection
of data. Previously, they were combined into a custom data class. This
change removes that and uses an `Array<Any>`. That works better with
the follow up change to switch to using field injection.

Bug: 302290849
Test: ./gradlew
Change-Id: Iba2e049bd75e5a14e52b2c97344db40606254190