Generalize FlaggedApiTest.checkFlaggedApis
Previously, it had separate parameters for different surfaces and with
and without flagged APIs as needed. That made it difficult to add tests
for other surfaces (e.g. module lib) and required each test to check
the same surface/flagged combinations.

This change extracts the `Expectations` used internally, into an
`expectationsList` parameter that the tests provide. It will then find
the version of that which matches the current test configuration and
use that if available, or just skip the test otherwise.

Bug: 306242007
Test: ./gradlew
Change-Id: Id91f68bcbe73eaa9bfccdb9607ca13e9e43ad066