Add test to show incorrect stubs when hiding flagged apis
Hiding the flagged APIs should prevent those APIs from appearing in any
API surface but it does not hide them from an extending API surface.
e,g. a system API that is hidden does not appear in the system API but
does appear in the module-lib API that extends it.

Bug: 306242007
Test: ./gradlew
Change-Id: I870eed4e132cb2b8c4aee66875fa4fedd0e6beed