Return ServiceSpecificException from Netd mdns service
UnsupportedOperationException is not caught in MDnsManager.java. It
will throw the exception and cause a fatal exception on system_server.
The exception will be caught by MDnsManager in 2024-02 module release.
But the latest droidfood image uses a legacy mainline prebuilt with
the latest platform code. Returning an exception that the MDnsManager
can catch is required.

Bug: 312092770
Bug: 312932263
Bug: 298594687
Test: atest CtsNetTestCasesMaxTargetSdk30, no fatal exception
Change-Id: Ib61c0fe35e14378e002d4568590175617a06838f