Group AnnotationFilterEntry by qualifiedName
Previously, if the same annotation class was used multiple times with
the same annotation pattern option, then they would be treated
separately which made it difficult to add checks to ensure that the set
of patterns for a single annotation class made sense.

This change groups the `AnnotationFilterEntry`s in an
`AnnotationFilter` by qualified name making that easier and making
matching when there are lots of entries in a filter more efficient.

Bug: 313398274
Test: ./gradlew
Change-Id: I3c89521d54fe551554b1102130338632296ef65f