Sandbox apexd test genrules
gen_corrupt_rebootless_apex already worked with sandboxing.

gen_key_mismatch_capex had a diff in an "original_apex" file, which
I beleive is an unused output of the apex listed in the sources.
This original_apex was in $(genDir), so when we zipped up gendir we
included it. Instead, create a temporary directory to unzip/rezip the
input apex to, so that the rezipping doesn't also pick up the
original_apex.

Other genrules in this directory solve the problem by rm'ing
original_apex.

Bug: 307824623
Test: ./build/soong/tests/genrule_sandbox_test.py gen_corrupt_rebootless_apex gen_key_mismatch_capex
Change-Id: I3f5976da5bc2b4f257b3abad0442a4220af608ec