Make AnnotationFilter.getIncludedAnnotationNames return a Set
Previously, it returned a `List` which worked ok because there were no
repetitions of the same annotation with different values in the same
filter. However, in future there will be lost of repetitions of
`android.annotation.FlaggedApi("<feature flag>")` to expose the APIs
of different feature flags.

This change switches it to a sorted set because having a consistent
and stable order does matter.

Bug: 313398274
Test: ./gradlew
Change-Id: I6b1e54d1f5251284f4d1135e9ab7235b4cf0e8ad