Autoupdate of 2 submodules
*build/soong updated to 122767d27cc6ebe590f2a8cf0af040a5a3248b52
Merge "Sandbox apexd test genrules" into main
*system/apex updated to d52a47dfdfb7829a111d70ee3fe0b3f9e01eea63
Sandbox apexd test genrules

gen_corrupt_rebootless_apex already worked with sandboxing.

gen_key_mismatch_capex had a diff in an "original_apex" file, which
I beleive is an unused output of the apex listed in the sources.
This original_apex was in $(genDir), so when we zipped up gendir we
included it. Instead, create a temporary directory to unzip/rezip the
input apex to, so that the rezipping doesn't also pick up the
original_apex.

Other genrules in this directory solve the problem by rm'ing
original_apex.

Bug: 307824623
Test: ./build/soong/tests/genrule_sandbox_test.py gen_corrupt_rebootless_apex gen_key_mismatch_capex
Change-Id: I3f5976da5bc2b4f257b3abad0442a4220af608ec