Refactor the handling of the released-flagged-apis.txt
Previously, it was intended to just be a list of Metalava options that
are passed to Metalava when building SDK snapshots that need to control
which flagged APIs, if any, are included in the snapshot. However, that
would be quite tedious to maintain with lots of duplication and direct
dependency on Metalava making it useless for other purposes. This
change switches it to be a list of flags to release and a script will
convert that into Metalava options.

It also adds `com.android.net.flags.register_nsd_offload_engine_api` to
the set of flags which were previously released in SDK extension #10.

Ignore-AOSP-First: Needed in `udc-mainline-prod` only at the moment.

Bug: 313398274
Test: UNBUNDLED_BUILD_SDKS_FROM_SOURCE=true vendor/google/build/mainline_modules_sdks.sh --build-release=next
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:b4d70a6bdd5d35544a0f1d25adf27c0ac9d0e968)
Merged-In: I5e9c00175b8c32eea0ad7fdfb30ce19487f286cb
Change-Id: I5e9c00175b8c32eea0ad7fdfb30ce19487f286cb