Don't use symlinks in genrules
A change is being made to sbox so that symlinks aren't resolved
before being copied into the sandbox. This requires that the targets
of the symlinks are also added to the sandbox.

In this case we can depend directly on the file that the symlink is
pointing to instead of the symlink itself.

Bug: 307824623
Test: m art-run-test-2263-method-trace-jit-expected-stdout with aosp/2850015
Change-Id: Ifd6e8f9a95bff5c91779a1751342ea9b78f54607