Remove workarounds for pore users.
It turns out I only needed this because pore's repo compat wasn't
quite complete. I fixed that with
https://github.com/jmgao/pore/pull/109, so once that lands we don't
need this cruft any more. I verified that I can still run
`./updater.sh update --no-upload rust/crates/android_logger` in both a
repo and a pore tree (the update itself fails, but because of a merge
conflict with the patches to be applied).

Bug: None
Test: ./updater.sh update --no-upload rust/crates/android_logger
Change-Id: I017b47f70694d15d516ccbe289c85f90f209127d