[Thread Backport] Exempt thread package from Android TV HAL VTS test.
Since ThreadNetwork AOSP AIDL interface will be backported on the new
TV device, we want to exempt the package from the VTS test needed for
Android 14 certification.

The implementation is referenced from aosp/1178808.

Use `android.software.leanback` to indicate that the device is for
Android TV instead of `android.hardware.type.television` as it is
deprecated.
Ref: https://developer.android.com/guide/topics/manifest/uses-feature-element#device-ui-hw-features

Bug: 314024337
Test: Run vts_treble_vintf_vendor_test on the target device.
Change-Id: I873cac6fc7f2aa883ef965351d8c0052694eeab6