Build and update the device in test module execution plan
Add a flag that builds the `sync` target and updates the device with
`adevice` before running tests. This is a commonly used workflow and
eliminates a few steps for developers that would have otherwise needed
to edit, build, update, and then test.

This feature also coalesces the product and test build commands to
eliminate the build system's startup cost of about 10s.

Note that this feature cannot currently handle updates that exceed the
configured `adevice` max file limit.

Test: atest-dev --update-device hello_world_test
Test: atest-dev hello_world_test
Bug: 314339352
Change-Id: I170ff1f1cdeed2fc2c83a5e3b78eda11a789e856