Dedup Predicate<Item> creation in ApiType and ApiLint
The `ApiType` class creates a number of `Predicate<item>`s and combines
them in a number of different ways. Previously, the `PUBLIC_API` and
`REMOVED` both duplicated the contents of `getReferenceFilter()` in
their `getEmitFilter()` and `PUBLIC_API` and `ApiLint` both duplicated
the value of `apiFilter` from `PUBLIC_API.getEmitFilter()`.

This change dedups that by:
1. Calling `getReferenceFilter()` from `getEmitFilter()`.
2. Extracting `getNonElidingFilter()` from `getEmitFilter()`.
3. Calling `getNonElidingFilter()` from `ApiLint`.

Bug: 311690008
Test: ./gradlew
Change-Id: I7482720711f66eea5c4ee22e92c7293a4cc37c5a