Only export unwindstack:: functions.
Since the compressed section support requires some pulling in of
other whole libraries, make sure we only export the unwindstack
functions.

This change also decreases the size of libunwindstack.so by ~100,000
bytes on 64 bit and ~70,000 bytes on 32 bit.

The change to add compression will increase the size back to about
the original.

Bug: 309857311

Test: Builds.
Test: Ran nm -CD on the library and verified only unwindstack:: are exported.
Change-Id: I0035f5f18a6a3c5188584a0d66c6bd18bcbe5a30