Add error when in-development aidl_interface module is frozen: true
If an AIDL interface has_development is true, then it can't be frozen.

This is an uncommon state, because building a modified aidl_interface
module that is marked `frozen: true` will fail to build until
`m foo-update-api` is run to update the current aidl_api dump. This will
modify the aidl_interface module with `frozen: false`.
So we only end up in this state if the aidl_interface change in the
Android.bp file is removed after `m foo-update-api` is run.

Test: add fake new method to the frozen android.hardware.biometrics.common
      $ m android.hardware.biometrics.common-update-api
      git restore Android.bp
      $ m
      # observe failure
Bug: 314305674
Change-Id: I50d27342c0e130411c73833eca822a574702c910