Revert^2 "Only export unwindstack:: functions."
Since the compressed section support requires some pulling in of
other whole libraries, make sure we only export the unwindstack
functions.

This change also decreases the size of libunwindstack.so by ~100,000
bytes on 64 bit and ~70,000 bytes on 32 bit.

The change to add compression will increase the size back to about
the original.

The reason this is a revert^2 is the previous change moved the AsmGetRegs
function to a new name. This caused problems where an art test tries to
load a perfetto library that uses that function. So you get the art
apex libunwindstack.so that is missing the symbol, and the test cannot
link.

I thought that AsmGetRegs was only used internally, but that's not true.
But this only exists for x86/x86_64 since it's an inline for all other
architectures.

Bug: 309857311

Test: Builds.
Test: Ran nm -CD on the library and verified only unwindstack:: and AsmGetRegs
Test: are exported.
Change-Id: I8be50292e10839e6db17f4dcdc729fa2ccdcfee4