Remove TextMemberItem.modifiers property
`TextMemberItem` extends `TextItem` which already has a `modifiers`
property which is initialized in the super constructor call in
`TextMemberItem` so there is no need to have a
`TextMemberItem.modifiers` property as well. In fact having one makes
it impossible to use the `modifiers` property in `TextItem`
initialization as when that gets the value it will try and retrieve the
value of `TextMemberItem.modifiers` which will be `null` as it will
only be set in `TextMemberItem` initialization which occurs after
`TextItem` initialization.

So, this change removes it.

Bug: 315206833
Test: ./gradlew
Change-Id: Id6b3a1bb61d6d757008e9c645d0243716f4ccf96