Remove overrides of TurbineItem codebase and modifiers properties
Classes which extend `TurbineItem` do not need to provide their own
overrides of its `codebase` and `modifiers` property as they already
pass them to its constructor. Overriding them also prevents their use
in `TurbineItem` initialization as it will attempt to retrieve the
value from the subclass override which will not yet have been
initialized.

So, this change removes the overrides. It also ensures consistency in
the structure of the super call to `TurbineItem` constructor and the
subclass of `Item` that is implemented.

Bug: 315206833
Test: ./gradlew
Change-Id: I74377063c533432967e29423c446d79d966fe068