Support @Deprecated on PsiParameterItem
Previously, the `PsiModifierItem` would correctly determine whether a
parameter was deprecated for Kotlin but not for Java. This change adds
a check for the presence of the `Deprecated` annotation in the list of
`AnnotationItem`.

That revealed a number of other issues that also needed to be fixed as
part of this:
* Usually, the `@Deprecated` annotation MUST be accompanied by an
  `@deprecated` javadoc tag for Java but there is no way to specify
  that for a parameter so that check was disabled for parameters.
* A number of tests failed because they started writing out
  `@Deprecated` on parameters. They were fixed by only writing out
  `@Deprecated` on parameters that were explicitly deprecated.

Bug: 315206833
Test: ./gradlew
Change-Id: I6380931c0d2b1e1adbef05065268963c97dc9ed4