Allow adding and removing type-use annotations from modifiers
This is set as unsupported for the text model because text type parsing
is cached, so one text type may be shared between several API elements.
This makes it unsafe to mutate the modifiers. If there is a later need
for mutating text type modifiers, the initialization of text types will
need to be updated.

Bug: 300081840
Test: CommonTypeModifiersTest
Change-Id: Ie99e538b58b3adfa76cb2161859f36383908c01b