Add test for inheriting into deprecated class
Inheriting a method from a hidden but not deprecated class into a
visible but deprecated class should cause the method to be treated as
deprecated. This change adds a test to verify that.

This tests behavior relied upon by the Android build and ensures that
follow up changes do not break it.

Bug: 315206833
Test: ./gradlew
Change-Id: I483f74cd6840c2896e8efc0ae0447825fdb90ffa