Treat DeprecatedForSdk as Deprecated when creating PsiModifierItem
Previously, the `DeprecatedForSdk` processing was done during
`ApiAnalyzer.performChecks` and would only set the `Item.deprecated`
property to `true`. This change takes advantage of the fact that any
`DeprecatedForSdk` annotations that do not apply to the API surface
being generated are filtered out before construction of
`PsiModifierItem` to treat `DeprecatedForSdk` like a normal
`Deprecated` annotation and call `DefaultModifierList.setDeprecated()`
if present.

That allows the later setting of `Item.deprecated` to be removed and
ensures that `DeprecatedForSdk` propagates through the hierarchy
in the same way as `Deprecated`.

Bug: 315206833
Test: ./gradlew
Change-Id: Ie404a81cc6a8078b744fe23b0766577d1c80b0ca