Use effectivelyDeprecated when writing api-versions.xml files
Previously, the `Item.deprecated` property was used to determine the
deprecated status when extracting information from a `Codebase` for use
in an `api-version.xml` file. There are two different sources for those
`Codebase`s:

1. Source files, in which case `ApiAnalyzer.performChecks()` is called
   and so `Item.deprecated` and `Item.effectivelyDeprecated` are
   equivalent.

2. API signature files, in which case while
   `ApiAnalyzer.performChecks()` is not performed after reading it was
   called on the `Codebase` that was written to the API signature files
   and so again `Item.deprecated` and `Item.effectivelyDeprecated` are
   equivalent.

When creating `api-versions.xml` files from jar files the jar files are
read using `JarReaderUtils` and not as a `Codebase` and so this has no
effect on that.

Bug: 315206833
Test: ./gradlew
Change-Id: I313284aad0043f29102ea9ecb9f1f417fcefdd6a