Use originallyDeprecated when writing JDiff
Previously, the `Item.deprecated` property was used to determine
the deprecated status when writing a JDiff file. However, as the file
is written out from a `Codebase` that has not had
`ApiAnalyzer.performChecks()` run on it then the `deprecated` property
has not been updated based on the containing class and so is still set
to its initial value of `originallyDeprecated`. So, this change uses
that directly.

Bug: 315206833
Test: ./gradlew
Change-Id: Ib82900ffa388b57a661eb6eb9c82b7e047a74be9