Disable Car tests that require stubgraph for riscv
There is no implementation of libstubgraphimpl.so for RISC-V and we
don't plan on creating new variants for this purpose yet. This can be
re-visited if we ever need to enable computepipe in the future, but
there's no use-case for the forseeable future.

Disable the test that requires this library, but only for the RISC-V
architecture, so that 'device-tests' builds properly.

Bug: 291108108
Test: lunch aosp_cf_x86_64_phone-staging-userdebug && m device-tests
Change-Id: Ida838abbc74154e40b58acdfd8bd54cc3b60919f
Merged-In: Ida838abbc74154e40b58acdfd8bd54cc3b60919f