Always inline values for FlaggedApi annotations
Annotation values that reference constant static fields that are
removed or hidden or not public have the values of those fields inlined
into the signature files so as to not add a dependency onto a class
that is not part of the API. Unfortunately, that does not work for the
way that `FlaggedApi` annotations are used because while the field that
is referenced is marked as `/** @hide */` it is provided as a `.class`
file so the javadoc is missing.

This change works around that issue by always inlining the value for
`FlaggedApi` annotations, where possible. It also modifies an existing
test to verify that behavior.

Bug: 314319629
Test: ./gradlew
Change-Id: Id59d51ea3ceabe520147a4391460a1033c28b56f