Rename thread that does OOM perfetto dump
The perfetto hprof plugin is invoked in two ways:
1. The user asks for a heap dump, perfetto send a signal to the process.
   The thread "perfetto_hprof_listener" receives the thread, it forks
   and dump the graph.
2. On OOM, the process forks and starts dumping the data to perfetto.

In the first case, the thread that does the dump is called
"perfetto_hprof_" (the first 15 chars of "perfetto_hprof_listener"),
which is set by art::SetThreadName called by AttachCurrentThread.

In the second case, the thread that does the dump doesn't have a
specific name.

This commit sets a name for the thread that does the dump in the OOM
case. Since only the first 15 characters will be used the thread will be
called "perfetto_oome_h".

Change-Id: I5440f838ceedbfafb06179732b02ed0786e8944f