Add --revert-annotation for use with @FlaggedApi
The `--hide-annotation` option no longer matches the expected semantics
needed for `@FlaggedApi`, so add a `--revert-annotation` option to
replace it. The `--hide-annotation` will continue to be treated as
`--revert-annotation` when passed the `android.annotation.FlaggedApi`
annotation until the build has been updated at which point support for
that will be removed.

Bug: 314196587
Test: ./gradlew
Change-Id: I3d265d7c69825a62fe67706f619339756c9588ca