Allow for revert annotation to differ in behavior by item
A revert annotation that is applied to an API that existed in the
previously released API will be treated as a show annotation but
when applied to an API that did not exist in the previously released
API will be treated as a hide annotation. However, it is not a show
annotation. Previously the `UnhiddenSystemApi` check assumed that if
`item.isShowAnnotation()` returned true then
`item.modifiers.annotations` would include at least one annotation for
which `AnnotationItem.isShowAnnotation()` would return `true` but with
revert annotations that is no longer true. This change prevents the
check from throwing an exception when the revert annotation is treated
as a show annotation on an item.

Bug: 314196587
Test: ./gradlew
Change-Id: If062dec93caedffa465edbd115d66b7fb9e1116a