Provide previously released API to DefaultAnnotationManager
In order to support the revert semantics of `--revert-annotation` it is
necessary for `DefaultAnnotationManager` to retrieve the previously
released APIs (both current and removed) to determine what reverting a
specific item actually means. This change provides the `Codebase`s for
them to `DefaultAnnotationManager`.

There was a bit of a catch-22 situation because in order to load the
APIs as `Codebase`s, reusing existing `Codebase` when available, the
`DefaultAnnotationManager` needed to use `SignatureFileCache` but
`SignatureFileCache` depends on `DefaultAnnotationManager`.

The solution was to lazily initialize the `Codebase`s and pass
providers for them rather than the `Codebase`s themselves. It was also
necessary to move `SignatureFileCache` into `Options` so that the same
instance could be used for initializing the `Codebase`s in addition to
its current use in `Driver.kt`.

Bug: 314196587
Test: ./gradlew
Change-Id: I5cebd18748cf81d6813f5626942ae1efb9c0de71