Support caching of revert annotations
The previous change that added support for `--revert-annotation` did
not add the revert annotations to the set of all filters used to
compute the `annotationNameToKeyFactory` which meant that the cache
would just use the name of any revert annotations as the key, ignoring
the values which meant that all `FlaggedApi` annotations would be
treated the same, irrespective of the flag name passed.

This change corrects that.

Bug: 314196587
Test: ./gradlew
Change-Id: Idb00f3ea27891426733e9a0b668c0d50fb5bac9f