Merge annotations from multiple java files for the same class
Previously, all the `.java` files for all the annotation directories
were merged together into a single set of sources before processing
into a `SourceCodebase` which mean that if multiple `.java` files
provided annotations for a class then only one of the files would be
used.

This change merges each annotation directory separately so that the
annotations will be merged from all `.java` files.

It also removes `@test.annotation.Hide` from the `cShown` method in the
first java file from which annotations will be merged. It could not do
that before because it would result in an error as all items with show
annotations must be hidden but the hide annotation from the second java
file was being ignored.

Bug: 316212602
Test: ./gradlew
Change-Id: I88b02d19fbb8cc6d21816ec4cf9c9ea37664c580