Only copy hidden state with annotations
Previously, `mergeInclusionAnnotationsFromCodebase` would copy the
`hidden` and `originallyHidden` states from an item in the external
codebase even if it did not have any annotations. That is not the
purpose of that method.

This change first gets the list of annotations to copy and only if
it is not empty will it then copy the annotations and any `hidden`
or `originallyHidden` state.

Bug: 316212602
Test: ./gradlew
Change-Id: I2afd71a73ecccb5038822347c1009a4833df780d