Use TraversingVisitor for merging inclusion annotations
Previously, it used a `CodebaseComparator` to perform the merging of
inclusion annotations from the external codebase. That was a very
inefficient way to complete that task. The `CodebaseComparator` creates
a special tree structure for each `Codebase` it is merging and then
compares that structure. It also has support for detecting when
differences between the `Codebase`s is due to fields/methods being
moved up/down the class hierarchy and then synthesize temporary
`FieldItem` and `MethodItem` objects to ignore those differences.

All of that is unnecessary for merging inclusion annotations. This
change replaces the use of the `CodebaseComparator` with a simple
`TraversingVisitor` that is used to visit the items in the external
`Codebase` which will be small compared to the main `Codebase`. It will
find the item in the main `Codebase` corresponding to the item in the
external `Codebase` and copy the annotations and `hidden` and
`originallyHidden` state across.

Bug: 316212602
Test: ./gradlew
Change-Id: I9f363bcbcaac417a5d2dcc7c0c14d39c883a06e8