Autoupdate of 2 submodules
*art updated to 0630862d505c0ef2168690199ee0f0ac17193abc
Add the targets of symlinks to compile_data

sbox is being updated to sandbox symlinks, meaning you have to declare
the target of the symlink in the android.bp file to avoid a dangling
symlink when the action runs.

Do that for the existing art tests that use symlinks.

Bug: 307824623
Test: m lint-check
Change-Id: I51f52f24abb8fe15ab84d6d8233a01f137a3cd70

*build/soong updated to c01104234ef69d13c2360c0434b14ea13f2fecf5
Merge "Allow compile_data on all java modules" into main