Avoid generating source representation of annotation when matching
Previously, when matching an `AnnotationItem` the `AnnotationFilter`
would first check to see if its `qualifiedName` was `null` or the
filter was empty and if either was `true` then return `false`.
Otherwise, it would generate the source representation of the
`AnnotationItem` and match against that.

The problem with that is generating the source representation of an
annotation could involve resolving any field references in the
annotation to their literal value and that could require checking
showability which could re-enter the filter matching. If that
encountered a cycle then it would result in a stack overflow.

This change avoids that by first checking whether an annotation's
qualified name was in the set that the filter could match before
generating the source representation. That means only a very limited
number of annotations will ever resolve their values which greatly
reduces the chances of cycles. Although, it probably does not
completely eliminate them.

Bug: 313398274
Test: ./gradlew
Change-Id: Ieaf04cd2da8b05b4d76a2cc8a2271b57816e79a6