Autoupdate of 2 submodules
*external/uwb updated to 6f937c588488a14d217a74af2e3a9c87af8642f9
Fragment UCI DATA SND packets based upon UWBS capabilities.

Currently the data packets sent by Host to UWBS are fragmented with a
fixed payload size of 255 bytes.

This CL changes the behavior to use the UWBS capability value "MAX_DATA_PACKET_PAYLOAD_SIZE" to determine
the max payload size of each data packet fragment. It will still use the
value 255 as the default (in case the capability couldn't be read from
the UWBS).

Test: atest libuwb_uci_jni_rust_tests libuwb_core_tests libuwb_uci_packet_tests
Bug: 301834443
Change-Id: I36866609ac8cbc091d66a134a17f923d7d330232

*packages/modules/Uwb updated to 48530f034e533f01c05098708778588f0e9f1d93
Define new rust library target for unit tests.

Define a new library target that compiles the AConfig flags in "test"
mode, for the Rust unit tests. With this mode, the library contains
setter methods, that can be used in unit tests to set the flag values.

Bug: 30183443, 315194046
Test: atest libuwb_uci_jni_rust_tests libuwb_core_tests libuwb_uci_packet_tests
Change-Id: If7c0213f4efc49ee27ad2fbcb2183dae4e9233a6