Add baseline testing to VisiblySynchronizedTest
Previously, it was not possible to test a mixture of baseline and
issues being reported in `DriverTest` because testing baseline
automatically switched it to `silentUpdate=true` mode which suppresses
issues from being reported.

This change allows a test to disable `silentUpdate` and makes use of
it in `VisiblySynchronizedTest`. That test was chosen because of some
issues found while testing on Android build. The baseline interacts
with `VisiblySynchronized` issues in a nasty way.

That reveals a few issues with the test and baseline handling:
1. The `CheckSynchronization.errorMethod2()` was repeated and while
   despite that it previously reported an issue with both, switching
   on baseline testing suppressed the second. So, the test was fixed
   to make it more realistic and stop duplicating `errorMethod2()`.
2. Some other issues were suppressed when they should not be.
3. The element id of some of the baseline entries are very generic
   and expose some implementation details.

Follow up changes will fix the issues.

Bug: 316456211
Test: ./gradlew
Change-Id: I96958773fed2b34e7329e7eb272fdaf565ee5d66