Correct reporting when both item and location are provided
Previously, when both `item` and `location` were provided it would
store both in the baseline which is not what was expected. That had two
effects on `VisiblySynchronized` issues (the only issue that uses
both):
1. If two issue reports were caused by syncing on the same object, e.g.
   `this` or `Foo.class` then the second one would be suppressed in the
   output (but both would be recorded in the baseline).
2. The location would be recorded in the baseline using some Psi
   implementation string.

This change ensures that either `item` or `location` is used but not
both. It also restructures the expressions for getting the location for
reporting issues and the location for baselining issues and makes it
clear why they behave slightly differently.

Bug: 316456211
Test: ./gradlew
Change-Id: Ibdcefe48e3daf076a23e24872335702110dc60a8