Uptesting is moving folks to /external/robolectric from /external/robolectric-shadows
We maintain /external/robolectric (keep it in sync with github/google3) and are getting ready to delete /external/robolectric-shadows
It looks as if these tests co-incided with changes to robolectric shadows that were never attempted to be upstreamed.
These shadows have been pulled out and made use of for just these tests, and made to live next to the tests themselves.

Bug: 314757990
Test: atest, 8 failing before upgrade, 11 after.
Change-Id: I0f67d04af747407d663962a7a1f3f46dc01a4c23