sign_virt_apex: add --do_not_validate_avb_version flag
The CTS binary for 24Q1 release is built from the different branch that
the system image. Unfortunately these 2 branches have a different avb
version, which results in MicrodroidHostTests failures.

This patch adds a flag to skip the avb_version validation when running
the MicrodroidHostTests. It is just a workaround for CTS failures.

Bug: 315405099
Test: atest MicrodroidHostTestCases
Change-Id: Ice9457428575c6d0c91894e4ec67c3aab673c192