Autoupdate of 2 submodules
*build/blueprint updated to ed49204e850f58556da5e95a0b6fff1b7dfdbf18
Use generics for providers API

Using generics for the providers API allows a type to be associated
with a ProviderKey, resulting in a type-safe API without that doesn't
require runtime type assertions by every caller.

Unfortunately, Go does not allow generic types in methods, only in
functions [1].  This prevents a type-safe API on ModuleContext, and
requires moving the API to be functions that take a ModuleContext as
a parameter.

[1] https://go.googlesource.com/proposal/+/refs/heads/master/design/43651-type-parameters.md#no-parameterized-methods)

Bug: 316410648
Test: provider_test.go
Change-Id: Ide91de9f2a2a7d075b05e287c7cc86b395db0edb

*system/tools/aidl updated to d43fe2497f82b431c37ee2ffcee5e7e2e6a123d8
Update system/tools/aidl to use generic providers API

Update system/tools/aidl to use the new type-safe providers API.

Bug: 316410648
Test: builds
Change-Id: Id99c7e494a1260df32e9d3e235d83d13b05f9f0f