ANDROID: disable readonly memslot usage on aarch64 (for pKVM)
crosvm started using readonly memslots to implement an optimization in
https://crrev.com/c/4891756, however, pKVM doesn't support them. It
causes VM boot to fail. Since there isn't an immediately obvious way to
detect pKVM, I'm disabling it for all aarch64 builds in AOSP. Once we
find a better way, we can move the fix upstream and revert this commit.

Bug: 316956218, 307396704
Test: .
Change-Id: Ie30526f0ff17d2c804fe4344659b325b8da58d20