Add RestorePerModule to always restore device
Add flag to always restore the device before every module.
This is strictly for development.
We would like to gather more data with restore, and currently a restore
only occurs when the device crashes.
This flag will allow us to restore before every module, providing more
data on the behavior and stability of the device post-restore.

Bug: 303845019
Test: m tradefed-all
Change-Id: I8778e9ef1ea3ea865f7aeea3951eb17de4fb9b19