Optionally skip constraint from a dice policy
If optional is set to true for a ConstraintSpec, skip it if the
key/value is missing in the dice chain.

Why? Some constraints (such as security_version/subcomponent's versions)
are not expected to be present in each dice node of all whole chain.
This parameter can bew used to construct appropriate spec.

Test: Added unit tests.
Test: Test on the given compos chain for security_version
Bug: 291238565
Change-Id: I2cc7f355bfc0bdbbd7fa08ab84b9d0d828084ff4