[x509] Reduce serial number array length to 16 bytes
This cl reduces the length of the x509 certificate's serial number
array from 20 bytes to 16 bytes to circumvent en encoding issue
related to MSB.

During the encoding process, a leading zero can be added to the
array when MSB is 1. Consequently, a 20-byte serial number could
exceed the maximum length of 20 bytes defined in RFC 5280 if the
first element of the serial number array is greater than 128.

Bug: 317065113
Test: atest rialto_test --iterations 10
Change-Id: Iac1655182d1706659f88afcf4ba0c0700c918a8d