tp: add a simple wrapper binary around RPC protocol
This CL adds a very simple wrapper binary around the trace processor RPC
protocol; all this bianry does is read stdin, push the bytes to the rpc
class and push the responses to stdout.

The purpose of this is to use this in Bigtrace in a gVisor sandbox: the
simplicity is a boon as it makes random failures inside the sandbox much
less likely.

Also add a simple test just to make sure there are no obvious errors.

Change-Id: I421ab958d0e75c3e1b4c9eec3f22b8c2b13243ed