Autoupdate of 14 submodules
*device/google/atv updated to 766738ac1f10dd4f7cac8531dfa4c5f045f1cd30
List baseline_filename on modules that are implititly using it

lintable modules currently pick up files named "lint-baseline.xml" to use as the
lint baseline implicitly. This is confusing because you could end up using the
baseline files in more modules than intended. Lint also has a feature where it
requests you remove unnecessary findings from the baseline file, so something
could be necessary for one module, but unnecessary for another that accidentally
picked up the baseline.

We're removing the implicit detection of the baseline file, which requires
all modules using it to list the baseline file explicitly.

Bug: 272769514
Test: Presubmits
Change-Id: I606d8cdfeb01c1ff42141ae37bfd95f20c558211

*cts updated to 7907a6a0f43f589cea86f2365e9b5c19a60367ae
Merge "List baseline_filename on modules that are implititly using it" into main
*development updated to 7d17bb7fea853669401bf665c770067e267517ee
Merge "List baseline_filename on modules that are implititly using it" into main
*external/conscrypt updated to 10684f7d81c6bf993e80d89b3e01576ae0aa4d29
List baseline_filename on modules that are implititly using it

lintable modules currently pick up files named "lint-baseline.xml" to use as the
lint baseline implicitly. This is confusing because you could end up using the
baseline files in more modules than intended. Lint also has a feature where it
requests you remove unnecessary findings from the baseline file, so something
could be necessary for one module, but unnecessary for another that accidentally
picked up the baseline.

We're removing the implicit detection of the baseline file, which requires
all modules using it to list the baseline file explicitly.

Bug: 272769514
Test: Presubmits
Change-Id: I7ba1589d127f194fc6956c36b41dec2bde0d1d8b

*external/setupcompat updated to 6f1b1d3013fad63f768850953b971579061fca1b
List baseline_filename on modules that are implititly using it

lintable modules currently pick up files named "lint-baseline.xml" to use as the
lint baseline implicitly. This is confusing because you could end up using the
baseline files in more modules than intended. Lint also has a feature where it
requests you remove unnecessary findings from the baseline file, so something
could be necessary for one module, but unnecessary for another that accidentally
picked up the baseline.

We're removing the implicit detection of the baseline file, which requires
all modules using it to list the baseline file explicitly.

Bug: 272769514
Test: Presubmits
Change-Id: Ibf84772602764d301da1f406d921c1b45d3b914d

*external/setupdesign updated to b07899ff0ae407963d9c745cf97c994ca06f8c29
List baseline_filename on modules that are implititly using it

lintable modules currently pick up files named "lint-baseline.xml" to use as the
lint baseline implicitly. This is confusing because you could end up using the
baseline files in more modules than intended. Lint also has a feature where it
requests you remove unnecessary findings from the baseline file, so something
could be necessary for one module, but unnecessary for another that accidentally
picked up the baseline.

We're removing the implicit detection of the baseline file, which requires
all modules using it to list the baseline file explicitly.

Bug: 272769514
Test: Presubmits
Change-Id: Ia0fd37552da2f660b89a95833384253d4acd5ef5

*packages/apps/Contacts updated to 72cc4897fa7aef52abf41b9a569cf576b866b92e
Merge "List baseline_filename on modules that are implititly using it" into main
*packages/apps/DevCamera updated to d4b6231508ace2312004a5520a4e21658b8b1fde
List baseline_filename on modules that are implititly using it

lintable modules currently pick up files named "lint-baseline.xml" to use as the
lint baseline implicitly. This is confusing because you could end up using the
baseline files in more modules than intended. Lint also has a feature where it
requests you remove unnecessary findings from the baseline file, so something
could be necessary for one module, but unnecessary for another that accidentally
picked up the baseline.

We're removing the implicit detection of the baseline file, which requires
all modules using it to list the baseline file explicitly.

Bug: 272769514
Test: Presubmits
Change-Id: Ie029f595ad81a101b167d35fc421412af6aedd75

*packages/apps/DocumentsUI updated to 563afa53b1f7c17a7fecc1a0f03411c0cac8d108
List baseline_filename on modules that are implititly using it

lintable modules currently pick up files named "lint-baseline.xml" to use as the
lint baseline implicitly. This is confusing because you could end up using the
baseline files in more modules than intended. Lint also has a feature where it
requests you remove unnecessary findings from the baseline file, so something
could be necessary for one module, but unnecessary for another that accidentally
picked up the baseline.

We're removing the implicit detection of the baseline file, which requires
all modules using it to list the baseline file explicitly.

Bug: 272769514
Test: Presubmits
Change-Id: I896dee62cc8931fea7a5aae0d62f5da527a9f949

*packages/apps/TV updated to cc05a7ada65bdbbed4eecf7f4ca8d8d29282ee9e
List baseline_filename on modules that are implititly using it

lintable modules currently pick up files named "lint-baseline.xml" to use as the
lint baseline implicitly. This is confusing because you could end up using the
baseline files in more modules than intended. Lint also has a feature where it
requests you remove unnecessary findings from the baseline file, so something
could be necessary for one module, but unnecessary for another that accidentally
picked up the baseline.

We're removing the implicit detection of the baseline file, which requires
all modules using it to list the baseline file explicitly.

Bug: 272769514
Test: Presubmits
Change-Id: I8e33eea2bf3134ec6f9595bd17f362d22998e900

*packages/inputmethods/LatinIME updated to 66093bf509ea92fa31d796326d5f30a8d9582ffe
Merge "List baseline_filename on modules that are implititly using it" into main
*packages/modules/NetworkStack updated to e597d85c02bc069fbe8fa961bb357f1ae9bf02f3
Merge "List baseline_filename on modules that are implititly using it" into main
*packages/providers/MediaProvider updated to ec6af815a0976b22c87a930b876b090188f340a0
Merge "List baseline_filename on modules that are implititly using it" into main
*packages/services/Telephony updated to be44be63914950b7387bce9449a7fb34a5dafe8d
Merge "List baseline_filename on modules that are implititly using it" into main