always_allow_unfrozen
Add a way to allow unfrozen interfaces, even
when RELEASE_AIDL_USE_UNFROZEN is set. This
will avoid the build check and behavior changes,
but there would still be an associated failure
in vts_treble_vintf_vendor_test. Still, this
is requested by some partners and teams to ease
the development process.

Bug: 317387425
Test: added in CL (as soong build test)
Test: libaidlmetadata_test
Change-Id: I85532c719972cc971e7a595f118ba4acc9b24ba9