Always follow DecrementSuspendCount by broadcast
In one case we were not doing so; in another case we make it more
obvious that we are.

This appears to have been a long-standing issue, which may have been
hidden by the previous complex transitioning_to_runnable logic, and
was then exposed by a recent CL that deleted it:
https://android-review.googlesource.com/c/platform/art/+/2883289

Unfortunately, this looks quite random, and has low occurrence
probability. I don't think there is a real way to test for this
without somehow controlling scheduling.

Bug: 317036326
Bug: 297973401
Test: Host run tests, TreeHugger
Change-Id: I8f9b55d2edd878072f8486dc7e564ea901be5ec8