Consolidate Traceur permissions checks
This change consolidates the checks for verifying that Traceur usage is
allowed (developer options are enabled, the current user is an admin
user, and debugging features are not disallowed). This will help keep
all startup/permissions checks in the same place, in case they need to
be updated in the future.

This change also removes SafetyNet logging, which is not actively
monitored for Traceur.

Test: Verify that Traceur cannot be interacted with for each case:
      - Developer options are not enabled
      - The device is using a secondary/non-admin user
      - Debugging features are disabled with pm set-user-restriction
Bug: 276333488
Change-Id: I2aac4c457ba147595445c343e5ef7d842aaa4862