casimir: Remove Mutexes
Make all functions take &self mut so that we no longer need mutexes.
This is made possible by removing the usage of join in Controller::run
by reading channels and the TcpSocket while sleeping.

nci::Reader cancellation safety is provided using a Stream that will
persist the state of nci::Reader::read.
The usage of a dyn Stream makes it also more easily mockable in tests.

Test: None
Change-Id: I37b529ec8f94f046cd4f3b2e15c363436d14c080