Integration test: Add back ANDROID_HOST_OUT_TESTCASES env key
ANDROID_HOST_OUT_TESTCASES is deprecated and replaced by HOST_OUT_TESTCASES, and we can't get the value of it from soong_ui.bash which would throw an error. Therefore we merged a CL previously to replace it with HOST_OUT_TESTCASES. However, ANDROID_HOST_OUT_TESTCASES is still populated when we call lunch command and it is required by atest/tradefed to run the tests. Without it the test would run as pass but no actual test cases were excuted. This CL fixes this issue.

Test: tools/asuite/atest/integration_tests/atest_ci_tests.sh --artifacts_dir /tmp/a -b
Change-Id: I373079d95037c67144998c6bacf0e1ded083cc69