BACKPORT: lib: TST_EXP_FAIL: Add array variants
For certain cases there is a possibility of a failure with more than one
errno, for instance testcases with invalid fd may return either EINVAL
or EBADFD and in many cases either one is fine, at least that was the
feedback from kernel devs.

This change also adds a tst_errno_in_set() function that is now the
single place to validate errno for all TST_EXP_FAIL*() variants. That is
intentional since this allows us to implement code to relax the
conditions if needed, e.g. we had requests to allow additional errnos
for systems with SELinux where failures may be caused by the SELinux
policies and the errors may differ.

Change-Id: I1f695a85a10ac56cefcfd6e85ba432471b515ef4
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
(cherry picked from commit 1cfe6142898204890e61429a66bd782cfcf8c23c)
Test: atest vts_ltp_test_arm_64
Signed-off-by: Terence Tritton <ttritton@google.com>