Revert "Skip these lazy tests when HIDL is not supported"
It's now skipped when the services aren't registered, so this isn't
needed.

This reverts commit d0b8ff93f54d535433c662c590da6635f309b35d.
This keeps the original Android.bp changes in that commit that moves the
generated HIDL libraries from shared to static to allow the test to be
run on a device that doesn't include the services and their
dependencies.

Test: atest hidl_lazy_test
Bug: 218588089
Change-Id: I88c00c233b89dc0feecffbfafad29eef94cf94e9