Solve type conversion between char and int8_t
Apparently, while in amd64 Cuttlefish the sync() DittoBinder transaction
was working fine, in arm64 (tried in Pixel 7), some kind of type
conversion is transforming a signed integer into an unsigned integer.
This results in an error such as:

dittosuite/src/binder_request.cpp:35: FATAL: Wrong result, expected: -2, but got: 254

Apparently changing char to int8_t in the requester code is enough to
prevent this conversion.

Test: Pixel 7 and CF: dittobench example/android/binder.ditto
Bug: 318651715
Change-Id: Ic05d70d72419fef089287781c5eed8247bf05763
Signed-off-by: Alessio Balsini <balsini@google.com>