genvk: options.filename --> args.target
Previously, for gfxstream the filename was "CMakeLists.txt",
but that was changed with the prior commit.  Not having a
filename doesn't affect codegen, but does throw an error
afterwards when timing statistics are thrown out.

Luckily, args.target and options.filename are more or less
synonymous in genvk.py, so use that.  This should be fairly
upstreamable too if we choose to do so.

BUG=296903491
TEST=run genvk.py

Change-Id: Ib90b64ea2e6ed5860239b0ac3e644454f94d2778