Clean up ApiVisitor comparator properties
Previously, `ApiVisitor` had two nullable properties, `fieldComparator`
and `methodComparator`. According to the javadoc `null` meant to use
source order but in fact it meant use a deterministic order. Also, the
`fieldComparator` property always had (or behaved) as if it was set to
`FieldItem.comparator`. This change removes the unneeeded
`fieldComparator` property and makes `methodComparator` non-null and
default to `MethodItem.comparator`.

Bug: 319107678
Test: ./gradlew
      # cherry pick to AOSP/main and run m checkapi
Change-Id: Icbb35d567ba67c7027fba7207c25f1e052b323a9