Avoid replacing framework.jar in odsign_e2e_tests.
In a test case, in order to simulate an OTA update and test
odrefresh's behavior on the OTA, we do a hack: we replace a
bootclasspath jar /system/framework/framework.jar by bind-mounting a
test jar at this path, and then we umount in tearDown. Normally, this
should not be disruptive because Zygote loads framework.jar during
boot and keeps the FD, and apps fork from Zygote. However, there are
commands such as `svc power reboot` that brings up a separate runtime,
not forked from Zygote, and therefore needs to read framework.jar from
disk. Such commands will crash if they hit the fake framework.jar.

If the umount fails, the fake framework.jar will stay, and break all
subsequent commands such as `svc power reboot`.

In this CL, I'm changing the test case to replace
framework-graphics.jar instead, to make it less disruptive when the
umount fails.

Bug: 318706268
Change-Id: I1ba76412484da07e3c6521a2f6bd40bc45c70a7e
Test: ABTD (http://go/abtd-run/L98800030001127223)