Nest VisitCandidate inside ApiVisitor
`VisitCandidate` is created by, used exclusively by and accesses
properties from within `ApiVisitor` so it makes sense for it to be an
inner class of `ApiVisitor`. This change nests it and a follow up
change will make it an `inner` class. The reason for splitting it into
two changes is  to simplify the diffs.

Bug: 319107678
Test: ./gradlew
Change-Id: Ie4a2e76662739b46e219d8d87c7f9dfa46d0b5ca