Simplify handling of fields in VisitCandidate
Previously, the set of `FieldItem`s would be partitioned into two
`Sequence`s, one for enum constants and one for other types of field.
That is so that the enum constants will appear together before all
other fields in the API signature file. This change combines them as
a single `Sequence` and uses a `Comparator` to make sure that the
enum constants are first.

Bug: 319107678
Test: ./gradlew
Change-Id: If26edebf743a55084e80b4e21cda601d3737447f