Add additional overrides test for superMethods edge case
The behavior of `superMethods()` has an edge case when determining
whether to elide a method that is inherited from both a hidden super
class and a default interface method. This affects the behavior of
Metalava when it processes the `java.lang.StringBuilder` and
`java.lang.StringBuffer` classes.

The issue is that the `superMethods()` behavior interacts with a
`PsiMethodItem` duplicated from a hidden super class resulting in
inconsistent behavior compared with an equivalent `PsiMethodItem` that
was created from source. That causes methods to be kept that would
otherwise be elided by the `ElidingPredicate` when adding additional
overrides.

This change adds a test and documents the issue with the duplicated
`PsiMethodItem`.

Bug: 317564701
Test: ./gradlew
Change-Id: I63dc10cb6518d444e86499ffe29c80badefd9e49