[scudo] Condition variable can be disabled by setting the flag to off (#77532)
To enable the condition variable, you have to define both
UseConditionVariable and the ConditionVariableT. Otherwise, it'll be
disabled. However, you may want to disable the condition variable by
setting UseConditionVariable=false, for example, while measuring the
performance and you want to turn it off temporarily. Instead of
requiring the removal of the variable, examining its value makes more
sense.

GitOrigin-RevId: e6c2952eb51a422e17f002d97b0ea467be4d325b
Change-Id: I51c9b36920d8f93f2a3567b2c1240919cbbbb2fc