Remove PsiItem.element property
Previously, each `Psi*Item` class inherited a `PsiItem.element`
property of type `PsiElement` and then also provided another property
containing the same value but of a more specific type and name. e.g.
`PsiClassItem` provided a `psiClass` property of type `PsiClass`. That
was an unnecessary duplication of the property.

This change removes the `PsiItem.element` property, while keeping the
initialization parameter for use when initializing `sourcePsi` and
replaces any reference to it with a call to `psi()`. Each sub-class
that did not previously implement `psi()` had an implementation added.

This approach was chosen as it preserved the type specific names of the
property, e.g. `psiClass`, `psiMethod` etc. as that improves
readability. The alternative of making `PsiItem` take a type parameter
that was a sub-class of `PsiElement` would have replaced all the type
specific names with `element`.

Bug: 317564701
Test: ./gradlew
Change-Id: Ia77a7669c8068c10569da8dbba6c3839c23a3f64