Compute inheritedMethod in terms of inheritedFrom
Previously, `inheritedMethod` was a concrete var with a backing field.
However, that field was not strictly necessary as `inheritedMethod` was
usually set to `true` when `inheritedFrom` was set to a non-null value
and so this change makes `inheritedMethod` a computed property that is
`true` iff `inheritedFrom` is non-null and so it needs no backing
field.

The one case when `inheritedMethod` was not set to `true` when
`inheritedFrom` was set to a non-null value is
`PsiMethodItem.duplicate()`. The method would set `inheritedFrom` but
not `inheritedMethod`. However, as `duplicate()` is only called from
`CodebaseComparator` the resulting objects are only short lived and
discarded once visited and `inheritedMethod` is not checked on those
objects.

Bug: 317564701
Test: ./gradlew
      # cherry pick to AOSP/main and run m checkapi
Change-Id: I165da6963b1ceaeba9e96ebd080eea55848789dc