Compute inheritedField in terms of inheritedFrom
Previously, `inheritedField` was a concrete var with a backing field.
However, that field was not strictly necessary as `inheritedField` was
set to `true` iff `inheritedFrom` was set to a non-null value and so
this change makes `inheritedField` a computed property that is `true`
iff `inheritedFrom` is non-null.

This also switches the `FieldItem.inheritedField` property from var to
val as it is not modified through the model API.

Bug: 317564701
Test: ./gradlew
      # cherry pick to AOSP/main and run m checkapi
Change-Id: I461f6b93320c07db59a8af7b07e8d3a5cda433cd