Clarify behavior of the ClassItem.createMethod(MethodItem) method
The `createMethod(MethodItem)` is only used by `ApiAnalyzer` to inherit
a `MethodItem` from a class that is hidden/inaccessible and so not part
of the API into a class that is part of the API. Previously, the
`ApiAnalyzer` would set `MethodItem.inheritedFrom` on the inherited
`MethodItem` to be the class from where it was inherited. That was the
only write to `MethodItem.inheritedFrom` through the API that required
it to be a var.

This change:
* Pushes the setting of `MethodItem.inheritedFrom` into
  `createMethod(MethodItem)`.
* Renames it to `inheritMethodFromNonApiAncestor(MethodItem)`.
* Makes `MethodItem.inheritedFrom` a val.
* Adds a testsuite test for it.

Bug: 317564701
Test: ./gradlew
Change-Id: I5618fbdac4a16ff7f35e26c309c620c79ee23357