Remove unnecessary `isCloned()` method
Previously, the only use for `isCloned()` was to filter out items which
are copied from another class when generating a file of dex signatures,
presumably to avoid adding entries which do not exist in the actual dex
files. However, that does not actually do anything as dex signatures
are only written from API signature files and the text model always
returned false for `isCloned()` as it was never cloned.

This change, removes that usage of `isCloned()` along with the
definition and all the implementations.

Bug: 317564701
Test: ./gradlew
Change-Id: I1e8cdd8034d69f6df58bd657c2257248048ce62d