Revert "Restrict the use of ConstantFolding's VisitIf"
This reverts commit 5eb1fd0dae3832ceee2102613bb08c291daca6f3.

Reason for revert: In aosp/2903248 we implemented a faster way of
doing `ReplaceUsesDominatedBy` which is used by `VisitIf`. The impact
of `VisitIf` is now small enough that running VisitIf in all passes
is faster that the previous implementation running some of the time.

This CLs re-enables the optimization in all constant folding passes
because:
A) Lets this optimization (and others that can use the result) kick in
   earlier
B) Run it for callee graphs in the inliner (which has been disabled as
   of CL aosp/2543831)
C) Consistency of the ConstantFolding pass, which helps to have a
   simpler mental model

Bug: 278626992
Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b
Test: Locally compiled GMS and compared time to compile
Change-Id: I5dc5f591557c8de0bc4d23dbfd0b91b5b7e56ab5