Revisit audio routing policy when connected
If A2DP only or HFP only devices are connected, don't activate them
automatically.
If there is an active stream via a remote device, do not route the
newly connected devices. For example, when a phone is playing a music
via connected earbuds and then it is connected a speaker, the music
stays in the earbuds.

Bug: 299023147
Bug: 300174072
Test: atest BluetoothInstrumentationTests
Change-Id: Ie63b278bbf6bf93b75141bd705907ecdccd6faf2