Remove ModifierList.write args matching default parameter value
The `ModifierList.write` method has quite a few parameters that have
default values. This change removes any constant argument passed,
directly or indirectly that matches the default parameter value. That
makes it easier to see which parameters are needed. As part of that it
adds a default value for the `removeAbstract` parameter of a method
that calls `ModifierList.write` that matches the default value of the
`removeAbstract` parameter on `ModifierList.write`.

Bug: 320457204
Test: ./gradlew
Change-Id: I5a16d1efabe58425b10bc6c56b60ffb08dd3861c