Migrate --check-compatibility:base to Clikt
Preparation for work on the check compatibility options. Clikt does not
allow the `:` character in option names as standard so this has to add
support for that by circumventing the Clikt option name check and
replacing it with a Metalava specific one that allows the `:` character
to be provided.

Clikt does not directly support providing a custom option name check so
this has to work around that. The check is performed in the
`OptionWithValues.provideDelegate()` method so this adds a new
function that will return a new `StructuredOptionName` class which has
its own `provideDelegate()` method that tried to replicate what Clikt
does but with its own check.

Bug: 320235600
Test: ./gradlew
Change-Id: I14786d0627a1e7a118d99bff9f0ee9d14a7fecc0