Remove ineffective code for fixing up error messages
When initially added in https://r.android.com/734770 the
`CompatibilityCheck.base` property was set to the new `Codebase` being
checked and was used to map errors against items from intermediate
signature files to the source file to make the errors more useful.

However, change https://r.android.com/1587433 made it so that it could
be a signature file and https://r.android.com/1587433 made it so that
it was always a signature file if specified. That meant that it was
effectively useless for its intended purpose. Given that no issues have
been raised about error messages it seems as though it was not doing
anything anyway, possibly because we no longer run the check against an
intermediate signature file.

If it does become necessary in future then a better way for handling
that is to create a delegating `Reporter` implementation that will do
the mapping back to the source, separate from the `CompatibilityCheck`.

Bug: 320235600
Test: ./gradlew
Change-Id: I63d2f44777e9ce88dd3ba45fc107900082b59355