Separate the fast path check from the test being performed
Previously, the fast path check (i.e. the part that set and checked the
`fastPathCheckResult` state) was integrated into the code that also ran
the test (i.e. called `check(...)`). That meant that input to the test
had to propagate down to the `check(...)` making the fast path check
more complicated, e.g. it had to handle being given source files
or a jar.

This separates those two aspects so the fast path check is just given
a lambda for the test to run. That makes future changes simpler.

Bug: 320435861
Test: ./gradlew
Change-Id: Ie3e6b4d91b0f981e4037ca336e3a16a476a1c52a