Remove modifiers from ModifierList.write(...)
Previously, the `ModifierList` to be written was passed in separately
to the owning `Item`. Even though the `ModifierList` that was passed in
was always `Item.modifiers`. This change removes the separate
`ModifierList`. In a couple of methods where the `modifiers` were being
used in the same method as the call to `write`, the getting of the
`modifiers` was moved closer to its first use.

Bug: 320457204
Test: ./gradlew
Change-Id: I1848dd18c8607b5e63f104cc17674875b3eaa4a0