Change previously released API options to return CheckRequest?
Previously, they returned a `File?` and then later code mapped that to
a `CheckRequest?`. This change moves that mapping into the options.

The combination of `allowStructuredOptionName()` and `map()` revealed a
minor issue with the former's `OptionDelegate.provideDelegate(...)`
implementation. It should have returned itself, not the wrapped
delegate so this change fixes that.

Bug: 320245343
Test: ./gradlew
Change-Id: Ic73813c7f244b3b76ecdd0010b9859da71f5aece