Autoupdate of 2 submodules
*frameworks/support updated to ae808ba0167082568ca0c7bf7b1a1136a62c2f07
Merge "Temporarily disable androidx_with_metalava" into androidx-main
*tools/metalava updated to fd78f6f69ac82dcc01220e9de77437418f80b70a
Enable mapTypeVariables for Kotlin

The old implementation of this method didn't work for Kotlin so I
temporarily left it broken because this will require updates to
AndroidX signature files (see aosp/2881486).

The signature file changes are because this method enables lowering
a method from a hidden parent by mapping type variables of the parent
class to the types used by the child. Since it didn't work for Kotlin,
there are some signature files which reference a type parameter of a
hidden parent instead of the real type used by the child.

Bug: 259931298
Test: CommonClassItemTest and CommonTypeItemTest
Change-Id: I5d2b2b550fd8b6077a441ca5e766319c2e468a94