Appease mypy in git_utils.is_ancestor.
mypy can't reason that the check_returncode() in the previous
implementation is actually no-return due to the fact that 0 (the only
case in which check_returncode() is not no-return) has already been
checked. mypy doesn't have an elegant way to mark code as unreachable
short of `raise AssertionError("unreachable")`, so this approach is
probably less likely to confuse tools.

Bug: None
Test: pytest tests/test_git_tuils.py
Change-Id: If538bba55e4dc7f817f1e181b7fe6408068e653d