Add test for auto generated parameter name
Some binary class files contain parameter names and some do not. This
change adds a test to show the behavior of `ParameterItem`'s `name()`
and `publicName()` methods.

The current behavior of `publicName()` is incorrect because while
`name()` must always return a name, so one is synthesized, that
synthesized name can never be something that is relied upon and so
should not be used as the `publicName()`. Instead that should return
`null`. A follow up change will fix that.

Bug: 321109677
Test: ./gradlew
Change-Id: Ibc5d2c11d8594f8616824253cb59ed522c237ba6