Delete unused NetworkObserverRegistry from IpClientLinkObserver.
After the cleanup of IPCLIENT_PARSE_NETLINK_EVENTS_FORCE_DISABLE flag
IpClientLinkObserver never depends on the netd callbacks any more to
receive the netlink update, therefore we can also delete the usage of
NetworkObserver and NetworkObserverRegistry class from there.

This update also changes the params in the IpClient constructor, so we
also need to update the IpClient constructor in unit/integration test
and NetworkStackService side.

Bug: 318764217
Test: atest NetworkStackIntegrationTests
Test: atest NetworkStackRootTests
Test: atest NetworkStackTests
Change-Id: Ic7de1d093ff59d904947889b2033d8c7315d897e