Move classResolver to TextCodebase
Previously, the `classResolver` was part of the `ResolverContext` but
that meant it was only available from the `ApiFile` which meant it
could not be used for resolving types lazily such as in
`TextClassTypeItem.asClass()`. This change moves the `classResolver` to
`TextCodebase` where follow up changes can make use of it.

Bug: 322147308
Test: ./gradlew
Change-Id: Id5ba72779a8bd575e255f114da8e0b3139590b7f