apf_checksum: sort-of add support for ICMPv4
This is minimal, and hard to use, but only costs us 4 arm32 instructions.
It is *still* much easier to use then fully calculating an ICMPv4 checksum,
since it takes care of summing the entire (possibly unknown length)
icmp payload.

Most of the precalculation can be handled in apf generator,
this is because 5 * 0xFFFF - 1 - (srcip >> 16) - (srcip & 0xFFFF)
is a constant known to the APF generator.

This leaves 'constant - (dstip >> 16) - (dstip & 0xFFFF)' + fold.

This potentially makes it feasible for APFv6 to offload reply
to ICMPv4 ping request (if we ever want to do that).

Test: ./mk; atest apf_dns_test apf_checksum_test
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: Ifb127be30a153899e038ce212b1852922adc5f20