Rename innerClass parameter to isOuterClass
The `innerClass` parameter of `TextCodebase.getOrCreateClass(...)` is
poorly named and the documentation is confusing. The `innerClass` is
set to true when searching for a reference from an inner class to its
outer class which means that the class being searched/created is
actually an outer class not an inner class. This changes it to
`isOuterClass` to match the `isInterface` parameter which indicates
whether the created class should an interface or not.

Bug: 322147308
Test: ./gradlew
Change-Id: Ia79e4b08a1e217aa5f72f769f84d67cb06e93b19