Test super class handling when loading multiple signature files
Change https://r.android.com/2704325 was part of some changes that
meant that when loading a single `TextCodebase` from multiple API
signature files the order mattered and the files had to be from
narrowest API (e.g. public) to widest API (e.g. module-lib). That
change made it so that the last specified super class was the one that
would win. Unfortunately, it did not add any tests for it so this
change adds those tests.

Bug: 322147308
Test: ./gradlew
Change-Id: I495a2cea0e6228da635729e8daf0f7ac72e32eba