Remove use of TextTypeParameterList.create from createStubClass
The `TextClassItem.createStubClass(...)` is only called from
`TextCodebase.getOrCreateClass(...)`, the name passed in to the latter
is passed to the former. Previously, the latter accepted a name with
type arguments but a previous change resulted in only raw class names
being passed. The `TextTypeParameterList.create(...)` method was called
to convert those generic arguments into a TextTypeParameterList` but
that path is no longer taken so this change removes that path.

Bug: 321993837
Test: ./gradlew
Change-Id: If4344ec46741512c0d682990244b70d8f584b8ab