Stop PsiTypeParameterItem extending PsiClassItem
There are a number of `Item` methods that are not needed and so this
change makes them throw an exception. A follow up change will clean
them up.

It was necessary for the `PsiVariableTypeItem` to use
`findOrCreateClass(...)` instead of `findClass(...)` to find the
`ClassItem` for `java.lang.Object` as previously the
`PsiTypeParameterItem.finishInitialization()` would call
`PsiClassItem.finishFinalization()` which would
`initializeSuperClasses()` which would end up calling
`findOrCreateClass(...)` for the super class of the type parameter item
which was always `java.lang.Object`.

Calling `PsiClassItem.initializeSuperClasses()`, either directly or
indirectly seemed like the wrong thing to do as it does more work than
is necessary and `TypeParameterItem` does not have a super class, its
bounds are similar but different.

Bug: 321075216
Test: ./gradlew
Change-Id: I340c579a8b4cb9d7bfc06c423f89e6c0867c5ba9