Create ThrowableType wrappers around ClassItem/TypeParameterItem
Replaces the typealias `ThrowableType` and extension functions and
properties with an interface that provides just the methods needed by
the users of the `ThrowableType`, along with a wrapper classes that
implement the interface and delegate to the wrapped `ClassItem` or
`TypeParameterItem`. It also removes the imports for the extension
functions and properties.

Bug: 322499518
Test: ./gradlew
Change-Id: I905e62716452ac72570ba37884098c47a912e9d3