Remove duplicate check for same class in MethodItem.throws()
Previously, the method would first iterate over the `throwsTypes()`
and call `ClassItem.extends(qualifiedName)` which would check if the
class extended the `qualifiedName` type. That would also check to see
if the class was the `qualifiedName` type. Then it would iterate over
the `throwsTypes()` again checking if the class was the `qualifiedName`
type again. This change removes the second loop.

Bug: 322499518
Test: ./gradlew
Change-Id: Id1cb31ec8e3fd1b575487cb4b7e56cd9bbe8a8c1