Stop TextClassItem from being equal to TextTypeParameterItem
When parsing a throws list that includes a type parameter the `ApiFile`
will create a `TextClassItem` instead of a `TextTypeParameterItem`.
Previously, it would treat them as equal which made creating a common
test that would detect the issue difficult. This change fixes the
`TextClassItem.equals()` method to disallow comparing equal to a
different class.

Bug: 322499518
Test: ./gradlew
Change-Id: I900ed5a042faeb2cdd6785e3489f6bf2672743c6