Add ofClass() and ofTypeParameter() factory methods for ThrowableType
Adds `ofClass()` and `ofTypeParameter()` factory methods to
`ThrowableType` and changes code to use the appropriate one. Currently,
they check to make sure that the supplied value is the expected type
and then return the supplied value through but a following change will
replace them with functions that return an actual `ThrowableType`
object.

Bug: 322499518
Test: ./gradlew
Change-Id: Ic64ed5601e70dff2eb467b8cd380e25e8e177a43