Bypass Thread HAL SendSpinelFrame test on TV devices
The Thread HAL VTS test is built from the staging folder on the
android14-tests-dev branch. And the HAL interface doesn't have
@VintfStability which means it is unstable interface. Thread HAL
interface has been removed from the staging folder and added the
@VintfStability on main branch. It reports following errors when
runing the VTS that froms android14-tests-dev branch on main branch.

`BpBinder: Cannot do a user transaction on a system stability
binder in a vendor stability context`

This CL bypasses the `SendSpinelFrame` test case of the Thread
VTS tests on TV devices.

Bug: b/322731574
Test: Run `run vts -m VtsHalThreadNetworkTargetTest` on the TV device
Change-Id: Ibf7fe96aafa58cee7ae55a0ef613f0be67548a80