Provide with_log_buffer() function
The goal is to make it easier to update the 3p android_logger crate.

This adds the upstream with_log_buffer() function. Android currently
carries a local patch which provides the same functionality under a
different name - with_log_id(). This is a temporary addition to allow
updating individual parts of the platform to match the upstream
crate to make the android_logger update more seemless. Normally we
would update everything in one topic, but the number of changes is
so large that this makes it simpler to get individual changes in.

Bug: 322718401
Test: build and run CF with the change.
Test: m aosp_cf_x86_64_phone
Change-Id: Icc63571d640cad0a0580e574df56853288380442