ui: create plugin for prototyping client-server view of Android
This CL introduces a plugin to prototype a client-server view of Android
by focusing on binder transactions which go between processes. By
framing it in this way, we can draw parallel to micro-services and
distributed tracing.

For now, this plugin simply adds a command which takes a slice id and
recursively computes the client-server graph and displays them as a list
of tracks. In the future, we should add even more contextual info about
why servers are waiting for example.

Change-Id: Ibb38677d8d1c86260451e85951bcffdd5f3dcfab