Update IPv4 link address lifetime on success DHCP Renew/Rebind.
So far the IPv4 link address lifetime is only updated on installation
with RTM_NEWADDR netlink message and doesn't refresh even after a
success DHCP renew/rebind, that results in the IPv4 address to be
removed from the interface when the lifetime expires finally. Previously
why we don't need to care about the link address lifetime is because
InterfaceController#setIpv4Address ends up calling netd API to send a
RTM_NEWADDR netlink message without IFA_CACHEINFO attribute to kernel,
thus kernel chooses to use infinite lifetime by default.

Before the DhcpClient state machine transits to BoundState after a
success Renew/Rebind process, unfortunately BoundState doesn't notify
IpClient to update the IPv4 link address and just schedules a new lease
expiry based on the DHCP lease duration. To refresh the IPv4 link
address lifetime, we have to transit to ConfiguringInterface state after
a success Renew/Rebind where we create a RTM_NEWADDR netlink message to
and send it to kernel to add a new IPv4 address or update the target
IPv4 address (e.g. lifetime). The old workflow would be:

- DhcpClient enters ReacquiringState to extend the address lifetime
    - send unicast/broadcast DHCPREQUEST
    - receive the DHCPACK from the DHCP server
    - acceptDhcpResults
    - notifySuccess
        - trigger onNewDhcpResults callback
        - trigger onLinkPropertiesChange callback
    - DhcpClient transit to BoundState
        - schedule new lease expiry

And the new workflow would be:
- DhcpClient enters ReacquiringState to extend the address lifetime
    - send unicast/broadcast DHCPREQUEST
    - receive the DHCPACK from the DHCP server
    - acceptDhcpResults
    - transit to ConfiguringInterface state
    - notifySuccess
        - trigger onNewDhcpResults callback
	- trigger onLinkPropertiesChange callback
    - update IPv4 link address via netlink message
    - DhcpClient transits to BoundState
        - schedule new lease expiry
    - IpClientLinkObserver sees the netlink message from kernel, and
      sync up with IpClient about the update in
      handleLinkProperitesChange and trigger onLinkPropertiesChange
      callback.

The only difference between the old workflow and new workflow is that
DhcpClient is forced to transit to ConfiguringInterface state to refresh
the IPv4 address before going to BoundState, and a
onLinkPropertiesChange callback with the updated IPv4 link address(e.g.
deprecationTime and expirationTime) would be triggered.

This change is guarded by an experiment flag and it's disabled by
default, should be a no-op change.

Bug: 129660498
Test: atest testPopulateLinkAddressLifetime_onDhcpRenew
Change-Id: Ia1649506c7732041da8e0d4ed5a2f01be921722d