Autoupdate of 2 submodules
*build/soong updated to 7acf94b22180b94c244cdaa6c13e46e5679496b6
Merge "Find LLNDK dumps in prebuilts/abi-dumps/vndk" into main
*prebuilts/abi-dumps/vndk updated to bd513c1fe6fd6516679c8d06ae729eedc36f1cb5
Link LLNDK ABI dumps from vndk directory to platform and ndk

LLNDK will not be finalized at the same time as APEX and NDK. The
dumps have to be placed in a different directory. The new LLNDK dumps
will be created in "vndk" directory. To keep the cross-version check
effective, this commit creates links from vndk to platform and ndk.

Test: lunch aosp_x86_64-next-eng ; make
Bug: 314010764
Change-Id: I5465031976e4cbc87a03351b3aa37bcd262b5cbb