Add TypeParameterScope containing TypeParameterItems in scope
Dedups the searching code and renames variables and parameters to
clarify their purpose. It also removes the default parameter value for
it in `TextTypeParser.obtainTypeFromString(...)` as the only code that
relies on that is a test and it will ensure that future changes do not
forget to pass one in.

Bug: 322960279
Test: ./gradlew
Change-Id: I1444f21ac1494e6a2a893e099683ad59e8767453