Use gatherTypeParams(...) consistently
Previously, some code would construct the `List<TypeParameterItem>`
manually, and others would use `gatherTypeParams(...)`. This change
makes them all use the latter.

That necessitated adding support for specifying an additional local
`TypeParameterItem` list parameter that, if provided, is prepended to
the start of the returned list.

That parameter was added before the current `owner` to match the order
in which they appear in the returned list. That did require existing
calls to use named parameters which caused some churn but it does make
the code slightly easier to read.

Bug: 322960279
Test: ./gradlew
Change-Id: I739b05480c2e1fd6e21c4a3dee94a4678714d913