Speed up HConstantFoldingVisitor::PropagateValue
We can speed it up in two ways:
1) Don't call it if it has exactly one element, as we will never
   be able to replace its use in the if clause
2) Lazily compute the dominated blocks when needed

Compiling locally GMS, HConstantFoldingVisitor::VisitIf goes down
from 1.8% of the compile time to 0.7%. Most of this improvement
(90%+) is coming from the `1)` optimization. This is because there
are many cases where we have only one use (the if), which is in the
same block so we compute the domination to always end up not doing
the optimization.

Bug: 278626992
Test: Locally compile gms
Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b
Change-Id: Ic17b4b44840c7efa0224504031bf635584850ced