Autoupdate of 2 submodules
*art updated to 516020a3fbfe3db43f7faf0ac3daf5a45dbeeb6b
Fix SDK visibility of methods in an interface class

SDK visiblity of methods in super-interfaces needs to be
checked to determine the result.

Example: BlockingDeque -> Deque -> SequencedCollection
Deque.getLast() is an existing public API, but
SequencedCollection and its getLast() method are hidden.

Bug: 322063300
Test: atest libcore-sdk34-test
Change-Id: I1b022d7b35e8a35e0240297eab14eb03f12f5330

*libcore updated to f30a6e6e4de1d94d3e1974956473b3d74650ee36
Merge "Add a regression test to check SDK visiblity of Deque.getLast()" into main