Add ClassCharacteristics to abstract the merge class compatibility check
When loading a `TextCodebase` from multiple files it is necessary to
merge classes that appear in both, e.g.:

```
  public class Foo {
     field public int publicFoo;
  }
```

```
  public class Foo {
     field public int systemFoo;
  }
```

As part of that merging there is a very simple compatibility check that
is performed to ensure that they are consistent, i.e. generated from
the same sources at the same time. Previously, that was done directly
on the `TextClassItem` classes but that necessitated creating a new
`TextClassItem`.

This change creates `ClassCharacteristics` to encapsulate the
information needed to perform that check and rewrites the check to
use that.

Bug: 323168612
Test: ./gradlew
Change-Id: Ib95aa02746c74195dbefca55892090e075fbeffc