Only create a new TextClassItem when needed
Previously, when parsing a signature file class definition a new
`TextClassItem` class was always created even though it might have been
discarded and the class definition merged into an existing class. It
greatly complicated the logic and made it extremely hard to reason
about. This change only creates the new `TextClassItem` if there is no
existing class.

The `tryMergingIntoExistingClass()` method does not bother calling
`pkg.addClass(existingClass)`, even though previously that call would
have been made with what was `foundClass`. That is because the existing
class would already have been added to the same `PackageItem` and
calling it twice is a noop.

Bug: 322960279
Test: ./gradlew
Change-Id: Id8069ec2e5ffe7aba06d67e561abd3c6525a3bd2