Fix findClass() tests for throwable and object
The behavior of these tests were inconsistent between models due to
subtle differences in their behavior and how they handled implicit
super types. This change side steps those inconsistencies by adding
explicit dependencies on the classes to make sure that they are
resolved properly.

This reveals another slight difference between the text model and the
other models which causes a couple of the tests to start failing in the
text model. They will be fixed in a follow up.

Bug: 322147308
Test: ./gradlew
Change-Id: I96d4d8ecd965acf444120c7c9a459b14afa94c4e