Make TextCodebase.findClass(String) consistent with other models
Previously, `findClass(String)` would only return classes that were
part of the [TextCodebase] and not those returned by `classResolver`.
This change fixes that and adds a new `findClassInCodebase(String)` for
those existing uses of `TextCodebase.findClass(String)` that relied on
the old behavior (which was all uses from within the text model
itself).

Bug: 322147308
Test: ./gradlew
Change-Id: Iaf3d0153c44235ae266f78d80c7256257a21dd39