Register TextPackageItem and TextClassItem on creation
Previously, when parsing `TextClassItem`s were not registered they were
added to the `TextPackageItem` and relied on them being registered when
the package was added to the `TextCodebase`. That happened after
parsing the `package` block. That meant it was not possible to find
either during parsing.

While that is generally not a good idea as the class or package may not
be created when they were being searched for, it is necessary for
resolving the outer class of a nested class. That is safe because the
outer class is always defined before the nested class, even when it is
in a separate file.

This change moves the registering to the point when the new items are
created.

Bug: 322960279
Test: ./gradlew
Change-Id: I187f495225ba2b87cff66a20dd17689ed78979dd