Remove owner from TextTypeParameterItem.bounds()
The `owner` was used to look up a type parameter name to see if it was
defined in a containing item. However, while that was tested by some
unit tests it was not actually relied upon any more because references
to a type parameter are always represented by a `VariableTypeItem`
which will never use this code.

The `TextTypeItemTest` class was removed in its entirety as after
removing all the code that was needed to support the `owner` the tests
either become empty or are covered by other tests in the test suite.

Bug: 323270493
Test: ./gradlew
Change-Id: Id1d58de89332c1e4a6bc7a5f928a141da7032ab6