Move boundsStringList parsing into TextTypeParameterList.create()
Previously, the parsing of the `boundsStringList` into a
`List<TypeItem>` was performed lazily in `TestTypeParameterItem`'s
`typeBounds()` method. However, that required the use of the
`TextCodebase.typeResolver` property which is being removed. So, this
change moves it into `TextTypeParameterList.create()`.

In order to do that it has to handle possible dependencies from the
bounds type back to the type parameters in the list being created,
causing a possible cycle. It resolves that by:
1. First creating a list of `TextTypeParameterItem` without a
   `bounds`.
2. Adding that to a nested scope to make them available when parsing
   `TypeItem`s.
3. Parses the bounds and sets `TextTypeParameterItem.bounds`.

This stopped using `boundsStringList` when generating the `toString()`
and has to use `TypeItem.toString()` instead. That revealed a slight
issue with a test which needed fixing.

Bug: 323270493
Test: ./gradlew
Change-Id: I4f9f86203cbf48b86ac7704df08ddd9c8f30b413