Move type parameter related parsing code into ApiFile
Previously, there was code for handling the parsing of type parameters
in `TextTypeParameterItem` and `TestTypeParameterList`. That code is
really part of the parsing code and so belongs alongside the rest of it
in `ApiFile` so this change moves it.

It also moves related tests from `TextTypeParameterItemTest` to
`ApiFileTest` and removes `TextTypeParameterItemTest`.

Bug: 323270493
Test: ./gradlew
Change-Id: Iaebe7b9b4793c0ece3b8593cda58493bb371b4ed