Create interface TypeItems early and remove resolveInterfaces()
Previously, interface type strings were kept and parsed as types after
all files had been parsed. This change moves the call to
`parseDeclaredClassType(...)` which creates the `typeParameterScope`
before the interfaces are parsed. That allows them to be parsed into
`TypeItem`s immediately, removing the need to defer the processing.

Cleans up the `TextClassItem`'s handling of interface types a little
as `addInterface(...)` was no longer needed.

Bug: 323358420
Test: ./gradlew
Change-Id: I2016852fc5157c20a51d0340f66bc41c98910a8b