Stop resolving the super class when constructing text codebase
Previously, the `TextClassItem` required its `superClass` to be set
alongside the `superClassType` which necessitated that the type be
resolved to a `ClassItem`. This change removes that and just sets the
`superClassType` and implements `superClass` as
`superClassType?.asClass()`.

To minimize any performance hit caused by that, this change caches the
result of the `TextClassTypeItem.asClass()` method.

Bug: 323358420
Test: ./gradlew
Change-Id: I2c3beca10565dd35e9d15aad00acf46434f24551