Avoid ConcurrencyModificationException in ApiAnalyzer.addConstructors()
The exception was caused by resolving a super class type to a
`ClassItem` for a class not in the codebase which required adding a new
one to the `PackageItem` whose list of classes was being iterated over.
This did not happen prior to aosp/2945320 because those classes were
resolved during codebase construction.

This change avoids the issue by resolving the interface and super class
types during construction as was done previously. A better solution
will be added in future but this is needed to unblock other work.

Added some tests that break without this change but work with it.

Bug: 323516595
Test: ./gradlew
      # Merge into AOSP and run `m checkapi`
Change-Id: I6dbe2d8a12ebfb8ba225a1c03131e54c3797c41c