Fix super type nullability for text model
Adds `TypeUse` enum to differentiate between different uses that may
affect the parsing of a type and the resulting `TypeItem`. Added that
to the key and made `TypeUse.SUPER_TYPE` cause the type to always be
non-null. A `getSuperType(...)` method was added to use that and the
existing code that resolves interface types and super classes was
changed to use that new method.

The `TypeUse` was added to the model as it will be used by similar
fixes in Psi and Turbine.

Bug: 323522029
Test: ./gradlew
Change-Id: I0fcb75b0aef9e69ee9c7b06821cc71cde0ea4686