Reduce atest argparser complexity and make atest --help menu easier to view
Previously when we want to edit an argument in atest we need to modify multiple places. This CL removes the epilog printing and utilize the argparser's built in help functionality.

Also previously atest --help displays the top of the help message first, and users have to scroll down to view the rest of the message by pressing keyboard Down so many times. Now it's all printed out at once, and users view the message by scrolling Up either with keyboard or mouse, which is easier.

Diff result in the resulting --help print: https://diff.googleplex.com/#key=U3CcZop5voX7

Test: atest-dev --help
Bug: 323567688
Change-Id: I3bd0869593f775c5440088ab33aad14e5d7df780