Fix super type nullability for turbine model
Reuses the `TypeUse` added for the text model. One of the test fails
but that is due to a separate issue. Namely, the incorrect handling of
nested classes loaded from the classpath. Instead of the `ClassTy`
containing `java.util.Map` and `Entry` as separate `SimpleClassTy`s it
contains one `SimpleClassTy` whose name is `java.util.Map$Entry`.

That will be fixed in the future.

Bug: 323522029
Test: ./gradlew
Change-Id: I7409ead3770a0b1ddaad6d54d5ee44cc53162de4