Add cachedParseType() for use within TextTypeParser
Previously, the `obtainTypeFromString()` method was used from both
outside `TextTypeParser` and within. In order to satisfy the latter it
was necessary to add an additional `annotations` parameter to handle
annotations on array components, That lead to some confusing, resulting
in `getSuperType()` also having an `annotations` parameter even though
it is only called from outside.

This change adds a new `cachedParseType()` that is used for internal
use only and the `annotations` parameter is removed from
`obtainTypeFromString()` and `getSuperType()`.

Bug: 323614314
Test: ./gradlew
Change-Id: I659ac33e7ebb72b36eb9ff79c8055516cfa4e76e