Autoupdate of 2 submodules
*art updated to c0c012c2e85616540865bf793cc423f5ba42aff3
Report ArtDeviceStatus as a pulled atom

The current implementation is reporting a pushed atom called
ArtDeviceDatumReported containing the the boot image status as soon as
the runtime initialization has completed. One drawback with this
approach is that, if the device is not rebooted in a long time, no atoms
will be pushed. This CL introduces a new pulled atom called
ArtDeviceStatus which will be pulled rather than pushed with a fixed
schedule set on the server side - the plan is to set the schedule to
report every 24 hours, in order to receive a device status on a daily
basis for continuous analysis.

Bug: 257028435
Test: atest ArtGtestsTargetChroot
Test: statsd_testdrive 10205
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:c34231a3bd9165caa5c4b721c479a6635790ac19)
Merged-In: I68a0dfd413271c458fe930a3e6c4d17ed49f8666
Change-Id: I68a0dfd413271c458fe930a3e6c4d17ed49f8666

*system/security updated to a57741efb97e5d60e0727a2332dae8a0ebb3b78a
Merge "Add libstatspull and libstatssocker to odsign" into main