Add pathological tests where a type parameter shadows int
The tests are pathological because this would never be allowed,
certainly not in an API for Android or AndroidX. However, it is a
possibility and it is better to have the cache be able to deal with
these cases than try and debug an issue where it did not.

The tests fail in the text model. A follow up change will fix that.

Bug: 323614314
Test: ./gradlew
Change-Id: I70e1d60c327561748b356f5f54d79f4dcacfdbf6