Integrate TypeUse.SUPER_TYPE handling with kotlinStyleNulls
They both have the same effect of treating a plain class type, e.g.
`MyClass` or `MyClass<String>` as being `TypeNullability.NONNULL` so
rather than pass `TypeUse` all the way down through the parser this
just takes advantage of the existing support for managing nullability.

It also does the same with the key. Previously, `MyClass` for
`TypeUse.SUPER_TYPE` would be cached separately to `MyClass` for
`TypeUse.GENERAL` when `kotlinStyleNulls=true` even though they would
have the exact same `TypeItem`. This change will cause them both to be
cached using `forceClassToBeNonNull` reducing the number of entries in
the cache slightly, as can be seen by the change in the cache stats
test.

Bug: 323614314
Test: ./gradlew
Change-Id: Ia12c29a5b5f8e2f6ba15af26d02d2c25c4b5e7fa