Write trace entries as a block in preperation for the new format
We used to use the same buffer to hold the method info entries and also
the trace event entries. This was okay for the current format since each
entry is its own record and it is okay to interleave them. In the new
format we dump the trace events as a block. To support this, we now
flush the method entries directly to the file. It is possible to use a
buffer for this but we don't expect this to be performance critical.
This also has an added advantage that we can check for the overflow at
the start of the loop and can avoid a few checks in the loop which
improves performance.

Bug: 259258187
Test: art/test.py --trace --stream
Change-Id: Ib0ff03d3bb84571f1b0c7daea68fb5151495fd1f