Support basic type caching with a non-empty TypeParamaterScope
Previously, `TextTypeParser` would not cache any types when the
supplied `TypeParameterScope` was not empty. This change adds basic
support for caching those types. It simply creates a map from
`TypeParameterScope` to `TextTypeItem`.

This does reduce the number of `TextTypeItem`s created by almost 5000
items but also increases the size of the cache by a similar amount.
That is because it will cache something like `int` once in every
non-empty `TypeParameterScope` in which it appears. Follow up changes
will improve that. By taking advantage of the hierarchical nature of
the scopes.

Bug: 323614314
Test: ./gradlew
Change-Id: I7991653a86f5f94e1ee7efafc2ff55913ca4def1