Create initial implementation of LambdaTypeItem for psi
While this change does make progress to the bug's desired goal that is
not its primary purpose. It was added to provide a simple abstraction
that can be used to avoid downcasting to Psi specific classes elsewhere
in Metalava.

It was necessary to fix an issue with the creation of
`PsiWildcardTypeItem` as part of this. Previously, it would pass the
`KotlinTypeInfo` through to `createBound(...)` unchanged. That was
wrong as the `KotlinTypeInfo` only applies to an explicit bound, not
the implicit bound. That did not cause problems before as it only
used the `KotlinTypeInfo` for nullability so did not care about the
specific `KtType` passed though. However, this makes use of information
only available in the `KtFunctionalType` that is passed through and it
must match the `PsiType`. So, passing `KtFunctionalType` through to a
type that has an implicit `extends Object` causes problems because
`java.lang.Object` is not a `KtFunctionalType`.

This change fixes that by making sure to only pass `KotlinTypeInfo`
through for the bounds to which it actually applies.

Bug: 169798041
Test: ./gradlew
Change-Id: Ib064d5f2c0f4c5c10b4800cd65be40ba3e0fffc8