Add DefaultTypeModifers and replace TurbineTypeModifiers
This does move the logic for finding nullability from the annotations
into the `create` factory method instead of doing it based on the live
annotations list which could have been modified. However, this behavior
matches what Psi already does and so as Psi is used by the rest of
Metalava that is the behavior that it expects.

Bug: 324196754
Test: ./gradlew
Change-Id: Idcbe512f742acea40524252c31506d597c5360d1