Fix recursive archives due to shared in/out dir
The generated sbox command ends up specifying the same directory
for input and output for soong_zip, resulting in including
a partial copy of the archive in itself.

This causes cache misses in accelerated builds.

Specifying a sub-directory for the input resolves the issues.

Requires: https://r.android.com/2954138

Bug: b/322788229
Test: Ran a couple of builds and confirmed no cache misses from xsdc.
Also manually checked an archive and confirmed the
abcence of partial self-copies.

Change-Id: Ic50212cfa4bfc0495af157121c56ce2d63a58942