Switch Turbine to use TypeParameterScope
Previously, it used a simple map from `TyVarSymbol` (the underlying
Turbine representation of a type parameter) to `TypeParameterItem`.
That worked well and was simple. However, using a consistent approach
across all models will allow for more code reuse and will help make the
behavior consistent. Also, the `TypeParameterScope` gives better error
messages when it fails to resolve a type parameter name.

So, while it is a little more complex than the previous solution this
change switches to use `TypeParameterScope` which is already used for
the text model and a follow up change will do the same for Psi.

Bug: 324317136
Test: ./gradlew
Change-Id: I43b488a5c2d713a6dacb98b6220924fdc0d929b4