Defer resolving super types until requested
Previously, the super class and interfaces would be resolved
immediately. However, that created a very complex, reentrant
interaction during construction, particularly, when inner classes
extended other inner classes of the outer class.

This change follows the approach taken in the text model and just
creates, and sets the `superClassType` and implements `superClass()`
as `superClassType.asClass()`. That necessitated a few changes to
some types from turbine specific classes to general model classes.

It does the same for interface types. As a result there was no need
for the `TurbineClassItem.directInterfaces` property to hold the
interfaces' `TurbineClassItem`s created during initialization. It was
only ever used by `allInterfaces()` and that value is cached so the
logic for accessing what was `directInterfaces` was moved into
`allInterfaces()`.

That broken one test which relied on the early resolution of the super
class so it was slightly restructured to explicitly force the super
class to be resolved at the appropriate time.

Bug: 324317136
Test: ./gradlew
Change-Id: Ie9d1b27001099e99fe84fa18cf825b90aa188d2c