Remove Turbine support for reentrant TurbineClassItem creation
Historically, during codebase construction some class types were
resolved to their `TurbineClassItem`, i.e. super classes, interfaces
and exception types. That required the codebase to support reentrant
creation of `TurbineClassItem` which made it very complicated.

Recent changes have deferred the resolution of those class types so
that it is no longer necessary to support that. Therefore, this change
removes that support and clarifies the behavior.

Bug: 324317136
Test: ./gradlew
Change-Id: If46eb1de4e42071f84dd2faf6ada1e3a65a33c99