Change ThrowableClassTypeItem to ThrowableExceptionTypeItem
Previously, when `ThrowableType.ofExceptionType..(...)` was given a
`VariableTypeItem` it would just unwrap the `TypeParameterItem` and
pass that to `ThrowableTypeParameterItem`. That was fine for Text and
Turbine models as they have already resolved the `TypeParameterItem`.
However, it broke Psi model as it relies on them being resolved lazily.
While the plan is to switch Psi to work like the other models this is
blocking that work.

This change fixes that by encapsulating an `ExceptionTypeItem` instead
of a `ClassTypeItem` in `ThrowableClassTypeItem`. That approach was
taken rather than adding a new class or maybe have
`ThrowableTypeParameterItem` take a `TypeParameterItem` supplier as the
long term plan is to remove `ThrowableType` wrappers and replace them
with `ExceptionTypeItem` and some extension/default functions that
would have to handle both type simultaneously so this fits in with that
plan.

Bug: 323678811
Test: ./gradlew
Change-Id: Ib5d98edb9d1a9f01670b0c352068bccaf558ef5c