Autoupdate of 2 submodules
*frameworks/support updated to bd45c245b1b644561c0bad268b38c3e58cdab28b
Merge "Temporarily disable androidx_with_metalava" into androidx-main
*tools/metalava updated to a8e9a11852a5fdcd7b3d2fa870e48acbf030928c
Use KtType to work around the retainAll(Collection<E>), etc. bug

Converting the `KtType` to `PsiType` provides the correct name for the
type parameter which is an improvement that fixes the test and avoids
breakages downstream.

The correct type for the resulting `PsiType` shoud include a upper
bound, e.g. something like `Collection<? extends Z>` but that broke the
AndroidX build and requires some API changes. Initially the plan was to
continue to use the current behavior of `Collection<E>` but that broke
AndroidX in different ways. Therefore, this change fixes it properly
and will update the AndroidX build in one go.

A follow up change will fix this properly by integrating the
`TypeParameterScope` used for resolving type parameters in Text and
Turbine models into Psi.

Bug: 324317136
Test: ./gradlew
Change-Id: I4be7169dd2c1af8f4a69cd021fbcf871c577ec6e