Merge remote-tracking branch 'aosp/main' into aosp/metalava-main
Merge changes to Android files back into aosp/main. This is needed
for two reasons. Firstly, because it makes good sense for them to be as
in sync as possible. Secondly, because without them checking out a
commit from `aoso/metalava-main` in `aosp/main` (or other downstream
branch) and then trying to run `m checkapi` will fail. That makes it
very difficult to run `git bisect` on to find out which change out of a
stack of changes is the cause of an Android build breakage.

Bug: N/A
Test: N/A
Change-Id: Icd527161dbec6eb1a027cf1fcacbdc5de813e593