Resolve super and interface classes in finishInitialization
A recent change deferred the resolution of the super class types and
interface types to their relevant `ClassItem` to simplify
initialization. Resolving involves creating `ClassItem`s and adding
them to the list of `ClassItem`s in the `Codebase` so deferring it
meant that the list was smaller. It still contained all the
`ClassItem`s for classes in the source code but did not contain the
dependent classes from the classpath.

Unfortunately, some code that needed to visit all the `ClassItem`s did
not take into account that more `ClassItem`s could be added afterwards
and so did not visit any of the `ClassItem`s created during resolution
which lead to some weird failures downstream.

This change works around that issue by resolving them in
`PsiClassItem.finishInitialization()` which is called after
initialization has been finished and just before returning. It does
handle additional `ClassItem`s being added during iteration and will
iterate over any new ones added for as many rounds as necessary.

Raised b/324561266 to track sorting this out properly.

Bug: 324317136
Test: ./gradlew
      # merged downstream and ran `m checkapi`
Change-Id: I6e77d6f1ebb7bbae54db449c20f46e34a98ceb11