Add test to show correct ClassKind for stub interface class
Previously, when the text model found a reference to an interface that
did not exist it would create a stub class with a `classKind` of
`INTERFACE` not the default `CLASS`. Unfortunately, a previous
refactoring lost that behavior. So, this change adds a test for it and
follow up changes will revert that regression.

Bug: 323678811
Test: ./gradlew
Change-Id: I6702f4e5bbb80481fa37358816b5417f271a6843