Store information about required StubKind in TextCodebase
Previously, the `StubKind` that was applied when creating a stub class
was determined by the `isInterface` parameter passed to
`getOrCreateClass(...)`. The problem with that is that it requires the
caller to pass the information but a previous change to defer resolving
classes until requested through a `ClassTypeItem` meant that the caller
no longer had the context to provide the information which is what
broke the historical behavior.

This change adds a `requiredStubKindForClass` map to the `TextCodebase`
that contains the required `StubKind` for each as yet unseen but not
yet created class (identified by its qualified name). When an
interface type is produced by `TypeItemFactory.getInterfaceType(...)`
it makes a request to record that the type is expected to be a
`StubKind.INTERFACE` in that map.

If a class has already been seen then it is not added to the map and
as soon as the class is created (whether real or a stub) then the
recorded `StubKind` is removed.

This approach was taken rather than try and store the information in
the `ClassTypeItem` and pass it in when requested because that would
greatly complicate the type caching and require changes to
`ClassTypeItem` that is not required for other models that do not
fabricate `ClassItem`s just in case it is a real class.

This change fixes the test that was added in a preceding change to
show a regression in the handling of interface stub classes.

Bug: 323678811
Test: ./gradlew
Change-Id: If112c484df4c7471fe63a869f8f2c4cb98ee7e73