Warn when duplicate classes are detected in the sources
Historically, Metalava has just ignored duplicate source definitions of
a class. Previously, a change (https://r.android.com/2956354) was made
to fail if that happened but it turns out that some targets downstream
provide duplicate classes to the source on purpose. Rather than delay
Metalava updates to Android while that is fixed this change simply
reports a warning when that happens and ignores the duplicate
definition.

Bug: 324317136
Test: ./gradlew
Change-Id: I577c78d2ca9a0a68abf1c6c31ac17cfe681b9f01