Exclude merge changes from bug list
Previously, bugs from merge changes would be included in the bug list
which meant that a merge from `aosp/metalava-main` that included a
merge back into `aosp/metalava-main` would list all the bugs of all the
changes that had already been merged from `aosp/metalava-main`.

This change adds the `--no-merges` to the command that generates the
list of bugs, to match the `--no-merges` already passed to the command
line that generates the list of changes.

Bug: N/A
Test: scripts/merge-from-metalava-main.sh
Change-Id: I57cd01a7de1aad399853d2ab102a5d51cb3107a0