PsiTypeItemFactory implements TypeItemFactory
Changes `PsiTypeItemFactory` to implement `TypeItemFactory`. It does
not yet support `TypeParameterScope` so those methods are stubbed out
for now and will be implemented properly in a follow up change.

As `TypeItemFactory` takes a single value for the underlying model
type and Psi uses both the `PsiType` and an optional `PsiElement?` it
was necessary to create a `PsiTypeInfo` wrapper around them to
represent the underlying model type.

Existing uses of `PsiTypeItemFactory` that require specific types, e.g.
`BoundsTypeItem` have been switched to use the appropriate
`TypeItemFactory` method but other uses have been left as is for now.

Bug: 323678811
Test: ./gradlew
Change-Id: Ic4e3e1aee5ade51134002c079ea60716a1fdfade