Resolve PsiTypeParameterItems when creating PsiVariableTypeItem
Previously, the `PsiVariableTypeItem` would defer the resolving of a
`PsiTypeParameterItem` until `asTypeParameter` was accessed. That was
necessary because within a `TypeParameterList` the bounds can form
cycles within and between type parameters in the list. This change
replicates the text model change in https://r.android.com/2944113 and
resolves the `PsiTypeParameterItem`s before creation of the
corresponding `PsiVariableTypeItem`.

This fixes an issue with a test that did not resolve a type parameter
correctly. It would get get the actual `TypeParameterItem` and would
use a `ClassItem` instead. It did have the correct name so signature
files were generated correctly but the model was wrong. After this
change it is correctly resolved to a `TypeParameterItem`.

Bug: 323678811
Test: ./gradlew
Change-Id: I5576b3449e47693aee72c3d08eb902a550547a58