pvmfw: Get vcpufreq by compat & NOP unused OPPs
If the input DT doesn't contain an OPP table, remove it from the
template instead of passing the placeholder table to the guest.

Instead of hard-coding the path to the cpufreq node in the VMM-provided
DT, locate it by compatible (as the guest kernel would do) and warn
about any extra node (similar to other nodes found by compatible).

Test: m pvmfw
Bug: 284369518
Change-Id: I9009ee0bb8eea932482582d8f005de35ec118920