Implement ThrowableType.throwableClass using TypeParameterItem.erasedType()
Previously, the implementation would return `null` if the
`TypeParameterItem` did not have an explicit upper bound. This change
switches to use `TypeParameterItem.erasedType()` which will return
`java.lang.Object` as the default upper bound. That necessitated a
change to a test that relied on `null` being returned to make it
more explicit and check that instead of no upper bound that the upper
bound is not a sub-class of `Throwable`.

Bug: 323678811
Test: ./gradlew
Change-Id: I3d09afa601ad93f81443180ffe98873fbb97eec0