Move `duplicate(withNullability)` into TypeItem
Previously, this was defined in `TextTypeItem` to return a
`TextTypeItem`. This change pulls it up into `TypeItem` and defines it
using `TypeItem.duplicate()` and `TypeModifiers.setNullability(...)`.
The latter did not use to work on `TextTypeModifiers` but does now
work as long as the `DefaultTypeModifiers` is mutable, or has been
duplicated.

That required that the existing `TextTypeItem` subclass implementations
of `duplicate()` needed to return a duplicate version of the modifiers
which they did not do before.

The `TypeItem.duplicate()` returns a model specialization of `TypeItem`
and not a `TextTypeItem` so it was necessary to replace `TextTypeItem`
with `TypeItem` in lots of different places as part of this change.

Bug: 324196754
Test: ./gradlew
Change-Id: I05f3bacaf431f3a223560aa7c81bc2140846648b