Add TypeItemFactory.getFieldType(...) and implement for Psi
Adds `TypeItemFactory.getFieldType(...)` to be used by models when
creating a type for a field. Its parameters define the information that
each model needs to consider when determining the nullability of a type
for a field. It is currently only implemented for for Psi but follow-up
changes will implement it for other models.

This moves functionality from `PsiFieldItem.implicitNullness()` and
copies functionality from `FieldItem.implicitNullness(`) into the
`PsiTypeItemFactory`. The `FieldItem.implicitNullness()` cannot be
removed yet as it is still relied upon by Turbine.

There is no need to call the Psi specific extension function
`TypeItem.finishInitialization()` on the field type anymore so the
`PsiFieldItem.finishInitialization()` method was also removed.

Bug: 325454211
Test: ./gradlew
Change-Id: I75a95a7fdc70c0babdfb783352496d6a1dcca930