Extract PsiFieldValue to provide access to the initial value
Previously, the only way to access the initial value of a `PsiField`
was to create a `PsiFieldItem` and call `initialValue(Boolean)`. This
change extracts that functionality into a separate `PsiFieldValue` so
that a follow-up change can access the initial value earlier so it can
be factored into the nullability of the fields' `TypeItem`.

Bug: 325454211
Test: ./gradlew
Change-Id: I60b08314ecb428dd50dc0d8ba109ed111f23d010