Use deconstructing assertions in CommonTypeModifiersTest
Previously, it used a combination of `isInstanceOf(...)` and down casts
to test the different types. This converts it to use the deconstructing
assertions.

Some changes were made to the assertions, a new one was added for
primitive, the existing ones added a default value of `null` for their
`body` parameters, the receiver type was made nullable to allow for
easier testing of `TypeItem`s that may be null, e.g. `superClassType`
and an otpional `body` was added to `assertReferencesTypeParameter`.

Bug: 325454211
Test: ./gradlew
Change-Id: Ic88217d33bc0e43ec9013ccb485827662dc55a5e