Remove unnecessary bssl-sys-raw / bssl-sys indirection in Trusty
This split is unnecessary and is a consequence of some Android Rust
limitations that do not apply to Trusty Rust.

BoringSSL, following standard Rust practice, uses an environment
variable to locate the bindgen output, which is then include!ed. Android
Rust is incompatible with this standard practice and, until their
limitations are fixed, use an (unsupported by BoringSSL) mode where they
need to import a separate crate and also sed out a line in BoringSSL.

However, Trusty Rust is perfectly compatible with this standard
practice. It doesn't need an indirection. Thus we can unwind all this
mess and simply build bssl-sys unmodified.

The one wriggle is that Trusty doesn't allow configuring the name of the
environment variable. Rust standard practice never settled on one (the
docs use OUT_DIR, which doesn't really work well), and it seems Chromium
and Trusty ended up picking slightly different canonical names. To avoid
that mess, this CL adds a MODULE_BINDGEN_OUTPUT_ENV_VAR feature to
bindgen.mk that allows the build file to customize this.

While I'm here, remove the workaround for the directory shuffling. That
was only ever temporarily needed, and was just a workaround from these
build files living in the Trusty repository rather than the
external/boringssl repository. (We generally expect the build files to
be kept together.) We can probably move this now, but I've left this for
a separate change.

Bug: 291253039
Change-Id: I0f6663b4b4ea69aa0f10c76cf7c4ee48524efa49