Reland "[scudo] resize stack depot for allocation ring buffer" (#81028)
First commit of the stack is a clean reland, second is  the fix.

There was a typo in the `static_assert` that meant we were asserting the
size of the pointer, not the struct.

Also changed `alignas` to be more intuitive, but that is NFC.

Ran builds in Android here: https://r.android.com/2954411

GitOrigin-RevId: aff6cb4957e60931a3c2205a5fe8e6ea59447bbb
Change-Id: Ib29c38cd74a3a326dde12bd9e554a97267309d01