Create AnnotationItems immediately when parsing signature files
Previously, the annotations would initially be collated as `String`s
containing the source representation of the annotation and then
converted into an actual `AnnotationItem` while being added to the
`DefaultModifierList`. However, during that period when they were
strings it was necessary to check to see if they were a nullability
annotation.

This change creates them as `AnnotationItem`s straight away avoiding
the need to manage them as strings and taking advantage of the
`AnnotationItem` methods to check for nullability annotations which
cache the result in an `AnnotationInfo` for later use.

Bug: 325691395
Test: ./gradlew
Change-Id: Ic79bc26f62c1008fff37dc51bfe947f70015cd1f