Simplify copying of PsiModifierItem
Previously, copying a `PsiModifierItem` while inheriting a method from
another class would make copies of all the `PsiAnnotationItem` and
`UAnnotationItem`s as well. However, that was unnecessary because the
`AnnotationItem`s are immutable and while they can have a dependency on
the `Codebase` in this case they are being copied within the same
`Codebase`.

So, this change simplifies the duplication by just copying the mutable
list of `AnnotationItem`s but reusing the `AnnotationItem`s themselves.

Bug: 325691395
Test: ./gradlew
Change-Id: I2d6b14bffe1335bdf4e02ef570fc62ed1a899ec1