Use @PublishedApi instead of @Deprecation
Previously, the code used `@Deprecation` to mark a method that was only
intended for use by an `inline` function and prevent it from being used
directly from outside the package. However, Kotlin provides
`@PublishedApi` annotation to support that directly. A method annotated
with `@PublishedApi` can be used by an `inline` function from outside
the module but cannot be used directly.

Bug: 325691395
Test: ./gradlew
Change-Id: Ifc8387e7b45f21dd243f1537f1561d08693f913f