Create ContextNullability to replace TypeUse
Previously, `TypeUse` was added to allow for type creation,
specifically the setting of `TypeNullability` but also other aspects to
be dependent on where the type is being used. The intent was that it
would be expanded to cover different uses over time. However, it turns
out that there are lots of different factors that affect the selection
of the appropriate `TypeNullability`.

Rather than factor them all of those different factors into many
different `TypeUse` enum constants the change creates a new
`ContextNullability` class that encapsulates all the information from
the context where the type is used that, in combination with
optional information provided by Kotlin and annotations, can be used
to compute the `TypeNullability` that must be used.

This switches the PsiTypeItemFactory to make use of it and similar
changes will be made to other models when needed.

The `ContextNullability` is passed down to all types that can have a
user supplied nullability throughw annotations, i.e. everuthing except
primitives (which are always non-null) and wildcard (which is always
undefined).

Bug: 325454211
Test: ./gradlew

Change-Id: I62a30508f3847dc785ed79400b7ab824ec7f8fe3