BugFix: Always set PRIVATE_PROVIDED_VNDK_VERSIONS
Previous change tried to set PRIVATE_PROVIDED_VNDK_VERSIONS
conditionally if it is not empty. However, defining
PRIVATE_PROVIDED_VNDK_VERSIONS is part of build action, so it is not
accessible with ifeq command. Also as this is env variable, build works
even when it is empty. This change would again set
PRIVATE_PROVIDED_VNDK_VERSIONS even if it is empty to cover any build
errors with PRODUCT_EXTRA_VNDK_VERSIONS.

Bug: 316829758
Test: AOSP CF build succeeded
Test: AOSP Husky build succeeded
Change-Id: Ie86996f15a928d048bfe983f4c94be5719ce02e7