Update tests to explicitly ensure JITed code
941-recursive-obsolete-jit and 943-private-recursive-jit test that
deoptimizing a method and marking it obsolete while transforming the
same method works fine. Since we don't deoptimize through runtime
frames, we ensure there are no runtime frames in between. This means
that we should ensure the methods called by the test method should be
JITed as well. Earlier this was ensured because we loop till this
happens. This causes timeouts in some cases so explicitly ensure we have
JITed code.

Test: art/test.py -t 941 / 943 --gcstress
Change-Id: I04f83cb2a55fdae44cb8ebd0511fc4ef72810881