Improve handling of TurbineFieldItem.isEnumConstant()
Previously, it checked that the field was of the same type as its
containing class and the containing class was an enum but that is not
necessarily correct as there could be a non-enum constant field in
the class that references an actual enum constant.

This change just checks the access flags for `ACC_ENUM` and passes that
into `TurbineFieldItem` constructor.

Bug: 325454211
Test: ./gradlew
Change-Id: Ide46cce86ac205bc42369c3f0011f248c89cd7f5