Extract TurbineFieldValue to provide access to the initial value
Previously, the only way to access the initial value of a Turbine
`FieldInfo` was to create a `TurbineFieldItem` and call
`initialValue(Boolean)`. This change extracts that functionality into a
separate `TurbineFieldValue` so that a follow-up change can access the
initial value earlier so it can be factored into the nullability of the
fields' `TypeItem`.

Bug: 325454211
Test: ./gradlew
Change-Id: I1478d92dc9ac2b7fe60d5ea5167eb180edcc0475