Move work around for KTIJ-19087 into PsiModifierItem
The work around for the Psi issue should not be exposed in the model.
This change started out removing the work around for the Psi issue as
https://youtrack.jetbrains.com/issue/KTIJ-19087 has been marked as
fixed. Unfortunately, there still seems to be an issue with inherited
nullability on reified type parameters.

Rather than just leave the work around where it was this moves it out
of the model and into the Psi code where it belongs.

Bug: 325454211
Test: ./gradlew
      # Tested in AndroidX and Android
Change-Id: Ia04472affc412188fd0659d3a9ee92925ba32885