Remove unnecessary final modifier from enum methods
Metalava removes `final` modifier from methods in `final` classes to
normalize the model, and ensure consistency as historically the
underlying Java/Kotlin parsing code was inconsistent in how it
handled implicit or inherited modifiers. Unfortunately, due to a bug
in Metalava that normalization was not applied to methods in an `enum`
class (which is itself implicitly `final`). That bug has been fixed
and this is updating the few places where that affected these signature
files.

Bug: 326310840
Test: m checkapi
Change-Id: Ie1b6486b0416f0f2fcb1430ec11da916c671bf56