Autoupdate of 2 submodules
*frameworks/support updated to 0d5267e638b0917d07aaf819bee5fbf4b4fac16b
Merge "Temporarily disable androidx_with_metalava" into androidx-main
*tools/metalava updated to 7881480c2ca027022267cb2a6cf65daa5eb18b65
Clear final on enum methods in a final class

Previously, it would clear the `final` modifier on methods that are
not "in an enum" and not "an enum synthetic method" but it should have
cleared it on methods that are not "in an enum and are synthetic". As
the `isEnumSyntheticMethod()` already checks to make sure it is in an
enum class that can be simplified to just `!isEnumSyntheticMethod()`.
This change corrects that and updates the affected tests.

This will have a minor impact on API signature files downstream.

Bug: 326310840
Test: ./gradlew
Change-Id: Ie00e092562f2d9960f579e17300a955fb9a5d857