Autoupdate of 2 submodules
*build/make updated to 0a8fd1404e443f81a99c7f0d1f15aa7e77f0fb20
Merge changes from topics "product_runtime_apis", "product_var_hidden_api" into main

* changes:
  Introduce the product sdk_with_runtime_apis
  Introduce the product variable PRODUCT_EXPORT_RUNTIME_APIS
  Introduce product variable PRODUCT_HIDDEN_API_EXPORTABLE_STUBS

*build/soong updated to 35349466cd725513a997a2f7e1cd38e90f0397b8
Convert the build flag RELEASE_EXPORT_RUNTIME_APIS to a product variable

RELEASE_EXPORT_RUNTIME_APIS build flag is currently used to modify the
filter condition for the "exportable" stubs, so that it exposes wider
set of flagged apis (ENABLED && READ_ONLY vs ENABLED || READ_WRITE).
This flag was originally introduced as a build flag to be enabled in
specific release configurations, but instead this is used as a product
variable that will be enabled in the product `sdk_with_runtime_apis`,
which will be introduced in a future change.

Test: m nothing --no-skip-soong-tests
Bug: 326312949
Change-Id: Ieb3cee517dd6cedcca844a9eab270aedf8199d4e