Support model specific options
Previously, there was no general way to pass model specific options
through to the model which is why there was a Psi specific `useK2Uast`
parameter to `EnvironmentManager.createSourceParser(...)`. This change
adds a new `ModelOptions` class that can be used to pass options
through to the model in a type safe manner.

Although the `ModelOptions` is only used in `metalava-model-source` at
the moment it is added to `metalava-model` as it will be used by the
text model too in following changes.

Bug: 326559516
Test: ./gradlew
Change-Id: Ided61448ed9f12c3adf7a6f8b470114430552308