ApfFilter: drop ARP REQUESTs when we do not have an IPv4 address
There's no reason to wake for arp requests when we don't yet
have an IPv4 address assigned, as the kernel will just ignore
them anyway.

(This *does* rely on the stack knowing if we have an IPv4
address assigned, but presumably if it didn't lots of other
stuff [like routing] would be broken anyway)

I've observed wake-ups on GoogleGuest wifi network
when operating in IPv6-only mode due to incoming ARP request
spam.  This fixes that.

Test: TreeHugger, on a device on GoogleGuest
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: Ifbfeed287ee2e3ef52aeb9d4a84e2e9974d4f2b3