Make BasePsiTest extend BaseModelTest
Start of the work to switch the Psi specific tests, i.e. not the model
test suites to run against K1 and K2. This change is part of the work
to migrate the custom Psi tests to use the same testing infrastructure
as the model testsuite so it can take advantage of its baseline files
just in case there are differences between K1 and K2.

Bug: 326559516
Test: ./gradlew
Change-Id: I572cff6ec44b252ec3bc2d5906d708ef355a826a