Add support in BaseModelTest for specifying common sources
The Psi model currently uses the K1 parser. However, it is being
transitioned to the new K2 parser. Part of that work involves migrating
to the project structure that K2 supports which is more restrictive
then K1. It mandates that when working in a multiplatform project the
code that is common to all the platforms must be in a separate common
directory to the platform specific code. While Android does not use
that AndroidX does.

Support has already been added to the `SourceModelProvider` APIs for
the common source directory and this is adding support for it to the
`ModelSuiteRunner` API. It also migrates the `PsiParameterItemTest`
to use it.

Bug: 326559516
Test: ./gradlew
Change-Id: I041ab9186aa5cb1f7edf5a3d964043df7e1917d2