Run metalava tests for both PSI K1 and K2
Previously, they were only run for Psi K1 which means that a lot of
functionality that Metalava relies upon was not being tested properly.
This change makes metalava tests run against K1 and K2 so that any
issues could be found.

Some tests are already K1/K2 aware so they have been annotated with
`@FilterByProvider("psi", "k1"|"k2", action = EXCLUDE)` to ensure that
they only run on K1 or K2 as appropriate.

One existing, not K1/K2 aware test fails on K2 so that was added to the
baseline file.

Bug: 326649406
Test: ./gradlew
Change-Id: I9f41c3608c392eeae482bd5baae61a747a0e1286