Dedup SourceModelTestInfo and ModelProviderTestInfo
The `SourceModelTestInfo` and `ModelProviderTestInfo` perform a similar
function when running `DriverTest`s and model test suite tests, i.e.
encapsulating the information necessary for creating a `Codebase`.
Whether that `Codebase` creator was a `SourceModelProvider` or a
`ModelSuiteRunner` respectively.

This change combines the two into a single `CodebaseCreatorConfig`
object that is parameterized by the `Codebase` creator. It was
necessary for `SourceModelProvider` implements to return their
`providerName` as part of the `toString()` method as that is used to
generate the test name.

Bug: 326649406
Test: ./gradlew
Change-Id: I49361b26d4f85b4c287f14c784a6d896e0a1af99