Remove workaround for @Ignore failing with DriverTestRunner
Previously, the `DriverTestRunner` would create a `Description` for
`CheckAndroidApisShard4Test` that looked something like this:

```
    CheckAndroidApisShard4Test - #1
        CheckAndroidApisShard4Test - #2
            Test All Android API levels[1 to 2]
  	    Test All Android API levels[3 to 4]
	    ...
```

The #1 layer was created by the `CustomizableParameterizedRunner`
(actually by its super class `ParentRunner`) and the #2 layer was
created by the wrapped `parameterized`. That caused a stack overlow
error in Gradle's handling of `@Ignore`.

Basically, when `@Ignore` is specified on the class it only fires a
single `testIgnored` event for the whole class but Gradle needs one per
method.

When JUnit fires a `testIgnored` event for #1. Gradle receives that,
 and because the class is annotated with `@Ignore` it looks like this,
 i.e. no children:
```
    CheckAndroidApisShard4Test - #3
```

Gradle detects that it is a class `Description`, and attempts to
retrieve a `Description` for the class without taking into account the
`@Ignore` annotation that will contain children for which it can fire
`testIgnored` events. That results in a `Description` that looks like

That is where the problem likes because despite the fact that the
`Description` already has children Gradle treats it as if it was an
event fired from JUnit and again retrieves a `Description` that looks
like #1. It repeats that until it runs out of stack.

This change avoids that by simply delegating the `getDescription()`
method to `parameterized` removing layer #1 added by the `ParentRunner`
super class. So, the description now looks like this:

```
    CheckAndroidApisShard4Test
        Test All Android API levels[1 to 2]
        Test All Android API levels[3 to 4]
        ...
```

Bug: 326649406
Test: ./gradlew
Change-Id: Idfa08adf53c940746d9cdedf9b61d5d801c40048