Fix another issue with DriverTestRunner and @Ignore
Previously, a test that was run with `DriverTestRunner` and had no
additional parameters would end up creating a `Description` something
like this:
```
    FooTest - #1
        FooTest - #2
	    method1
	    method2
```

That is similar to the structure in https://r.android.com/2982073, and
would cause the same stack overflow issue if the class was annotated
with `@Ignore`.

This change uses `[]` as the name of the runner so the description
would look like this:
```
    FooTest - #1
        []
	    method1
	    method2
```

That avoids the stack overflow.

Bug: 326649406
Test: ./gradlew
Change-Id: Iec523665cad32d1ddc882e2c986f2780c9ad97a5