Remove unnecessary arguments from test name suffix
Previously, the test name suffix was ignored and replaced with an empty
string if it matched "[psi]". This change improves that support to
remove other extraneous arguments from the test name suffix.

It moves the logic for removing arguments from `InstanceRunnerFactory`
to `CodebaseCreatorConfig` allowing the `BaseModelProviderRunner`
subclasses to control how that behaves. It adds support for ignoring
not just the provider name but also the optional `InputFormat`.

The `ModelTestSuiteRunner` always ignores provider name (as there is
only ever one provider each time it runs) and ignores `InputFormat` if
the provider only supports one of them.

The `DriverTestRunner` just ignores the provider name if it matches the
default provider of "psi".

Bug: 326649406
Test: ./gradlew
Change-Id: Ife8b45ad5235de596ebaaac878d5c069c11819c6