Embed additional argument sets inside ModelProviderWrapper
Previously, a test that has its own `@Parameters` annotated function to
provide its own additional argument sets to run tests over would end up
with a set of arguments in `TestWithParameters` that looked something
like this:
```
    listOf(ModelProviderWrapper(...), argument)
```

The problem with that is it relied on the `Parameterized` class to
generate the suffix for the test name and if `ModelProviderWrapper`
returned an empty string then the suffix would be `[,argument]` which
is confusing.

This change moves additional arguments inside `ModelProviderWrapper`
so the `TestWithParameters` arguments looks like:
```
    listOf(ModelProviderWrapper(..., listOf(argument))
```

`ModelProviderWrapper` can then ensure that the suffix does not end up
with a leading `,`.

Bug: 326649406
Test: ./gradlew
Change-Id: I320713f415a142eebe2af5c214c5bc25d66c6bb6