skip if entry has been added
Before if an entry name exists, an error will be thrown. This change
will check whether the existing processed entry is the same entry as the
one currently processing. If they are the same, then the code will
continue to process the next entry. Thus this entry will be only added
once. If they are not the same, then the original error will be thrown.

Test: presbumit
Bug: 327404701
Change-Id: I36c6f13c7899a28534de6d56bcbc8896a60cfd47