Change the way we find the latest tag
We used to find the latest tag with git describe but it's not always
accurate because alpha, beta, or rc tags are not what we are looking
for. Instead, we find the the latest tag by parsing the current tag and
matching the list of all references against the pattern.

Test: updater.sh check python/cpython3
Change-Id: Ia8ab1848bc245ff5f94b4a75a97e030759544914