Replace ApiLintTest usage of uastCheck(...) with check(...)
Now that the tests are run with K1 and K2 it is no longer necessary to
explicitly manage that in the test. This change fixes up `ApiLintTest`
use of `uastCheck(...)` to just use the standard `check(...)` function.
That did reveal that the `Options.useK2Uast` was not being set
when passing the `ModelOptions` through the `TestEnvironment` as it
would be when passing the command line option. So, this change fixes
that too.

Bug: 326649406
Test: ./gradlew
Change-Id: I508841b92b7be879d800b17b7c6437cf8ce0755c