Remove UastTestK* methods that delegate to UastTestBase methods
They provide no value. One of the K2 methods was marked as `@Ignore`
and that was converted into an `@FilterByProvider` annotation that
prevents the test being run for K2.

The methods in `UastTestBase` where changed to be public test methods
but the `UastTestK*` classes were kept as it simplifies running the
K1 and K2 tests and allows for better sharding when running tests.

Bug: 326649406
Test: ./gradlew
Change-Id: I92e6867abc4768c9b4b48495fc81aa78a7856523