Push choice between K1/K2 behavior up into UastTestBase
Most of the tests in `UastTestK*` either have the same behavior in both
or choose the expected behavior in `UastTestBase` depending on the
value of the `DriverTest.isK2` property. However, there are a couple of
tests which provide the expected behavior in the `UastTestK*` function
by passing it as the `api` parameter to the `UastTestBase` method.

This change removes the `api` parameter from those test methods and
moves the choice into the `UastTestBase` method.

Bug: 326649406
Test: ./gradlew
Change-Id: I579f282605a5d54597324c6698ff9fc7261a1f2f