Add Item.baselineKey
This moves the construction of the `BaselineKey` for `Item` from
`Location` to `Item` removing the dependency from `Location` on `Item`.
This will make it possible to move `Location` into `metalava-reporter`.

There was a slight issue with `PropertyItem` as it did not have its own
case in `getElementIdForItem(...)` and fell back to `describe(...)`.
This change preserves that behavior as otherwise it would break
AndroidX. It will be fixed in a follow-up change.

Bug: 327562507
Test: ./gradlew
Change-Id: I8bcb64fc12cb15ab97c1b8aaf86299852647055f