Revert "libnativeloader_test: link non-public libs statically" and remove libnativeloader_test from CTS instead.
This reverts commit b2b067192ac047b99c0d03104c5d7dbf79192ef1 which made
the CTS test no longer work for API coverage. It also (effectively)
reverts commit 996fa745c3ad3e3150363c85816ad2db3a00f5bd which put it in
CTS in the first place for that purpose.

It's clear libnativeloader_test won't work for API coverage due to the
mocking it does, so we'd have to add new tests if there are gaps.
However there doesn't seem to be any.

Also do a few other cleanups in the Android.bp file.

Test: atest libnativeloader_e2e_tests \
        libnativeloader_test libnativeloader_lazy_test
Bug: 327407752
Change-Id: I488c3d3c5143b7694a8456c3b56edfd511c3c610