Add Reportable.fileLocation
`IssueLocation`s created for `Item`s use a `baselineKey` that is common
for all models. It is only the `FileLocation` that differs. This change
adds a `Reportable.fileLocation` property and then implements
`Item.issueLocation` in terms of that and `Item.baselineKey`.

This fixes a few tests on Turbine as previously the default
implementation of `Item.issueLocation` returned an `IssueLocation` with
`FileLocation.UNKNOWN` and `BaselineKey.UNKNOWN`. However, this change
causes it to return an `IssueLocation` with `FileLocation.UNKNOWN` and
a correct `BaselineKey` for the `Item`.

Bug: 325962341
Test: ./gradlew
Change-Id: I6d6353f8f9bb2c4472315f13e833d45f981f972a