Add config to hide hidden parent methods
The lint for @RestrictTo considers a method hidden if it is defined on or overridden by a hidden parent and is not overridden by a visible class lower in the hierarchy. This changes the default dackka behavior to match RestrictTo but leaves an option for the previous behavior of including hidden parent methods/fields because that is more similar to what metalava does.

This revealed an issue with class resolution of some kotlin classes for KMP common sourcesets (b/327630237), where APIs from unresolved parent classes were appearing as methods defined on the class instead of inherited. This leaves them in, but they still appear in the wrong section.

Bug: 323045178
Test: Updated existing tests, added test for non-default behavior. Preview of AndroidX changes in cl/611118494

Change-Id: Ib1433b0220e885864c66dc9e58965e79b97a1bb0