Default wpa2 network when scan result unavailable
Sometimes scan results does not cantain information about a network.
when this happens, we fail to get the networks type and fail to connect.
Instead of failing, default to wpa2 network since they are more common
in lab setup and if the default network type wpa2 is still wrong, it
will fail to connect anyway.

Bug: 311053528
Test: Presubmit
Change-Id: I9fc2f546f1c1d3fd1d28a0214dfbd94f49868ee8