Run all metalava tests against turbine
Enabling the tests for Turbine results in lots of test failures so
this creates a baseline file that can be used to track the impact of
following changes.

Some of those test failures resulted in stack traces and other messages
being dumped into stderr. So, this change stops just printing the error
messages to stderr and instead reports it as an `InternalError`.

Bug: 326649406
Test: ./gradlew
Change-Id: Ie5947e7c82bf1ec9d69ce0a3e963824dbb0bbbff