Autoupdate of 2 submodules
*external/python/cffi updated to de77edbe02d4cc29c5f218144e41d4e708244566
Empty external/python/cffi

This module contained prebuilt binaries in its source. With the current
Python support, it is not possible to build this module with Soong.

Bug: 205265538
Test: presubmit
Change-Id: Ib0bcd46f6e91ce37057ac36b9880968ffd72a5ed

*external/python/cryptography updated to 4bdc0923eadc5c0978999c18a8c155116099c520
Empty external/python/cryptography

This module contained prebuilt binaries in its source. With the current
Python support, it is not possible to build this module with Soong.

Bug: 205265538
Test: presubmit
Change-Id: I11ea12c9de6688a0256edfaa785b897fa55bca06