Add minimum capabilities needed by a BaseModelProviderRunner
Adds a `Capability` enum containing a set of capabilities that the
provider may provide and specifies that for each of the providers. The
`BaseModelProviderRunner` can require a minimum set of capabilities
from providers in order to run the tests.

Sets the minimum set for the `ModelTestSuiteRunner` to empty and for
the `DriverTestRunner` to `JAVA` and `DOCUMENTATION`.

Bug: 326649406
Test: ./gradlew
Change-Id: I56d993183bdd31c890bd3ee750f0ad2cd1ff13d8