Bugfix: pass country code to Config even it's invalid.
When country code is switching between invalid(00) and valid.
UWB should be toggled between enabled/disable accordingly.
But it didn't happen as Config detect it hasn't changed.

To fix this, always pass country code even if it's invalid ones,
so Config detects it was changed. And forcing uwb disabled whenever
the country code is invalid in COUNTRY_CODE_CAPS and ExtCal
handlings.

Bug: 327504423
Test: Check uwb status with country switching 'KR-00-KR'

Change-Id: I50ab874bf640b2cd20d1f05a6da7b7831d1d8da1