Fix the issue with nullability of non-last Kotlin vararg parameter
A previous change to fix up the nullability of Kotlin vararg parameters
only worked if the vararg was the last parameter because it relied on
the Java view (through `Psi...` classes) of the vararg status and that
only supports the last parameter being a vararg. This change uses the
`ModifierList.isVarArg()` to check the vararg status as that shows the
Kotlin view.

Bug: 325454211
Test: ./gradlew
Change-Id: I802ff760b233f6adf40306a251c873e45f717e97