Get correct KtType for synthetic setter methods on data class
Previously, the code for getting the `KtType` for a synthetic property
setter method assumed that the `sourcePsi` backing the method was the
`KtProperty`. However, in the case for properties defined as
constructor parameters, e.g. in a data class, that is not true. They
are backed by a `KtParameter` instead. This change adds support for
that.

This eliminates a discrepancy between the nullability annotations on
the parameter, i.e. `@Nullable` (which is correct) and the type
nullability retrieved from the `KtType`, i.e. `NONNULL` (which is
incorrect). That does not currently cause an issue as in
`TypeItem.fixUpTypeNullability(...)` an `@Nullable` annotation on the
parameter overrides `TypeNullability.NONNULL` from the `KtType`.
However, the discrepancy causes confusion and is preventing further
work to remove `TypeItem.fixUpTypeNullability(...)`.

Bug: 325454211
Test: ./gradlew
Change-Id: I47b858edeb4d68e2a0b3e6d5b5613abb0ae7ee4b