Get correct type for synthetic data class copy(...) method
The parameters on the synthetic data class `copy(...)` method currently
have a backing `sourcePsi` of the `KtClass`. It seems like it would
make sense for them to be backed by the corresponding `KtParameter` in
the primary constructor but it does not. This change works around that
by getting the corresponding `KtParameter` from the primary constructor
explicitly and then using the type from that.

This eliminates a discrepancy between the nullability annotations on
the parameter, i.e. `@Nullable` (which is correct) and the type
nullability retrieved from the `KtType`, i.e. `NONNULL` (which is
incorrect). That does not currently cause an issue as in
`TypeItem.fixUpTypeNullability(...)` an `@Nullable` annotation on the
parameter overrides `TypeNullability.NONNULL` from the `KtType`.
However, the discrepancy causes confusion and is preventing further
work to remove `TypeItem.fixUpTypeNullability(...)`.

Bug: 325454211
Test: ./gradlew
Change-Id: Ifbb3e619ad20544942d9a9caa9872b1dbe8abfa1