Split ApiGeneratorTest into separate tests
Previously, `ApiGeneratorTest` had 4 tests that were basically
integration tests that ran across all the previously released jars in
`prebuilts/sdk` and took about 15-20 seconds each which is quite a long
time for unit tests. As they run sequentially the test was taking well
over a minute to run. Recent changes to run the tests against K1, K2
and Turbine almost tripled the run time, adding it to the critical path
for the tests and so significantly increased the run time for the tests
as a whole.

This change splits each of the 4 integration tests out into their own
test class which allows them to be run in parallel taking them out of
the critical path.

Bug: 326649406
Test: ./gradlew
Change-Id: Ibbafd71e0f00748d6b47acf96788da7f4ddc38d5