Improve test coverage of nullability of type parameters
Uses of type parameters that are unbounded, or are explicitly bounded
by `Any?` and which do not explicitly specify that it is non-null have
an unknown nullability as their nullability depends on the nullability
of type that is assigned to them. The same is true for any nullable
bound. That requires some special handling in determining the
nullability of type parameter uses.

Previously, there was only one test that checked for an unbounded type
parameter to make sure it was correctly assigned `UNDEFINED`. This
change adds tests for other cases, like when its bound is not nullable
or its use is explicitly nullable or its bound is nullable but not
`Any?`.

Bug: 325454211
Test: ./gradlew
Change-Id: I4738b3722e3c413d17ffab8f790a433cdcd92565