Add kotlin stdlibs to the classpath
This is needed to allow for testing of `suspend` functions. It also
fixes an issue with a property delegate test that was breaking on "k2".
The property delegate test was setup incorrectly and instead of
accessing the `ReadWriteProperty` class it used `java.lang.Object` as
the former was not available. That test was failing on "k2" as it
appears to be stricter and uses an error class instead which is why it
was failing previously. Adding the Kotlin stdlibs to the class gives it
access to the `ReadWriteProperty` class which works in both "k1" and
"k2".

Bug: 325454211
Test: ./gradlew
Change-Id: I2e45625ecb02a6172b623dcfdac84c16618d4597