Stop assuming that Kotlin types are nonnull
Previously, `KotlinTypeInfo.nullability()` assumed that if no `ktType`
was found that the type should default to `TypeNullability.NONNULL`.
While that is true more often than not it is not true all the time and
when it is wrong, it can confuse API reviews and/or hide issues where
Metalava does not correctly understand the Kotlin structure.

This change is the culmination of lots of changes to fix the existing
issues in Metalava and corrects the AndroidX API signature files.

Bug: 325454211
Test: ./gradlew
Change-Id: I5556abb7b1991600e102e7b08b17cdbc485f2d7c