Remove DEFINE_2OP_ARITHMETIC_INTRINSIC_VS intrinsics.
As a result of refactoring when masking was moved partialy to separate
intrinsics and partially to the interpreter we ended up with pretty
bizzare intrinsics that take two scalars and produce the desires result
in all elements of vector. And then interpreter takes one element from
that result.

Such intrinsics technically work but are entirely useless and to design
sensible handling of reduction instructions we first need to run real
apps on the emulator and collect statistics: what kinds of reductions
are in actual use, do they use masking or not, etc.

Test: berberis_all

Change-Id: I34929a3f38ca06e4dad3683e3052c236dc673616