Avoid tag recommendation if repo has no tags.
get_latest_stable_release_tag raises ValueError when passed an empty
list. In this case we should be returning None to signify that there
is no latest tag.

This was caught by tests/endtoend/test_check.py, but those can't be
run in CI because we can't run repo in CI, so we can't set up the test
directories. For the same reason it's rather difficult to set up a
unit test to cover this, since GitUpdater.latest_tag_of_upstream
assumes that it has a git repo that's been set up with the
update_origin remote. With some effort that could probably be hacked
around, but it's beyond the scope of this fix.

Bug: None
Test: make test
Change-Id: I817f46543a38777b109a22f2064580061598ead2