Set COMPILER_RT_HWASAN_WITH_INTERCEPTORS=OFF in NDK builds as well.
The NDK version of the HWASan runtime is built with interceptors
enabled. This is a holdover from when we previously intended that apps
using HWASan would ship their own copy of the runtime with interceptors
enabled. But now the runtime is expected to live on the device so
apps shouldn't be linked expecting the interceptors to be there.

This wasn't a problem until recently when more interceptors were added
to the runtime, when among other symbols, _ZnwmSt11align_val_t was
added to the interceptor list, causing apps to try to link against
a non-existent symbol of that name on the device and fail.

Fix it by also disabling interceptors in the NDK build of the runtime.
Confirmed that with this change there are no differences in the
list of defined exported symbols:

$ nm -DU out/stage2-install/runtimes_ndk_cxx/libclang_rt.hwasan-aarch64-android.so | cut -c17- > /tmp/1
$ nm -DU out/stage2-install/lib/clang/18/lib/linux/libclang_rt.hwasan-aarch64-android.so | cut -c17- > /tmp/2
$ diff -u /tmp/{1,2}
[no output]

Bug: 328681977
Change-Id: Id003a1c1124978288eed5b1b8293e28dc006028e