Autoupdate of 2 submodules
*system/logging updated to 81bacf97371d8ef5b2fa0e76601686903d3ddb9a
Make the old structured log API internal.

The new macro simplifies the old API, but requires a backend to
log_event_list. If we merge both into a single implementation, we
essentially implement the same helpers as the existing API provides,
i.e. we use it for the macro generation, but mark it internal.

Bug: 290589708

Test: Run keystore, integration test, and new sdv sample
Change-Id: Iaddff12b6139c492185b3526837f96f65c7e5863

*system/security updated to e2ce4fd6428cf719a1b28d649d0c739b98492684
Merge "Migrate structured logging for audit logging to the Rust macro." into main