Ensure consistent handling of typealias types in K1 and K2
K1 appears to always fully expand `typealias`es, i.e. replace them with
the type for which they are an alias but K2 does not. That caused a
difference in behavior between K1 and K2 and meant when using K2 that
the nullability of type arguments of the aliased types was handled
incorrectly.

This change fixes that by ensuring that `typealias`es are always fully
expanded.

Bug: 325454211
Test: ./gradlew
Change-Id: Ida681430740d7599fc26e0858778331f254d2922