Show "vm" output on failure
If the "vm run-app" command fails, we get an IOException while trying
to parse the CID from the output. But then there is nothing in the
logs to indicate why it failed, which makes it very hard to diagnose.

So, collect the output, and include it in the error message on
failure.

While I'm here, make sure we clean up the spawned process.

Bug: 328610193
Test: Induce failure, see error message in the exception
  atest MicrodroidHostTestCases
Change-Id: I30f0a18f7885bd7de791d63a93a1d423d86eb13f