Move handling of suspend functions return type to KotlinTypeInfo
Previously, `KotlinTypeInfo` would return the actual return type for
the suspend function (not the `Any?` to which it is transformed). As a
result it could have the wrong nullability but that was fixed up by
applying the `@Nullable` annotation that Kotlin added to the suspend
function. Care had to be taken in `PsiModifierItem` to ensure that the
annotation was not removed.

This change updates `KotlinTypeInfo` to use the correct return type
of `Any?` in the first place and removes the special handling of
suspend functions in `PsiModifierItem`.

Bug: 325454211
Test: ./gradlew
Change-Id: I5667ace11f4bdf1a173df5fdd9830c3cff870cb7