Make `Init()` and `Deinit()` reentrant for `HalInterface`
Thread API `leave()` relies on the ot-daemon's `Application` class
to de-initialize and re-initialize itself. As a subcomponent,
`HalInterface` needs to support this.

It seems that `ABinderProcess_setupPolling` can only be called once
throughout the lifetime of the service. Therefore I'll move this step
from the `Init()` to the constructor. Also in `Deinit()`, the file
descriptor won't be closed.

Bug: 323301831

Change-Id: I7ebae3f8d31d5a959bc33b2ffda80db08aadf54d