Update content uploader to handle chunk fallback
The name of the chunked version of an artifact has the prefix "_chunked_", to distinguish from the regular version (unchunked).

For an artifact that has `chunk` and `chunk_fallback` true, there will be one entry in the CAS digest for each version. The downloader should check CAS digest for the availability of the versions and download the desired one.

Added logic to avoid redundant upload when multiple ArtifactConfigs share files.

Bug: 320561670
Test: TH
Change-Id: Ia3df058b3dad6e74d348d909074378679f846929