Preserve nullability information from type parameter use if needed
If a type that is being substituted for a type parameter has platform
nullability, i.e. provides no information about whether values can be
null or not, then use the nullability information from the type use
instead.

e.g. in the following, after substituting `String` (which has platform
nullability) in for `T` in `T foo()` it will have
`TypeNullability.NONNULL` from the `@NonNull` annotation.
```
class Hidden<T> {
  @NonNull
  public T foo() {...}
}

public class Foo extends Hidden<String> {
}
```

This does not change the existing behavior as the
`TypeItem.fixUpTypeNullability(...)` method will apply the `@NonNull`
annotation to `String` after being substituted for `T` in `T foo()`.
However, this does allow for the `TypeItem.fixUpTypeNullability(...)`
method to be removed in the future.

Bug: 325454211
Test: ./gradlew
Change-Id: Ia81e56fd0cdff07655861bb4f3bfc4787f4fa812