trusty: mmutest: restore run_x test, and add run_wx test
Restore the run_x test, which shows code can execute from a dynamically
allocated region.  Update the mm_test_execute() function that creates
the executable region so that it does not violate W^X semantics.

Add a test for run_wx, to check WXN works on ARM64 where it is
supported.

Bug: 328206729
Test: build.py qemu-generic-arm64-test-debug --test com.android.kernel.mmutest
Change-Id: I8519aad2bd547f5eca4bcadc7fcb007f98996b71