Remove PsiConstructorItem override of fileLocation
Previously, in order to fallback to the containing class's
`fileLocation` property if needed the `PsiConstructorItem` overrode the
`fileLocation` property. This change removes that override and instead
passes the `fileLocation` as an optional constructor parameter to
`PsiConstructorItem` and its super class' constructors. The
`PsiConstructorItem` created by `createDefaultConstructor(...)` passes
in the containing class's `fileLocation` property directly removing the
need for the `FallbackFileLocation` class altogether. That matches the
behavior of Turbine.

Bug: 325962341
Test: ./gradlew
Change-Id: I00a81827645a1f514e701d11ca3aa5894d0a1b38