Improve Native Coverage latency
* The current implementation works by sending signal 37 to
the testing process and busy-looping for 60 seconds or until it
has witnessed the testing process act upon the signal. Sometimes
we miss the signal completely which leads to a 60 second overhead
to the test runtime. Changing the deadline is infeasible as we
don't have any estimation on how long does it take to dump the coverage.

* This CL improves the current implementation by completely eliminating both
problems, we don't busy loop instead we call directly the method to dump
the profile data to the disk and then we exit, there is an introduced
latency for loading the native library but that should be negligble.

* Maintained backward compatibility with previous use-case by falling
back to the old implementation if we fail to load the native library.

* The next step should be migrating everyone over to the new way and
incorprating this change directly into Soong so that it would add
the native library + java dependency to all of the targets.

Bug: 329680904
Test: atest NetHttpCoverageTests && confirmed that coverage did not change
Change-Id: I966f4e2923b75258831857af3d68bd72685d93f6