Don't run tests that write to storage
aosp/2958037 re-enabled various unit tests for the KeyMint TA, but
failed to notice that some of those tests write to secure storage
and thus are destructive.

So disable any tests that look like they might affect the state of the
device under test. (This includes some tests that were already enabled
before aosp/2958037 landed, just in case. These are probably OK because
they write to the secure storage for the test app, not the TA itself,
but best to err on the side of safety.)

Test protocol:
- Run VtsAidlKeyMintTargetTest --gtest_filter="*EcAttestKeyChain*default"
  beforehand (which checks a batch key attestation chain)
- Run `trusty-ut-ctrl com.android.trusty.rust.keymint.test` to exercise
  these unit tests.  Check which tests are included in the list.
- Run VtsAidlKeyMintTargetTest --gtest_filter="*EcAttestKeyChain*default"
  again afterwards.  Prior to this CL, this would fail because the unit
  test run had damaged the attestation keyboxes.

Bug: 326564087
Test: see above
Change-Id: I3a62f4ee6d02f31ddbdfa747c7d944c366a4fbe2