Attempt to deflake HeapprofdJavaCtsTest Oom tests
The perfetto ART hprof plugin works by:

* Forking a process with an ART instance.
* Connecting to perfetto in the forked process.
* Iterating the java heap.
* Writing info to perfetto.

The process is not bulletproof and may fail. If another thread is
holding a lock while the main thread is forking, the forked process
might try to acquire said lock and be stuck there.

Other threads can hold all sorts of locks, including ART internal locks,
bionic internal locks and so on.

The code tries to pause other threads as much as possible, but it
doesn't have a way to completely prevent the problem.

This is a known limitation of the ART perfetto hprof plugin.

Even if the feature is known not to work 100% of the times, we still
need a way to test it reliably.

The perfetto hprof plugin can start (and therefore fork) in two cases:
1. When the user explicitly requests a snapshot by sending a signal.
2. When an Oom happens (and a tracing configuration is active).

The testcases for 1 always worked reliably, because, when the test
explicitly requests a snapshot, the app is not doing much, therefore
it's very unlikely that other threads are holding locks.

The testcases for 2 on the other hand, fail more often (~0.5% of the
times on an emulator on my machine, but probably more often in CI).

The problem appears to be that the test app for 2 causes an OOM
immediately on Activity.onCreate(), as soon as the app start. There are
very likely a lot of other threads running while the app is starting, so
it's possible that they will hold lock while the OOM thread is forking.

Waiting for 500ms appears to give other threads a chance to calm down.

The solution is far from perfect: testing a feature that by design
doesn't work reliably makes this especially hard.

We could implement retries instead, but it would require complicating
the tests a little. Also, we would still need an arbitrary timeout to
decide that the forked process is stuck. We might revisit this in the
future.

Bug: 329124210
Change-Id: I8562ca96d663ecb6c679c6069d67e4edae84917b