Add embedded_launcher: true to jni_generator
The jni_generator writes the path to the script into its output files.
When using a non-embedded-launcher soong python binary, soong will
extract the binary to a nondeterministic temporary directory before
running it. These two points together means that the output files
of jni_generator are non-deterministic.

Adding embedded_launcher: true causes all the python sources to be
precompiled as well, so we start getting syntax errors for code paths
that are still in python2 but we weren't running before. Fix those
syntax errors.

Test: Presubmits
Change-Id: I0a856f8cc98b11841c07bd6188d9722980428b5f