Upload oriole divice images with chunking only
Updated logic so that if the chunked version is uplaoded without fallback, the regular version is skipped even if it matches other artifact configs.

This way "cas-chunk" flag doesn't need to be specified when there is
only a single verion of an artifact is available in the digest. In other
words, the "cas-chunk" is only needed for verifcation of rollout of new
artifacts.

Upadted oriole device image artifact pattern to apply to oriole_hwasan.

Test: Uploaded with various chunk and chunk-fallback flag combinations.
Bug: 320561670
Change-Id: I3a3be195e3ec0d85661a33fca0f72eb587cc975d