Cleanup: use actual buffer size in read()
phTmlUwb_spi_read() calls read() always with 4 buffer count
disregarding the actual buffer size(4200).

The current scheme works because both the kernel driver and
HAL assume the actual buffer is 4200 bytes, even if the given
length is 4.

Fix it to provide the actual buffer size to the kernel.
The kernel driver anyway doesn't check the count but it also
should be fixed.

Additionally, print errno on errors from read()/write()
instead of its return value.

Bug: 321604848
Test: uwb hal init
Change-Id: If0597b578836d9a952cf3140c0174805ce0d0b0c