aconfig storage daemon: handle new storage registeration socket request
1, Update aconfig storage daemon service into two parts (see aconfigd.rc).
aconfigd-init as the first service target that is meant to be blocking
and will get executed first by init. Then the service aconfigd will get
started by init in a non blocking fashion. So we don't need to fork like
before.

2, Update the socket setup to use stream socket type. Also, added code
logic to handle an incoming request of a new storage file set.

3, Add an integration test to send a new storage request to aconfigd
socket and lock down the new storage registration.

Bug: b/312444587
Test: m, avd and then atest

Change-Id: Ide9b8d7829e7716bca230f79b6fe9e62486f9e46