Add support for an APF read adb command
This change adds an APF read adb command that reads the currently
installed APF bytecode and returns it as a hex string.

The read command requires to wait on mApfDataSnapshotComplete without
blocking the handler thread. Hence, the read command is triggered in a
thread-unsafe manner from outside of the handler thread (this really
should not pose an issue, since these commands are solely used in a
controlled test environment).

Reaching into ApfFilter to get this information is not a great solution.
Changing the program under ApfFilter's feet might cause all sorts of
weird behavior (for one, the counters it knows about may not exist
anymore etc.)
In a perfect world, "apf pause" destroys the APF filter and registers a
barebone test instance that subsequent adb commands can reach into.
Furthermore, IP provisioning / IpClient should be restarted after
running the tests, so ApfFilter can sync back up with the state of the
kernel (this might not actually cause a problem, since eventually, they
should get back in sync).

Unfortunately, mostly due to the existence of LegacyApfFilter, our hands
are currently tied. There is simply no (good) way to make the
changes mentioned above (one problem being the RA receive thread that
lives inside ApfFilter, which currently makes it hard to synchronize APF
filter operations with IpClient).

Test: adb shell cmd network_stack apf wlan0 read
Change-Id: Ie6c7bd6154a44cf3a6bd7b7d8b8fdecf6e547573