input_helper: Modify the check range of the input event number
If the input event number is greater than 32, the newly added event ID
will be 256 or later.
When there were already 32 input events in the device, this test program
only checked 0~99, so it failed because it could not find a new input
event ID.

In order to eliminate this failure, we changed the event ID range to be
checked by this test program to 0~1023.
(https://github.com/torvalds/linux/commit/7f8d4cad1e4e11a45d02bd6e024cc2812963c38a, it looks like the upper limit should be 1023.)

Bug: 330269984
Test: ltp
Change-Id: I1b53008b15641fcdccf214af9183de1c29a3309c