Autoupdate of 2 submodules
*external/ot-br-posix updated to 97d639a9d58fdeae7a3fd1bce5728142c446334a
[Thread] add terminate() API to force stop ot-daemon

The Thread HAL supports only one client for a single Thread chip, so
ot-daemon needs to be force stopped in VTS tests because the testapp
needs to connect to the HAL service to perform tests.

This commit adds a new terminate() API to ot-daemon to stop itself.

Bug: 328538612
atest: atest ThreadNetworkIntegrationTests:android.net.thread.ThreadNetworkShellCommandTest
Change-Id: Ic6000eaf2a912e3869048efdc0816b217b81e970

*packages/modules/Connectivity updated to 82599d3e21130ba4b642fc5a4552b3fbe02de2c1
[Thread] add command to force stop ot-daemon

The Thread HAL supports only one client for a single Thread chip, so
ot-daemon needs to be force stopped in VTS tests because the testapp
needs to connect to the HAL service to perform tests.

Bug: 328538612
Change-Id: I5288e5747ab3957050d0e85c1ea684c6ead6c87d