First commit for libstorage testlib
Move TestBlockDevice into a standalone module and export it as a new rust_library.

Implement a builder pattern for flexible customizations with
reasonable defaults.

Remove duplicated implementations of TestBlockDevice in other test code
and replace uses with centralized implementation.

Note: the consumption and re-export of storage lib traits is an
unfortunate workaround for using Vec for block storage and scratch buffer.

Tests: all unit tests pass

Change-Id: I936a75dc2a4d30c7cab30af7001362010f0617db