Let exceptions escape onCommand()
Exceptions in this function are caught by its caller and their content
is written to stderr. When using any of these commands in tests,
returning an error code is not enough as
SystemUtil#runShellCommandOrThrow (and any other run shell command
utilities in Android, see
UiAutomationConnection#executeShellCommandWithStderr) ignores the error
code and only checks whether anything was written to stderr.

Test: tested on cuttlefish
Change-Id: I1c09cba2e2a6b152d1907e968a120a48e89b7a34