Include FlaggedApi annotations in the stub files
Previously, they were not included as they were only `SOURCE` retention
and so would have been useless in stub files. Since then change
https://r.android.com/3000696 switched them to `CLASS` retention so
that they are available to Android lint when running against compiled
classes so it makes sense to include them in the stub files.

Adds `android.annotation.FlaggedApi` to `stub-annotations` so that the
stubs will compile correctly.

Bug: 331251171
Test: ./gradlew
Change-Id: I8e0b8dc5e030dd074868a34dcd66b0bc0dc8263c