libfmq: add mode where message size is only known at runtime
we do this by specializing on a sentinel enum payload type, MQErased. in other cases the compile-time knowledge should still be available to the compiler.

this is safe to do because everywhere else, the MQDescriptor's notion of the quantum is used, and in this file we assert that the static (sizeof(T)) and dynamic (mDesc->getQuantum()) sizes are equal.

Bug: 294698705

Test: fmq_unit_tests and fmq_test
Change-Id: Icc938ca9f6508865584d9b26059d4d2cf0109a10