Do not duplicate a field from an interface onto itself
The `ClassItem.filteredFields()` method will search for fields to
include in the class in the transitive set of interfaces that it
implements. Previously, it would try and duplicate the field into
itself because when `allInterfaces()` is called on an interface the
interface is included in the returned list. That duplication is
unnecessary as the interface's fields are handled separately. Also, it
runs the risk of introducing bugs due to differences in the `FieldItem`
returned by `FieldItem.duplicate()`.

This is a follow up for the fix to the issue revealed in
https://r.android.com/3027423. While https://r.android.com/3027424 was
a better fix as it is more general this change would also have fixed
that issue because it was caused by a difference in the `FieldItem`
returned by `FieldItem.duplicate()`, i.e. the duplicated `FieldItem`
did not set `hidden` correctly because `originallyHidden` was not set
correctly. Change https://r.android.com/3027424 corrected the setting
of `originallyHidden` in both.

Bug: 332611311
Test: ./gradlew
Change-Id: I7f66c96dcfa688749b682a2e06abe64d760b785b