Android patch: intern() LSR loaded from langinfo.txt
Previously, LSRs loaded from langinfo.txt were
a list of language, script and region strings. My theory is
that they were linked to same string value in the .dat file,
and thus can be shared when loaded into ART.
However, it is renamed to be lsrnum, and becomes an intvector,
and requires decoding via String.substring() and thus needs to
intern() manually to reduce the zygote heap usage.

It saves ~1MB zygote heap.

This patch is applicable on other runtimes and can be upstreamed to ICU.

Bug: 331291118
Test: am dumpheap and inspect the zyogte heap usage
Change-Id: Ic652477c4f82698d078dea343f9188bc350a4cd6