protozero: Fix decoding repeated fields with high max id
The current implementation of TypedProtoDecoder is supposed to store
repeated fields **after** all the field.

This worked for messages where MAX_FIELD_ID is small. For messages where
MAX_FIELD_ID >= PROTOZERO_DECODER_INITIAL_STACK_CAPACITY -1, the
implementation pushed the repeated field at the end of the stack buffer,
instead that at the end of the all the fields.

This commit fixes the problem.

We don't have any proto currently where MAX_FIELD_ID >
PROTOZERO_DECODER_INITIAL_STACK_CAPACITY and repeated fields are used,
so that's not a problem for now, but in future commit I might introduce
some.

Change-Id: I99eb646d5107f758e0360d44d06f29bfafa16667