Autoupdate of 2 submodules
*build/blueprint updated to 09fe90e407fcc2ac93b1641792cff5c95d2f8318
Rename Evaluate() to Get() and add GetDefault()

Part of the design of property structs is that they were easy to access.
In keeping with that spirit, use a shorter and easier to spell name
for the getter, and add GetDefault() so that you don't need to pass
the result of Get() to one of the
proptools.StringDefault/BoolDefault/etc functions.

Bug: 323382414
Test: m nothing --no-skip-soong-tests
Change-Id: Ib9a69dcf2ab56a758935a461f37fe46bc0e17e27

*build/soong updated to 42a8b256afa320f29ad2aa0ad92e3ad27cb385ba
Merge changes from topics "replace_instead_of_append", "selects_get_and_get_default" into main

* changes:
  Rename Evaluate() to Get() and add GetDefault()
  Add tests for android:replace_instead_of_append