Fix import order disagreement (kotlinx)
For checkstyle, 'kotlinx' is not specified, meaning it is expected
at the end, in alphabetical order with all other imports which lack
a defined ordering.

This disagrees with google-java-formatter and all IDE code formatter
configurations in use. This leads to confusion, manual correction or
in many cases ignored checkstyle warnings. This is in part due to
unhelpful checkstyle output:

Given:

    import kotlin.Pair;

    import kotlinx.coroutines.CoroutineDispatcher;

    import java.util.ArrayList;
    import java.util.Arrays;

Checkstyle reports:

    .../Foo.java:16:1: Wrong order for java.util.ArrayList import. Use
    Ctrl+Shift+O (Eclipse) or Ctrl+Alt+O (Intellij) to sort imports.
    https://source.android.com/setup/code-style#order-import-statements

Bug: 333533940
Test: manual
Change-Id: Ife87e1f92bdc620031cf2e90add6b688aef8b5ef