[scudo] Add EnableContiguousRegions mode (#85149)
This releases the requirement that we need to preserve the memory for
all regions at the beginning. It needs a huge amount of contiguous pages
and which may be a challenge in certain cases. Therefore, adding a new
flag, EnableContiguousRegions, to indicate whether we want to allocate
all the regions next to each other.

Note that once the EnableContiguousRegions is disabled,
EnableRandomOffset becomes irrelevant because the base of each region is
already random.

GitOrigin-RevId: bab0507ff2679d2bbfa34921eeed4ff1cadbe7e2
Change-Id: I4cc4edebf4601a1f90584b2d28596fc07c7019aa