FlaggedApiTest: Support multiple previously released APIs
Previously, the `FlaggedApiTest` took a single `previouslyReleasedApi`
parameter that was used as the API surface for each surface that was
being tested, i.e. if the test had expectations for public, system and
module-lib then the `previouslyReleasedApi` would be use as the
previously released public API, the previously released system API and
the previously released module-lib API respectively.

That made it quite confusing as some tests relies on that behavior
while others did not. It also meant that it was not possible to test
behavior that affected multiple surfaces.

This change replaces the `previouslyReleasedApi` parameter into a map
from `Surface` to the previously released API string.

The two tests that required the `previouslyReleasedApi` to be used for
each surface under test made that explicit by using `config.surface` as
the key instead of a fixed key like `Surface.PUBLIC`.

Equivalent changes were made to the  `previouslyReleasedRemovedApi`
parameter.

Bug: 333394978
Test: ./gradlew
Change-Id: If58ad29ae0ffc7956890b37633fea8670f3ece7c