FlaggedApiTest: Pass the whole previously released API
Previously, the `FlaggedApiTest` just passed a single previously
released API that was specific to the `Surface` being tested to the
`check(...)` method.

As part of the work to support specifying multiple files this change
constructs a list of all the previously released APIs that affected a
particular surface. e.g. when the surface being tested is public it
will just contain the previously released public API but when the
surface being tested is system it will include public and system.

This revealed an issue with the handling of multiple `Codebase`s in
the `--revert-annotation` processing so a temporary change was made to
just use the last `Codebase` for now.

Bug: 333394978
Test: ./gradlew
Change-Id: I7a7cc7a7578ac8d5a8845b7dc7dde607d859b3a1