Use same ApiVisitor.Config when traversing added items
When `ComparatorVisitor` detects that an `Item` with contents, e.g.
`ClassItem` has been added then it will optionally visit the contents
when `visitAddedItemsRecursively` is `true`. Previously, it used the
default `options.apiVisitorConfig` but this change switches it to use
its `apiVisitorConfig` property. This is effectively a no-op as the
only `ComparatorVisitor` subclass that set `visitAddedItemsRecursively`
to `true` is `NullnessMigration` and it sets `apiVisitorConfig` to
`options.apiVisitorConfig` anyway.

Bug: 333394978
Test: ./gradlew
Change-Id: I0b30dde70ebda206baad0fc7ab5dca352d0f1c71