Stop ApiVisitor depending on options.apiVisitorConfig
Removes the default value of `config = options.apiVisitorConfig` from
the `ApiVisitor` constructor parameters and passes it in from the
callers.

All one but of the callers just pass `options.apiVisitorConfig` but
one does not. `ConvertJarsToSignatureFiles` passes in an empty config
as it does not initialize `options` so it is empty anyway.

Bug: 333394978
Test: ./gradlew
Change-Id: Ib7e5c81f92433e3b6912b2d38ee5b966fa0ae818