Trace Redaction - Create share-able packet filter
Right now, there is "filter by allow-list" which makes it easy
to remove whole packets based on an allow-list. However, there
are some conditional filters coming down the line.

This creates a transform that allows multiple filters to be
registered. By changing some primitives into filters, it cuts
down on the number of serialization operations.

Bug: 318576499
Change-Id: Ib6107dd9a8e0b2d7d606449dd9bbfbb0e1e9b8a3