Remove unnnecessary preFiltered parameter
The `SourceParser.loadFromJar(...)`'s `preFiltered` parameter is always
`false`. That is used to set the `DefaultCodebase.preFiltered` property
but only the Turbine and Psi models implement `SourceParser` and their
subclasses of `DefaultCodebase` initialize `preFiltered` to `false`
anyway so the parameter is unnecessary.

This change removes the parameter and switches
`DefaultCodebase.preFiltered` from `var` to `val`.

Bug: 333394978
Test: ./gradlew
Change-Id: I409e378fc3da655810c6f18d77f691040223bf16