Correct some unrealistic compatibility tests
A number of compatibility tests that use `signatureSource` are missing
a `public` modifier on the `class` or `interface`. They work because
the `CheckCompatibility` will visit everything when `preFiltered=true`
but it results in unrealistic tests. This change makes the tests more
realistic by adding the missing `public` modifier.

Most of the tests were unaffected by a couple of the tests reported an
additional issue.

Bug: 333394978
Test: ./gradlew
Change-Id: I49701c3a1f6456ff3426b201f9f458465b761743