Remove Psi overrides of DefaultItem.emit
Previously, a number of `Psi*Item` classes overrode `DefaultItem.emit`
in order to set it to `!modifiers.isExpect()`. This change removes
the overrides and simply initializes `emit` in an `init {...}` block
instead. That allows `DefaultItem.emit` to be marked as `final`. The
`init {...}` block is added in `PsiClassItem` and `PsiMemberItem` to
remove some duplicates.

Bug: 333394978
Test: ./gradlew
Change-Id: If04c202d46903075295265c88588051026787029