Trace Redaction - Remove comm values from all process free events
Process free events contain a pid and a comm. The timeline uses an
exclusive end bound, so it is not possible to use the event's
timestamp to see of the pid belongs to the package. Instead of using
a fudge value (e.g. time - 1), all process free events will lose their
comm values.

Bug: 318576499

Change-Id: Iceb2adb7f1618a46fe47c8dc5f44b6f68f73701d