Trace Redaction - Use Begin/End with CollectTimelineEvents
CollectTimlineEvents was one of the reason Begin/End was added to the
collect primitive. CollectTimelineEvents needs to be able to
create the timeline instance (smart pointer) once, events multiple
timelines, and then finalize it by sorting the events. The first
behaviours were possible using the old model, but the last option was
hard to support.

Having the timeline be search-able in build phase is required to
filter the frame timeline. However, previously the timeline was
converted from write-only to read-only, another build primitive
(optimize timeline). This would create a dependency between primitives
in the same phase - we want to avoid that.

Bug: 333404689

Change-Id: I3504e309d234e280345fce9771b0918b452e8071