Enable art cover for A2DP source profile
Problem: AAOS needs to support art cover when acting as AVRCP target

Solution: Bluetooth stack supports art cover when acting as a AVRCP target.
It can send the peer sink device the art cover for the corresponding track.
The stack needs to set its avrcp version as 1.6 in order to support art cover.

Bug: 334996714
Test: m seahawk-trunk-userfood
Test: Connect sink only enabled AAOS seahawk and validate art cover is
      supported
Change-Id: I851bbdc08ee7e9b498fc41b4dec953a8d770e011