Autoupdate of 2 submodules
*bionic updated to e7f4506bcb0f2f33d81601a5e0ae3b900fd62f04
Merge "Add LINE_MAX." into main
*system/memory/lmkd updated to c2d8bfdf59ac20795e4d43fdad304911150e25b3
Don't define LINE_MAX.

It's a POSIX mistake we'd been avoiding, but finally added to
bionic.

It's unclear to me that this code shouldn't just be using fscanf()
and getline() to simplify the code and avoid arbitrary hard-coded
limits -- since the code does do some allocation already -- but
that's a bigger cleanup for a code owner.

Change-Id: Ibab28f082a084c1f933ee6a39aa4ee3fa7868e5a