Disallow mixing jar and signature files for compatibility check
Previously, Metalava would allow a mixture of jar and signature files
for the compatibility check but that had a number of issues:
1. It makes little sense to mix them as they have very different uses.
   e.g. Signature files can be a delta, Jar files are a complete API.
2. Jar files are unsafe, as they contain additional non-API code, like
   private constructors in order to compile, and are not reviewed like
   API signature files are.

This change adds a `PreviouslyReleasedApi` abstraction, and creates
separate `JarBasedApi` and `SignatureBasedApi` and modifies
`CheckRequest` to use that instead of a `List<File>`. Creation of
`CheckRequest` ensures that the input files are exclusively jar files
or signature files, creating the appropriate `PreviouslyReleasedApi`
implementation.

Bug: 333394978
Test: ./gradlew
Change-Id: Icbf16e157b3e3ffd303add663a36686563f336d7