Clarify that previouslyReleasedCodebases(...) does not support jars
Previously, if the `CompatibilityCheckOptions` method
`previouslyReleasedCodebases(...)` was called with a jar file then it
would attempt to parse that jar file as a signature file resulting in
a confusing parse error. That method is used for `--revert-annotation`.
This change makes it explicit that it does not work with jar files.

Bug: 333394978
Test: ./gradlew
Change-Id: I44beabcf1626db68d7338cb14bcd0da7d4953a33