Perform compatibility checks against the whole previously released API
Previously, compatibility checks were only performed on the `Codebase`
containing the previously released API for the API surface currently
being generated, even if that `Codebase` was a partial, incomplete
delta on another API surface. e.g. if generating the system API it
would only run compatibility checks against the previously released
system API delta of the public API and not the public API itself.

This change switches it to operate on the whole `Codebase`. Note, that
does not change the behavior of only reporting issues on the part of
that `Codebase` corresponding to the API surface being generated.

While this has no effect on the compatibility issues that are reported
it will ensure consistent behavior between compatibility checks and API
lint.

Bug: 333394978
Test: ./gradlew
Change-Id: Id4049b082efc52ba760177704c12b5bb62e8ee59