Remove current platform ABI dumps
The "current" directory was added before trunk stable. The purpose was
to automerge the dumps from the dev branch to the master branch while
the dev branch was being finalized. It is not necessary for the main
branch. The SDK finalization process will generate ABI dumps in the
directories for each SDK version. The release configurations in
different versions can cover cross-version ABI checks.

Test: make
Bug: 333532038
Change-Id: I90136ec8e80b877b3a55320aecfd9c799200c8ea