Fix overwriting of compatibility check files during testing
Previously, if multiple signature file contents were passed to
`checkCompatibilityApiReleasedList` then they would all be written to
the same file (so the last one written would overwrite all the others)
and then that same file would be passed in the command line arguments
once for each input. That was hiding an issue with the check
compatibility behavior.

This change fixes that by checking to see whether the file already
exists before writing it and if it does then adding a numerical suffix
before the extension (or if none at the end of the file name). It also
fixes the tests that are broken.

The new behavior of one of those tests now shows the incorrect behavior
which will be fixed in a follow-up change.

Bug: 333394978
Test: ./gradlew
Change-Id: Id7b3c246c3b626bbba492dc2c618638167a8aaac