Only report compatibility issues against API being generated
Prior to the switch to support passing the complete previously released
API (which may consist of multiple files) only a single file of the
previously released API was supplied and that corresponded directly to
the API surface being generated. e.g. If the public API was being
generated then the previously released public API was provided. If the
system API was being generated then the previously released system API
delta on top of the public API was provided.

That had the effect of only reporting compatibility issues that were
found in the system API part. It would not report compatibility issues
in the public API part. That is because compatibility issues are only
reported against `Item`s whose `emit` property is `true`. While the
`Codebase` for the system API would include classes from the public API
they were created as stubs and so had `emit` set to `false`.

Comparing against all the `Codebase`s from the previously released API
would report issues with the public API even when generating the system
API because the `Item`s from the public API would have `emit` set to
`true`.

This change just checks compatibility against the last `Codebase`, as
that corresponds to the API being generated. e.g. When generating the
public API there is only one `Codebase` in the list. When generating
the system API there will be two, the first will be public and the
last will be the system API delta.

Bug: 333394978
Test: ./gradlew
Change-Id: I908df20b84a19a59b92c10b6dcacc918f3c271f9