Migrate JDiffXmlTest to use SignatureToJDiffCommandTest
Previously, it uses the `DriverTest.check(...)` method relying on the
`apiXml` parameter to specify the expected output XML. However, that
relies on `--api-xml` option which will be removed in a follow-up
change. Also, the formatting, specificalling the indenting, of the
input API and expected XML was very inconsistent.

This change migrates the tests to use the testing support provided by
`SignatureToJDiffCommandTest` and cleans up the formatting and
indentation to match that.

Bug: 335616501
Test: ./gradlew
Change-Id: I22ef3fa8b99397e40b43dc1e4482af4c6cf961ce