Include reverted item in stubs for extending API surface
Previously, when generating an API surface that extends another, if an
`Item` in the extended API surface was reverted back to a previous
version (e.g. because it had changed) then it would not appear at all
in the extending API surface's stubs. That is because when generating
the extending API it did not have access to the previously released API
for the extended API surface. e.g. When generating stubs for system it
would not know about the previously released public API.

This change fixes that by making `previouslyReleasedCodebases(...)`
return all the previously released `Codebase`s not just the one for the
API currently being generated.

On its own that would cause items from an extended API (e.g. public) to
be considered part of the extending API (e.g. system). That would
result in the system API including items from public API in its
signature files. So, this change uses the `Item.emit` property to
differentiate between the extended and extending APIs to tweak the
behavior to avoid that as follows:
1. Reverting a change to an item will only show the item if the item is
   part of the API currently being generated, i.e. when `Item.emit` is
   `true`.
2. `forStubsOnly` will be set to `NO_EFFECT` (as previously) if the
   item is part of the API currently being generated, i.e. when
   `Item.emit` is `true`, and `SHOW` if the item is part of an extended
   API, i.e. when `Item.emit` is `false`.

Bug: 319874764
Test: ./gradlew
Change-Id: I4d1ea7fc4916336a58e64090700b54ec8296845e