Support reverting changes to modifiers
Previously, reverting a change to an `Item'`s modifiers did not work,
it used the changed modifiers. This change fixes that for stubs and
signature files by using the modifiers of the `Item` from the
previously released API.

This is a temporary solution as it is quite fragile and limited as it
requires every place where the modifiers are used to be modified to use
`Item.actualItem`. However, it will suffice until a better, more
comprehensive solution can be created.

Bug: 319874764
Test: ./gradlew
Change-Id: Ie7ad5cd5b6849f62536bde2098135f10ac072a04