Add --api-lint-previous-api
The existing `--api-lint` option has an optional argument. That is
difficult to process and makes the command line generation fragile.
It has caused issues in the past where the order of command line
arguments has been changed, resulting in a non-option being added after
an `--api-lint` without the optional argument resulting in that
non-option being unintentionally consumed by the `--api-lint` option.
That caused weird errors seemingly unrelated to the change, wasting a
lot of time to debug and fix. See https://r.android.com/2977083 for
more information.

This change adds a new `--api-lint-previous-api` option that will
replace the optional argument.

Bug: 327019023
Test: ./gradlew
Change-Id: I2ecdc67898d8e8e2b0111322d0a1cf73916378af