ui: remove arg autocompletition in pivot table
This feature materialized the distinct key of all arguments in the args
table in the UI. This is both very slow and expensive.

The fuzzy matching should have been implemented in SQL instead of in
Typescript but that's water under the bridge.

For now, delete the feature as it's hurting performance on large traces
and this is a very niche feature. It can be reimplemented at a future
date if this turns out to be useful.

Change-Id: I918196ae9ced7f0625a8f442a2ad76757e5db625