Explicitly specify whether a device is needed in each integration test
Previously integration tests didn't need to specify whether a device is needed and the device serial is always appended to the atest commands when available. This works in the past even for deviceless tests running with '--host' argument. Now due to some change either in atest or tradefed, running atest with --host and --serial results in tradefed timeout waiting for device. While this could potentially be a regrssion of atest/tradefed or being intended, the integration test framework should not do this by default unless in a test case that's written to test this feature. This CL makes all tests calling atest command specify whether device serial should be appended.

Note that we still support running atest commands without serial argument even when serial argument is set to being required. This is to support easier local test run via the 'python XX.py' command (vs 'atest XX').

Test: run 'atest atest_integration_tests' without ANDROID_BUILD_TOP env
Fix: 333483685
Change-Id: Ifea46942125360e72a15724400046b518d442218