Autoupdate of 2 submodules
*external/rust/crates/oneshot-uniffi updated to 1559a0d87a4f3ea7029fe92513ce60fce292a444
Generate Android.bp for oneshot-uniffi

The crate is a bit special: it’s identical to the oneshot crate, but
with a dependency removed:
https://github.com/faern/oneshot/compare/main...bendk:oneshot:main

As such, the crate is called oneshot-uniffi, but the rest of the
UniFFI crates import it as just “oneshot”. The same applies to the
tests in the crate itself.

Patching the `crate_name` seems like the easiest way to accommodate
this. The library is visibility restricted so this will only be
visible to `uniffi_core`.

Bug: 330717829
Test: atest --host
Change-Id: I8dc3b9c39abcce5ca48bf0f7b8adeb7b149090b3

*external/rust/crates/uniffi_core updated to 69c166aea2541b6fdf8e185eff3dc78e0be6e0ce
Generate Android.bp for uniffi_core

Bug: 330712502
Test: atest --host
Change-Id: I0219616dca70d54de1bf28d0405d70c43d03ba77