Set cmake_minimum_required version to 3.22.1
On newer NDKs that make use of features like IN_LIST, older
cmake_minimum_required versions used in some of our CMakeLists.txt files
can cause build errors.

As the cmake version used for building is tied to the version
distributed by the NDK toolchain, bump the version across the repository
to the version the NDK supports. This will allow us to freely use the
features in the cmake binary available in the NDK.

Test: ./gradlew assembleRelease
Change-Id: I22a9182ccdaf41cb3a657fd47f35d343c41affbb