Improve windows support
This makes sure we include windows specific linker options for
executables and filters out unused includes.

We also prefix static libraries with the lib prefix. This is mainly to
make sure we do not get in a conflict situation with shared_libraries
where naming really matters.

The generated build files will mark the "includes" keys to be ignored by
buildifier, this will make sure the windows build will not be broken.

Change-Id: I58a78e7ca33ed52fc74a12c0fc555d7904bb0a38