ui: utilise segment forest to improve slice tracks on large traces
This CL introduces the use of the segment forest to significantly speed
up the queries of slice tracks on large traces. Specifically the data
structure is exposed to Typescript via the use of a new "SliceMipmap"
trace processor operator.

This CL only ports base_slice_track to the operator: other tracks will
be ported in followup CLs.

Change-Id: Icf5d5a620b01dece7b102f9bdae181db0ee145fc