Remove a debugging log
Previously the log uploader is disabled by default so we add this debug log so that when we intentionally enable it and it doesn't work, we know why. Now that it's enabled by default, and it's expected that some branches like aosp doesn't have the credentials, this logging no loner adds any value.

Test: unittest
Change-Id: Ie54e7c4d6d0121655a04fcf2b3644d81302c5b48