Add tests for when API is included in later SDK than dessert release
If an API is listed as being in an SDK extension of a dessert release
`X` where `X` is greater the dessert release in which the API was
included then Android lint seems to ignore that it was included in
a dessert release and complains even when it is used in an App that
sets `min_sdk_version` to the dessert release in which the API was
released.

e.g. The `ConnectivityManager.registerQosCallback(...)` method is
currently defined in the `api-versions.xml` file as:
```
<method name="registerQosCallback(Landroid/net/QosSocketInfo;Ljava/util/concurrent/Executor;Landroid/net/QosCallback;)V" since="31"/>
```

Android Lint handles that correctly.

However, change https://r.android.com/3061686 (which fixes an error
where system API sdk extension jars were not being passed to Metalava)
causes the definition to change to:
```
<method name="registerQosCallback(Landroid/net/QosSocketInfo;Ljava/util/concurrent/Executor;Landroid/net/QosCallback;)V" since="31" sdks="34:1,0:31"/>
```

That causes Android lint to report the following problem:
```
packages/services/Telephony/testapps/TestRcsApp/aosp_test_rcsclient/src/com/android/libraries/rcs/simpleclient/protocol/msrp/MsrpSession.java:118: Error: Call requires version 1 of the U Extensions SDK (current min is 0): android.net.ConnectivityManager#registerQosCallback [NewApi]
            connectivityManager.registerQosCallback(new QosSocketInfo(network, socket),
```

This change adds tests to highlight that issue and a follow-up change
will fix them. That will prevent the `sdks` attribute being added in
this case and so it will have the same behavior as without change
https://r.android.com/3061686.

Bug: 336993217
Test: ./gradlew
Change-Id: Ic93d9224eb79011047e38654890256704b1fe5af