riscv64: Change default calling conv to LP64D
Guest ABI allows selecting either the LP64 or LP64D calling convention.
This choice affects floating-point parameter passing. The default was
set to the simpler LP64. However, the bulk of the Berberis code base is
not written to allow for using a non-default ABI. For example, the JNI
trampolines are generated using GuestReturnReference with only the
default calling convention without a parameter to select a different
one. This causes floating-point values to be improperly marshaled on
LP64D platforms.

This change sets the default calling convention to LP64D, which is the
ABI we intend to initially support. This fixes failures related to
floating-point marshaling on all current device targets.

Bug: 330396865
Test: CtsJniTestCases and berberis_run_host_tests
Change-Id: I337e5c68516948ac3e18907cf873ad80d37c154f