Run ApiLint checks over the whole API not just new APIs
Previously, when a previously released API was provided to `ApiLint` it
would use `CodebaseComparator` to only run lint on the newly added
`Item`s. That was a problem because it prevented checks that are needed
from being run, e.g. it would not call `checkClass()` on a previously
released `ClassItem` but some of the checks performed in `checkClass()`
actually check class members so should have been run on new members but
are not.

This is of particular concern for the `@FlaggedApi` checks which need
to run on changed `Item`s as well.

This change removes the use of `CodebaseComparator` and runs ApiLint
over the whole API. So, even if a class was previously released it will
still run the checks on that class and all its members. On its own that
would mean Metalava would reports lots of new (in the sense they have
not been suppressed) errors that would break downstream builds.

To mitigate that `ApiLint` will downgrade the severity of all issues
reported on a previously released `Item` to `HIDDEN` to replicate the
previous behavior of the `CodebaseComparator` based ApiLint.

Future changes will stop that downgrading for specific issues to
allow them to be reported on changes to the API, e.g. The
`UnflaggedApi` issue should be reported on changes to existing APIs as
well as new APIs.

Bug: 311690008
Test: ./gradlew
Change-Id: I1afc456bea7aeb019e2964394f9b33f3198e3cef