Autoupdate of 2 submodules
*external/qemu updated to 1e5cef31c638c66f419128d2365ebf08fb575771
Merge "Remove need for a full build" into emu-dev
*external/meson updated to 07df48f2d4c55e7a070b04ae8fdaaa766c13bdb2
Remove requirement for a full build

We now only invoke ninja in the shadow build directory for custom
targets that are needed for the compilation to succeed.

This makes sure we do not have to do a full build before generating the
bazel files.

Change-Id: I095e91d03c9d860ca13510afd44f24ca09719831