Disallow jars in --source-path
It makes no sense for `--source-path` to accept jars. It seems as
though it was a copy and paste error from the code for handling
`--classpath`. Also, it seems as though the
`SourceSet.gatherSources(...)` which processes the source path
assumes that any files it finds are source files, i.e. not `.jar`
files.

Bug: 286023667
Test: ./gradlew
Change-Id: I5d6d353175c51af846d09fc5161447c779753375