Make SymlinkTest more realistic
The `--source-path` option is only specified once in the different
build environments within which Metalava is used but previously this
test specified it twice, once by the `check` method and once in the
`extraArguments`. The latter added an empty string to try and avoid
scanning the current working directory, however as `--source-path`
is already specified itis unnecessary. So, this change removes that
extra option.

Bug: 286023667
Test: ./gradlew
Change-Id: Idc8ee751f74da56f20d0dd90de35f0cc26647a71