Removes the windows specific timout marker
The custom windows only timeout marker is causing test failures,
and is not necessary. Instead we use the same timeout across for
all platforms.

Also removes all the timeout markers under 60 seconds, since
the default timeout per test is 60 seconds and will be enforced
if the timemout is not specified.

Test: passes sanity test suite
Change-Id: I7624a005ac0cf0e371c746538f80f25aabfa9706