Make linux cc toolchain compatible with rust rules
For some reason the rust rule does not honor the custom library linking
flags (e.g. -l:libc.so). Therefore we switch to a less customized way
that specifies sysroot and b_prefix instead. This works well as long as
the default libs linked by clang is provided by the system at runtime.

Test: building the rust process wrapper.
Bug: 335737457
Change-Id: If09a09c235ee46831779525eb5a6a08e9947e04f