gn: Use canonical target name for perfetto_proto_library metadata export
The perfetto_proto_library generates some metadata (used by the
gen_bazel and gen_android_bp scripts) that references some target.

In one case, it's generating something like
```
'//protos/perfetto/trace/android/:winscope_common:source_set'
```
which is not the canonical way to represent a target in GN (note the
'/:').

The problem is that we use `get_path_info`, which is for handling paths
and not GN target.

Switching to `get_label_info`, plus some string concatenation, produces
the correct target name.

This allows us to remove a workaround in gn_utils.py, that was
introduced just for this.

This is just an internal cleanup and it has no effect on the generated
build files.

Commit fe05d3a438ec("gen_bazel: handle target names containing '/:'")
bypassed the problem by special casing '/:', but that I think '/:' it's
not a valid gn target name.

Change-Id: I319e8bb7df50cc32ad3da5cd83651fe5a99c025e