ui: add fallback for user tracks without package name
aosp/2912626 attempted to add a fallback for UID tracks with
no package name, but it did not work because those tracks were
excluded by the inner join in the SQL query.

Having a fallback is useful because not all workloads have
a package name, for example benchmarks that work by uploading
and running a native binary on the device.

Bug: 325049734
Change-Id: I71fffa13410685df28b38bd99665984416f80798