AdapterService: delay native library loading
We want to delay loading the native to when we actually enable the
bluetooth. This remove operation from the onBind phase and decrease the
risk of binding failure.

Note that a failure during binding can never be catch by the manager
service as there is no binder returned yet. Reducing this "sensitive"
windows is always helping to increase the stability of Bluetooth

Bug: 339554081
Bug: 315241296
Test: atest pts-bot,avatar (presubmit) | Lifecycle is tested there
Test: manual, boot the bluetooth and check timestamp of log
      With this patch we can notice a 25% faster init on Pixel 7
      Metrics TBD for watchs devices
Flag: com.android.bluetooth.flags.avoid_static_loading_of_native
Change-Id: I3636bae16a8854fad50bfbb920b1dd12fbf8fb11