Add new warning severity that is treated as error in new code
This adds `Severity.WARNING_ERROR_WHEN_NEW` which is treated as a
warning but includes `(ErrorWhenNew)` at the end of the message, before
the `[issue-name]` that indicates if the issue appears in new code that
it should be treated as an error. The purpose of this is to make it
easier to transition an issue from a warning to an error by preventing
additional cases from being introduced into the code while the existing
issues are fixed.

This relies on the infrastructure for handling changes, i.e. the
"Build Warnings" Aye Aye Analyzer, to build the change and then check
the build output for the presence of the `(ErrorWhenNew)` in an issue
reported for one of the changed lines. If found then it will treat that
issue as an error, highlight it to the developers and prevent
submission without addressing it, or overriding it.

Bug: 339602728
Test: ./gradlew
Change-Id: Idb66398215c79e2d1e221206525ae04e3ca9cfc6