Autoupdate of 3 submodules
*frameworks/av updated to ec4a0953c70bf47282ada2f8580888d5237149d8
Merge "libaudiohal@aidl: Abort on data FMQ pointer corruption" into main
*hardware/interfaces updated to 2222e7d599802d2a87f3d150b3d07c3bd146864e
Merge "audio: Abort on data FMQ pointer corruption" into main
*system/libfmq updated to c6a3e1753c19f49f8c1ea5896a456ee74601edec
Improve FMQ pointer corruption diagnostics

Since sometimes an abort stack is the only information
available, it is useful to have suspicious pointer values
in the abort message. For that, it must be generated
by the FMQ implementation itself as the client does not
have access to the pointers.

Because of that, change the parameter introduced by
aosp/3073435 to an optional enum and an error string.

Bug: 338974476
Test: m
Change-Id: Ib9d61c787a9e1cd1197503aed698bc1e7e61bc87