Allow AvdWriter to save an AVD with a predetermined name.
Currently the AVD name created for a given api, tag, abi and device is
hardcoded as "{api}_{tag}_{abi}_{device_name}". This change will allow
using a non-default AVD name by setting the key "AvdId" of the
configuration dictionary.

This change is required for P0 test case scenario for verifying if
multiple user actions can be performed when two or more AVDs are
running.

BUG: b/337088629
Change-Id: I650b54d663a960a9ac1721be00ae22cd404d2122