ANDROID: Boot UEFI partitions if possible in the cuttlefish bootcmd
This sets the default boot command to first try to run an EFI bootloader on
the first FAT partition that appears on any disk. If that fails, it
continues to run the pre-defined boot command.

Explanation:

This enables the CONFIG_CMD_BOOTMENU setting in the cuttlefish
bootloader configuration, which enables the `bootmenu` command.
`bootmenu` is one of two callers of the
`efi_bootmgr_update_media_device_boot_option` function in u-boot.

https://android.googlesource.com/platform/external/u-boot/+/901bc28/lib/efi_loader/efi_bootmgr.c#641

This function populates the UEFI `BootOrder` variables based on the
`Boot Option Variables Default Boot Behavior` section of the UEFI
specification:

https://uefi.org/specs/UEFI/2.10/03_Boot_Manager.html#boot-option-variables-default-boot-behavior

The "firmware specific behavior" here appears to be to create entries
for every FAT partition ordered by disk number first then partition
number, trying to load a `efi\boot\boot{arch}.efi` file, following

https://uefi.org/specs/UEFI/2.10/03_Boot_Manager.html#removable-media-boot-behavior

`bootmenu` then waits for up to ${bootmenu_delay} seconds and tries to
boot the first option. If that option turns out to be invalid (e.g. no
file by the expected name exists), it will return a success exit code.

https://android.googlesource.com/platform/external/u-boot/+/901bc28/cmd/bootmenu.c#640

This should work as a default configuration for both cuttlefish and AVF
as Cuttlefish does not have EFI bootloaders floating around
unintentionally.

Bug: 325930313
Test: Run on x86_64 with cuttlefish
Test: Run with VmLauncherApp on a physical device
Change-Id: Ie039565417d91d1d73863e8a7efed5860996d451