Specify start_threshold to match period size and count
A start_threshold of 0 lets tinyalsa set it, which defaults to
(period_count * period_size / 2). On some hypervisors, the
virtio-snd virtual device expects a certain amount of periods
to be sent before the device is started. This change lengthens the
start_threshold to this value, which is equal to (period_count *
period_size).

Bug: 331600622
Test: Build, confirm sound plays on the device
Change-Id: I30d328b5ff2037ecc55c04d6dd0c182498065fbf