aconfigd: move current functions under aconfigd.cpp into the class Aconfigd
Currently the aconfigd code is not test friendly. To test, it will
actually write some mockup storage files on /metadata/aconfig dir, and
have to clean up after each test point. Instead, we should strcut the
code in a class so that we can provide a temp dir as a mock of
/metadata/aconfig. This change will allow socket request handler to be
tested without modifying /metadata/aconfig dir.

Bug: b/312444587
Test: atest aconfigd_socket_test
Change-Id: Ifa46183c465493c4559c257c74773fa7f01071f5