Extend ContentProviderHandler to handle non-current user.
ContentProviderHandler had an assumption that all tests were running as
the current user, which failed media tests if they run as visible
background user with `--user-type secondary_user_on_secondary_display`.
Remove the assumption and extend the class such that it can take a
userId as a parameter when constructed.
The new constructor will later be utilized by NativeDevice and
MediaPreparer to run the tests as visible background user.

Bug: 336880935

Test: atest android.media.player.cts.MediaPlayerSurfaceTest
Test: m -j tradefed-all && tools/tradefederation/core/javatests/run_tradefed_tests.sh \
      --class com.android.tradefed.device.contentprovider.ContentProviderHandlerTest

Change-Id: I12c5fa03800bec46edc4ddd2b86bea1eba552684