Autoupdate of 7 submodules
*external/boringssl updated to a4384672c6efd89c58ecd2524028f52967b5adc6
Update config to build with updated boringssl

This CL fixes some configuration errors that arose with the updated
boringssl. This includes some linkage configuration errors, and some
temporary fixes required while `static inline` is still unsupported
in `rust-bindgen`.
We have
added the `unsupported_inline_wrapped` `cfg` to the `bssl-sys`
Android.bp file, to include the functions gated by this flag.

This CL also modifies the linker configuration to properly point to the
generated bindings file.

This is the AOSP changes required; there are some upstream boringssl
changes that add wrapper functions for some functions that are `OPENSSL_INLINE`.
These boringssl changes need to be pulled in for this to work as intended.

Test: m checkbuild
Change-Id: I0dbcfbbc03db2e1c3257a29ed71bfe3bc53179df

*external/rust/beto-rust updated to 5d6bb694d12d7daa2d8c7f54a20ed75b76da706a
Project import generated by Copybara.

Manual changes:
* Modified Android.bp files to move liblock_adapter to common/

GitOrigin-RevId: c548f88b53228f43a4992c6d2ab05dcc602534bf
Change-Id: Ib96e462be4754493641f9f06045a0acb397030ee

*external/rust/crates/quiche updated to 73a569f9ab0be217fcf956364154dc3b36a23c99
Adding libc++ explicitly to the shared libraries of rust-only tests

There is an issue with the boringssl build configuration right now, where
libc++ is not made available to a rust-only client. This CL adds an explicit
link to libc++.

Test: m checkbuild
Bug: 340583190
Change-Id: Ibe51ef8cbd0836ddbf6021d2c6955ab3b89a3b0f

*frameworks/base updated to 650ef384fe829125afe97c6fc9baf0e5e894f0e9
Merge "Throw new exceptions in SecureChannel from beto-rust update" into main
*packages/modules/DnsResolver updated to 7c2967fb3b9c9e800f1f31590b1b737e499bdc44
Merge "Adding libc++ explicitly to the shared libraries of rust-only tests" into main
*system/keymint updated to eae66423133c605c23c1cbcf0cde129949e7b2d1
Merge "Wraps call to boringssl unsafe exports in `unsafe`" into main
*trusty/user/base updated to 4854517dbbb28c06b45752e2242f5cdebff1a101
Update to use new boringssl

This CL enables the cfg `unsupported_inline_wrappers`, which is needed
to allow access to some functions that are currently gated by this flag.
Note that this flag will be removed when inline is supported,

This change is necessary because trusty has its own build configuration
separate from the bssl-sys Android.bp file.

Test: trusty/vendor/google/aosp/scripts/build.py generic-arm64-test
Change-Id: Ie6c8056a8e85c2e95c9d7f7189dcdb9900a31718