Rework memory allocation with free policy
The memory allocation function was allocating a number of bytes in the
heap, traversing all of them page by page to make sure all the addresses
were mapped with physical pages, and the memory released with the
destructor. While the memory is allocated, there is no way of knowing if
the kernel swaps it out.
To remove this level of uncertainty, instead of overcoming the lazy page
allocation manually, ask the kernel to allocate a resident memory region
for us.
In addition to that, let the user decide at what stages of the
instruction the memory should be freed:
1. like before, until the destructor;
2. after every execution;
3. after the last "repeat".

Test: inspection with Perfetto
Bug: 339590280
Change-Id: If139971222af56ad3f5aaf2f96d09122a27f8259
Signed-off-by: Alessio Balsini <balsini@google.com>