Drain tun interfaces before every test
The tun interfaces were not drained after being set up, meaning that a
test that expects read() to succeed might do so due to the presence of
the initial MLD, NS, or RS packet after bringing up the interface.

Additionally, the tun interfaces in this test are static, so draining
them in SetUp() ensures there are no packets leftover from a previous
test.

Test: atest netd_integration_test
Change-Id: I7585b6dfd35e1964c28765f7624e4955b5f5cfb4