Fix WaitForDeviceDatetimePreparer
Since this preparer is used to correct the device datetime and the
expectation is that the device and host time will be within a 10s
margin, it makes sense to force the datetime into the device instead of
waiting for a service to kick in and correct the device datetime
automatically. The optional force-setup is deprecated, since tests
should not use the preparer, wait until the timeout and be okay with an
incorrect datetime. force-setup should always be enforced if the
preparer is used.

Bug: 327462702
Test: Presubmit
Change-Id: Ia1ff90710d91371a79a80238d1dcc3afe8d11b88