[libgbl] hook up libavb and remove sw_digest
Hook up enough of libavb to get the existing verification tests passing.
Just uses the base libavb for now, will add more tests and use the
libavb_cert extensions later on. The test data generation script has
also been updated to be more deterministic and provide the now-required
key data.

Also removes sw_digest; it turns out this is going to be pretty
difficult to integrate given libavb's somewhat awkward hashing API, and
the `ring` crate is not approved for production use in Android. We'll
revisit this later to try to figure out how to inject custom hashing
implementations, but for now it just uses the built-in libavb software
implementation.

Bug: b/334962570
Change-Id: Ie55ce0719c88855ce7032f2ba55018265385d99c