aconfigd: add aconfigd unit test using aconfigd and container mocks
Old aconfigd_test directly sending messages to actual aconfigd socket.
It would modify change the actual storage file dir (leaving temp file
and etc) and we have to clean it up. In the case of test error, these
temp files are not properly cleaned. Switch over to using mocks to avoid
this issue.

Also added a test mapping file to ensure that this is run properly.

Bug: b/312444587
Test: atest aconfigd_test
Change-Id: If3c34ac883d4e5002e26653fba8efb5fe92fcbb1