Disallow passing multiple jar files for previously released API
It is unnecessary to pass multiple jar files for a previously released
API, as each jar file must contain a complete API surface. e.g. the
system jar file must contain system and public. That is not the case
for signature files where the system signature files only defines a
delta on top of public and so the public signature files must be
provided in order to create a complete API surface.

No existing known uses pass more than one jar file so this should be
safe.

The main reason for doing this work is because it will be necessary to
change `PreviouslyReleasedApi.load(...)` to return a single `Codebase`
instead of `List<Codebase>`. That is needed because loading a single
`Codebase` from multiple signature files creates a complete and
coherent `Codebase` ensuring the code which introspects that will
work correctly. However, loading each file into its own `Codebase`
results in incomplete `Codebase`s.

e.g. If a class in the system API extends a class in the public API
and adds some members then loading them in to separate `Codebase`s
means that the system API extends a stub `ClassItem` and only includes
some members.

Bug: 341078916
Test: ./gradlew
Change-Id: Id497f7d0794038a4e984575a3bf0acbe186e1f58