Add PreviouslyReleasedApi/CheckRequest.lastSignatureFile
Avoids having to check the extension of a file to see if it is a
signature file when taking the fast path. It can just take advantage
of the checking already done.

This does mean that the fast path will be taken more often than before
as previously, it would only be taken if the extension was `.txt` but
now it will be take if the extension is NOT `.jar`. That could
marginally improve the Android build performance as it copies the
previously released file (which does have a `.txt` extension) into a
file which does not have it.

Bug: 341078916
Test: ./gradlew
Change-Id: I122994b35443041f00296556ff1795d1737be070