Autoupdate of 2 submodules
*build/bazel updated to 2d0c604b010a68e1e096760756fa7157b002980e
Add support for gRPC

This adds the gRPC package, which will allow us to create gRPC services.

Change-Id: I22f9f267b9b6f672fbc9e8a7966996d492007ce1

*external/grpc updated to 542e7d54b6dcafa14e30ac918ff87ca9ff27a9ad
Disable packages we do not need

gRPC contains a large set of features that we do not need, but require
a large set of dependencies if we would support them.

We do not want to bring in envoy, google cloud support or ios support.
The unused functionality has been commented out, so we can hopefully
merge down changes relatively easily.

Change-Id: I38c4e4e62e0f291a49cfc4e70ae35cd5947478d6