Add a pile of missing tests to TEST_MAPPING
It's 2024 and running tests on Android is still extremely difficult. It
seems we need to do this in two steps, first adding it to "postsubmit"
so they run at all and then "presubmit" once they're found to pass. I
already ran treehugger to see if they pass as presubmits, but I can't
figure out how to read Android's very confusing output, so I don't
actually know if they run.

Absent useful tools, let's just add these to postsubmits and see how it
goes. We'll just have to figure this out the slow way from the
dashboard.

Test: treehugger
Change-Id: Ifa5bb5624ab8474ec8a141c15a4608071bb8ca9c