Update to clang-r522817
The clang-r522817 prebuilt includes a libunwind.so.  For some reason
this is being linked into `libLLVM.so` despite the llvm-libunwind value
being set to 'in-tree'.  For now we will use the prebuilt libunwind and
resolve the bug later.

This change is required to make use of LLVM perfraw version 9, which is
used in the in-tree LLVM.

Bug: 333887339
Test: ./tools/build.py --lto thin --llvm-linkage shared
Change-Id: I578b1f604c5543487c7c3695c9e79e9b1734e259