Load all previously released APIs into a single Codebase
Previously, `PreviouslyReleasedApi.load(...)` would load each
`SignatureFile` into its own `Codebase`. That meant that a class which
was in the public API and had some extra methods in the system API
would have a separate `ClassItem` in each `Codebase`. That would cause
problems for code that introspected each `Codebase` as whichever
`ClassItem` it looked at would be incomplete. Similarly, a class that
was only in the system API but which extended a class in the public
API would not be able to access its super class in the public API
(which would be in the public `Codebase`) and instead would only
retrieve an empty stub `ClassItem` in the system API.

This loads all the files into a single `Codebase` using the text
model's ability to merge a class's contributions from each file into a
complete `ClassItem` in a coherent `Codebase`.

Bug: 341078916
Test: ./gradlew
Change-Id: Ib90c74882defddf95fa06a298bac29c8988a9632