Split deprecated/modifier FlaggedApiTest method into two
Previously, the test method tested the behavior of reverting deprecated
and modifier changes to the API. That test method was quite complicated
and while the functional changes are similar there is some special code
needed for handling changes to an API's deprecated status that made the
joint deprecated/modifiers test more complicated.

This change splits the test into two, one for deprecated and one for
modifiers. Part of the reason why they were originally combined was
that it used the deprecated status change to check the behavior of
fields. This change switches the field to use the `final` modifier in
the previously released API and then removes it to test the field
behavior instead.

Bug: 341672035
Test: ./gradlew
Change-Id: Id885984fa2bd403958bfc8545d8c9e0d11fc38f2