Show broken behavior when reverting change to native/abstract modifier
Reverting a change that removed the `abstract` modifier from a
previously released method causes invalid stubs to be generated as the
code that writes the modifiers uses the modifiers from the previously
released API, however the code that decides whether to write the body
uses the changed modifiers.

So, when the previously released API included `abstract` an `abstract`
modifier is written out, which means a body is not allowed, but it also
writes out the body, which results in a compilation error.

A similar, albeit complementary issue, arises with the `native`
modifier. Although the `native` modifier is not considered part of the
API surface (and so never appears in the previously released API) it is
still written out in the stubs (because that is currently relied upon
by the code coverage infrastructure). When reverting a change to that
method the `native` modifier is not written out in the stubs so it
should have a body but the code that determines whether to write out
the body thinks the method has a `native` modifier and so does not
write a body. That too results in a compilation error.

This change improves the modifier test to cover those cases and show
the invalid behavior.

Bug: 341672035
Test: ./gradlew
Change-Id: Idbb183403f98633ba08f6ce048f045e0f12202b4