Fix stubs for previously `abstract` methods and `native` methods
Previously, when reverting a change to a method, the code for writing
the modifiers would use the previous `Item`'s modifiers but the
`requiresMethodBodyInStubs(MethodItem)` method, which determines
whether the method body was written out would use the changed `Item`'s
modifiers resulting in uncompileable stubs being generated.

This change passes the previous `Item` to
`requiresMethodBodyInStubs(MethodItem)` to ensure the correct behavior
and updates the test to check for the correct behavior. As part of that
it was necessary to use generics for `Item.actualItem` so that when
called on a `MethodItem` it would return a `MethodItem` not an `Item`.

Bug: 341672035
Test: ./gradlew
Change-Id: I3976468a002b947e26dc969fe824609833f2123d