Autoupdate of 2 submodules
*test/suite_harness updated to 30edbdc2883b3c0cb5cd693f08134ade26566736
Update MediaPreparer to support non-current user.

Remove the assumption that MediaPreparer always runs as the current
user by setting the userId from the "RUN_TESTS_AS_USER" property.
This will enable media CTS to run with
`--user-type secondary_user_on_secondary_display`.

Bug: 336880935

Test: atest android.media.player.cts.MediaPlayerSurfaceTest
Test: atest --user-type secondary_user \
    android.media.player.cts.MediaPlayerSurfaceTest
Test: atest --no-bazel-mode \
    com.android.compatibility.common.tradefed.targetprep.MediaPreparerTest

Change-Id: I8c077c02fbac7fbc433582ca223ad31dc885e77e

*tools/tradefederation/core updated to 3198546039545e093e822dc6d077bc3404171f2f
Add a utility method to extract RUN_TESTS_AS_USER.

Add a new getTargetUserId() method to UserHelper, so it can be used to
extract the userId that the test runs as, set in the TestInformation.

Bug: 336880935

Test: m -j tradefed-all
Test: atest android.media.player.cts.MediaPlayerSurfaceTest

Change-Id: Icfd010342a069ca80232032567356e481f1abfe3