Override `uniffi_macros::util::mod_path` in AOSP
The original code depends on Cargo: it expects to parse the Cargo.toml
file to find the name of the current package or crate.

When building with Soong, there is no Cargo.toml file, but we can use
the CRATE_NAME environment variable instead. This works as long as the
package is named like the crate (which is the case for mls-rs-uniffi).

Bug: 330712798
Test: Presubmit plus local build of mls-rs-uniffi
Change-Id: Ia725df41d1c5dcf18ae7d7d9b698bc01bba580ae