Remove unnecessary checks
event.ip is of type InetAddress which is guaranteed to be an instance of either its
Inet4Address or Inet6Address subclass.
In addition, shouldIgnoreIncompleteNeighbor is called from handleNeighborLost(), which in turn
is called from NeighborEventConsumer which ignores all events for neighbors that are not in the
watchlist. So both checks are guaranteed to succeed and are therefore unnecessary.

Test: TH
Change-Id: I1a678ccda8f8c1beaf2509f3c8dabd395090db1f