Add super speed plus detection
The latest version of libusb supports SuperSpeed+ capability
descriptors. We can leverage this information to more accurately
infer the maximum supported speed.

However the new capability descriptor does not work with an enum but
with a mantissa + exponent which does not work well with the initial
design relying soly on enums.

Test: N/A
Bug: N/A
Change-Id: Ib6fc1fc929ef342e61329f035262ca6af9105981