Autoupdate of 2 submodules
*frameworks/av updated to e7b43745894b26e951c909a54efef632cd4481e7
Merge "libaudiohal@aidl: Allow onError callback in any state" into main
*hardware/interfaces updated to ab222c63bcbd6ebb1ecf7fe68cbf5534f7716f09
audio: Allow onError callback in any state

In practice, the AoC can crash while the stream is in
any state, thus this should be allowed for correct
handling of error conditions in the framework.

The list of states when `IStreamCallback.onError` can
be received is not enforced in VTS tests, thus no
test regressions are expected.

Bug: 339400414
Test: repro steps for simulating AoC crash in b/339400414
Change-Id: I9a52779f773307882109e159d715072de5af2ece