Flush frontend file after every message
If the buffer wasn't filled, the proto wouldn't be written out
to the frontend file till the next message. The "Analyzing Android.bp
files and writing build actions..." message is an action that takes
a long time, and is the only action running at that point, so it will
not be pushed out of the buffer by another message until it's finished
running.

Bug: 318434287
Test: m nothing
Change-Id: Ib3b9b4603c3cbc3f01d4ec9dfc807dbfe80beadc