Increase test timeout.
For the mentioned bug, there is a timeout. It's not clear if this is a
real deadlock happening in the test and/or code so increase the timeout
to see if the problem disappears.

I cannot reproduce this when running locally, so I'm hoping this is
just a case where the machine running the host tests is simply
running many different tests and occasionally it slows them down enough
to trip the timeout.

Bug: 341330386

Test: All unit tests pass.
Change-Id: Ia317b57815ca779864b8977631db7751b01d83cb