Remove unit tests from mts-adbd and mainline/adbd-mainline presubmits
These unit tests don't really benefit from being run on older mainline
releases, so remove them. The problem with running them on older
releases is that they depend on libc++, which won't match. If we move to
use a static libc++ (as seen in the link below), we also need to do
extra work to remove other symbols with a version script, so that the
other C++ users can still depend on the older platform libc++. This is
not worth the extra confusion as it provides low/no value.

https://android-review.googlesource.com/c/platform/packages/modules/adb/+/3101841

Bug: http://b/342542917
Test: TBD
Change-Id: I322d493f862a4ccea9478ec77e951423b314ba67