Fixed missing end time on previous runs.
Before this change, there was a bug with metrics reporting. When a job
was rerun, the end time of previous runs were unexpectedly cleared.

The root cause is that, when a job was cancelled and rerun,
`recordJobEnded` was called by a reporter inside chroot, and then
`recordJobStarted` was called by a reporter outside chroot. The latter
didn't have the job end time.

After this change, both `recordJobEnded` and `recordJobStarted` are
called by a reporter outside chroot. This change also brings a bonus:
the job duration now includes the time spend on teardown.

Bug: 336239721
Bug: 311377497
Test: Run Pre-reboot Dexopt twice.
Change-Id: Iaf8f574087e2ebbc99aef1004d862ae2777e80c7