Test duplicated/inherited parameters are part of the containing method
Expands the tests for duplicating and inheriting methods to check that
the newly created parameters reference the newly created method that
contains them. These changes reveal issues with the Turbine
`MethodItem.duplicate(...)` and `inheritMethodFromNonApiAncestor(...)`
methods not duplicating `ParameterItem`s correctly and temporarily
suppresses the tests. A follow up change will fix the issue.

Bug: 343430701
Test: ./gradlew
Change-Id: I6af0c652015ceeaf49bb5ab616e613d66f259791