Report UnflaggedApi issue on modifier changes to existing APIs
Previously, changes to modifiers of existing APIs were not required to
be annotated with `@FlaggedApi`. This change extends the existing
`UnflaggedApi` issue to require `@FlaggedApi` on changes to an
existing API's modifiers. The `UnflaggedApi` issues are reported as
`Severity.WARNING_ERROR_WHEN_NEW` until the existing issues are fixed.

Supporting that required a number of other changes:
1. The `computeMaximumSeverity(Item)` needed to be passed the `Issue`
   being reported so that it would not suppress `UnflaggedApi` issues.
2. The `ApiLint.report(...)` method needed a `maximumSeverity`
   parameter to allow the caller to reduce the configured severity of
   the issue.

Bug: 339166506
Test: ./gradlew
Change-Id: Ia955eebae29053d144be69e6e585c6236d2cfea7