DeviceSnapshotHandler: always throw exception for error responses
Error repsonses that can be deserialized are probably just string error
messages, so turn them into an exception and throw them. This simplifies
the callers and makes it easier to find the underlying error in test
failures.

The typical failure path was to throw an exception even before, so the
intricate code in callers to throw a new exception in a particular way
seems like a mistake (e.g. not all snapshot failures use
DeviceErrorIdentifier.DEVICE_FAILED_TO_SNAPSHOT). If any of that is
important, we should move it into DeviceSnapshotHandler.

Bug: 341991139
Test: m tradefed tradefed-tests && tools/tradefederation/core/tradefed.sh run singleCommand host --console-result-reporter:suppress-passed-tests --class com.android.tradefed.device.internal.DeviceSnapshotFeatureTest
Change-Id: Ic8b2cc6e126fa3afa1758b185649a7d9942285ee