Add tests for copying MethodItems
The `MethodItem.duplicate(ClassItem)` and
`ClassItem.inheritMethodFromNonApiAncestor(MethodItem)` methods do some
very similar work but are implemented differently under the covers.
This change adds some tests that run on both and ensure consistent
behavior, where needed.

It reveals an issue with the `default` modifier being copied from an
interface method even if the target is a class. That will be fixed in a
follow-up change.

Bug: 339166506
Test: ./gradlew
Change-Id: Id7de9c7d26f171064740c9a8adfa40a6447386f8