Disallow mixing abstract with static and default on methods
Previously, the `abstract` keyword would be written out alongside
`default` and `static` on methods even though they are mutually
exclusive. That is not an issue when reading from sources or real API
signature files but it does affect a test that incorrectly defines an
`abstract static` method in an API signature file. This change fixes
that.

The `abstract` and `static` keywords are allowed on a class.

Bug: 339166506
Test: ./gradlew
Change-Id: Ie6f3b3f6981107a81dc3602618165accdf136b38