Trace Redaction - Combine process event redactions
Outside of scheduling information, there are three events that must
be redacted: new task, process free, and task rename.

Each event can be redacted using the scheduling filter and modifier
handlers, making it easier for the scheduling and event redaction
to share redaction logic.

This moves the different redactions into one class. This was
previously avoided in an effort to make the system more modular.
However, to avoid issues around thread merging, this change should
decrease risk, improve readibility, and decrease computational the
number of packets decoded and encoded.

Bug: 318576499
Change-Id: I919266e02d250cf64154fd9c110d197d4d385cbc