Fix ApexCompressionTests
The test is supposed to verify the fallback behavior when the data
active apex was invalid. However, the test code doesn't trigger the
fallback behavior when the device-under-test is not writable yet and it
just passes for a different reason.

In this change, make sure the device-under-test is writable so that the
test correctly verifies the expected fallback behavior.

Bug: 342824092
Test: atest apex_compression_platform_tests:com.android.tests.apex.host.ApexCompressionTests#testFailsToActivateApexOnDataFallbacksToPreInstalled
Change-Id: I3f1396672992cf7ef959f9de8da461d568934fbb