Add support for normalizing final in ModifierListWriter.writeKeywords
The preceding change revealed some inconsistencies in the handling
of the `final` keyword between the models. This change adds a
`normalize` parameter to `ModifierListWriter.writeKeywords(..)` that if
true will cause `final` to not be written for methods if the containing
class is itself `final`. This is done conditionally because there are
some inconsistencies in the existing signature files where some methods
(specifically those that are explicitly `final` and inherited from a
hidden class) are marked as `final` even though the containing class is
`final`.

That inconsistency will be resolved in the future but in the meantime
this makes the normalization conditional.

Bug: 339166506
Test: ./gradlew
Change-Id: Iebc005d247c8453f9328b2dccd259b997945f9bf