Require a change in deprecated status to be flagged
Previously, a change in deprecated status was not required to be
flagged. This change updates the `UnflaggedApi` check to check
the deprecated status.

It was necessary as part of this change to modify the `ApiLint`'s
`FilteringReporter`, which previously would suppress an issue on a
deprecated `Item`, to only suppress issues on deprecated `Item`s if it
was not previously deprecated. Otherwise, the `UnflaggedApi` report
would be suppressed.

Bug: 339166506
Test: ./gradlew
Change-Id: I26c8331502356417e57535c022ceba8b74affeb9