Duplicate super methods into corresponding class item in ApiLint
Previously, the `UnflaggedApi` check could compare the modifiers of a
super method from the previously released API and an overriding method.
That caused some false positives due to differences in the meaning of
modifiers caused by differences in the containing classes. e.g.
`default` on an interface is the same as no modifier (including no
`static` or `abstract` modifiers) on a normal class.

This change uses the `MethodItem.duplicate(ClassItem)` to create a copy
of the super method on the class corresponding to the class containing
the overriding method. That ensures consistency in the modifiers.

Bug: 339166506
Test: ./gradlew
Change-Id: I819cd7cf98b7a87743d60f094410f8a375f446af