Add tests to show deprecated status after copying
This change adds tests to show the effect copying a field or method has
on their deprecated status, i.e. the `originallyDeprecated` and
`deprecated` properties (a follow-up change will expand this to cover
`effectivelyDeprecated`. It shows that there are a number of issues.
* `TextMethodItem.duplicate(ClassItem)` treats `deprecated` differently
  to the other models, i.e. it copies `deprecated` from the source
  `MethodItem` ignoring the target `ClassItem`.
* None of the models update the deprecated status correctly in
  `FieldItem.duplicate(ClassItem)`.
* The `inheritMethodFromNonApiAncestor(...)` method does not update the
  `deprecated` property at all; it relies on the
  `ApiAnalyzer.performChecks()` method to do it later.

In order to test the `deprecated` property properly it was necessary to
replicate the part of `ApiAnalyzer.performChecks()` that propagates the
`deprecated` property from `ClassItem` to `MemberItem`. That will be
removed when the `deprecated` property is removed.

Bug: 315206833
Test: ./gradlew
Change-Id: I4b8cb42f83f22f83a170db6a97a7836bd64b13a9