Compute effectivelyDeprecated on demand
Previously, `effectivelyDeprecated` was a `var` that was updated by
the `ApiAnalyzer.markAsDeprecatedIfNonPackageParentIsDeprecated()`
method. However, it was not updated when copying members. Rather than
try and update it whenever the state that affected it was changed this
change switches to compute it on demand. The cost of doing that is
quite low and it ensures that the value is always up to date.

This change also updates the deprecated assertions that previously only
checked `deprecated` to also check `effectivelyDeprecated` to ensure
that it is behaving consistently.

Bug: 315206833
Test: ./gradlew
Change-Id: I93db4942d8e0c162d14c2f5876ccc4091f0fb469