Add CommonCopyFieldItemTest to test FieldItem.duplicate()
This change adds common tests for `FieldItem.duplicate(ClassItem)`. It
is intended to be equivalent to `CommonCopyMethodItemTest` which is why
it is named `CommonCopyFieldItemTest` and not
`CommonDuplicateFieldItemTest`. The tests added in this change are just
testing working but previously untested functionality. Follow-up
changes will add additional tests.

Functionality of the `FieldItem.duplicate(ClassItem)` method that is
specific to source models is already tested in `SourceFieldItemTest`.

Bug: 315206833
Test: ./gradlew
Change-Id: Ie21d1844e268d339e38710c11c6cc7a90e88efc1