Lookup bindings from the build before the rule
I think this is an unintuitive lookup order, but it appears to be
what android's ninja uses:
https://android.googlesource.com/platform/external/ninja/+/8bdfb2cefbf380d69cde30bed1489c95987552d8/src/graph.cc#568

This doesn't seem to affect the actual build at all, but previously,
n2 was displaying different action descriptions than what ninja was.
This is somewhat surprising, but I guess it's relatively rare to have
the same binding on both the build and the rule, and it was mostly
localized to descriptions in our case.

Bug: 318434287
Test: Building android with n2, see proper descriptions
Change-Id: I0fd6d4d51352a5e086abd013fa3506702ee0ebca