Make inheritMethodFromNonApiAncestor(MethodItem) update deprecated
Previously, the `inheritMethodFromNonApiAncestor(MethodItem)`
implementations did not set the `deprecated` property of the newly
created `MethodItem` to `true` if its containing class was itself
`deprecated`. It relied on the `ApiAnalyzer.performChecks()` method
which is called later to do that.

This change makes the implementations consistent with the
`MethodItem.duplicate(ClassItem)` method which will allow the
`ApiAnalyzer.performChecks()` to stop mutating the `deprecated`
property.

It moves the code for updating the `deprecated` property from the
`MethodItem.duplicate(ClassItem)` implementations into
`MethodItem.updateCopiedMethodState()` which is also called from
`ClassItem.inheritMethodFromNonApiAncestor(MethodItem)`
implementations.

Bug: 315206833
Test: ./gradlew
Change-Id: Ida4470f50304e7a982845335c65b7b4b9196516f