Update deprecated property in FieldItem.duplicate(ClassItem)
Previously, the `FieldItem.duplicate(ClassItem)` implementations did
not set the `deprecated` property if the target class was deprecated.
That was inconsistent with the `MethodItem.duplicate(ClassItem)`
implementations and could lead to different behavior.

This change updates the `FieldItem.duplicate(ClassItem)` to make it
consistent. Unfortunately, the inconsistency affected the signature
files in Android so in order to avoid causing costly churn in those
files It was also necessary to tweak the writing of `@Deprecated` to
preserve the inconsistencies in the signature files. They will be
cleaned up in future.

Bug: 315206833
Test: ./gradlew
Change-Id: Ia762c8a2ddd307ffa0fbdb184ecda5711c1187c9