Refactor deprecation writing to simplify future changes
There are quite a few differences in the handling of deprecated status
that leads to inconsistencies in the model and that results in
inconsistencies in signature and stub files. This change is in
preparation for moving those inconsistencies into the writer so that
the inconsistencies can be moved out of the model which will improve
the behavior of the rest of Metalava.

Bug: 315206833
Test: ./gradlew
Change-Id: I0d924f4440ad51a2ce0254dac8bfa997a39c3a28