Add tests for special deprecation behavior
There are quite a few differences in the handling of deprecated status
that leads to inconsistencies in the model and that results in
inconsistencies in signature and stub files. This change adds some
tests for those cases to ensure that future refactorings do not break
that behavior.

Longer term the inconsistencies in signature files will be cleaned up.

Bug: 315206833
Test: ./gradlew
Change-Id: I87775c6e5699e5ceee0a687f81454d07cc7396c2