Build libextservices_tests with static libc++ and a version script
This tests are run on older versions which causes them to fail because they're built with libc++.so from ToT AOSP and
run against libc++.so on older OS versions.  Similar to aosp/3092549,
build these tests with static libc++ and add a version script so symbols
from the executable don't override those from the system.

Bug: 345108414
Test: abtd
Change-Id: Ieae1e2a4ab640d598a6f938e4d0e4d344c28394d