Autoupdate of 2 submodules
*frameworks/support updated to 69919b88daaae3bceef544e76313969745234648
Merge "Temporarily disable androidx_with_metalava" into androidx-main
*tools/metalava updated to 9c86cda1a8c1b628e8cb1d9c759e869a71e8fef9
Use effectivelyDisabled when checking equivalence

Previously, when checking equivalence of modifiers and methods for the
purpose of determining whether a method could be elided from an API
signature file it checked `deprecated`. The behavior of that was
inconsistent depending on whether it was performed before the
`deprecated` property was propagated from deprecated classes onto its
members or afterwards. This change switches it to use
`effectivelyDeprecated` which is computed dynamically so always
accurate.

This has the potential to cause minor changes in API signature files.

Bug: 315206833
Test: ./gradlew
Change-Id: I87af6f3a0be6fa1be5228321b3c669bf69c91537