Move throw types filtering from SignatureWriter
Previously, throw types filtering, was performed by `SignatureWriter`
in the `writeThrowsList(MethodItem)` method. However, it did not write
type use annotations or filter them because that is not currently
supported. This change moves the throw type filtering into
the `FilteringApiVisitor` and applies `typeAnnotationFilter` so that
when type use annotations on throw types is properly supported that it
will work correctly.

Bug: 348701068
Test: ./gradlew
Change-Id: Ia9d5310ccbb0ae791aea26aff8a54c097b5385ba