Replace std::process::exit with ExitCode
std::process::exit's documentation warns about it having a unclean
shutdown. Particuarly it says destructors on other threads won't be
run, which I think could cause n2's detached threads to not be cleaned
up, and cause n2 to hang forever. It recommends using ExitCode instead
for better cleanup, switch to that to see if it fixes the ci flake.

Bug: 318434287
Test: cargo test
Change-Id: I9de83643d05e0f331b4b7d853e997810d148768f