Make SignatureWriter extend BaseItemVisitor
Previously, `SignatureWriter` extended `ApiVisitor`. This change
switches it to extend `BaseItemVisitor` instead. That required adding
a new `FilteringApiVisitor` that delegates to `SignatureWriter` and
which is constructed from the `SignatureWriter` by the new
`createFilteringVisitor(...)`. Some of the parameters to
`SignatureWriter` were only needed by the `ApiVisitor` and so were
moved to the `createFilteringVisitor(...)` method.

Follow-up changes will move filtering functionality from
`SignatureWriter` into `FilteringApiVisitor`.

Bug: 348701068
Test: ./gradlew
Change-Id: Ie1df94e1410936482ad9de69fa17462929149e20