Fix CtsAdminPackageInstallerTestCases module not able to finish on Android Auto or before Android T
Android Auto does not support the Main User, it uses the Primary user to identify the device owner. However it cannot be a primary user on a headless system as the primary user is a full system user, and the headless system separates the full user form the system user. The main user was introduced in Android T, therefore this CL provides backwards compatibility with API below 33 which takes Primary User instead of Main User. The change checks against feature:android.hardware.type.automotive and Android API version to provide the correct owner of the device.

Bug: b/342086201
Bug: b/343473656
Test: cts -m CtsAdminPackageInstallerTestCases
Change-Id: Iae2a79655a4f0369a4ee53c43ac382964094dba7