Prioritize method defined last from multiple signature files
Previously, when a method was defined multiple times for an API, e.g.
in both `public` and `system` signature files it would use the one
defined first, i.e. in `public`. That is wrong as the `system` method
could be different, e.g. it may extend the `throws` list or use system
only annotations that are not available in the `public` API.

This changes the behavior to replace the existing method with the later
one instead of ignoring it.

Bug: 347922041
Test: ./gradlew
Change-Id: I292036d865b0b9cd2e4c142d5611a186585778d4