Remove qualifiedName from Psi/UAnnotationItem.create(...)
Allowing the `qualifiedName` to be passed in provides no real benefit
apart from in one use case where it avoided getting the property value
again as the caller had already retrieved it. That was not worth the
extra complexity which made follow-on changes a little more complex.

This change removes the `qualifiedName` parameter and replaces it with
an `originalName` local variable. The name was changed to make it clear
that this is NOT the same as the `AnnotationItem.qualifiedName`.

Bug: 347944677
Test: ./gradlew
Change-Id: I7ed2e072c60a6679a134ac75ae5f5cafbb2254d9