Make Psi/UAnnotationItem's originalName/qualifiedName non-null
Previously, `PsiAnnotationItem` and `UAnnotationItem` would accept and
then pass a nullable `originalName` and `qualifiedName` to their super
`DefaultAnnotationItem` constructor and rely on the consumer to ignore
an `AnnotationItem` with a `null` `qualifiedName`. This change causes
their factory methods to just return a `null` `AnnotationItem` if
either of those are `null` and only accept non-null values for them in
their constructor.

It also fixes up the code that is broken when their `create` methods
switch from `AnnotationItem` to `AnnotationItem?`.

Bug: 347944677
Test: ./gradlew
Change-Id: Ife5880a2416e4e37bfe12331a6e8d01119f921cb