Reuse TurbineSourceFile where possible
Previously, a `TurbineSourceFile` was created for each top level class
which mean that if a file contained multiple top level classes in the
same `SourceFile` they would each get their own `TurbineSourceFile`,
with their own `LineMap`. This change registers the `TurbineSourceFile`
by path so that they can be reused. That registry will be used in a
follow-up change to get the `TurbineSourceFile` for an `AnnoInfo` and
create a `TurbineFileLocation` for an `AnnotationItem`.

Bug: 346530869
Test: ./gradlew
Change-Id: Idc4b89df46c7160ade7a28e5235bef5d620f5059