Use AnnotationItem.location for @Deprecated in UnflaggedApi check
Previously, when an `@Deprecated` annotation had been added and was not
flagged it would report the issue at the annotated item's location. If
that was on a separate line to the annotation then it would not be
treated as an error by the "Build Warnings" Aye Aye analyzer which
would allow new cases to be added to the sources.

This change reports the issue against the location of the annotation
itself ensuring that new cases will be treated as an error.

Bug: 319874764
Test: ./gradlew
Change-Id: I19ddd2390ed09d7ae7abbd48cafabfb577215152