Refactor method dispatching in ComparisonVisitor
Previously, it overloaded `visitAdded`, and `visitRemoved` methods to
do two separate jobs, check whether an item was added/removed or just
moved and dispatching to the appropriate `added(...)`, or
`removed(...)` method.

This change renames them to make it clearer and narrows the scope of
the `USELESS_CAST` suppression for casts which are actually useful.

Bug: 347885819
Test: ./gradlew
Change-Id: Iec686bcf395086f7625b535f54be1136318ba20f