Retry "Remove --check-compatibility:base"
The `--check-compatibility:base` option was a bit of a hack that was
added because `--check-compatibility:api:released` did not support
being specified multiple times. The only uses of the former have been
replaced with the latter and so it can now be removed.

A number of tests that either directly tested the behavior of the base
`Codebase`, or relied on it have been deleted after first checking
whether there was any point to migrating them to use multiple copies of
`--check-compatibility:api:released`.

Bug: 320245343
Test: ./gradlew
Change-Id: Iaba157e0857160b9f8b90c87f7c21858b8151ef8