Pull modifiers duplication out of TypeItem.duplicate(...)
Previously, the type specific `duplicate(...)` methods that take one or
more arguments (not the `TypeItem.duplicate(withNullability)` method)
would always duplicate the type's `TypeModifiers` even when
unnecessary. This change adds a new `modifiers: TypeModifiers` method
that lets the caller decide if it is necessary which makes the code
more efficient and flexible.

Most existing places that call one of those methods have been modified
to pass in a duplicate of the `TypeModifiers` to replicate the previous
behavior for now. There are two exceptions:
* `TextTypeItemFactory` - in this case there is no point in duplicating
  it because it is not being modified and the original owning
  `ClassTypeItem` is being discarded anyway.
* `CommonClassItemTest` - in this case the modifiers are not mutated
  afterwards.

In the other places there is either a slight possibility that the
modifiers would be mutated which would affect both the original and the
duplicate in weird ways that would be really hard to debug, or they
will be fixed in follow-up changes.

Bug: 348701068
Test: ./gradlew
Change-Id: I8c929aeb2603ba92e409319d98b6d8463978bc02