Add Item.setType(TypeItem) to allow Item.type() to be mutated
Previously, the `TypeModifiers` provided mutator methods which allowed
a `TypeItem` returned by `Item.type()` to be modified in-situ. This
change adds a new `Item.setType(TypeItem)` that instead allows the
`TypeItem` associated with an `Item` to be replaced in its entirety.
It then uses than in place of `setNullability(TypeNullability)`.

A follow-up change will removed the `TypeModifiers` mutators.

Bug: 325454211
Test: ./gradlew
Change-Id: Ifcf49b224fcca3de0265db9f191f126fa94431ac