Optimize TypeItem.convertType(...) to reuse instances
Previously, `TypeItem.convertType(TypeParameterBindings)` would always
create a new instance of the `TypeItem` on which it was called even
when that was not necessary. This change optimizes the methods so that
they only create a new instance when absolutely necessary.

It adds new `substitute(...)` methods for use by `convertType(...)`
implementations instead of `duplicate(...)`. They are like the existing
`duplicate(...)` methods except that they will only create a new
instance if the parameters are not the same as the properties of the
`TypeItem` on which they are called. If they are the same then they
will just return the instance on which they are called.

The `PrimitiveTypeItem` and `VariableTypeItem` do not need a
`substitute(...)` method but were also optimized.

The test was updated to show the new, improved behavior.

Bug: 325454211
Test: ./gradlew
Change-Id: I939e0e49698d4c87f575acf83382b935ba1d9ad3