Add a permission file to apex
An APEX com.android.adservices includes SdkSandbox app that requires
com.android.adservices.sdksandbox.xml to be installed in
etc/permissions.
Although the app is included in the APEX, the xml file is installed
in the system partition instead of the APEX because APEXes do not
follow the "required" dependencies.
Add the xml file to the APEX as prebuilts and remove the "required"
property from SdkSandbox to install the xml file inside the APEX
instead of system/etc/permissions.

Bug: 349495890
Test: com.android.adservices.sdksandbox.xml is moved
        from /system/etc/permissions
        to /apex/com.android.adservices/etc/permissions
Change-Id: I6b125a0f9e2a6730f69222188dd1394e18603464