Disallow changing type property on ClassItem
Change https://r.android.com/3146376 added support for changing the
`type()` property on `ClassItem`. Some concerns were raised in review
that it was not necessary or even meaningful but the changes were
submitted anyway as they replicated the existing behavior. After
investigating the current usage made of that mechanism in the Android
build and looking into the nullability annotations themselves it was
determined that it is not necessary or meaningful. So, this change
removes support for them.

Bug: 325454211
Test: ./gradlew
Change-Id: I560787dd1456bf6e65d60a360e84982a011dda38