Redirect dex2oat logs to stderr for VM tests.
VM environment is somewhere in-between host and target: we consider VM
a target device use the target builds for it, but it's Linux rather than
Android, so it has no logcat / tombstoned, and all the logs must go to
stderr, or else they are lost. This patch redirects dex2oat logs to
stderr (dalvikvm logs were already redirected).

Test: Ran VM tests on LUCI with this patch (using led), observed dex2oat
      logs on stderr (some tests are currently failing).
Bug: 283085397
Change-Id: I37ae66feb8a8d51aef3abb908affd15bbf7ede23