Move interface type list filtering from SignatureWriter
Previously, interface type list filtering was performed by
`SignatureWriter`. This change moves that into the `FilteringClassItem`
in `FilteringApiVisitor` and moves the `getInterfacesInOrder(...)`
method from `SignatureWriter.kt` into `FilteringApiVisitor.kt`.

As the order of the interfaces is very particular to `SignatureWriter`
an `interfaceListAccessor` lambda parameter was added to all that
behavior to be customized.

As all callers of `writeExtendsOrImplementsType(ClassTypeItem)` now
filter type annotations before calling it there is no need for it to
pass `filter = filterReference` so that was removed.

Bug: 348701068
Test: ./gradlew
Change-Id: I577415706888a7aebf5e5abfbb657edbecb78b12