Add tests for type use annotation filtering
The `SignatureWriter` depends upon the type use annotation filtering
performed by `TypeItem.toTypeString(...)` when it is passed a non-null
`filter` parameter. Unfortunately, previously that was not well tested.
This change adds `TypeUseAnnotationFilteringTest` to improve the test
coverage. The behavior of one of the tests is inconsistent between Psi
and Turbine and while in one aspect Turbine is more correct than Psi
(it does not duplicate type use annotations) it also leaves out the
following from the signature file:
```
@java.lang.annotation.Retention(java.lang.annotation.RetentionPolicy.CLASS)
```

That behavior is relied upon by the Android build whereas it does not
yet use type use annotations so Psi was chosen as the expected behavior.

Bug: 348701068
Test: ./gradlew
Change-Id: I686697ec13f79ee99b03771aea14fbe923da39c3