Remove KotlinStubWriter
The `KotlinStubWriter` is not currently used by AndroidX and it does
not work. It generates invalid and unusable Kotlin code and has poor
test coverage. At best it does nothing and at worst it is a drain on
engineering time to maintain it when it is broken by changes made
elsewhere.

If/when we need to generate Kotlin stubs it will be simpler to start
from scratch, iteratively adding tests.

Bug: 349774933
Test: ./gradlew
Change-Id: I654ec30997850c3ea1dce49b6b06ff7ef4b7f90a