Move stub constructor dispatching from JavaStubWriter
Previously, `JavaStubWriter` would synthesize a `visitConstructor(...)`
call for the stub constructor if it did not appear in the filtered set
of constructors that would be visited by the `JavaStubWriter`. That was
needed for stub constructors that were inaccessible through the API and
so would usually be ignored, e.g. package private constructors added
simply to allow the stubs to compile properly.

This change moves that code into the `FilteringApiVisitor`, controlled
by a new `visitStubsConstructorIfNeeded` property which defaults to
`false` as it is only needed for stubs.

Bug: 348701068
Test: ./gradlew
Change-Id: Ie9f4d5702035a065b3883c46c409f29d0e301721