Move filtering of super constructor's siblings from JavaStubWriter
Previously, the `JavaStubWriter` would execute the following to
determine whether a super constructor had any siblings and so casts
would be needed to ensure that the `super(...)` call was unambiguous.

```
val includeCasts =
    superConstructor
        .containingClass()
        .constructors()
        .filter { filterReference.test(it) }
        .size > 1
```

In order to remove the `filter(...)` call it was necessary to add
support for the following into the `Filtering*Item` classes.
* `ConstructorItem.superConstructor` - this does not need filtering
  as it has already been taken into account when initializing it.
* `ConstructorItem.containingClass()` - this needs to return a
  `FilteringClassItem` to access the filtered constructors.
* `ClassItem.constructors()` - this needs to filter the constructors
  (to replace the `filter(...)` call above, but it does not strictly
  need to return `FilteringConstructorItem`s but it does for
  consistency.

The above does mean that a `FilteringClassItem` will be constructed
multiple times for the same `ClassItem` but the cost is minimal and as
yet there is no need for a `ClassItem` to always be wrapped in the
same `FilteringClassItem`. Bearing that in a mind it was decided that a
cache would be an unnecessary complication that would increase the
lifespan of each `FilteringClassItem` making garbage collection more
expensive.

Bug: 348701068
Test: ./gradlew
Change-Id: If87d0c576c14e9e0e6e09a42c4d6001ec5331ce0