Autoupdate of 3 submodules
*build/blueprint updated to 3d060a1903562086803dc337312177f00d5f32fb
Separate blueprint parsing and evaluating

Before this cl, blueprint expressions were evaluated as they were
parsed. We want to add a feature to select statements where we can
bind the value of soome value from soong into a blueprint variable,
that then can be used like a regular variable in the .bp file. This
means that select statements need to hold whole unevalated expression
trees, and have the ability to evaluate them later on when the value
of the bound variable is known.

This cl doesn't implement the new select syntax, but it does split
blueprint's parsing and evaluating into two separate stages. We also
store expressions in selects and evaluate them when the select is
resolved.

I didn't do extensive performance evaluation, but a simple comparison
of the time of `touch Android.bp && m nothing` before/after this cl
showed a 1 second speedup. (That was probably just noise)

Bug: 323382414
Test: m nothing --no-skip-soong-tests
Change-Id: I12f373719991afeb4aec76517153f32229d97ff2

*build/soong updated to b7bd9a17ecc060ee6ae4108f41e8e4f860c34e0a
Merge "Update selects_test for selects with deferred expressions" into main
*development updated to b234331dccb8a1845a62b421d22c7b8ed1fa6b64
Merge "Update bpflatten for blueprint changes" into main