Fix handling invalid header length
The previous fix failed to exit the while loop when
encountering an invalid header length,
resulting in the processing of subsequent headers,
which could be incorrect.

This update addresses this issue by throwing an IOException
to immediately exit the while loop upon
detecting an invalid length.
(As the function is marked to throw IOException,
all the callers are handling IOException already)

This ensures proper error handling and prevents
the processing of potentially corrupted data.

Additionally, it handles the case of an empty name header,
eliminating unnecessary error logs.
Pratically it does not change the logic at all.

Bug: 338905642
Test: atest BluetoothInstrumentationTests

Change-Id: I33df2f6040791ed5cc405eb7d804c88e308eac2c