Refactor FilteringApiVisitor handling of interface lists
Previously, it took a single `interfaceListAccessor` parameter that was
responsible for filtering and sorting. That was because the
`SignatureWriter` has a special order that requires access to both the
filtered and unfiltered interface lists. Unfortunately, that meant that
every other user of `FilteringApiVisitor` will need to duplicate that
even when their order just requires providing a `Comparator`.

This change adds an optional `interfaceListComparator` and makes the
`interfaceListAccessor` optional. If the latter is not provided then
the `FilteringApiVisitor` will do the filtering. If the former is not
provided then `FilteringApiVisitor` will not sort the list.

This allows the current behavior of `SignatureWriter` to be kept by
simply not providing a `interfaceListComparator`. However, other usages
can just provide a `interfaceListComparator` and not provide a
`interfaceListAccessor`.

Bug: 348701068
Test: ./gradlew
Change-Id: Ic1d0fb0717ac5beb95af9aa7f7240957804b9af5