Do not fetch_artifacts when executed via Gantry.
When executed via gantry, it is expected that the module sdk artifacts
are already downloaded and are present in the --gantry_download_dir.

Hence, do not fetch the artifacts again.
The artifacts are fetched outside this script because finalize_sdk uses
x20 fetch_artifact, which is failing to work in rapid production.

Bug: 321970705
Test: packages/modules/common/tools/finalize_sdk.py -f 14 -b 256848771 -c mainline -r 'Finalization Test' -g /tmp/tmpmx5ncaoq  11965776
Change-Id: I5c40569a571ac9ef055b97b8fbf0d67a120f7657