Remove unnecessary use of filterEmit from DexApiWriter
The `filterEmit` that is passed in to `DexApiWriter` is for the
`ApiType.ALL` which filters out anything for which `Item.emit` is
`false`. However, `ApiVisitor.shouldEmitClass(ClassItem)` which
determines which `ClassItem`s will be visited also ignores any
`ClassItem` whose `emit` property is `false`. So, the
`visitClass(ClassItem)` method will never be called on a `ClassItem`
for which `filterEmit.test(Item)` will return false. So, this change
removes it.

Bug: 348701068
Test: ./gradlew
      # Double checked on Android build by building
      # out/soong/hiddenapi/hiddenapi-flags.csv with and without this
      # change and verifying that there was no change.
      # Also modified the code to throw an error when
      # `filterEmit.test(cls)` returned false and that was never
      # thrown.
Change-Id: I154127f523f9e921173d3c6a85db8e5bb9ae4e87