Handle nestInnerClasses=false by flattening class nesting
Previously, `nestInnerClasses=false` would be handled by iterating over
the class nesting hierarchy as normal but visiting the nested classes
after the call to `afterVisitClass(ClassItem)` for the parent class.
The problem with that is while it works fine in the very simple case
when every package and class are visited, it very quickly becomes very
complicated when the classes will be visited conditionally and when the
nested classes need visiting even if the containing class is not.

e.g. take the following case:
```
public class Foo {
    public void method() {}
    public class Bar {}
}
```

For both values of `nestInnerClasses` they would be visited as follows:

```
  visit(Foo)
  |
  +--> visit(Foo.Bar)
```

With that approach the only way to visit `Bar` is to first visit `Foo`
so the logic as to whether `Foo` should be visited depends on whether
`Bar` should be visited. Similarly, the logic as to whether `method()`
should be visited will depend on whether `Foo` was being visited
solely to visit `Bar` (in which case `method()` should NOT be visited)
or if `Foo` was being visited for its own sake (in which case `method()`
SHOULD be visited).

This change simplifies that support by flattening the class nesting
before iteration, so for the above example the classes would be
visited as before when `nestInnerClasses=true` but like this when
`nestInnerClasses=false`:

```
  visit(Foo)

  visit(Foo.Bar)
```

So, the decision as to whether to visit `Foo.Bar` is independent of
whether `Foo` should be visited which simplifies the logic.

Follow-up changes will remove the now unnecessary complexity. They will
do it in very small steps, to make them easier to understand and will
be tested using `gather-android-metalava-artifacts.py` to make sure
that they do not change the Android artifacts.

Bug: 350130387
Test: ./gradlew
Change-Id: I954b11db9121b95a0dbbceebee6aacf0780bca3c