Remove VisitCandidate.innerClasses property
Previously, the `innerClasses` property was needed to check whether an
outer class needed visiting in order to reach the inner class. However,
thanks to the preceding change that is no longer necessary and so this
change removes the `innerClasses` property and just inlines it into the
code to nest visits to the inner classes inside visits to the
containing class and removes related functions like
`shouldEmitInnerClasses(...)` and `shouldEmitAnyClass(...)` which are
no longer needed.

Bug: 350130387
Test: ./gradlew
Change-Id: Ic659ea2e7ea9d3fa0df5111eccf5206fd4506b13