Add tool to ease testing impact of changes on Android build
Metalava is used extensively within the Android build to create a
number of different artifacts and care must be taken when changing
Metalava that it does not cause any unexpected changes to those files.

The build prevents any unexpected changes to signature files but does
not do so for others.

At best it may do a partial indirect test by consuming them in other
parts of the build. e.g. stubs are checked that they will compile and
can be compiled against. However, while that will detect all
compilation errors it will only detect if an API element is missing if
it is used. That goes for things like missing methods as well as more
subtle things like the presence and type of nullability annotations
used, e.g. `@Nullable`, `@NonNull` vs `@RecentlyNullable`,
`@RecentlyNonNull`.

Similarly, for `api-versions.xml` files the continuous builds will only
detect syntax/structural errors in the generated files but not whether
the contents are correct.

This change adds a `gather-android-metalava-artifacts.py` that makes it
easier to detect unexpected changes. When run it will build and gather
some artifacts and copy them into a directoy. It can be run before and
after making a change, copying the artifacts into two different
directories  that can then be easily compared to detect any
differences.

This was added because of changes to the `ApiVisitor` class which
impacts lots of Metalava. While the Metalava tests can will detect lots
of changes they are not representative of how it is used. They could
not cover some behavior, so will not detect breaking changes. In that
case when a difference is found in Android the tests need improving.
Conversely, they could detect changes in legacy behavior that is no
longer relied upon in which case they can be changed or removed.

Bug: 350130387
Test: ./gradlew
Change-Id: Ic8b07fd935f0eb41129f6c00b8c88e4f7f6fd2bd