Remove shouldEmitClass(...) as it always return true
Previously, `shouldEmitClass(vc)` returned the result of calling
`shouldEmitClassBody(vc)` so could be inlined. Doing so would produce
the following:
```
    if (!shouldEmitClassBody(this)) {
        return
    }

    val emitThis = shouldEmitClassBody(this)
```

However, `shouldEmitClassBody(this)` is implemented so that the second
call will always returns return the same value as the first call. Which
means that the second call is always `true` as that is the only value
from the first call to reach the second call. So, the second call can
be replaced with `true`.

That is true of the `ComparisonVisitor` override which always returns
`true`.

So, this change removes `shouldEmitBody(...)` and replaces the call to
it with `true` and then simplifies the code to remove the now redundant
`if` statements.

Bug: 350130387
Test: ./gradlew
Change-Id: I8c591f6bd179da32edaf40fd64e4d46c7df8fb60