Add getVisitCandidateIfNeeded to create VisitCandidate if needed
This combines the filtering (done by `include(ClassItem)` and
`shouldEmitClassBody(VisitCandidate)`) and the creation of a
`VisitCandidate` into a single method. If the filters reject the class
then the method will return `null`, otherwise it will return a newly
create `VisitCandidate` that wraps the class and provides access to its
filtered members.

The purpose of this is to allow the follow-up change to support
filtering the classes before visiting any to allow removal of the
lazy package visiting.

Bug: 350130387
Test: ./gradlew
Change-Id: I571995cb86d1b8f39c6b96c2433c904ad4f09e3e