Filter classes early to avoid visiting packages lazily
Previously, the classes in a package that would be emitted was not
known at the time the package was visited. That was a problem as empty
classes should be visited. To handle that `visitPackage(PackageItem)`
was called just before the first class to be emitted and
`afterVisitPackage(PackageItem)` was called after visiting all the
children and only if at least one class had been emitted.

This change determines the list of classes that will be emitted in
the `visit(PackageItem)` method and uses that to simplify the package
visiting.

Bug: 350130387
Test: ./gradlew
Change-Id: I0d53e98b629ad4e34559499a6ae873a4c933c701