Remove unnecessary BaseItemVisitor.afterVisit... methods
Apart from some delegating to another `BaseItemVisitor` the following
methods were unused, so this change removed them.
* `afterVisitConstructor(...)`
* `afterVisitMethod(...)`
* `afterVisitField(...)`
* `afterVisitParameter(...)`
* `afterVisitProperty(...)`

Apart from the first two none of them have any internal structure that
is visited after the `visit...(...)` so any work that would need to be
in the `afterVisit...(...)` method is just being done in the
`visit...(...)` method.

As for the first two, they do have parameters that are visited after
the method/constructor but they are intrinsically part of that and
cannot be used alone so any work that needed to be done after the
parameters were visited is just being done in the method/constructor
visit method. In fact very few implementations of `BaseItemVisitor`
override the `visitParameter(...)` method,

Bug: 348701068
Test: ./gradlew
Change-Id: I8ee391fc2449a7308ee0e25da13304953ec29657