implement tests for GameTextInput
This CL adds initial implementation of tests for GameTextInput.
This tests are currently non-stable (sometimes nullptr access in
onActivity happens). I believe this is a minor issue which could be
fixed later.

In general, I added androidTest/ directory, which contains two tests.
ExampleInstrumentedTest is a simple check that tests work.
InputTest is a real instrumented test that uses Espresso and Android
Instrumentation. It also uses a special test Activity class called
MainActivity (it only exists in test environment and is not included
into library).

This also adds a new command in build.sh script. To run all tests, do
./build.sh tests

Bug: 343915764
Test: ./build.sh tests
Change-Id: I177d82db3b5ab9d76745cd23a25da9f852e60f38