Make BaseItemVisitor package visiting consistent with ApiVisitor
Previously, `BaseItemVisitor` would preserve class nesting while
visiting classes and implement `nestInnerClasses` by visiting nested
classes before or after calling `afterVisitClass(ClassItem)` on the
containing class depending on whether `nestInnerClasses` was `true` or
`false` respectively. That was the same way that `ApiVisitor` did it
but it caused problems when the nested classes needed to be visited
even if the containing classes did not. So, it was changed to flatten
the classes into a single list using `PackageItem.allClasses()`.

This change makes a similar change to `BaseItemVisitor` and pulls some
of the `ApiVisitor` code up to allow it to be reused.

Bug: 350130387
Test: ./gradlew
Change-Id: Ic70742cec1fe299088b93e9d1f4d89a0ff6da5f6