Inline the shouldEmitClassBody(VisitCandidate)
Previously, the `shouldEmitClassBody(VisitCandidate)` performed a
number of checks, some of which required a `VisitCandidate` object and
some of which did not. This change inlines the checks and does the ones
that do not require a `VisitCandidate` object first to avoid having to
create a `VisitCandidate` object when not needed.

That does slightly change the order of the `filterReference` and the
(newly renamed and inverted) `containsNoEmittableMembers` check.
Previously, the latter was done first but because it requires a
`VisitCandidate` instance but `filterReference` does not they were
swapped.

Swapping the order does not change the logic as previously they were
essentially AND-ed together. It may affect performance, and memory
usage but any differences were lost in the noise.

Bug: 350130387
Test: ./gradlew
Change-Id: If58dd216951cda3e6f7fb298893ae5b42f617afc