Pass arguments to resolvePartialDownload
Ensure we put enable-resolve-sym-link in the arg map for the
partialDownloadFeature.
I also added it to the unused codepath of doing partialDownload without
the feature.

The key is to ensure that everywhere that calls
`resolvePartialDownloadZip` first call resovler.setExtraArgs() with
resolve-enable-sym-link values.

This should help us launch using symlinks in suite zips.

Test: abtd && presubmit
Change-Id: Ibbb44f5f48cd10022c59df44516a99bd52be97d1