Merge remote-tracking branch 'aosp/upstream-main' into HEAD
Just syncing with upstream.

We already have "drm_hwcomposer: Remove hardcoded PAGE_SIZE usage in
gralloc_helper" merged in aosp, so this merge only brings in one fix
"drm_hwcomposer: Use the global default cpp_std", which was separately
submitted here: https://r.android.com/3115073

This is an unflagged merge, but as it only contains the one fix, I
think syncing with upstream is preferable to merging the fix
separately and eventually having two copies of it in the history.

* aosp/upstream-main:
  drm_hwcomposer: Use the global default cpp_std.
  drm_hwcomposer: Remove hardcoded PAGE_SIZE usage in gralloc_helper

Signed-off-by: John Stultz <jstultz@google.com>
Change-Id: Ibf0496a8a3434ec946edf8d866a0e55a33cfbfa3