Correct mis-use of inner when referring to nested classes
Replaces any use (outside tests) of "inner" to refer to nested classes
in general with "nested". That is done in symbols, text and comments.
It does not fix tests as there are too many to check manually.

The `ClassItem.isInnerClass()` is used outside Metalava to refer to
actual inner classes (it qualifies the test by ensuring that it is not
static). Adding a new `isInnerClass()` method that does the same was
considered but rejected as it would silently break any existing code
that makes use of it to refer to all nested classes.

Instead the `isInnerClass()` is kept but deprecated.

Bug: 325470146
Test: ./gradlew
Change-Id: I744e678dec3fcdb769748399067e48c9485fe920