Explicitly list the math .c files we need.
Normally wildcards in external projects are fine, but given that this library is only used to implement parts of libm, and any addition here would probably require a removal there (to avoid an ODR violation), being explicit is safer. And, as we've seen in the recent "where does expf() come from?" confusion, less unclear.

Change-Id: Iffb651725a157a09e5bf15697e578503aff1321a