Remove unnecessary mutation of originallyHidden/hidden
Prior to https://r.android.com/2875675 the method
`mergeInclusionAnnotationsFromCodebase()` used a `ComparisonVisitor`,
which used an `ApiVisitor`, which used `ApiPredicate`s which checked
the `hidden` field which depended on the value of `originallyHidden`.
The result of that was that the lazily initialized `hidden` and
`originallyHidden` fields of the source codebase had been initialized
before the additional inclusion annotations had been added and so
would not reflect the addition of those annotations. As a result it
was necessary for them to be mutable so they could be updated after
merging in the annotations.

However, https://r.android.com/2875675 switched to use a much
simpler `TraversingVisitor` that does not check the `hidden` and
`originallyHidden` properties so they are no longer initialized before
the annotations are added which means that when they are initialized
they will take into account the merged annotations.

That allowed this change to remove the properties.

To ensure that this does not cause issues it was also tested in the
Android build to make sure that it did not break `m checkapi`. First,
this change was tested on its own and `m checkapi` passed. Then the
`mainItem.originallyHidden` was initialized before merging the
annotations to show that would cause failures in `m checkapi`, and
`./gradlew` which it did. That confirmed that the explanation above was
correct and that any future changes which accidentally initialize
`originallyHidden` before that call will be detected.

Bug: 350927039
Test: ./gradlew
Test: # Android build as described above.
Change-Id: I140574dad9d6c6f0f043f0986bf896eec50c4044