Fix edge case in detection of package-info.java file in Turbine
Previously, Turbine used `unit.decls().isEmpty()` to identify a
`package-info.java` but that was not quite accurate as it would also
match any empty java file and use its comment as the documentation for
the package which was incorrect.

This adds a `fun CompUnit.isPackageInfo()` extension function that
checks the source path.

Bug: 350927039
Test: ./gradlew
Change-Id: I1375c9d69049c623a1d385542773c63340bf5db0