Limit the embedded profile feature to V+.
The embedded profile feature can't work on U due to an SELinux
restriction, so this is a no-op change.

Limiting this feature to V+ instead of fixing it for U is more ideal,
because it makes it easier for app developers to expect the behavior:
the feature is available iff the device has SDK level 35, regardless
of the ART module version.

Bug: 350978283
Test: atest ArtServiceTests
Test: atest CtsCompilationTestCases
Change-Id: I21b533709b018683b859e53eb3b13e9bf09a2f93