Autoupdate of 2 submodules
*trusty/device/arm/generic-arm64 updated to c2016c195fb99eaa5154fd487f780c6568314d8f
project/qemu: Support full wipe and non-wipe reboots

The previous `reboot` and `shutdown` are always factory resets, which
stopped us from supporting composite tests which reboot and either want
RPMB wiped as well, or want no storage wiped.

To do this, set `snapshot=off` in the drives' args (so that changes will
be written back to the drives' image files) and pass qemu a path to a
copy of the image instead of the original. This way we can revert back
to the original when we want to wipe any changes.

Bug: 343502615
Change-Id: Icdbc0ccb11637d2f8d7d28362fbd888e445b5971

*trusty/vendor/google/aosp updated to 00c81a5acd31321f3931487c2651e141af7f1152
scripts: Add support for reboots that don't factory reset

Allows tests to use reboot instead of using storageproxyd restarts to
simulate non-wiping reboots. Also adds support for RPMB wiping and
printing messaged from composite tests.

Bug: 343502615
Change-Id: Ia3048cea49491b7c65ba04945d0cc94f9e441420