Autoupdate of 2 submodules
*system/libvintf updated to 9bac23386671e35bc4ce6e8f0f23c348523b9585
Add next year's release Level

This is needed for to support devices that are launching next year.
The numbers and letters used for the release are just placeholders for
right now and may be changed in the future.

The flag protects the user of this change. This change only allows the
new behavior to be supported.
Flag: build.RELEASE_AIDL_USE_UNFROZEN
Test: atest -p
Bug: 346861728

Change-Id: I6a1840b93a573d4bfa18863b12a0a4317d007d05

*test/vts-testcase/hal updated to eca82bda0c22ac552114f64a428863ac49a9f4aa
Merge "Expect next year's Android devices in VTS" into main