Fix inconsistency between Turbine and PSI in hidden packages
The previous change adds a test and explains in detail the problem this
change is fixing but basically there was an ordering issue where a
hidden package would be unhidden by a class that has a show annotation
BEFORE the package's hidden property was inherited onto a nested
package.

This change fixes that by propagating `hidden` (and similar properties)
through the package hierarchy first before visiting any of the classes.
That ensures consistent behavior between PSI and Turbine, although it
is still slightly broken.

Bug: 351118278
Test: ./gradlew
Change-Id: I02f23546c3029adc85543b7c2ad032f0bc9849bd