Add test to show problems in propagation of hidden in packages
Given a file structure like this:
* `test.a` - /** @hide */
  * `test.a.A`
  * `test.a.B` - annotated with show annotation
  * `test.a.C`
* `test.a.b`
  * `test.a.b.A`
  * `test.a.b.B` - annotated with show annotation
  * `test.a.b.C`

What `propagateHiddenRemovedAndDocOnly() is supposed to do is produce a
structure like this:
* `test.a` - not hidden as `test.a.B`'s show annotation unhides this
  * `test.a.A` - inherited hidden
  * `test.a.B` - not hidden as show annotation overrides inheritance
  * `test.a.C` - inherited hidden
* `test.a.b` - not hidden as `test.a.b.B`'s show annotation unhides this
  * `test.a.b.A` - inherited hidden
  * `test.a.b.B` - not hidden as show annotation overrides inheritance
  * `test.a.b.C` - inherited hidden

However, what it actually produces is a structure like this:
* `test.a` - not hidden as `test.a.B`'s show annotation unhides this
  * `test.a.A` - inherited hidden
  * `test.a.B` - not hidden as show annotation overrides inheritance
  * `test.a.C` - inherited hidden
* `test.a.b` - not hidden as when this inherits `test.a` is not hidden
  * `test.a.b.A` - not hidden as when this inherites `test.a.b` is not hidden
  * `test.a.b.B` - ditto
  * `test.a.b.C` - ditto

Due to some complex logic in PSI it will actually hide `test.a.b` on
creation so in PSI (but not Turbine) `test.a.b` will be hidden at the
time `test.a.b.A` inherits so it will be hidden in PSI an not Turbine.
However, in both cases `test.a.b.C` will not be hidden when it should
because at that point `test.a.b` will have been unhidden by
`test.a.b.B`'s show annotation.

The reason why `test.a.C` is hidden even though at the time it inherits
`test.a` has been unhidden by `test.a.B`'s show annotation is the
propagation will also update a class's `hidden` field if its
containing package's `originallyHidden` field is true. That is true in
the case of the `test.a` package which has `/** @hide */` but not true
of `test.a.b` as it inherited its hidden status.

Follow-up changes will fix this issue.

Bug: 351118278
Test: ./gradlew
Change-Id: I6c78ec16c5038acba0f5d093b355271f8b1ff68f