Propagate hidden from package to class before visiting classes
Previously, the propagation of hidden status from package to classes
was done at the same time as visiting the class. Unfortunately, a
class that has a show annotation will cause a previously hidden
package to be unhidden BEFORE it is propagated to other classes which
results in classes not being hidden when they should.

This change moves the propagation from package to contained classes
before the code that may unhide the package which fixes the issue.

Bug: 351118278
Test: ./gradlew
Change-Id: I723afa68c7de8848645030ead52201047419964a