Propagate docOnly/removed from package to class alongside hidden
Previously, the `hidden` property propagation from package to top level
class was moved out of the visitor into the loop iterating over
packages. This change moves the `docOnly` and `removed` properties
alongside to keep it consistent. As the `else` block the code was moved
from was only executed if `showability.show()` and `showability.hide()`
were both `false` it required an extra check. That check was wrapped
around `hidden` as well as there is no point in setting that if either
of those is `true` as the visitor will set it anyway.

Bug: 351118278
Test: ./gradlew
Change-Id: Ib433a3986e0731e217ca26721a212092f59ca65b