Add minimal ItemDocumentation abstraction and use in text model
The first of a sequence of changes to add an `ItemDocumentation`
abstraction to replace the use of `String`. This will make it easier to
make Psi and Turbine consistent in how they handle the javadoc and also
allow for future improvements, e.g. conditional tags in javadoc for
documenting flagged apis.

Bug: 350944340
Test: ./gradlew
Change-Id: I258d528134e681d166dffab18076a2f985795ad2