Add ApiVariantSelectors to contain originallyHidden/hidden for TextItem
Previously, `TextItem` would define its own `originallyHidden` and
`hidden` properties to always return `false`. This change adds an
`ApiVariantSelectors` class to contain those properties and uses that
to migrate those properties `TextItem` into `DefaultItem`. This is part
of a series of changes to improve consistency across the models, and
allow for deduping of the text and turbine model classes.

At the moment the only implementation of `ApiVariantSelectors` is an
immutable one that is used by the text model as that does not require
those properties to be used. However, follow-up changes will add an
implementation that is suitable for use in Psi and Turbine which do
require those properties are used.

In preparation for that the `ApiVariantSelectors` is created by
`DefaultItem` through a factory that is supplied rather than being
passed in directly. While that is not needed at the moment as the
factory always returns the same object, follow-up changes will need
that as they will need to create an instance that depends on the `Item`
that references it.

Bug: 351118278
Test: ./gradlew
Change-Id: Ife37a5a16def6d3dfb957cceb237d5959ea1e15d