Add workAroundJavaDocSummaryTruncationIssue() to ItemDocumentation
Previously, the code would extract `text` from `ItemDocumentation`,
check to see if it needed fixing up, and if it did would fix it up,
wrap it back in an `ItemDocumentation` and then update
`Item.documentation`.

This change moves that functionality into `ItemDocumentation` so it can
fix it up in place. In order to do that this change also needed to
create specializations of `ItemDocumentation`, one for `NONE` which is
immutable and a default one which is mutable.

This allows the `Item.documentation` to be switched from a `var` to a
`val`.

Bug: 350944340
Test: ./gradlew
Change-Id: I7983f61f422c92b187334f658086a614d8d850a7