Use a current entry pointer instead of index for the method trace buffer
For the thread local method trace buffers we used to store the pointer
to the buffer and the current index of the next free space. We compute
the address of the location to store the entry in the generated JITed
code when storing the method trace entry. Instead we could use a pointer
to the entry to avoid computing the address in the generated code. This
doesn't have a noticeable impact for the regular method tracing but is
better for the upcoming changes that will introduce an experimental
feature for on-demand method tracing.

Bug: 259258187
Test: art/test.py

Change-Id: I7e39e686bee62ee91c651a1cb3ff242470010cb6