Prepare for ItemDocumentation being mutable
Adds `duplicate()` method and calls them when creating a new `Item`
instance from an existing one. This will ensure that the same
`ItemDocumentation` is not referenced by multiple `Item` instances.

It also stops use `NONE` when converting an empty string to an
`ItemDocumentation` as just because it is empty to begin with does not
mean that it will not be modified later.

Bug: 350944340
Test: ./gradlew
Change-Id: If6fece1148b895b9b4497a7fb4ed637b1221ada9