Pull codebase property into DefaultItem
Previously, `TextItem` had a `codebase` property of type `TextCodebase`
and `TurbineItem` had one of type `TurbineBasedCodebase`. Those were
pulled up into `DefaultItem` as a type `DefaultCodebase` and the types
of the `codebase` parameter on each of the subclasses was also switched
to `DefaultCodebase`. The code did actually compile if it used the
`Codebase` interface but `DefaultCodebase` was chosen as being the most
flexible.

Bug: 351410134
Test: ./gradlew
Change-Id: Ia00d3afd91f5c5a0eda44f0bc37bcc5a33b68a85