Replace TextTypeParameterItem with DefaultTypeParameterItem
Generalizes, `TextTypeParameterItem` for use with both text and turbine
(in following change). That necessitated adding a couple of extra
parameters to cover the differences between the two. That is something
that will need to be done when creating the other `Default*Item`
classes. Rather than duplicate the model specific values of those
parameters on every creation of a `Default*Item` this adds a
`DefaultItemFactory` that allows those model specific values to be set
once.

Bug: 351410134
Test: ./gradlew
Change-Id: I650e8d495c8fa333cd96064a728b97e5b531e7d7