dev: virtio: Add mapping from vsock ports to tipc service names
The vsock-tipc bridge expects vsock connections to send the name of the
tipc service to connect to in the first packet. Add a mapping from vsock
port numbers to tipc services such that the port number identifies the
tipc service to connect to. This means we can start connecting to tipc
as soon as we get a request (rather than waiting for the first packet).

Vsock ports zero and one retains the functionality where the tipc service
name is provided in the first packet.

Bug: 298705967
Test: boot aosp_cf_x86_64_phone_trusty_vm-trunk_staging-userdebug
Change-Id: I061e50f5f5dcbea729fc9ce4fc4b9eb5406deae4