Defer creation of parameter items to TextMethodItem initialization
Previously, the list of `ParameterItem`s was created before the
`TextMethodItem` was created which meant they were created in an
incomplete state and required that their `containingMethod` reference
was updated later.

This uses a `ParameterItemsFactory` lambda to defer the creation of
that list until during the `TextMethodItem` initialization so that each
`ParameterItem` can be given a reference to its containing method when
it is created. That avoids having to modify the objects afterwards.

It reorganizes the constructot parameters to make the consistent with
the `DefaultMember` constructor and uses named arguments to prevent the
formatter from moving the `parameterItemsFactory` outside the
parentheses which makes no sense in this case.

Bug: 351410134
Test: ./gradlew
Change-Id: Ia541a118a76a591dd3d2551c667131fd846d86cd