Make MethodItem.equals(Any?) and hashCode() consistent across models
Previously, `TextMethodItem` implemented equality in terms of a
method's containing class, name, parameters (number and type), type
parameters (number and type bounds) but not return type or parameter
names. However, Psi and Turbine both implemented it in terms of the
equality of its underlying method object. The latter was incredibly
difficult to reason about because duplicated instances shared the same
underlying object which meant that equality depended on how the method
was created.

As the implementations were inconsistent they were also asymmetric
which meant that a text method might compare equal to a psi method when
using the former's `equals` method but not if using the latter's.

The `hashCode()` methods were similar.

This change follows a similar pattern to that used in `TypeItem` to
pull the `equals(Any?)` and `hashCode()` method implementations into
the `MethodItem` and define it in the same way as for text model.

Fortunately, this change has little impact on the behavior of Metalava
as other comparison functions are often used instead. However, it did
cause a problem with the `RemovedMethod` check. There was an issue with
the logic that broke the test added for https://r.android.com/853533.
Unfortunately, it was not possible to separate that from this change.

Bug: 351604449
Test: ./gradlew
Change-Id: I63dcc0d707711ed0fcd0c94fdd753c082b3a152d