Dedup check for synthetic enum methods
Synthetic enum methods are not included in the model because they add
no value. Previously, each model implemented their own check because
apart from Turbine, they had to operate on the model specific types
because they were checked before creating a `MethodItem` as a minor
optimization. This change adds a method to `MethodItem` and reuses
that instead of the model specific implementation.

That comes at the cost of Psi and Text creating a `MethodItem` only
to have it immediately discarded. The cost of that is small as enums
are not commonly use in Android due to performance issues anyway.

Bug: 351410134
Test: ./gradlew
Change-Id: Ic5c6894b55890425abd1f562f5708c836d15e21f