Make PackageItem.containingPackage() consistent across models
Previously, `PsiPackageItem` would find the closest enclosing,
non-empty, package. It does that by searching up through the package
hierarchy until it finds a package in the `Codebase`. However, Turbine
and Text would both return the result of looking for the immediately
enclosing package. Psi and Turbine would both cache the result but
Text did not.

This makes them all behave as `PsiPackageItem`.

Bug: 351410134
Test: ./gradlew
Change-Id: I2aeb75cbfc697806e49d4444370c0d9f6c0471ef