Add PublicNameProvider abstraction
Previously, `TextParameterItem` was supplied with a `publicName` at
initialization because the name, if any, in the signature file is the
public name. However, the `TurbineParameterItem` would find it on
demand by checking for a `ParameterName` annotation of some sort. This
change unifies the two by adding a `PublicNameProvider` abstraction
that supports both cases.

Bug: 351410134
Test: ./gradlew
Change-Id: I873c94bda17b2e5f72d9eb1fdfe14c168bda1bc5