Make PsiMethodItem.duplicate(...) consistent with other models
Previously, it was implemented by reconstructing a new `PsiMethodItem`
from the underlying `PsiMethod` object. That prevented reuse of
existing model objects and was inconsistent with the other models which
copied the `MethodItem` implementation directly. This change switches
to be consistent with the other models. While the resulting code is a
little more verbose it should be simpler to understand and more
efficient.

Bug: 351410134
Test: ./gradlew
Change-Id: I8cc30e906e8492f5bfb47e2e3b1bdd2190c38784