Defer creation of parameter items to PsiMethodItem initialization
Previously, the list of `ParameterItem`s was created before the
`PsiMethodItem` was created which meant they were created in an
incomplete state and required that their `containingMethod` reference
was updated later.

This uses a `ParameterItemsFactory` lambda to defer the creation of
that list until during the `PsiMethodItem` initialization so that each
`ParameterItem` can be given a reference to its containing method when
it is created. That avoids having to modify the objects afterwards.

This implemented the `PsiParameterItem.duplicate(...)` (replacing the
existing `create(...)` method that implemented basically the same
function).

Bug: 351410134
Test: ./gradlew
Change-Id: I4b4dbe68853c978c90809e020dda54cfef23a815