Set hidden, etc. in inheritMethodFromNonApiAncestor(...)
Previously, the `inheritMethodFromNonApiAncestor(...)` implementations
did not set the `hidden`, `docOnly` and `removed` properties on the
inherited method based on its new `containingClass` as is done in
`MethodItem.duplicate(...)` implementations. This corrects that to
allow a follow-up change to implement it in terms of `duplicate(...)`.

Generally, this should be a no-op as this method should only be called
to inherit a method from a hidden class into a non-hidden class. Even
if it was not, marking a method as hidden if it is in a hidden class
seems like the correct thing to do. Ditto with the other two properties.

It was necessary as part of this to fix a test but it is not clear that
the test is particularly realistic. However, it was left as it is
because it is still providing test coverage.

Bug: 351410134
Test: ./gradlew
Change-Id: I0a097b34369b7f2fdcab1b64f05b48301b99a153