Generate outer stub TextClassItem before inner stub class
Previously, when the `TextCodebase` needed to create a stub
`TextClassItem` for a class that could not be it would first create the
innermost class and then its containing class. That order is opposite
to the order they are created elsewhere in the text model and other
models. This change switches the order to allow a follow-up change to
supply the `containingClass` to the `TextClassItem` constructor instead
of setting it later.

Bug: 351410134
Test: ./gradlew
Change-Id: Ib5fe1fcb16cd1a58128695f13bda529d4e061890