Offload `onStartJob` and `onStopJob` calls from the main thread.
The job scheduler calls `onStartJob` and `onStopJob` on system_server's
main thread. Our handlers can take long, to synchronize on the lock and
to wait for cancellation to be done, causing device freezing.

This change is a low-risk change that offloads the calls from the main
thread while keeping the execution order as the main thread does.

Bug: 350999413
Test: atest ArtServiceTests:com.android.server.art.PreRebootDexoptJobTest --iterations 10
Test: Run Pre-reboot Dexopt and cancel it.
Change-Id: I270de362f69a1f89e843e219f76b2e1f8b5bda61