Move nestedClasses into DefaultClassItem
Previously, Turbine set it to an `internal lateinit var` of type
`List<TurbineClassItem>` and set the whole list at once, while Text set
it to a `private val` of type `MutableList<ClassItem>` and added
classes to it one at a time using `addNestedClass(...)`.

This change basically copies the Text approach and fixes Turbine to use
that.

Bug: 351410134
Test: ./gradlew
Change-Id: I3e41692dd18f7b3fc3a557318bbfdf8bb0cd734b