Autoupdate of 2 submodules
*external/googletest updated to 9f835bb95a850a60125ce533e53f8ee78d408882
Sync gtest tests with BUILD.bazel

test_per_src is being removed, it never worked well and is incompatible
with all of the test infrastructure.  Switch to a test structure that
matches with how the tests are defined in BUILD.bazel.

Test: atest
Change-Id: I95efe8e0ac9f40e2785de298df666e935ac3e732

*prebuilts/build-tools updated to ccbf6ad6af4430b013ddf35285f532f0bf535be9
Update path to ckati find_test

The path to ckati's find_test is changing due to the removal of
test_per_src, update the path in build-prebuilts.sh to the new
location.

Test: prebuilts/build-tools/build-prebuilts.sh
Change-Id: Ibdb115c98ab43b77bcb7a4cbae221115297713f6