Correct documentation for signature-to-jdiff --strip option
The original code that implemented "stripping" did not explain what it
meant and so its intended behavior had to be determined from what it
did. Unfortunately, that determination was wrong, it does not elide
methods, it elides references to types not defined in the input
signature file. This change corrects that.

Bug: 351410134
Test: ./gradlew
Change-Id: Ia39ca5bb2e3a4305922926ab3123cff1754c4260