Extract findConstructor(String) from findMethod(String, String)
Previously, the `findMethod(String, String)` would search for
constructors if the first parameter, the method name was the simple
name of the `ClassItem`. However, it returns `MethodItem` for both
and in future `ConstructorItem` will not extend `MethodItem` so that
will not work. This change extracts a `findConstructor(String)` method
that can be used instead.

The `AnnotationsMerger` actually relied on `findMethod(...)` to return
both constructors and methods so it was necessary to add
`findCallable(String, String)` that calls either of the other two
depending on the method name.

Bug: 352481366
Test: ./gradlew
Change-Id: I95d9294a20bea0e4087b7c54f9ecff8b2c6798fe