Autoupdate of 4 submodules
*frameworks/base updated to d7a24e9a89bf02f357ef823cf06421321c1210bd
Merge "fix dependency on modules-utils-build" into main
*packages/modules/Bluetooth updated to c8d0714669430301fd0960d6ef01c3c12a45a69e
Merge "BluetoothManager: override context deviceId" into main
*packages/modules/Connectivity updated to 63b1d4561c267e9e569f177b7ce4b0efcdea6d9b
Merge "fix dependency on modules-utils-build" into main
*packages/modules/Wifi updated to 7a61a15b58e6002de124bca2ea297cc26432abe9
fix dependency on modules-utils-build

net-utils-services-common used to include modules-utils-build as a
static lib, but this has the issues of forcing lib depending on it to
also have the modules-utils-build.
When trying to add this lib to a framework library within the same apex,
it fail the StrictJavaPackagesTest.
By moving the responsibilities to link it to the caller we no longer
have the initial issue

Bug: 343121936
Test: Build target using net-utils-services-common
Test: atest StrictJavaPackagesTest
Flag: Exempt simple dependency fix
Change-Id: Ic198d522bb87aff78e12d4dfb08351daa521b522