Add Report data class to represent a reported issue
This change simplifies the reporting logic in `DefaultReporter` and is
useful for follow-up changes.

The `FileLocation` is relativized into a `String` before creating the
`Report` as that requires knowledge the `DefaultReporter` has and
follow-up changes will pass the `Report` outside the `DefaultReporter`.
Similarly, the `line` is kept separately because follow-up changes will
need to sort on that.

Bug: 352868158
Test: ./gradlew
Change-Id: I7049c392b17b993127b805ce77cdfa147fbc9e4a