Save Reports away in a list in DefaultReporter
It is necessary to record all the `Report`s that are reported in order
to sort them. Instead of writing them out immediately they are stored
away and only output when `writeSavedReports()` is called. That made it
necessary to call that explicitly in a couple of tests that checked the
written output.

Bug: 352868158
Test: ./gradlew
Change-Id: I2fa58e4ac09a3546e179d0034fbf2973ebfc930a