Change BaseItemVisitor.visitConstructorsAsMethods default
In order to allow `ConstructorItem` to extend `CallableItem` instead of
`MethodItem` it is necessary to switch visitor implementations that
currently set `visitConstructorsAsMethods = true` and rely on
`MethodItem` representing methods and constructors, to set
`visitConstructorsAsMethods = false` and use `CallableItem` instead.

This change makes it easy to see which visitors need work by changing
the default of `visitConstructorsAsMethods` from `true` to `false` and
removing any setting of `visitConstructorsAsMethods = false` as
unnecessary. As each visitor is switched the setting of
`visitConstructorsAsMethods = true` will be removed and once all uses
have been removed then `visitConstructorsAsMethods` will be removed.

Those visitor implementations that do not override `visitMethod(...)`
or only use it for accessing methods not constructors are not affected
by `visitConstructorsAsMethods` and so they are untouched by this
change.

Bug: 352481366
Test: ./gradlew
Change-Id: I666957f3b36c0df5be3800c44ce469200ab1cab0