ClassItem.findMethod(); Return methods not constructors
Previously, it returned `MethodItem` so could return both constructors
and methods. However, that will not be true in future so this change
stops it from returning constructors. It seems as though that behavior
was not actually relied upon anyway as no code seems to call it with
a `ConstructorItem` as the `template` property and no tests failed.

Bug: 352481366
Test: ./gradlew
Change-Id: Icd7ff301ac164f047cf37296868a7897aa20faf3