ApiLint: Switch method and constructor checks to CallableItem
Previously, checks that applied to methods and constructors used
`MethodItem` as that is currently the super type of `ConstructorItem`
but in future it will not be and instead they will both be
`CallableItem`. This change switches those checks, i.e. the ones called
from `visitCallable(CallableItem)`, to use `CallableItem` instead of
`MethodItem`.

This is quite a large change but it is pretty simple type change and
variable renames, with a couple of extra type checks to handle the
differences between `MethodItem` and `CallableItem`.

Bug: 352481366
Test: ./gradlew
Change-Id: Id28929e00e213fdcf05523a1d4df4bd186270125