API Writer classes: Differentiate MethodItem/ConstructorItem
Previously, `MethodItem` was used to represent both methods and
constructors, however that will not be possible in future. So, this
change makes `*Writer` classes treat them differently and use
`CallableItem` for anything that needs to handle both.

`FilteringApiVisitor` and `ModifierListWriter` changes were included
in this as they participate in the writing of the different API
formats.

Bug: 352481366
Test: ./gradlew
Change-Id: Iaa1bde85af108fb01de3f0d99767a3ba06d34187