OGKI: Add config for approved OGKI builds
This config will act as the source of truth for the OGKI builds which
have been approved. The config is indexed on Android branches and ids
for the builds.

The build ids are SHA256 hashes of `uname -r` for the approved builds.

There are plans to add more information about the builds to this config
in the future.

Bug: 342094847
Bug: 352795077
Test: adding config, no test required
Change-Id: Ia759c0a563354b533ed4feff3bfde8dd495cfb67