Add ItemDocumentationFactory and use in AbstractItem
The Psi model has the most support for handling item documentation and
one of the things that it does is report issues with it. In order to do
that it needs to have an `Item` against which the issues can be
reported. At the moment the `ItemDocumentation` classes cannot have a
reference to the `Item` because they are created before the `Item`.
This change adds `ItemDocumentationFactory` to defer the creation of
the `ItemDocumentation` until after the `Item` has been created, during
its initialization.

Bug: 350944340
Test: ./gradlew
Change-Id: If68a6cdb4f1ded5458f9c895600a594d07e0ff90