Switch psi model over to using ItemDocumentationFactory
As part of this it was necessary to add `Item` to
`ItemDocumentation.duplicate()` as that needs to create a copy of the
`ItemDocumentation` for use in the duplicated `Item`. That meant that
in order to separate this out from similar changes to the default model
it was necessary to pass `this` to the calls to `duplicate(Item)` made
within the default model. That is not strictly correct but is safe as
currently `DefaultItemDocumentation.duplicate(Item)` does nothing with
the `Item`. A follow-up change will clean that up when it switches the
default model to use ItemDocumentationFactory.

Bug: 350944340
Test: ./gradlew
Change-Id: I18299cf609b7ae0feecbe5f2d7be284304921fa5