PsiItemDocumentation: Store PsiItem instead of PsiBasedCodebase
Previously, it would store the `PsiBasedCodebase`. The problem with
that is that when it is duplicated it stores the `PsiBasedCodebase` of
the `PsiItemDocumentation` that is being duplicated and NOT the
`PsiBasedCodebase` for the new `PsiItem` to which it refers. This
change fixes that by storing the `PsiItem` for which it was created and
retrieving the `PsiBasedCodebase`.

Bug: 350944340
Test: ./gradlew
Change-Id: Icfc463fc2364366dc99eec1350d9b9a6fd3464bd