Select appropriate ItemDocumentationFactory for PsiElement
Previously, the `javadocAsItemDocumentationFactory` was documented to
return `ItemDocumentation.NONE` when `allowReadingComments=false` to
optimize that behavior. However, it did not actually do that and even
if it did it would have been wrong.

This change does optimize the path while ensuring that package hiding
(i.e. adding `@hide` in the javadoc of a `package-info.java` file)
still works.

Bug: 350944340
Test: ./gradlew
Change-Id: Idf4de9df0a8413dc312f3d4235bcabef3004660f