Move findThrownExceptions into new CallableBody
Neither turbine or text models provide access to callable bodies and
so the `DefaultCallableItem` always returns `CallableBody.UNAVAILABLE`.
The psi model does provide access to them so it returns its own custom
implementation.

The existing `CallableItem.findThrownExceptions()` method is moved to
the new `CallableBody` implementation as that is a better place for it
to be.

Bug: 353421288
Test: ./gradlew
Change-Id: I192ca4057e5d373e832ededde00261ef16ae00bd