Refactor checkSynchronized to simply extraction of psi specific code
Previously, the psi specific code for scanning the method body was
mixed with the model independenct code for handling the reporting. This
change splits them apart to make it easier to move the psi specific
code to `PsiCallableBody`.

The model independent and psi specific code both produce a list of
`FileLocation`s for where the synchronization is visible outside the
method. That is then used to report the issue.

Bug: 353421288
Test: ./gradlew
Change-Id: I7430066eb1407e5f5b86f04023f1e3430ee4f785