ApiLint: Start differentiating MethodItem/ConstructorItem
Previously, `MethodItem` was used to represent both methods and
constructors, however that will not be possible in future. So, this
change starts the process of making `ApiLint` treat them differently
and use `CallableItem` for anything that needs to handle both.

Removes `visitConstructorsAsMethods = true` and switches from
`visitMethod(MethodItem)` to `visitCallable(CallableItem)`. It casts
down to `MethodItem` where needed to make it compile. This is a
temporary measure to allow `ApiLint` to be migrated in multiple steps
as it is too big to do in one change.

Bug: 352481366
Test: ./gradlew
Change-Id: Iec10c4385903bb8eab669411292834faac362197