Add Codebase.reporter
Makes Turbine consistent with Psi as that already has a `reporter` in
its `PsiBasedCodebase` class. This is needed for follow-up changes but
will be generally useful to improve consistency of handling errors.

The `TextCodebase` currently throws an error if `reporter` is accessed
as it does not have one. That should not be an issue for the follow-up
changes as they do not affect a `TextCodebase` loaded from a signature
file.

Bug: 351118278
Test: ./gradlew
Change-Id: Iae322d92f4a82697ce54f9584b892e0d773881ef