ApiAnalyzer: Separate ensureParentVisible(...) from propagation
Previously, the `ensureParentVisit(...)` method call was mixed in with
the propagation of the `hidden` property. This separates it out so it
runs after the propagation has been completed. It does mean that it is
doing slightly more work as it is repeating some of the same work done
by the propagation, albeit accessing information that has been cached.
However, it does allow the propagation to be moved to
`ApiVariantSelectors` which will allow further enhancements.

Bug: 351118278
Test: ./gradlew
Change-Id: Ibbbeea8b2eaf9ec61838ad6d566adc031325c732