Stop ApiVariantSelectors.Mutable extending ItemVisitor
Previously, it used `ItemVisitor` to dispatch to the correct type to
minimize the changes needed when moving the code from `ApiAnalyzer`'s
`propagateHiddenRemovedAndDocOnly()` method. However, that is not
necessary and limits the ability to optimize and dedup.

This change replaces the use of `ItemVisitor` with a simple type
checking `when` statement that dispatches to a type specific method.
Although the methods are specific to a particular type they do not
take an appropriately typed parameter. Instead they use smart cast to
treat the `item` property as the expected type within the body.

Follow-up changes will merge those type specific methods together at
which point the smart cast will no longer be needed. In preparation for
that a minor refactoring change was made to `inheritIntoClass()` to
make it more similar to `inheritIntoCallable()`.

Bug: 351118278
Test: ./gradlew
Change-Id: I1059d3dac6210aadaea593435c8d3f1d844400fb