Remove unnecessary redirection through Item
The `item` property in `ApiVariantsSelectors` refers to an `Item` whose
`variantSelectors` property refers back to `item`. So, `item.hidden` in
`ApiVariantsSelectors.inheritInto()` delegates to `this.hidden`. This
change removes that unnecessary delegation by replacing `item.hidden`
with `hidden`. It also makes similar changes for `docOnly`, `removed`
and `showability`.

The few other places in `ApiVariantsSelector` and `ApiAnalyzer`'s
`propagateHiddenRemovedAndDocOnly()` method that access the properties
through `Item` have also been change to access them from
`ApiVariantSelectors` directly.

This is needed for a follow-up change that will add a new property
to `ApiVariantsSelectors` and doing this avoids having to add a
delegate for that property to `Item`.

Bug: 351118278
Test: ./gradlew
Change-Id: I937681af4ade09b71ceb7239e1d0a7bbcabea7c5