Move showability caching from Item to ApiVariantSelectors
Previously, `showability` was lazily initialized on every `Item`. That
had two issues:
1. It meant that `Item`s loaded from a signature file had to keep track
   of and initialize it when it was accessed despite signature files
   not containing any show annotations.
2. The `lazy {...}` has quite a high overhead as it synchronizes access
   even though it is not multi-threaded.

This change moves the storage of `showability` into
`ApiVariantSelectors`. The `Immutable` implementation (which is used by
the text model) always returns `Showability.NO_EFFECT`. The `Mutable`
implementation caches it as a `lateinit var` saving time and storage.

The main reason for moving it is that is really where it belongs,
alongside the other similar properties.

Bug: 351118278
Test: ./gradlew
Change-Id: Icda193d2b840ee803909222e9eeed5d62e3d9d45