Add ApiVariantSelectors.inheritableHidden
Previously, the `hidden` property fulfilled two related but separate
purposes:
1. Determining whether an `Item` is `hidden.
2. Determining whether an `Item`'s contents are hidden.

Unfortunately, that did not work well in two cases:
1. When a shown top level class unhid its containing package it created
   an ordering issue that required two passes, see
   https://r.android.com/3159907.

2. When a class had a `--show-single-annotation` it set
   `hidden = false` but to prevent that from inheriting down to its
   contents (which is not supposed to happen) there needed to be an
   extra check in those contents to inherit from `originallyHidden`.

This change adds an `inheritableHidden` property that is basically
identical to what `hidden` was, including its initializer. The `hidden`
property's initializer is changed to use the `inheritableHidden` value.

The `inheritableHidden` property replaces `hidden` in almost all the
inheritance related code with one exception and that is to fix case #1..
It was necessary to fix it as part of this change because without it
some tests break.

Case #2 will be fixed in a follow-up change.

Bug: 352481366
Test: ./gradlew
Change-Id: Ica212481c1d9129712f816f02a9c85f385c8d07f