Autoupdate of 2 submodules
*packages/modules/Virtualization updated to 6734b0da9fe31865b910eaa5c83e9645be33d04e
Drop system shared uid in ferrochrome app

it doesn't need to use that shared uid and platform signing key.
and also remove unnecessary sleep logic.

but still need to remove sysprop set to move the app to product
partition.

Bug: 348113995
Test: run ferrochrome app
Change-Id: I5fcac560ab91eda8124f55685084487c6ebb1732

*system/sepolicy updated to e96a52c34a06830ac896d4fc8fdd9b236e0f9aba
Define sepolicy for ferrocrhome app

it needs to use debug sysprop for a while

Bug: 348113995
Test: run ferrochrome app
Change-Id: Ibc518c71a4533ad846514af02cbf5bffb9916d81