Autoupdate of 2 submodules
*art updated to 1fe9dc85a7111f7e678c47ed48b2611b536b9dca
Add stubs to libdexfiled, to make the build system add it to
provideNativeLibs in the debug APEX.

This fixes an issue when the debug APEX com.google.art.debug is in use,
typically in `eng` builds: In that case external (i.e. non-ART) users
of libdexfile_support load libdexfile.so instead of the correct
libdexfiled.so, leading to crashes.

Test: lunch aosp_cf_x86_64_phone-trunk_staging-eng
      Build and boot
      adb shell cat /linkerconfig/ld.config.txt
  Verify that com_android_art.shared_libs lines list libdexfiled.so
Bug: 349767294
Change-Id: I042b6126b082009346bfd576d98179857666971b

*build/soong updated to 6485cf11dc9b83d110815f285ac4287cb7fa6497
Merge "Add libdexfiled to the stub library multi APEX exception list." into main