Move package -> top level class inheritance into ApiVariantSelectors
Previously, it was done in a separate pass in
`ApiAnalyzer.propagateHiddenRemovedAndDocOnly()` to work around the
ordering issue tested and documented in https://r.android.com/3159906.
However, that issue has now been resolved so it can be integrated back
in with the other inheritance code, albeit in `ApiVariantSelectors`
instead of `ApiAnalyzer`.

Bug: 351118278
Test: ./gradlew
Change-Id: I8809f097df6702fb1719a597563f53a135782776