Inherit properties automatically in ApiVariantSelectors
Previously, the properties were inherited only at one point in the
processing by `propagateHiddenRemovedAndDocOnly()`. That lead to later
stages of the processing, e.g. `duplicate(...)` methods having to fix
up the properties to their correct state because the `Item`s were
created after the inheritance was done. This change does the
inheritance automatically when the properties are accessed.

Note: It is still necessary for `propagateHiddenRemovedAndDocOnly()`
to traverse over all the `Item`s and call `inheritInto` because calling
`inheritInfo` on a `ClassItem` can unhide the containing `PackageItem`.
Without that code that checks the `PackageItem`s `hidden` property
before that `ClassItem` would think that the `PackageItem` was hidden
when it was not.

This change had to correct the expected values of the properties in a
number of testsuite tests because they did not inherit them correctly
because `propagateHiddenRemovedAndDocOnly()` was not run because it is
in the `metalava` project so is inaccessible to the
`metalava-model-testsuite` project.

Follow-up changes will clean up those areas which can be simplified
after this.

Bug: 351118278
Test: ./gradlew
Change-Id: Iebc8404bc1b61cecb1a136ef8d9a82915fc6af56