[scudo] Add static vector functionality. (#98986)
The scudo vector implementation maintains static local data before
switching to dynamically allocated data as the array size grows.
Users of the vector must now specify the size of the static local data
through the vector template (the default size has been removed).
If 0 is specified for the size of the static local data, an assertion
will
be triggered.

GitOrigin-RevId: d08527ee3ee2dc1e90d2afcc6e5982d0997dad20
Change-Id: I66e9b836b8120570381c532b8873ef4cab46631f