Trim duplicate function symbols.
In the symbol table, there can be multiple functions at the same
offset. There is no need to keep these duplicate symbols since the
unwinder will only ever recognize one of them.

Therefore, keep only a single function for each offset, and choose the
symbol with the longest function size and the shortest name to save as
much space as possible.

Also, reorganize the code slightly so that only functions from the
dynamic symbol table are recorded and their offsets. This is now
used to skip any symbols at the same offset instead of using the
symbol name.

NOTE: The auto-formatter moved the include files around.

Test: Ran with the old algorithm and new algorithm and verified there
Test: are no duplicates at the same offset.
Test: Ran the unwind_symbols tool on these libraries and verified
Test: before the change and after the change there are the same number
Test: of symbols found.
Test: Built the entire tree and observed many system libraries are
Test: smaller in size.
Change-Id: I0c772fe2dfd335b971a9b58e75b6ea4a26d0d116