Remove unnecessary mutations of ApiVariantSelectors properties
Previously, the `duplicate(...)` methods mutated `ApiVariantSelectors`
properties to ensure that inherited their values correctly. However,
that is no longer necessary now that they are inherited automatically
when accessed. This change removes those mutations, simplify the
methods and stop the properties from being externally mutable.

There was one case where the package's `hidden` property needed to be
mutated. That required a cast. Although it could theoretically fail,
e.g. if a `PsiClassItem` was added to a `DefaultPackageItem` from a
signature file that is not supported. The previous code would also have
failed in that case just with a slightly different error so it does not
change the behavior. The better encapsulation and easier reasoning that
making them immutable provides is worth that minor yuckiness.

Bug: 351118278
Test: ./gradlew
Change-Id: I475618059b15d73160585f1ebd67143860a55370