Run snapshot tests against Turbine model as well
Previously, the snapshot tests just ran against a text model. This
change extends it to work against a turbine model as well. Turbine was
chosen instead of Psi because while it is slightly less capable then
Psi it is MUCH faster and that will improve development speed.

This change had to make a number of changes to make this work:
1. Add a dependency onto `metalava-model-turbine` to get the Turbine
   `SourceModelProvider` service.
2. Add a dependency onto test fixtures for `metalava-model-source` to
   get the `SourceModelSuiteRunner` which runs tests against any
   provided `SourceModelProvider` services.
3. Modify `SourceModelSuiteRunner` to apply the `CodebaseTransformer`
   which needed a dependency on `metalava-model` test fixtures.
4. Modifying `ModelTestSuiteRunner` to support multiple
   `ModelSuiteRunner` making sure to include provide name in the test
   name to differentiate the same test being run against different
   providers.

The latter change is why the baseline file was completely rewritten
to add `text` to the existing lines when it was refreshed.

Bug: 353737744
Test: ./gradlew
Change-Id: I0f8fcf264933e1f07d1458fa979e5d2f2945b811