Add CategoryRedirectingReporter for use by Codebase.reporter
Recent refactorings have moved some model specific code behind some
new abstractions, e.g. `ItemDocumentation`. That code sometimes needs
to report issues that need to be targetd to a specific `Reporter`.
That is needed because the API Lint `Reporter` will output a message to
explain how to fix issues by baselining or using `@SuppressWarnings`
and that will not work if the issue is an `InvalidSyntax` or
`UnresolvedImport` error.

However, tunneling at least two `Reporter` instances through to the
`Codebase` seems like an unnecessary complication so this change takes
advantage of `Issue`s have a `Category` that correlates with the target
`Reporter` to wrap the 3 reporters and redirect the report to the
appropriate one based on the reported `Issue`s `Category`.

At the moment this is just passed to `createSourceParser(Reporter, ...)`
but in future it will most likely be used everywhere to avoid having to
worry about which one to use and stop passing 3 `Reporter`s around.

Bug: 351118278
Test: ./gradlew
Change-Id: Ia8f3b89acd8e9eb8d78f3c2db8977740f8683c49