Add basic tests for ApiVariantSelectors
Previously, tests were not added for `ApiVariantSelectors` and instead
it relied on the end-to-end tests to exercise its behavior. That was
largely because it was not clear what `ApiVariantSelectors` would need
to do and unit tests are more sensitive to refactorings so time would
be spent maintaining them as it was refactored.

However, now that it is clearer how it will behave it makes sense to
add some as it is difficult to test all the low level behavior, e.g.
making sure that lazy properties are not initialized unless needed
which is important.

So, this change adds some basic tests for it. During testing a bug was
found in the `toString()` method because it was using the bit position,
i.e. bit 1, bit 4 etc. instead of the mask `0b001`, `0b100` etc. This
change fixes that and also breaks the properties onto separate lines as
it was becoming difficult to see the differences when a test did fail.

Additional tests will be added as needed.

Bug: 351118278
Test: ./gradlew
Change-Id: I2fbd15a1f08197975985fd43602edf5b9191f380