Stop the metrics thread in teardown in MetricsReporterTest
We have observed a heap-use-after-free in ASAN tests, and the root cause
has been found in a race condition for which the background thread that
reports metrics can try to update a member variable within the
ArtMetrics object which has already been freed due to the object
destruction ordering.

This change stops the background thread explicitly as part of the test
teardown phase so the race condition is avoided.

Bug: 344766711
Test: TARGET_PRODUCT=armv8 TARGET_RELEASE=trunk_staging TARGET_BUILD_VARIANT=eng ./prebuilts/build-tools/path/linux-x86/python3 ./art/test/testrunner/run_build_test_target.py -j80 art-gtest-heap-poisoning
Change-Id: I17c5ee2937d6a2898c0808a348ef5f338619764a