Pass containingPackage in DefaultClassItem constructor
Prevents it from being forgotten about, like it was in snapshot code.
It does have a slight change in behavior because Turbine was only
setting it for top level classes so nested classes had to check their
containing class. However, this means it is also set for nested
classes. That allowed the `containingPackage()` method to be
simplified.

Bug: 353737744
Test: ./gradlew
Change-Id: Ia987aa5efe917120285990d8d4203e44bf4f6349