Simplify docOnly handling in ApiPredicate
There is no need to traverse up the hierarchy to find if any
containing `Item` is marked as `@doconly` as that work has already been
done. This change removes that logic and returns `false` immediately if
`docOnly` `Item`s are excluded and the `Item` is `docOnly`. It returns
after other early exits to avoid changing behavior as some of them can
return `true`.

Bug: 351118278
Test: ./gradlew
Change-Id: Id3f4bc33898360ee27e14b3794969ac520846ee7