Separate docOnly inheritance from hidden
The `docOnly` property is only used by `ApiPredicate` and it will
traverse up the class nesting hierarchy OR-ing the value at each level
to get the actual value it uses. That is slightly different behavior
to that implemented in `ApiVariantSelectors` which will not inherit
it if the `Item` has a show or hide annotation.

The `ApiPredicate` behavior is what wins out and it also makes the most
sense. It should be possible to use `@doconly` on an `Item` with a
show annotation (e.g. `@SystemApi`) and have it be included in docs but
not stubs or signature files.

So, this change makes the `ApiVariantSelectors` behave like
`ApiPredicate` by handling it completely separately to the other
properties in `inheritInto()`, moving the logic into its
`initialValueProvider`. That dedups the inheritance logic, makes it
simpler and cleaner and avoids having to update the other properties.

Bug: 351118278
Test: ./gradlew
Change-Id: Ibe180d666bdd8a0ed61f4a1cad378b9ef6761490