ui: Improve Wattson by process/package runtime
Since the prerequisite tables to generate the Wattson by process and
Wattson by package UI are nearly identical to the tables generated by
Wattson by thread, create the Wattson by thread tables with the
additional data that process/package need, and then skip re-creating the
tables in the process/package step. This is based on the assumption that
iterating through a Typescript array will be a FIFO, and that the
developer-defined push onto the array will stay consistent (thread
before process/package).

In a 45min trace when selecting over the entire trace, the Wattson by
process tab is produced instantaneously after the Wattson by thread,
shortening the time from 21s -> 15s.

UI Profiling (cumulative sum):
           before baseline -> .. -> prev -> curr
  CPU by thread:       ~7s -> .. ->  ~6s ->  ~6s
  CPU by process:     ~10s -> .. ->  ~7s ->  ~7s
  Wattson by thread:  ~16s -> .. -> ~14s -> ~15s
  Wattson by process: ~30s -> .. -> ~21s -> ~15s

Bug: 354256369
Change-Id: I9508c1f8615c66aad47aeabbe5046ffccb19c87b
Signed-off-by: Samuel Wu <wusamuel@google.com>