Add ApiVariantSelectors.accessible
The `ApiPredicate` (and a number of other places in the code which are
not affected by this change but will be cleaned up later) check the
accessibility (called `visible` in `ApiPredicate`) of `Item`s, i.e.
they make sure it is `public`, `protected` or `internal` with a show
annotation. `ApiPredicate` will traverse up the class nesting hierarchy
to make sure that an `Item` is accessible every time it is called.

This change moves that check behind an `ApiVariantSelectors.accessible`
property which caches it and avoids having to repeatedly traverse the
class nesting hierarchy for it.

Bug: 351118278
Test: ./gradlew
Change-Id: I6510ceddb44ddd9c9046f4290331e0dbe637035a