Clean up tests of ApiVariantSelectors properties
The checks of the properties in the `duplicate()` methods were removed
as they add no value. The `duplicate()` methods no longer have any code
to set them (and cannot as they are immutable from outside). The
behavior on the duplicated method or field is identical to that of a
non-duplicated method or field.

The other tests use `ApiVariantSelectors` directly instead of the
delegates in `Item` as the intent is that they will be removed.

Bug: 351118278
Test: ./gradlew
Change-Id: I8c7c7a34aa2e1bf026719026315bfbb1e1ed2799