Remove cast to SourceCodebase that was breaking snapshot tests
Previously, a few snapshot tests were breaking because it attempted to
cast a `DefaultCodebase` used by snapshot to a `SourceCodebase`. As
`SourceCodebase` only had one method (which is not used by any of the
tests) the cast could just be removed.

Follow-up changes will pull the `SourceCodebase` method up into
`Codebase` and remove `SourceCodebase` altogether.

Bug: 353737744
Test: ./gradlew
Change-Id: Ia684489e926427f05af84b5c435e4712ca3a7524