Dedup creation of DefaultTypeParameterList
Previously, all the models would call `DefaultTypeParameterList`s
`createTypeParameterItemsAndFactory(...)` method to create a
`TypeParametersAndFactory` and extract the `List<TypeParameterItem`
and `...TypeItemFactory` and wrap the former in a
`DefaultTypeParameterList` and then pair it with the latter.

They did that because at the time each model had their own
implementation of `TypeParameterList` but now they all use the same
`DefaultTypeParameterList` so this change pushes the creation of
that into `createTypeParameterItemsAndFactory(...)` avoiding the
need to unwrap, rewrap and pair.

As `TypeParameterListAndFactory` is used in a few more places it was
moved to the top level to simplify the importing.

Bug: 353737744
Test: ./gradlew
Change-Id: I4c7a31c01943e19a9f19a64e63311bff34af3978