Stop Type/ParameterItem using ApiVariantSelectors.Mutable
The `ApiVariantSelectors.Mutable` class supports tracking and updating
the properties like `hidden` and `removed` for an `Item`. That is
pointless for `TypeParameterItem` and `ParameterItem` as whether they
are `hidden` or `removed` is determined solely by their owner. This
change forces them to always use the `Immutable` object instead and
changes `Mutable.item` from `Item` to `SelectableItem` to enforce that.

Bug: 351118278
Test: ./gradlew
Change-Id: I7e26d865e7a73935e8bffbb9b3f46961609beeb5