Fix NPE in Turbine in annotation with implicit value attribute
Previously, an instance of an annotation class with a `value` attribute
that did not specify the attribute name caused a NPE which broke an
annotation merger test. This change fixes that.

Although, the issue was a problem with annotations in general it was
found during, and is being fixed as part of, work on default values.

Bug: 353737744
Test: ./gradlew
Change-Id: If021c6dc411f3a21832394096b64489ad6fd642b