Move immediate from AssemblerRiscV to separate riscv namespace.
The main advantage is that this makes immediates from different
assembler types compatible.

Most instructions and algorithsm don't even need to know the exact type
of RISC V implementation used.

You still need to know that ShiftImmediate is different on RV32 and
RV64, but that's difference in hardware it couldn't be hidden.

And even there handling of Shift32/Shift64 is assembler-independent, the
actual difference is that RV32 maps ShiftImmediate to Shift32Immediate,
while RV64 maps ShiftImmediate to Shift64Immediate.

Bug: http://b/232598137

Test: m berberis_all

Change-Id: I8573138088e9fc7f3595734861d5c57294de07eb