Combine Text/TurbineFieldValue together into FixedFieldValue
Previously, both the `TextFieldValue` and `TurbineFieldValue` were just
wrappers around some fixed value(s), 1 for the former (for both
constant and non-constant values), 2 for the latter (one each for
constant and non-constant values). This change combines them into a
single `FixedFieldValue`. That will be used in a follow-up change to
contain a snapshot of a `FieldValue`.

Bug: 353737744
Test: ./gradlew
Change-Id: Idf48b829020b34ab4183faa32c333edfad49e428