ApiPredicate: stop checking selectability of non-selectable items
Previously, `ApiPredicate.test(Item)` recognized that there was no
point in running the gamut of the selection algorithm on
`TypeParameterItem`s because they were not individually selectable.
They would be selected if and only if their owner (class, callable) was
selected. So, `ApiPredicate` returned `true` immediately. However, it
did not do the same for `ParameterItem` even though the exact same
conditions apply to them.

This change modifies the condition to check for items that are not
`SelectableItem`s which treats them both the same.

Bug: 351118278
Test: ./gradlew
Change-Id: Ib0517d74d0d3667f86af84e976b44ad22c87232b