Split findOrCreateClass(...) into separate find/create parts
Previously, `TurbineCodebaseInitialiser.findOrCreateClass(...)` would
first search for the class in the current codebase and if it could not
be found would then attempt to create one from the underlying Turbine's
classes, if available. This change moves the searching part into the
`TurbineBasedCodebase.resolveClass(...)` method and just leaves the
creating part in `TurbineCodebaseInitialiser`, in the renamed
`createClassFromUnderlyingModel(...)` method.

Bug: 353737744
Test: ./gradlew
Change-Id: I2aa9393040bb16f9b5f1ef234d58de0e7887afcd