ui: move root_type from name to unaggregatable column in flamegraph
This CL moves the root_type from being concatenated to the name of the
nodes in the flamegraph to instead being an unaggregatable property, a
newly added concept. By doing this, we remove the confusing suffix from
the name of the nodes and instead make it show up in a clear manner
when a node in the flamegraph is clicked on.

Change-Id: I1b7d906ad5e3711c28430b3d67569fece9c18217