Add --config-file option that supports multiple files
Future work, e.g. selecting Turbine for use with certain modules
requires being able to supply a configuration file with the rules that
determine which modules should use Turbine and which should use Psi
(the default). Other work needs something similar, e.g. to provide a
full definition of the API surfaces that are being generated.

Those configuration files will be orthogonal to one another so rather
than having a single file that forces all the information to be
included in one monolithic file this adds support for multiple files to
be provided. The exact meaning of the file will depend on its contents.

XML was chosen for the content format as while it can be a little
verbose it has good support for validating the content using XML Schema
files (which will help when editing) as well as good support in Kotlin
for deserializing.

At the moment this just adds a simple schema with a single `config`
root element. Future changes will extend that as needed.

Bug: 354699349
Test: ./gradlew
Change-Id: I8c876ffe6a4edda2b19274583861d2801e8cf3aa