Remove PsiParameterItem type check from KotlinInteropChecks
Previously, the check was needed as `isSamCompatibleOrKotlinLambda()`
would throw an exception if it was called for a non-`PsiParameterItem`.
However, this change makes the default implementation return `false`
(which was effectively what it did before) and adds a TODO with an
associated bug to make sure it did not get forgotten.

Bug: 352480646
Test: ./gradlew
Change-Id: I9291eb08b457a9fb4aa15556ddd445a0f7d4e4dc