Add CallableBody.verifyReturnConstants(...)
Previously, the `ExtractAnnotations.verifyReturnConstants(...)` method
had to dig inside the internals of a `PsiMethodItem` in order to make
sure that a method body annotated with a type def annotation used the
correct constants. This change moves that functionality behind the new
`CallableBody.verifyReturnConstants(...)` which allowed the Psi
specific code to be in `PsiCallableBody`.

Bug: 351118278
Test: ./gradlew
Change-Id: Ie021b73076bb9e75ca72f3f59381b19f3936cec0