Add FieldItem.ensureCompanionFieldJvmField()
This removes the dependency from `KotlinInteropChecks` onto
`PsiFieldItem`. Ideally, the check should be rewritten to use the model
abstraction, adding new ones if necessary. However, that will require a
lot more work. This is a pragmatic solution to remove the last few
dependencies on `Psi*Item` classes from the `metalava` project to allow
those `Psi*Item` classes to be marked as `internal` preventing any new
dependencies being added. The abstractions can be cleaned up later on.

Bug: 352480646
Test: ./gradlew
Change-Id: I8fc6c3d726b53be7330a7605cbf2bf3f3f5d42c2