Add tests for Callable.getCallableParameterDescriptorUsingDots()
This renames the `getApiLookupMethodDescription()` extension function
to `getCallableParameterDescriptorUsingDots()` and moves it into
`PsiCallableItem.kt` so that it can be tested from within the
`metalava-model-psi` project. The intent is that the Psi specific
implementation will be replaced with a model independent implementation
that will work with the other models. At that point the test will be
moved into the `metalava-model-testsuite` project.

The test cannot be written in `metalava` project, e.g. as a
`DriverTest`, because that has no access to the `Codebase` that was
created and no way to run a test on specific parts of that.

Bug: 352480646
Test: ./gradlew
Change-Id: I4b31e2512e8a938a6f9b6183b1c7787e4ebbd2a9