Change PsiItem from class to interface
This is part of the work to change the `PsiItem` class hierarchy so
that instead of `Psi*Item` classes extending a `PsiItem` class and
implementing an `Item` interface, they extend a corresponding
`Default*Item` class and implement the `PsiItem` interface. That
structure is better because it allows greater sharing of code, i.e.
`Item` interfaces are bigger and more costly to implement than
`PsiItem` is.

This change switches `PsiItem` to an interface and adds a temporary
`AbstractPsiItem` class for the `Psi*Item` classes to extend for now.
Follow-up changes will migrate the `Psi*Item` classes to extend
`Default*Item` and eventuaally `AbstractPsiItem` will be removed.

Bug: 352480646
Test: ./gradlew
Change-Id: Ie69239ba24dfda2814044b538bf3c40ea4fd8932