Generalize initializing of emit to false for expect declarations
Previously, `emit` was initialized to `!modifiers.isExpect()` in
`Psi*Item` classes only. This change moves it to `AbstractItem`. While
this should have no effect, as neither Turbine or Text support the
`expect` modifier (so it will always be `false`), this does make the
models more consistent and simplifies the process of having `Psi*Item`
classes extend `Default*Item` classes.

Bug: 352480646
Test: ./gradlew
Change-Id: I475d3af943085751bb45e3a5970114096fd2bd82