Refine HostOrchestrator initializatio
For no_wait_for_boot, it needs to refine TF logic for running HO service
to poll the device status after successfully lease a device. Scan the
logic, there is no way to pass ITestDevice in the lease call, so we need
to adjust it accordingly.

Since we plan to migrate this class to avd util, it needs to deprecate
TestDeviceOptions to reduce the dependency.

Bug:335049249
Bug:328328646

Test: unittest
Change-Id: Idefbb57a267118560e36a4f7bc44aac4be5e77aa