Use CategoryRedirectingReporter everywhere
Previously, `Options` provided access to three different reporters
requiring any code that wants to report an issue to choose the correct
reporter. This change hides them away behind a
`CategoryRedirectingReporter` so there is only a single reporter for
all the reporting code to use. The slight downside of that is that it
means that the issue category is a lot more important.

As issues are redirected to the correct reporter based on their
category there is no point in each reporter checking the category of
the issues reported through it as they will always be correct. So, this
removes the enforcement mechanism added in the previous change.

It was also necessary to remove the call to `getBaselineDescription()`
on `reporterApiLint` from a progress status string as `reporterApiLint`
was no longer accessible. That is not a significant loss as the name
of the baseline file was not a useful piece of information. As that was
the only use of `getBaselineDescription()` the method was removed
completely.

Bug: 351118278
Test: ./gradlew
Change-Id: Ia14b37f18b6130bbf9529f116eb8f1193cfa709f