Enforce issues have the correct category for their reporter
Temporarily enforces that issues must only be reported via the reporter
corresponding to the category. Change https://r.android.com/3182981
added a `CategoryRedirectingReporter` that redirects an issue to the
reporter that matches its category. The purpose of that was to avoid
having to pass multiple reporters into the `Codebase`. This change is
preparation for using that everywhere instead of the separate
`reporter`, `reporterApiLint`, etc.

This change temporarily (it will be removed in a follow-up change)
enforces that each issue is reported to the reporter that corresponds
to their category. The enforcement assumes the same mapping as is
currently used by `CategoryRedirectingReporter`. This was an easy way
of verifying that switching to `CategoryRedirectingReporter` will not
change the behavior, e.g. fail tests, print wrong error messages,
due to an issue being reported through a different reporter.

A number of issues were found to be in the wrong category and they were
corrected.

Bug: 351118278
Test: ./gradlew
Change-Id: I8c0fd3ead71a326a7dd5edd709bdfd77298f945c