Add test to show behavior of AnnotationItem.resolve()
The test is failing because `AnnotationItem.resolve()` uses
`Codebase.findClass(...)` and not `Codebase.resolveClass(...)`. The
difference between the two is the former will only find classes from
the class path that have already been resolved, whereas the latter will
resolve the classes.

The tests works on Psi but not on others because Psi seems to have
already resolved the `java.lang.annotation.RetentionPolicy` name to a
class on the class path, whereas the other models have not. The
follow-up change will fix it.

Bug: 352480646
Test: ./gradlew
Change-Id: I1ebe0d83c99ecf7be5d68b69a7800612e44c5199