Make PsiCallableItem extend DefaultCallableItem
Allows for more reuse of code. In addition to the necessary changes to
remove `PsiCallableItem`'s now unnecessary overrides, this change
includes some minor clean ups:
* Change the order of `PsiCallableItem` constructor parameters to match
  those of `DefaultCallableItem`.
* Moved the `psi()` method to the start of the class to be consistent
  with other `Psi*Item` classes that now extend `Default*Item` classes.

This change was slightly complicated because previously the
`parameterList(...)` method was called during `PsiCallableItem`'s
initialization AFTER the `codebase` and `psiMethod` properties were
initialized so could retrieve them from `containingCallable`. However,
this change swaps the order which means those properties are no longer
initialized at the time `parameterList(...)` is called so they needed
to be passed in explicitly.

There was also the need to add a down cast to `PsiParameterItem` as the
`parameters()` override that was removed by this returned
`List<PsiParameterItem>` instead of `List<ParameterItem>`.

Bug: 352480646
Test: ./gradlew
Change-Id: I4b1c99570e13033032d5e6844ab39aacbfb11956