Defer DefaultValue creation to DefaultParameterItem initialization
Previously, a `DefaultValue` was just passed as a constructor parameter
to `DefaultParameterItem`. That was enough for text and turbine models
as their `DefaultValue` implementations were not coupled to the owning
`DefaultParameterItem` instance. However, that is not true for the psi
model. This change follows the standard Metalava pattern of adding a
`DefaultValueFactory` type, along with matching `duplicate(...)` and
`snapshot(...)` methods. The latter already existed but needed a
`ParameterItem` parameter added.

Bug: 352480646
Test: ./gradlew
Change-Id: I6085996185e7d2ba44cd48fe612abbfda5995560