Move setting of varargs modifier to DefaultParameterItem
Previously, it was set to `true` in `ApiFile` if the type was a varargs
array. This change moves it into `DefaultParameterItem`'s initializer
to ensure consistent setting across models. It does not seem to be
needed by Turbine as the underlying model tracks the modifier and sets
the type's `isVarargs` property but this does no harm and keeps things
simple.

Bug: 352480646
Test: ./gradlew
Change-Id: I0e021fc580621486a1e3ab75b7e0540e01607744