Stop adding root package in TextCodebaseAssembler init block
Previously, the `TextCodebaseAssembler` broke the rules for the
`CodebaseAssemblerFactory` because it accessed the `Codebase` in its
initialization block. While that works when the `CodebaseAssembler` is
created in `TextCodebase` after `DefaultCodebase` has finished
initializing it will not work when that is done in `DefaultCodebase`.
This change replaces the `init` block within an `initialize()` method
that is called in the `TextCodebase` initializer, after
`DefaultCodebase` has finished initialization.

Bug: 353737744
Test: ./gradlew
Change-Id: Id2d2c2b621361eba90b1e9147420bb09819d7cbb