Move computeSuperMethods() et al into DefaultMethodItem
Previously, a number of `MethodItem` extensions functions were used to
share functionality across models. They are no longer necessary as all
models either just use `DefaultMethodItem` or extend it so this change
moves the extension functions into `DefaultMethodItem`.

The `isOverrideable()` extension function was kept as an extension
function because it is called on other `MethodItem`s but was moved into
`DefaultMethodItem.kt`.

Bug: 352480646
Test: ./gradlew
Change-Id: I1d2b57a8a3d16e6d515ca24728150b199a8a73dc