add aconfig_storage_reader_java for generated flag lib
This change adds aconfig_storage_reader_java for the generated flag
library to read value from new storage.

Car services and framework uses some flag for guarding the code. We are
working on switching the storage backend of the flag from DeviceConfig
to a new storage system. The flag libray will get the flag value through
this reader library.

The flag library was added as source file, so we need to explicitly add
the reader library here as dynamic dependency for building purpose.

Test: presubmit
Bug: 349874828
Change-Id: Ibe63b7f7423872cdcdbc6294a7e05751d41e21f6