Remove redundant code from hasImplicitDefaultConstructor(PsiClass)
Previously, it checked, among other things, if `psiClass.constructors`
was empty and if it was it called `PsiUtil.hasDefaultConstructor(...)`
and it that returned `true` then it returned `true`, otherwise it
dropped through to some other checks. However, the first thing that
`PsiUtil.hasDefaultConstructor(...)` does is check to see if
`psiClass.constructors` is empty and it it is then it returned `true`.
So, `hasImplicitDefaultConstructor(PsiClass)` always returned `true`
if it called `PsiUtil.hasDefaultConstructor(...)` making that call and
the following code redundant.

Bug: 352480646
Test: ./gradlew
Change-Id: I45c89c8f7b95ca6c46fdd770a28285a71fa830a6