Fix big negative dex2oatWallTimeMillis.
We calculate dex2oatWallTimeMillis by (uptime - starttime). To get a
reasonable result, both times need to be from the same clock. However,
it's unclear which clock the kernel uses for the latter.

proc_pid_stat(5) does not have anything helpful on this matter:

    (22) starttime  %llu
        The time the process started after system boot.
        Before Linux 2.6, this value was expressed in
        jiffies.  Since Linux 2.6, the value is expressed
        in clock ticks (divide by sysconf(_SC_CLK_TCK)).

        The format for this field was %lu before Linux 2.6.

To make sure we use the same clock, we obtain the uptime on process
start as the start time.

Bug: 315061143
Test: Presubmit
Change-Id: I5f344e953a945656474b3f625368c1ccb5aa52d8