Move DocQualifier code into PsiItemDocumentation
It was created and referenced from `PsiBasedCodebase` but only used
from within `PsiItemDocumentation` and relied on the state of that
class so it made sense to move the code into there.

The change was a simple move, a follow-up change will clean up the code
to take advantage of its new location.

Bug: 352480646
Test: ./gradlew
Change-Id: I1f214fb5ef16a5445b39d43bdbcd045067c3bc5b