Pass SourceSet into PsiBasedCodebase.initializeFromSources(...)
Previously, the work for getting `units`, i.e. `PsiFile`s and package
docs from the `SourceSet` was done in `PsiSourceParser` and then were
passed in to the `initializeFromSources(...)` method. However, that
information was not used in `PsiSourceParser` itself and the package
docs were updated in `initializeFromSources(...)`.

This change moves all that work inside `initializeFromSources(...)` to
allow the package docs to be collected in one go, consistent with how
Turbine does it.

Bug: 352480646
Test: ./gradlew
Change-Id: Iff5df699258f350d78872274050d469bacfea825