Pass SourceSet into TurbineCodebaseInitializer.initialize(...)
Previously, the work for getting `units`, i.e. `CompUnit`s and package
docs from the `SourceSet` was done in `TurbineSourceParser` and then
were passed in to the `initialize(...)` method. However, that
information was not used in `TurbineSourceParser` itself and the
package docs did not include `package-info.java` files.

This change moves all that work inside `initialize(...)` to allow the
use of `gatherPackageJavadoc(...)` like Psi.

Bug: 352480646
Test: ./gradlew
Change-Id: If40eed55fe874b5dbf6ce423d28bf89f945045b4