Add CodebaseContext.removeTestSpecificDirectories(...)
Previously, `BaseModelTest`s could not easily check location
information as it would contain test specific directories which would
be different on each run of the test. This change adds support for the
tests to replace the test specific directory with a placeholder making
the result consistent across test runs.

This will be used in a follow-up change.

Bug: 352480646
Test: ./gradlew
Change-Id: Id453f1226bd68e8652a39ba3b0f24356b6b98afd