Remove codebase from PackageList
The `codebase` property in `PackageList` created a cycle between the
`PackageList` (which was created by `Codebase` implementations) and
`Codebase`. That makes future refactorings more difficult. However, it
was only need so that the `BaseItemVisitor.visit(PackageList)` could
call `visitCodebase(Codebase)` and `afterVisitCodebase(Codebase)`. This
change avoids that by adding an `ItemVisitor.visit(Codebase)` method
which `BaseItemVisitor` can override instead.

Bug: 352480646
Test: ./gradlew
Change-Id: I469404c300cabc4cb44592997718a4f8ba096158