Make PackageDocs.comment an ItemDocumentationFactory
Previously, the `comment` property needed to be wrapped in a possibly
model specific implementation of `ItemDocumentationFactory` which meant
that the code for creating `PackageItem`s needed to be model specific.
This change does the wrapping in the model specific lambda passed to
`gatherPackageJavadoc(...)` which will allow `PackageItem` creation to
be model independent.

Bug: 352480646
Test: ./gradlew
Change-Id: Iee748567fe250e72072b2144001381b80201b902