Simplify TurbineCodebaseInitializer.createAllPackages(...)
Previously, it took a `classUnits` list with which it did two things:
1. Create packages for the classes
2. Save a mapping from `TyDecl` to `CompUnit` so that it can be used
   in `createClass(...)` to find a `CompUnit` for a specific `TyDecl`
   from which it can create a `TurbineSourceFile`.

The first point is not needed as the `createClass(...)` method creates
packages as needed. So, this change removes that.

The second point is not package related and should be done elsewhere.
So, this change moves it into `initialize(...)`.

Bug: 352480646
Test: ./gradlew
Change-Id: I3a343df0aae460795e58ad418170ecc9a96dd219