Remove TurbineCodebaseInitialiser.sourceClassMap
Previously, this was used in a couple of places but this change
refactors those uses to allow it to be removed.

The `initialize(...)` method wrapped it in a `SimpleEnv(...)` and then
combines that with the `classPathEnv()` to get a combined env for all
the classes on the source and class path. This change maintains that
but just uses a local `sourceClassMap` variable. It also stores the
combined env in the `envClassMap` property for use in
`createClass(...)` below.

The `createAllClasses(...)` method iterated over it to create all the
source classes. This change passes it as a parameter.

The `createClass(...)` method uses it to lookup the `ClassSymbol` to
find a `TypeBoundClass`. This change makes it use the combined env
created in `initialize(...)`.

Bug: 352480646
Test: ./gradlew
Change-Id: Ic4b44e2e62f7d15d88ea3389b97c853d611808d9