Combine SourceTypeBoundClass and CompUnit for package-info.java
Previously, the `package-info.java` file was represented purely as a
`CompUnit`. Unfortunately, that meant that package annotations were not
available as they are only present on the synthetic `package-info`
class that Turbine creates for a `package-info.java` file. The
annotations are available on a `SourceTypeBoundClass` but that does not
have access to the `CompUnit` which is necessary for determining the
file location.

So, this change adds a `PackageInfoClass` that will contain information
from them both. A follow-up change will add `SourceTypeBoundClass` and
use that to obtain the package annotations.

Bug: 352480646
Test: ./gradlew
Change-Id: I50d5448257a9261e2046510e14e89cbc9913be39