tracebox: add mode to use system sockets
This CL adds the --system-sockets to tracebox to force the use of the
default system sockets instead of the temporary sockets by the autostart
mode of tracebox. This makes it easy for e.g. users of the Perfetto SDK
to connect to the producer/consumer sockets which is not as easy with
the default behaviour

Fixes: https://github.com/google/perfetto/issues/850
Change-Id: I9e215760e9dc6d62349ea003f20733d6c393bc0a