Add findOrCreatePackage(...) in PackageTracker
Previously, callers had to find and then create packages separately,
this change adds the `findOrCreatePackage(...)` to allow that code to
be deduped.

It currently returns a `FindOrCreatePackageResult` object that contains
the `DefaultPackageItem` found/created and a `created` flag to indicate
whether it was created or not. That is needed because lots of callers
have different behavior depending on whether it is a newly created
package or no. Follow-up changes will gradually move that behavior into
the `findOrCreatePackage(...)` method at which point it will just
return the `DefaultPackageItem`.

Bug: 352480646
Test: ./gradlew
Change-Id: I6ade27ce4cb57da10907564f6c42047f31c0b82d