Stop using ItemDocumentation.NONE_FACTORY for DefaultPackageItems
The `NONE_FACTORY` creates an `ItemDocumentation.NONE` object which
does not allow for any documentation. It is intended for use by `Item`s
which either do not support documentation at all, i.e. `ParameterItem`
and `TypeParameterItem` or when javadoc is not being processed, i.e.
when `allowReadingComments` is `false`. Unfortunately, it was being
used by default for `DefaultPackageItem`s created by the
`DefaultItemFactory`. This change switches it to an empty (but mutable)
`ItemDocumentationFactory`.

Bug: 352480646
Test: ./gradlew
Change-Id: I8e8d2fbc927f5362aef9c59f7c3d426275f5a5fd