Add owner to ModifierList.equivalentTo(...)
Previously, the `DefaultModifierList.equivalentTo(...)` implementation
used the `owner` property to tweak the check to take into account the
context. However, that property is being removed in a follow-up change
so this adds an `owner: Item?` parameter to `equivalentTo(...)`. It is
optional because there are some places where it is not needed.

Bug: 356548977
Test: ./gradlew
Change-Id: I8741dc780d49d042640baf48d28e6456a0e12c28