Add general MutableModifierList.mutateAnnotations {...}
`MutableModifierList` provides some methods for mutating the
annotations list. Previusly, they were all implemented in
`DefaultModifierList`. This change adds a general method for mutating
the annotations list directly and implements the existing mutator
methods in terms of it. This will allow multiple mutation operations,
e.g. removing one annotation and adding another, to be performed within
a single mutation operation.

This is the first in a set of changes to make `DefaultModifierList`'s
`annotations` property a `List<AnnotationItem>`. By doing all the
mmutations inside of a specific method call it allows the underlying
representation to be changed while still preserving the atomicity of
multiple mutation operations.

Bug: 356548977
Test: ./gradlew
Change-Id: I66eb6dd7d73997c2b08c57da74d6de990850ac6f