Replace PsiPackageItem with DefaultPackageItem where possible
It was necessary to change `PsiPackageItem.addTopClass(...)` to take a
`ClassItem` (with conditional down cast) so that it overrides the
method in `DefaultPackageItem`. Otherwise, switching to
`DefaultPackageItem` ends up calling the wrong `addTopClass(...)`
method which means the `PsiClassItem.containingPackage` is not set.
This is only a short term measure. Longer term, the
`containingPackage` will be passed into `PsiClassItem` constructor like
it is for `DefaultClassItem`.

Bug: 352480646
Test: ./gradlew
Change-Id: I00b8cc19db7980e4d0dfea6912dc75ff94b201f0