Use findOrCreatePackage(...) in createInitialPackages(...)
Previously, the packages that were present in the source file as
`package-info.java` and `package.html` files were created separately
as part of the `createInitialPackages(PackageDocs)` method and the
`containingPackage` property was initialized on demand (or in the case
of `PsiBasedCodebase` explicitly wired together). However, future
changes to `findOrCreatePackage(...)` will cause it to create the
containing package before any contained packages so that they can be
wired up at creation time.

This change works towards that goal by using `findOrCreatePackage(...)`
in `createInitialPackages(...)` so that the order in which that
method processes the packages is irrelevant.

e.g if the source contains the following:
* `a/package-info.java`
* `a/b/package-info.java`

Then the `createInitialPackages(...)` will iterate over the packages
`a.b` and `a` and create them. If the `a.b` package is created first,
then the process of creating that will create the `a` package. However,
`createInitialPackages(...)` will still try and create `a` because it
is next in the list. Using `findOrCreatePackage(...)` will prevent that
from attempting to create the `a` package twice.

In order for that to work this change has to make `PackageDocs`
available to `findOrCreatePackage(...)` so it can retrieve any
additional information from that as needed, not just for the package
being created but its containing packages too, if it has any.

Bug: 352480646
Test: ./gradlew
Change-Id: I2113dd5dbcd9803f4ea5a49786d2b1a35d5418a3