Fix Turbine setting of PackageItem.emit
Previously Turbine would set `PackageItem.emit` to `false` if and only
if `isFromClassPath` was `true`. That was true whether or not the
`PackageItem.emit` was previously set to `true`. That caused some test
failures as a package item that had one `ClassItem` that was to be
emitted and another that was from a class path would have its `emit`
set to either `true` or `false` depending on which one was processed
last when it should always be `true`.

This change makes it consistent with the other models, so it will set a
newly created `PackageItem`'s `emit` property to `!isFromClassPath` but
if the package already exists then it will only set it to `true` if
`!isFromClassPath`, otherwise it will leave it alone. That means that
once a package has had its `emit` set to `true` it will never be set
back to `false`.

Bug: 352480646
Test: ./gradlew
Change-Id: If346c1fe2bc4d5e4d79544d9cc0b12220a451a41