Remove FindOrCreatePackageResult
Previously, `PackageTracker.findOrCreatePackage(...)` would return an
instance of `FindOrCreatePackageResult` to encapsulate the
`DefaultItemPackage` that was found or created and a `created` flag to
indicate which. That was needed as the callers did a variety of work
depending on the value of the `created` flag. However, all that work
is now down in `findOrCreatePackage(...)` itself so it is no longer
needed.

Bug: 352480646
Test: ./gradlew
Change-Id: I802ef175c82bdc81e7546d0a3f4349356f1ef0c3