Move ClassItem wiring code to PsiClassItem init block
Previously, it was duplicated in a number of places. This dedups it by
moving it into an `init` block which is consistent with how it is done
in `DefaultClassItem`. It did require some casting but it should be
safe and if it's not then someone did something very wrong.

This allowed the removal of the `packageClasses` which kept an
association between classes and the package name to which they belonged
so that they `PsiClassItem <-> PsiPackageItem` connections could be
added in `finishInitialization()`. However, that is no longer needed as
the wiring is done on the fly while creating `PsiClassItem`s.

Similarly, the `finishClassInitialization(PsiClassItem)` method was no
longer needed.

It was necessary as part of this to update `initializeFromJar(...)` to
call `PackageTracker.createInitialPackages(PackageDocs.EMPTY)` to
create the root package.

Bug: 352480646
Test: ./gradlew
Change-Id: Idf1dc69385b525c13e4e21e3853bf2fd1e1a6857