PackageTracker: Create full package containment hierarchy
Previously, the text and turbine models would only create a partial
package containment hierarchy. e.g. given the following files:
* `src/a/package-info.java`
* `src/a/b/c/Class.java`

Text and Turbine would create two packages, `a` and `a.b.c` and the
former would be the containing package for the latter. There would be
no `a.b` package. The problem with that is that it does not adapt to
discovering other packages during processing. e.g. If `a.b.OtherClass`
was a class on the class path that was only discovered after the above
structure was created then the package containment would be incorrect.

Now, it would be possible to go back and patch up the hierarchy but
that would greatly complicate the code. So, instead this change simply
creates all intermediate containing packages. They are created with
`emit = false` so will not be included in the API unless they also
contain API classes in which case `emit` will be set to `true`. That
more closely matches what Psi already does and the few extra packages
does not affect performance in a measurable way.

Despite not modifying the text and turbine code for creating packages
this change fixes the `Test package containment` test as
`DefaultPackageItem` (which is used by text and turbine) lazily
computes the containing package to be the closest enclosing package.

It was necessary to tweak `ApiFileTest.dumpPackageStructure(...)` as
it relied on the previous text behavior of not creating intermediate
containing packages.

Bug: 352480646
Test: ./gradlew
Change-Id: I8e55c1492e19892fd7c2a2efd647b9515292a19a