Correct snapshot package creation
Previously, `CodebaseSnapshotTaker` took a complete copy of the
`PackageItem` being snapshotted, including the `ApiVariantSelectors`.
However, that was incorrect as the snapshot, by definition, does not
track `ApiVariantSelectors` as it is supposed to be a snapshot of an
already filtered API surface.

This change fixes that by using the `CodebaseSnapshotTaker.itemFactory`
to create the `PackageItem` as that already handles using the correct
`ApiVariantSelectors`. It did break 5 tests which needed to be
baselined but that is ok as those tests are testing the tracking of
`ApiVariantSelectors` so are expected not to work with snapshot.

Bug: 352480646
Test: ./gradlew
Change-Id: Ic73881adf3402e687afb0b5c191f4719ffa4d1d5