Add ResourceFile wrapper around overview.html file
Previously, the contents of the `overview.html` file were read in and
passed around as a `String`. However, very few runs of Metalava
actually need to write out the `overview.html` file. This change is the
first in a number of changes that will only read the contents when
necessary. It starts off by creating a `ResourceFile` wrapper around
the contents.

Bug: 356799622
Test: ./gradlew
Change-Id: Idc6622144fbca574ad1becae78acc770312396bb