Test generating JDiff for class with implements list
This adds two tests for generating JDiff representation of a class with
an `implements` list. In the first one it generates a full JDiff file
and the `implements` list is included. However, in the second one it
generates a partial JDiff file that adds a field to a class with an
implements list and in that case the `implements` list is not included
which is an error.

A follow-up change will fix the issue.

Bug: 352480646
Test: ./gradlew
Change-Id: Ide2a0ba242506b082a1b04dbacff798cbc862d00