Avoid mutating ClassItem.createDefaultConstructor(...)'s modifiers
Previously, `ClassItem.createDefaultConstructor(...)` implementations
would set the visibility to `PACKAGE_PRIVATE`, then modify it to match
the containing class's visibility and then one of the users of it, i.e.
`ApiAnalyzer`, would mutate it and set it back to `PACKAGE_PRIVATE`.
This change avoids the mutation by allowing the caller to specify the
visibility but it defaults to the containing class's visibility.

The `PsiConstructorItem.createDefaultConstructor(...)` was called
directly but this change simplfies that by calling
`ClassItem.createDefaultConstructor(...)` instead.

Bug: 356830333
Test: ./gradlew
Change-Id: Ifb411bb4ca0c03f622e39e5143a8d419e9b8f132