Move files from verification/ to verifier/ to match Java package
Previously, the Java classes were in the `.../sdksandbox/verification/`
directory but internally use `package ...sdksandbox.verifier;`. This
change moves the files to match the Java package. Due to the mismatch
Metalava silently ignored these classes so they were hidden from the
API. To maintain that behavior now that the mismatch has been corrected
it was necessary to add `@hide` to a couple of the classes.

Bug: 357045589
Test: m checkapi
Change-Id: I605eea4a78a0bcf832cfe0e812d8d54241dfd259