Report package mismatch and carry on
Prior to some recent changes Metalava would ignore classes whose
package statement did not match the directory. Those recent changes
caused Metalava to start throwing an exception instead. This change
creates some fake `PsiPackage`s in order to carry on. In order to
preserve (in the short term) the previous Metalava behavior to ignore
classes with this mismatch such a class will have its `emit` set to
`false` so it will be excluded from the API.

However, rather than silently ignoring the issue it reports it so that
developers have a chance to fix it properly.

Bug: 352480646
Test: ./gradlew
Change-Id: I6820602b335e0a80a777da2e29eb86c77b3c86f1