AnnotationsMerger: Clarify purpose of methods
`AnnotationsMerger` supports merging two types of annotations,
inclusion and quantifier annotations. Inclusion annotations are those
that can be used in `--show-*`, `--hide-annotation` and
`--revert-annotation` options. They are treated specially as they have
to be applied early, before deciding what `Item`s belong in the APIs.
Qualifier annotations are all the rest, e.g. nullability, `@IntDef`,
etc.. It also supported multiple different sources, e.g. XML, jar and
signature files.

Previously, it had a number of `merge*Annotation*` methods that did not
indicate which type they were used for or the source of the annotation
which made it difficult to understand their purpose. This change
renames them to make their purpose clearer and make it easy to refactor
to address other issues.

It's not clear why the previously unused `findAnnotation` import was
removed from `ApiAnalyzer` by `./gradlew ktFormat` as part of this
change and not before.

Bug: 356830333
Test: ./gradlew
Change-Id: Ifdc697a5700ca2d74cdd4ccc758dfb79e48c3116