AnnotationsMerger: Separate inclusion annotation merging
Previously, inclusion annotation merging would use the
`mergeInclusionOrQualifierAnnotation(...)` method but that was
unnecessary as it did some post merge checking that is not needed for
inclusion annotations. So, this change rewrites the inclusion
annotation merging to use the `mutateAnnotations {...}` mechanism
directly and avoid using that method. The method is renamed to make it
clear that it is only use for merging qualified annotations.

Bug: 356830333
Test: ./gradlew
Change-Id: I082624c400a69e34416b0ce411f0ca971841838f