Report conflicting nullness when merging from source
Previously, while merging from XML did report conflicting nullness
annotations, merging from sources did not. This change rectifies that
discrepancy. The XML checks for conflicts as it merges but this does it
as a separate check. It might be slightly less efficient but it is
slightly simpler. It was necessary as part of this change to ignore
type annotations that were duplicates of item annotations to avoid
reporting the same issue twice. That is because annotations which are
both `TYPE_USE` and `FIELD` will appear twice, one on the `FieldItem`
and once in its `TypeItem`. That should not cause any behavioral
changes as only one annotation of a single type (i.e. `qualifiedName`)
is being merged anyway.

Bug: 356830333
Test: ./gradlew
Change-Id: I508d83c3f1f3950ef45302ef2acddefc03abdbc6