Refactor addDefaultRetentionPolicyAnnotation(...)
Previously, it took a `ClassItem` parameter which meant that it had to
be called after a `ClassItem` had been created. While that worked when
the `DefaultModifierList` that was supplied when constructing a
`ClassItem` was the one used by that `ClassItem` that will soon no
longer be the case. In future if a `MutableModifierList` is supplied
when constructing the `ClassItem` then it will take a copy.

That means that for `addDefaultRetentionPolicyAnnotation(...)` to work
it would either have to explicitly mutate the `ClassItem.modifiers`,
which would require extra work, or it will need to be called before the
`MutableModifierList` it is called upon is copied.

The latter is preferred so this change refactors it to support that. A
follow-up change will move the call (along with other similar) code.

Bug: 356830333
Test: ./gradlew
Change-Id: I948bf3b1ed2ad2cc1f8c3f2ff442561c7386c118