Cache DefaultModifierList instances with no annotations
Now that `DefaultModifierList` is immutable instances of it that whose
`annotations` list is empty can be shared across `Item`s even across
`Codebase`s. This change adds a cache for those instances.

Some testing was done on Android and it reduced the number of
`DefaultModifierList` instances by somewhere between 50%-75%. Small
libraries with a few source files saved about 50% (out of a few
thousand in total) with a cache size of ~10. Large libraries with lots
of source files saved about 75% (out of ~1,000,000 in total) with a
cache size of ~70.

Bug: 356830333
Test: ./gradlew
Change-Id: I300a57a07fbb38945932f242e22819a805b1aa45