Move tracking of external classes into DefaultCodebase
Previously, `TextCodebase` tracked external classes, i.e. classes
retrieved from the class path via a `ClassResolver` itself. This change
moves that logic into `DefaultCodebase` in preparation for moving more
functionality into `TextCodebaseAssembler`.

Rather than add a `registerExternalClass` that has to be explicitly
called when an external class is found the registration of external
classes is done automatically in `resolveClass(...)` if the class
returned by `CodeAssembler.createClassFromUnderlyingModel(...)` does
not belong to the calling `Codebase`.

The previous `TextCodebase.resolveClass(String)` override returned
`ClassItem` but the overridden method returned `ClassItem?`. That
necessitated the use of `!!` in `StubKind` to forcibly ignore `null`
but that is safe because while the return type has changed the actual
returned type is still `ClassItem`.

Bug: 352480646
Test: ./gradlew
Change-Id: I39b819f4e34ecbb295422bcb6c5594bbfc81bea8