base: Add base::Token class
This CL is a cherry-pick of the following CL: https://crrev.com/c/1320190
It adds token.h and token.cc introduced in the CL above. These files are
requried to build ARC with [Uuid] attribute in Mojo enabled.
This CL also replaces unguessable_token.h and unguessable_token.cc with
those in the original CL to let base::UnguessableToken to inherit
base::Token.

Bug: b:357737923, b:41420830
Test: m

Change-Id: I93cc0a588a4268416d3a3ba28e170992794e479d