Add CodebaseAssembler.createPackageItem(...)
Previously, the `itemFactory` was added to `CodebaseAssembler` so that
it can be used to create `DefaultPackageItem`s for `PackageTracker`.
However, that will not work for the Psi model because it cannot use
`DefaultItemFactory` to create its Psi specializations.

So, this change adds a `createPackageItem(...)` method and pushes the
implementatation used by the turbine and text models down into a new
`DefaultCodebaseAssembler` class.

Bug: 352480646
Test: ./gradlew
Change-Id: Ibd59fb0690a4c7976b83f92e9fc131b13ffcb2c4