StubWriter: Remove StubWriterConfig.apiVisitorConfig
Previously, `StubWriterConfig.apiVisitorConfig` was used to supply the
`ApiVisitor.Config` when creating `filterReference` in the
`StubWriter.createFilteringVisitor(...)` method but a separate
`apiVisitorConfig` parameter was used elsewhere. The
`StubWriterConfig.apiVisitorConfig` is left over from when `StubWriter`
was an `ApiVisitor` but was not cleaned up properly.

This change removes it and uses `createFilteringVisitor(...)`'s
`apiVisitorConfig` parameter instead.

It also corrects a copy/paste error in one of the comments.

Bug: 357833584
Test: ./gradlew
Change-Id: I607da79001b7fcadc71b4c3765ecc3d9aefe0cb2