Add CodebaseFragment to encapsulate Codebase and ItemVisitor
Previously, the `createReportFile(...)` method took a `Codebase` and a
`(Writer) -> ItemVisitor` lambda that given a `Writer` would return an
`ItemVisitor` that when applied to `Codebase` would write to the
`Writer`. Most of the uses of `createReportFile(...)` would create and
return a `FilteringApiVisitor` that determined which parts of the
`Codebase` would be visited and would delegate to a `DelegatedVisitor`
that did the actual writing.

This change extracts the `Codebase` and `FilteringApiVisitor` creation
aspects of that into a `CodebaseFragment`. The purpose of that is to
make it easy to visit a fragment of the `Codebase`. It adds another
`createReportFile(...)` method that consumes a `CodebaseFragment` and
uses that for writing out the current and remove API files.

Bug: 357833584
Test: ./gradlew
Change-Id: I72fdada051bb5f1e13828810b9a40840a7915471