Make DefaultItemFactory compatible with Default*Item constructors
The `DefaultItemFactory` was previously only used by text and turbine
models which meant that it did not have to support all the capabilities
of the corresponding `Default*Item` constructors. However, it will be
used by `CodebaseSnapshotTaker` to take snapshots of `PsiBasedCodebase`
and so does need all those capabilities.

This change adds:
* `ItemLanguage`
* `DefaultValueFactory` instead of `DefaultValue`
* `CallableBodyFactory`

Bug: 357833584
Test: ./gradlew
Change-Id: I8a13d9e89dd6daf4064e7ae6975fc47ff6338d61