Snapshot documentation correctly
Previously, if an `Item` was being reverted the snapshot `Item` would
use the previously released API's documentation but that would always
be empty as the previously released API does not have any
documentation.

This change uses the documentation from the `Item` being reverted. If
that includes `@deprecated` then that is removed if the previously
released `Item` was not already deprecated. That needs to be done as
part of this change as otherwise the `@deprecated` in the documentation
will force the snapshotted `Item` to be marked as being deprecated even
when it was not previously deprecated.

Bug: 357833584
Test: ./gradlew
Change-Id: I33c549950f279fb4a536f44e814dbbfddd8a37f1