Autoupdate of 2 submodules
*packages/modules/CaptivePortalLogin updated to 7b191d37c8620b48e1aa0680cf9ae1ed7f3ef62c
Exclude androidx package from jarjar rule.

Exclude "androidx" package prefix, otherwise, some classes within the
package such as "androidx.startup.InitializationProvider" cannot be
found when launching the CaptivePortalActivity. androidx is typically
used by apps as-is so it should not conflict with classes in the
bootclasspath.

Bug: 329361366
Test: atest testCaptivePortalIsNotDefaultNetwork --iterations
Test: manual test, install the CaptivePortalLogin module with this cl
      and connect to the AP, never see the ClassNotFoundException and
      the portal page can be shown correctly.
Change-Id: Iffb5a881907e5945fbea8ebe7a7c9bb927558a85

*packages/modules/common updated to a1cd73b95e4f2e8daa12e6d6107b2b9842407b82
Revert "Revert "Update allow_deps.txt for refactored CaptivePort..."

Reason for revert: found the root cause of cts failure, and fix it with a new change which should be on top of this topic, so I revert revert this CL first and apply the fix on top of this.

Reverted changes: /q/submissionid:3206651-revert-3196237-CaptivePortalLoginLib-VOXHIDWDUB

Restore the previous CL which updates allow_deps.txt for refactored CaptivePortalLoginLib.

Apex-Size-Increase: no size change, the source files in the lib
such as CaptivePortalLoginActivity.java are already in the module.
Previous-Platform-Support: Yes
Aosp-First: Yes
Test-Info: Part of CaptivePortalLoginTests

Bug: 329361366
Test: TH
Change-Id: I909bfa2a051bdcb7d354bc2e2e08953bf151399b