Do not visit packages that should not be emitted
Packages are special in that they only appear in APIs if they contain
at least one class which appears in APIs. If that is the case then
their `emit` property will have been set to `true`. So, if their `emit`
property is `false` then it means that they are either empty of classes
(e.g. because it was created solely as a containing package for another
package) or all its classes are from the class path.

This change avoids visiting those unemittable packages or their
classes as it is a waste of time.

Bug: 357833584
Test: ./gradlew
Change-Id: Iafa2ab05f6c2685ad565f4f7b351043770483acc