Compute fullName in DefaultClassItem
Previously, `fullName` was passed as a parameter to `DefaultClassItem`.
That was unnecessary as it could be computed from the other parameters
and opened up the possibility of inconsistencies arising, e.g. the
`fullName` being say `Outer.Inner` but `containingClass` being `null`.

This change removes the parameter and computes it from the `simpleName`
and the `fullName` of the `containingClass, if there is one.

After remove the call to `computeFullClassName(...)` its only remaining
use was from `PsiItemDocumentation` so it was moved into that file.

Bug: 357833584
Test: ./gradlew
Change-Id: I4bd23905fb60b2b85f444a1084db9f9a22b2b8dd