Use snapshot when generating stubs for reverted items
Previously, the `StubWriter` and `JavaStubWriter` had some special code
in to handle reverts. This change switches them to use the snapshot to
generate a `Codebase` that includes reverted items and removes the now
unnecessary special code for handling reverts.

The `revertDocumentationDeprecationChange(...)` method is not needed as
that is already handled by `CodebaseSnapshotTaker`.

Bug: 357833584
Test: ./gradlew
Change-Id: If33e0b3e0a961e909b0ac1330ce4c16cb55e68c0