Use correct modifiers to determine if field needs initializing
Previously, `JavaStubWriter` would always use the current `FieldItem`'s
modifiers to determine if the field needed initializing. That meant
that if the previously released `FieldItem` was `final` but the current
`FieldItem` had removed the `final` modifier the resulting stub would
not compile because it would correctly use the previously released
`final` modifier but would not provide an initializer.

This issue was found in a follow-up change that switched stub creation
over to use the snapshot and shows that the snapshot is a much better
solution than patching the `JavaStubWriter` as it ensures consistency
of the output. This change makes a temporary workaround and fixes the
test to avoid changing the tests in the follow-up change which will
removed this temporary workaround.

Bug: 357833584
Test: ./gradlew
Change-Id: Ie943131c673248b20544de0cf1dda5376830d4b4