Use correct parameter name when reverting
Previously, it would always use the `name()` of the `ParameterItem`
that is being snapshot but that does not work well if that is the
parameter from a previously released API as that does not always track
parameter names properly as parameter names are not considered
significant in Java (even though they are in Kotlin).

This change will fall back to use the `name()` of the `ParameterItem`
from the current `CallableItem` in that case.

Bug: 357833584
Test: ./gradlew
Change-Id: Ie613574af3848ea8e7ffef568312edc9ac0353d5