Make ApiFileTest.Test for current API surface more realistic
Previously, it did not provide a `ClassResolver` which meant that any
classes that would have been loaded from the resolver, e.g.
`java.lang.Object` were created as stubs which is not realistic. This
change adds a `ClassResolver`. It does not change the test behavior
but it will be affected by a follow-up change and this makes the result
of that change more consistent and realistic.

Bug: 357833584
Test: ./gradlew
Change-Id: Ic84ed39d0d218abf04905cdaaf0e378501348cdd