Simplify handling of PackageItem.emit
Logically, `PackageItem.emit` is `true` if and only if it contains at
least one `ClassItem` whose `emit` property is `true`. Previously, that
was mostly achieved by having callers of `findOrCreatePackage(...)`
passing in an appropriate value for `emit`. However, that was quite
complicated and there were some cases where it was not quite true.

This change simplifies the process by moving the logic for initializing
the `PackageItem.emit = false` to a `DefaultPackageItem` init block and
moving the logic for updating it if a emittable class is added to it to
the `DefaultClassItem` init block. The `emit` parameter in
`findOrCreatePackage(...)` is no longer used and will be cleaned up in
a follow-up change.

That fixed a couple of test issues:
* `ApiFileTest` - the `<root>` package was present in the previous
  expected  output as its `emit` property was incorrectly set to `true`
  despite containing no emittable classes.
* `130 - check classes from classpath` in
  `metalava-model-snapshot-testing` failed previously as the snapshot
  taker did not set `PackageItem.emit` to false for packages that only
  contained classes from the class path.

Bug: 357833584
Test: ./gradlew
Change-Id: Ia2676b1c3c59977659168412f9f455e9b0a4095c