eliminate ref-counting in AIDL generated client methods
`asBinder` returns a `SpAIBinder` by value. That type is ref counted and
so we end up with a `AIBinder_incStrong` and `AIBinder_decStrong` call
for each copy. It is called twice in each client method, so an 4 extra
function calls. Switching to the new `asBinderReference` function
eliminates those function calls and reduces the total binary size of a
sample AIDL .so file by 5%.

Test: m
Bug: 338458975
Change-Id: I830a459284c7bf6f205f3f5b92960935822a4af8