Treat stubs classes as if they came from the class path
Previously, stub classes were created with `isFromClassPath` set to
`false`. That was strictly speaking correct as the stubs were not from
the class path but in practice they behaved as if they should. In fact,
often the stub classes were only created because a `ClassResolver` was
not provided. If a `ClassResolver` was provided then the classes would
be loaded from the class path. It does not have any behavioral impact
as the class's `emit` property is set to `false` anyway.

This change sets `isFromClassPath = true` for stubs and changes the
corresponding test to match. This is needed now because a follow-up
change will use the `isFromClassPath` to set the `emit` property.

Bug: 357833584
Test: ./gradlew
Change-Id: I40a1f0d509bea32548c9aaec4da1762e66ca1d93