Move ApiAnalyzer.addConstructors(...) into a separate class
Previously, the code to add additional constructors needed when
generating stubs was part of `ApiAnalyzer`. However, they are only
needed for writing stubs so it makes no sense for it to be part of
`ApiAnalyzer`. This change moves that logic into its own
`StubConstructorManager` class. The move was very clean showing that it
did not really belong in `ApiAnalyzer`.

Bug: 357833584
Test: ./gradlew
Change-Id: Ie9864f40f48c904b8361ec8282903020f301d280