Always cast `null` in super constructor call
Previously, the code for generating a `super(...)` constructor call was
optimized to only generate casts when there was more than one possible
constructor that it could reference. The problem with that is it relied
on knowing exactly how many constructors would be present in the super
class in the generated stubs. Accessing that is complicated because it
depends on the behavior of the `filterEmit` predicate used to determine
which constructors to write out. That does not work well with the
`FilteringApiVisitor` and in fact https://r.android.com/3213738 caused
a slight change in the generated stubs output because it gave the
`wrong` number of constructors. Fortunately, it said that there were
more constructors than there actually was so it added extra casts which
did not cause any issues. However, if it said that there were less
constructors then it would remove casts that were needed.

Rather than continue to maintain the optimization this change removes
it and always adds casts even if they are not strictly needed. That
simplifies the code and the testing at the expense of slightly bigger
stub files. It just sets `includeCasts = true` to keep the change
small. Follow-up changes will clean up the code.

Bug: 357833584
Test: ./gradlew
Change-Id: Ic2d3d5ea2367226e375b3e8bcf47bfa6a6ed47c0