Add test for duplicate source classes
Reproduces an issue seen when merging changes to Android.

There are three tests:
1. `CommonDuplicateClassItemTest` - basic duplicate test, checks what
   can be seen of the duplicated class in the `Codebase`.
2. `CommonDuplicateClassItemTest` - as above but on a snapshot of the
   `Codebase` - this replicates down stream issue.
3. `DuplicateClassTest` - duplicate test that checks the issue
   reported.

All the tests fail in Turbine as Turbine itself throws an exception if
it finds duplicate source classes so the test will always fail.

Psi carries on and registers the class twice which results in
`DuplicateSourceClass` being reported. All the classes are added to the
package but only the first class is added to the list of all classes
held by `PsiBasedCodebase`.

The snapshot of Psi fails because it is given the duplicate classes but
it throws an exception instead of reporting the issue.

It is a bit of a mess but follow-up change(s) will clean it up.

Bug: 357833584
Test: ./gradlew
Change-Id: I8640c0b7768dcbbd9ae53e9c3b3001cffdc7dc2c