Stop snapshot throwing an exception on duplicate classes
This makes `DefaultCodebase.registerClass(...)` report an issue rather
than throw an exception so it will just carry on like Psi does.
However, the exception message was more useful than the Psi report so
this changes makes them both use that message in their report.

It was necessary as part of this to pass in a `non-null` `Reporter`
for the snapshot and for the text signature, Psi already passes in a
non-null `Reporter`,

Bug: 357833584
Test: ./gradlew
Change-Id: I3103cf0692b6ec8b839efdc8529fdd37d5d11780