Ignore duplicate class
Previously, a duplicate `ClassItem`, i.e. one registered with the same
name as an existing `ClassItem` would replace the existing one in the
`qualifiedName -> ClassItem` map and be added alongside it in the
`PackageItem.topLevelClasses()` list. That caused problems when taking
a snapshot as the duplicate `ClassItem`s were passed to the snapshot.

This change avoids that problem by ignoring the duplicate class
altogether.

Bug: 357833584
Test: ./gradlew
Change-Id: Ic0bee6d5a00c194c20a691f4f27007e6db2a083e