Add support for logging with abseil backend
This enables forwarding of the logging messages to the abseil backend.
Note that this is only enabled for QEMU8 and newer.

It will make sure that all the GFXStream messages end up on the same log
stream, and will use the same formatter.

- 'V' level will use VLOG(1)
- 'D' level will use VLOG(2)

Note that QEMU8 will never set any of the gfxstream callbacks as it will
be invoked through rutabaga.

Includes a set of unit tests to validate the behavior.

Bug: 358381877
Change-Id: Ia7aa835044d538960646a9e9b8d262828228f325