Autoupdate of 2 submodules
*kernel/build updated to c210336458dd2a99aa361a1dfd02fe7fd562c167
kleaf: Use platform-agnostic labels for prebuilts

For dependencies on //prebuilts/build-tools and
//prebuilts/kernel-build-tools, use the platform-agnostic
labels instead of inferring to a specific target. That way, we have a
central place (//build/kernel:prebuilt_*) to control the set of
binaries we want to use by changing the transition.

Bug: 325514706
Change-Id: I78ca229125d81c91a62b9b834feb8750a5e222f1

*prebuilts/build-tools updated to 02e0bd095ff2ab2a6dcd24a6d6b2d175b59dabcd
Merge "Add libs to data of all prebuilt_tool." into main