metric: Add Wattson thread attribution for trace
Add metric that attributes estimated power to individual threads for the
duration of the entire trace.

wattson_tasks_attribution.proto defines the output format for all power
attribution to tasks. Possible tasks include threads (this patch), and
potentially processes and packages in the future. The
wattson_tasks_attribution.sql is a helper metric that can be re-used in
the future for attributing power to other tasks.

Test: tools/diff_test_trace_processor.py out/linux/trace_processor_shell --name-filter '.*wattson.*'
Bug: 355070694
Change-Id: Ic323015c88d153e0cd0684b0b7cc37297ac50097
Signed-off-by: Samuel Wu <wusamuel@google.com>