Treat mismatching package statement and directory as error
Previously, it was a warning but treated as an error when new as there
were some examples of that in the source code. Those have now been
fixed so this is promoting it to an error to prevent any new cases
being added.

Historically, classes with mismatching package statement and directory
were ignored by Metalava. That behavior was maintained in the recent
refactoring while the existing uses were fixed as otherwise it would
have tried to add new classes to the API breaking the build. Now those
uses have been fixed this change no longer ignores the class.

Bug: 352480646
Test: ./gradlew
Change-Id: Ic9052b7cdcda80671103441c4a36a7d20e1d069c