Run metalava-model-snapshot-testing against Psi model
Previously, `metalava-model-snapshot-testing` was only run against text
and turbine but unfortunately, they do not provide all the same
capabilities as Psi and Psi is the one currently used in the Android
build. So, this change runs the tests against Psi, baselining all the
current failures.

It removes the `Test duplicate classes with snapshot` test as that is
no longer needed as this will run `Test duplicate classes` against the
snapshot anyway.

Bug: 353737744
Test: ./gradlew
Change-Id: Ie34e71e11eef61e1c544a013509daaba8606490b