Check `emit` property in assert(Resolved)Class(...)
The `emit` property is really important when it comes to determining
whether something will even be considered for an API. Differences in
the `emit` property across models have caused issues in the past. They
seem to have been resolved but to ensure consistency this change makes
the `assertClass(...)` and `assertResolvedClass(...)` check the `emit`
property of the `ClassItem` they retrieve. This will prevent divergence
across models.

One test used `assertResolvedClass(...)` for a class that did not need
resolving so it was changed to use `assertClass(...)` instead.

Bug: 357833584
Test: ./gradlew
Change-Id: I994aadf05dd4a754acd1faa767b365008c61cdca