Avoid taking a snapshot of unemittable items
Previously, `CodebaseSnapshotTaker` would take a snapshot of `Item`s
whose `emit` property was `false`. That is unnecessary and a waste of
time as unless those `Item`s are referenced from an `Item` that has
`emit` set to `true` they will never be used.

This change adds a `NonEmittableDelegatingVisitor` that will ignore
`Item`'s whose `emit` property is `false`.

This does not break any tests as they were written to handle
differences in behavior between models about which types are resolved
and which are not. Any test that requires access to a class that is not
in the source files (i.e. whose `emit` property is set to `false`) uses
`assertResolvedClass(...)` to ensure that the class is resolved if
necessary. Ignoring such classes in the snapshot just means that they
will be created when resolved.

Bug: 357833584
Test: ./gradlew
Change-Id: If8e6c865cbcc9a4dc8f638ffe54db3103beea76b