Remove PsiBasedCodebase.initializing
Previously, the `PsiBasedCodebase.initializing` property was used to
distinguish between classes defined in source files explicitly listed
source files and classes defined either on the class path or, contrary
to the name, elsewhere on the source path. Specifically, if
`initializing` was `true` then `ClassItem.isFromClassPath()` would be
set to the value of `PsiBasedCodebase.fromClasspath` and if
`initializing` was `false` then `ClassItem.isFromClassPath()` would be
set to `true`.

Using a mutable property in that was was confusing and hard to
understand so this change removes it and replaces it with an
`isFromClassPath` parameter that is passed to the
`PsiBasedCodebase.createClass(...)` method.

Bug: 352480646
Test: ./gradlew
Change-Id: Ibe22be59082df2952eeb651eaaa770d2278e1d01