Remove PsiBasedCodebase.findOrCreatePackage(PsiPackage)
The `findOrCreatePackage(PsiPackage)` method is only called from
`PsiBasedCodebase.createClass(...)` and it is unnecessary. The
`createClass(...)` method would call `findPsiPackage(...)` to find a
`PsiPackage` with a specific package name. It would then pass that
to `findOrCreatePackage(...)` which extracted its name and then passed
that to `packageTracker.findOrCreatePackage(String)`. This change just
calls `packageTracker.findOrCreatePackage(String)` directly.

The `findPsiPackage(...) call is left to detect and report package
mismatch errors but it no longer needs to create a fake `PsiPackage`.
The `PsiCodebaseAssembler.createPackageItem(...)` method will do that
if needed.

Bug: 352480646
Test: ./gradlew
Change-Id: I3df00f13560bd44ff63bcfb07302e8d81eebd341