Clean up some left overs from previous refactorings
The comment was removed as there is no longer any `initializing`
state that affects whether a class is assumed to be from the sources
or not. That is now explicitly supplied on class creation.

The setting of `emit` was removed as that is already set to
`!isFromClassPath` in a `DefaultClassItem` init block.

Bug: 352480646
Test: ./gradlew
Change-Id: Id866e9ba7e3fdd605ed978034e59ea835aae56b2