Defer creating ClassItem for inaccessible classes
Previously, during the `Codebase` initialization Psi and Turbine models
would create a `ClassItem` for every source class that is specified on
the command line, even if that source class could not contribute
directly to an API surface, e.g. because it is package private. This
change makes them both defer creating `ClassItem`s for those source
classes after initialization, until they are actually needed, e.g. when
resolved from a `ClassTypeItem`.

Deferring a class from being created during the `Codebase`
initialization has the effect of changing its `emit` property from
`true` (because it was `ClassOrigin.COMMAND_LINE`) to `false` (because
it will be treated as `ClassOrigin.SOURCE_PATH`). That required
changing some tests that checked that their `emit` was `true`.

Bug: 325341742
Test: ./gradlew
Change-Id: I4d7e358ec64cc893d99215b4bfe66192868ea6ca