Add ClassItem.origin of ClassOrigin type
The `ClassItem.isFromClassPath()` is inaccurate and misleading as it
implies that when it is `true` the `ClassItem` is from the class path
but that is not strictly true, at least not for Psi. What it actually
means in Psi is that the class was either from the class path or it
was from a source file that is on the source path but not in the list
of files supplied on the command line.

That causes a number of issues:
1. It is a source of inconsistencies with Turbine which does not yet
   handle searching the source path.
2. Some places in the code really needs to check to see if something is
   on the class path.

This change adds a `ClassItem.origin` property that is of `ClassOrigin`
enum type which can differentiate between the different origins of the
class item. At the moment it sits alongside `isFromClassPath()` but it
will eventually replace it altogether.

Bug: 359168852
Test: ./gradlew
Change-Id: I00bc376644b10ed01885b281c10e533b777f8f2d