Replace uses of isFromClassPath() with origin
Each use of `isFromClassPath()` has been replaced with an appropriate
check of `origin`. The exact check differs from place to place as some
of the code really needs to check whether something is on the class
path and some just needs to check whether something was not on the
command line.

Bug: 359168852
Test: ./gradlew
Change-Id: I7cf3b161e5de58eb2efeebe9fc00e9f3ba360160