Add test for isFromClassPath() for class on source path
Recently a number of refactorings have passed the test suite but failed
in the Android build. That has been due to the Android build passing
some source files explicitly to Metalava and others have been passed
implicitly via the source path but the tests almost always pass the
files explicitly.

That is also a big different between Psi and Turbine. Psi will search
the source path but Turbine will not.

This change adds test suite support for supplying additional files to
pass implicitly via the source path. It uses it to test the behavior
of `ClassItem.isFromClassPath()` on a class that is loaded implicitly
from the source path but will be used in future for more general
testing.

Bug: 359168852
Test: ./gradlew
Change-Id: Idbcb9bd47592a666916cb1d1bbe2e9f6708fe9ef