Add MTE-globals handling to procinfo mappings glue
MTE globals remaps the segments that contain global variables on
compatible devices. It does so by creating an anonymous mapping, copying
the contents, and renaming the mapping so the information about the file
and offset are not lost.

But, obviously procinfo mapping stuff needs to know about this and be
able to handle it. So, this patch adds the handling.

Test: atest libprocinfo_test
Bug: N/A
Change-Id: I1198fa8895831f75083a9e6c20d7cfd264cb1a89