[scudo] Support linking with index in IntrusiveList (#101262)
The nodes of list may be managed in an array. Instead of managing them
with pointers, using the array index will save some memory in some
cases.

When using the list linked with index, remember to call init() to set up
the base address of the array and a `EndOfListVal` which is nil of the
list.

GitOrigin-RevId: 8ea8f1f2fe501e0be5d0142d79651f490fb2ae2c
Change-Id: I36f22552c1ce6934c4cfc8d8a8e82a55e72bf14f