Fix MixKernelTargetPreparer to support uncompressed device images.
The MixKernelTargetPreparer was hardlinking the device image to the device, but this works only if the device image is a zip file. This CL changes the preparer to symlink the device image if it is a directory.

This fixes the device image linking issue. To support uncompressed device images, MixKernelTool also needs to be updated to support uncompressed images (a separate effort).

Bug: 358632998
Change-Id: I4eec8073360d902c85e582fbee86f13d00e424ca
Test: https://atp.googleplex.com/test_runs/438306675 (The test fails due to the mixtool as expected with force-disable-chunk-dir=false, but no more hardlinking issue)