Remove tap device conflict checking in host-sandboxing mode
This removes a `/bin/bash` intermediate when running in host-sandbox
mode. If the `egrep` executable or an equivalent was run in the sandbox,
it wouldn't see any other processes due to the constraints of the pidfd
namespace, so it couldn't do any conflict checking anyway.

See the `/proc and PID namespaces` section in
https://man7.org/linux/man-pages/man7/pid_namespaces.7.html

Bug: b/359316164
Change-Id: I1e27aed46a212a806a84370298d4e3ca20925b6e
Test: launch_cvd --enable_host_sandbox=true --resume=false