Registers musl toolchain.
The toolchain allows building CC binaries and libraries against
musl libc. The toolchain is selected if the target platform has
//build/kernel/kleaf/platforms/libc:musl constraint value, which is
not the default at the time when this patch is submitted.

Note: the support for cc_* rules are incomplete by this CL. This change
only allows building kernel with the musl toolchain.

Bug: 325514706
Change-Id: Ibcbb74d13e0ccf35f4d57ddaf5ac68cd169bba19