Add extension properties to access PsiClass names
Simplifies code that has to access the nullable `PsiClass` `name` and
`qualifiedName` properies. Some code checked `qualifiedName` and then
if that was `null` fell back to `name`. That was because those code paths
were also used type parameters, i.e. `PsiTypeParameter` classes, which
are a subclass of `PsiClass` but have a `null` `qualifiedName`.
However, type parameters are handled completely separately now so the
code for named classes no longer needs to handle a `null`
`qualifiedName`. That is why the `classQualifiedName` extension
property does not fall back to `simpleName`.

Bug: 325341742
Test: ./gradlew
Change-Id: I5c2e9f432604ccbf5813a8b5000903756b119c09