Inherit stub methods from top down
Previously, the methods and interfaces were inherited into classes
based on the order in which they appeared in the list of classes. That
was inconsistent as sometimes a super class would inherit stub methods
before the extending class and vice versa. As the set of methods that
are inherited depends on the set of methods in the super classes (i.e.
the `superMethods()`), the inconsistent order would lead to inconsistent
behavior.

That did not previously cause any problems because the value of
`superMethods()` was cached prior to the stub methods being added.
That is a bug that leads to it reporting incorrect values after stub
methods have been inherited. However, it actually ensures consistent
and correct behavior as the order in which the classes are processed
does not affect the value of `superMethods()`.

Future changes will fix the caching of `superMethods()` at which point
the order will become significan. So, to avoid that future change
causing inconsistent behavior this change enforces an order by making
sure to inherit stub methods into a super class before the extending
class.

Bug: 359862331
Test: ./gradlew
Change-Id: I415576c04851346e4476af3efc1ffd61ba70765d