pw_rpc: Handle pw::Status returns
Pigweed is enabling [[nodiscard]] checks for pw::Status by default.
pw::Status returns must be used to avoid compiler warnings. The
IgnoreError() may be used to silence these warnings if the status is not
important.

Ignore the status returned when closing a pw_rpc client or server, and
check the status from opening an RPC channel.

Change-Id: Id344554ef9ef03e5970255dce073f7d08df06f70