align snapshot/restore implementation with vhost-user spec
This is the second half of https://r.android.com/3182888.

It is more complex, mainly because the performing the data transfer
requires some form of concurrency to conform the spec correctly (the
initial IPC method is only supposed to negotiate the FD, not interact
with it), but also, to a lesser degree, because I wired it up into the
wmediumd specific code so that we might snapshot more data there
eventually.

Bug: 301269927
Bug: 354026289
Test: launch CF, then snapshot/restore
Change-Id: I10d6400ce62aa2f03ae696c9a5daae7b49b28dd8