cmdline_integrationtest: Correctly wait for start of trace
We observed a failure on CTS tests running
`PerfettoCmdlineTest.SaveAllForBugreport_LargeTrace`

```
fake_producer.cc:188 PERFETTO_CHECK(trace_writer_) (errno: 0, Success)
```

The test expects that `FakeProducer::StartDataSource()` is called before
it calls `FakeProducer::ProduceEventBatch()`.

This doesn't appear to happen, because the test is not waiting for the
tracing session to be started.

The test just waits for any tracing session to be started, but, if
there's another tracing session executing in the background, that's not
enough.

This commit changes the code to check for the specific tracing session,
like the test immediately above does.

Bug: 358042251
Change-Id: I3466b1fcbf1e8fc14d3c89b0897e841e563ae8cf