trusty:mmu_test: rework reserve test to fix various issues
Previously the reservation test would create multiple aspace's to
reserve most of the available pages in the system, but would not free
them.  A new probe function now does the same but does not leave pages
reserved after the test.

Also reduce the chunk of pages from 5000 to 500 so that the test won't
fail if 5000 free pages cannot be found.

Bug: 356975821
Test: build.py qemu-generic-arm64-test-debug --test com.android.kernel.mmutest
Test: on device (adb and Android with test TAs loaded)
Change-Id: I85d42637d4cf80b9d57ed23a68bc9de6d42b7fed