tp: recognise new field for ftrace cropping and work around b/344969928
The new timestamp is treated exactly the same as the previous one, so
the distinction is mostly for manual debugging at this point in time.

The workaround is hacky since it relies on a magical unique_session_name
field, but it is more than likely the only case where the pessimistic
cropping is relevant.

Bug: 344969928
Change-Id: I10a83f5bb8e8eb0fba90eb5b57fac565e6982f85