Skip running ART Checker tests during code coverage runs
We have observed some Checker tests failing during code coverage runs
for the v2/mainline-engprod/art/gce-coverage-v-build-module-on-v-platform-shard-00
config as the instrumentation code added to collect the coverage
information is changing the Control Flow Graph file, making some
assertions fail unlike the "normal" runs.

In order to avoid this failures, this CL aims to skip running Checker
whenever we identify the ArtRunTest is being run as part of a Java code
coverage run, based on the Tradefed configuration.

Bug: 356852324
Test: -
 1. acloud create --branch git_main --build_target cf_x86_64_phone-trunk_staging-userdebug_coverage
 2. atest art-run-test-482-checker-loop-back-edge-use
    --experimental-coverage
Test: m tradefed tradefed-tests \
        && tools/tradefederation/core/javatests/run_tradefed_tests.sh \
             --class com.android.tradefed.testtype.ArtRunTestTest
Change-Id: Iff91b39abfe78919664bd12a1436b00a6dffc602