Add an error unwrap macro OR_ASSERT_FAIL for use in gtests.
Similar to OR_RETURN and OR_FATAL, but it aborts the test with an
assertion failure on error.

Also introduce a generic macro UNWRAP_OR_DO, which collects the common
boilerplate in OR_<something> macros. It can also be useful directly in
user code to simplify unwrapping.

Also extend and correct a comment for the gtest-related macros in
result.h.

Test: atest libbase_test
Test: Success and provoked errors with https://r.android.com/3236844.
Bug: 360737799
Change-Id: Ib9132e82c0435488a4d5c3f73ef4c2ae3c913c4c