efi: adds libefi mock APIs and a few tests
Adds the beginning of a libefi_mocks library, and uses it in a few
console printing tests.

The mock API is a bit complicated due to the global nature of UEFI so we
may not be able to fully mock out all the surface, but this will at
least give us a starting point.

Bug: b/355436086
Change-Id: I0e12b858757c70ae35602cd164b7331cc4d09341