NDK Backend: use/export include_cpp from AIDL libs
include_cpp libs were shipped as part of the NDK, but this
caused problems as the NDK and AIDL were shipped separately.
In order to resolve this, we removed include_cpp from the
NDK and started shipping these includes in the SDK in
platforms/$PLATFORM_NAME/optional/libbinder_ndk_cpp/.

However, we never actually removed libbinder_ndk_cpp from
the NDK sysroot. So, in order to do that, we have to add it
back into the Android build.

Removing headers from the NDK sysroot does not remove them
from incremental builds, so it may be difficult to see why
this change is needed.

Bug: 357711733
Test: build apps in tree
Change-Id: I3b12d5d09c6888901831437c447f2f99c1f1947d