Autoupdate of 2 submodules
*external/ot-br-posix updated to 454da7f6c3af23cb2127e6945fc0ae33908a975b
Split the infra link state and configuration

The OtDaemonConfiguration was a combination of configuration and the
platform state. This made it harder to update only one of them at
ot-daemon. In this CL, the original config is splitted into
OtDaemonConfiguration and the InfraLinkState.

Bug: 362172947
Test: CI

Change-Id: I00cd361bd975023a9c08b6a4de4d0852c589a4c3

*packages/modules/Connectivity updated to 9309dc87c3e058eef89b235585887a368e3a80d3
Merge "[Thread] split the configuration and the infra link state" into main