trusty: Re-order modules so virtio modules load first
The device for the virtual RPMB virtio-serial device is expected to be
available at /dev/vport3p1, but loading the trusty modules first causes
this serial device to move to vport4p1. We want the ordering to be
consistent with the kernel built in the Trusty build system without
modules, so this change re-orders these modules to the same load order
as in that kernel build.

Test: m && run.py --android .
Bug: None
Change-Id: I86bcaa9cef24e87fb2675561a2624dddd5443cda