Pass TestInformation in SearchArtifactUtil
Some tests can set temporary working directory or test directory in the
TestInformation object which is not tracked by the execution files.
Instead of requiring tests to update execution files, pass the testinfo
object, so that we can collect the test directory or working directory
or the executionFiles directly from the testInfo.

Bug: 329338742
Test: ABTD, Presubmit
Change-Id: I455e32f6157a3f56bc631c963f9a160cfaecc74b