ui: Fix performance regression
Due to a bug in TrackManager, we currently schedule a microtask for
99.9% of ALL tracks every render cycle, which tanks performance.

This bug was introduced in aosp/3247216 but uncovered in aosp/3247998,
which changed the implementation of TrackDecider to call tick on all
tracks every frame, not only those which have been seen thusfar. Before
aosp/3247998 performance would have slowly deteriorated over time, but
now it's prevalent right from the get-go.

This CL fixes the issue at its core, and only schedules an onDestroy
once a track has reached a certain age, not after.

Change-Id: Iaf767b87804da2085e3eaca344a79f35d582adcf