ui: move CriticalPath buttons to its own plugin (part 2)
Moves the logic behind the critical path buttons into
the critical path plugin. This finishes the move of critical path
into its own core plugin.
This allows to get rid of some TODOs and awkward dependencies on
addDebugSliceTrack().
There is a little price to pay, that is, some inherent coupling
between thread_state_tab and the critical path plugin. But that
coupling is limited to a command name, which seems tolerable.

Change-Id: Ib35a6863fa412d5f440452a753bc37ffcd4a990d