[trace-processor] Add importer for Instruments XML files
This adds an importer for Instruments XML files, as exported from an
Instruments trace with `xtrace export` (see code comments for the exact
command line and XML format).

Specifically, the XML tokenizer parses the XML into time sample Row data
structures, which are inserted into the sorter and then semi-trivially
parsed after sorting into frames, callsites, and samples. Row timestamps
are relative to the start of Instruments profiling -- there is support
for clock sync events in the profile to sync these timestamps against
Perfetto's boottime clock.

The XML parsing requires pulling in a dependency on libexpat, which is a
third-party streaming XML parser.

Bug: 362665903
Change-Id: I7e638f0b3c7275214aef3410f6263fb75e3968ac