Avoid caching superMethods() on unfrozen class
Previously, `superMethods()` would compute its result on first call,
cache it, and then return that value on every subsequent call. The
problem with that is that additional methods could be added to super
classes after the first call which meant that the cached value was no
longer valid. That did not cause any test failures as code that used
that method relied on the cached value not reflecting changes made
since.

This change prevents caching the value unless the containing class is
frozen. That revealed a bug in `ApiAnalyzer.addInheritedStubsFrom(...)`
where it ended up adding the same method twice, breaking one of the
tests so this change also fixes that.

Bug: 359862331
Test: ./gradlew
Change-Id: Icfe0a7431c3bba2e4207e32660a4fc053525d20e