Restrict exported symbols from libnative{loader,bridge} lazy libs.
The shims in the lazy libs bypass the stubs and hence the vetted
exported APIs in the .map.txt files. To avoid inadvertently giving
external access to internal APIs that way, let's apply the .map.txt
files as version scripts to the lazy libs instead, so that any
functions added there just don't become available to external code
unless the .map.txt files get updated.

This hack doesn't work if the lazy libs are linked statically, so
restrict them to be shared libs (although ideally they should be static
instead to avoid the shared lib overheads).

Also drop the shims for the two functions added in
https://r.android.com/2689969 and https://r.android.com/3022103,
because they're only used internally in ART and don't go through the
shims.

Test: atest libnativebridge-tests art_libnativebridge_cts_tests \
            libnativebridge-lazy-tests libnativeloader_test \
            art_libnativeloader_cts_test libnativeloader_lazy_test
Bug: 124250621
Change-Id: I1e1eba3f656830f51b792ce672276f37b3f0d865