gbl: start migrating Android boot to libgbl
Starts migrating the Android boot logic to libgbl, so that all the
EFI-specific backends are handled by the generic GblOps interface.

The plan is to do most of the migration in-place, until the logic
entirely uses GblOps and is ready to move to libgbl. Hopefully this will
keep it from being too invasive to other concurrent changes.

Initially, this CL just switches all console printing away from
efi_println to gbl_println.

Bug: b/363074091
Change-Id: I1e8dd3952134d62f2f8f90a6050c8336513aed9d