RESTRICT AUTOMERGE: Skip when subgroupSize > 64 instead of fail
The legacy tests have artificial limitations against subgroupSize <= 64.
When the impl reports 128, lots of reconvergence tests would fail. The
fix (https://gerrit.khronos.org/c/vk-gl-cts/+/8944) is not applicable
for backport to Android 13. So intead of removing the tests, we could
workaround to skip those when an impl reports > 64 subgroupSize.

See also below for more contexts:
https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/29277#note_2420733

Bug: b/358360958
Test: dEQP-VK.reconvergence.*
Change-Id: Id7e62b63ed61f6fa91e8c1002b53e9d9aba88c56