Support secondary_user_on_secondary_display for CtsNetHttpTestCases
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on
devices that have config_multiuserVisibleBackgroundUsers enabled.
The main use case is Automotive's multi-display Whole Cabin experience
where passengers (modeled as visible background users) can interact
with the display in front of them concurrently with the driver
(modeled as the the current user) interacting with driver's display.

Bug: 364413340
Test: atest CtsNetHttpTestCases
Test: atest --user-type secondary_user CtsNetHttpTestCases
Test: atest --user-type secondary_user_on_secondary_display CtsNetHttpTestCases
Flag: TEST_ONLY

Change-Id: I0a2cf249b3b63b630c357ca473ab99bd5716e326