LUCI: Create couple (hidden) experimental builders.
Create a clone of four arm builders with a new name
(this the first step of trying to rename them in UI).

Create similar copies of six host builders,
and also run those on newer version of Ubuntu.

The new builders are hidden from console UI for now,
and they should not generate failure notifications.

Our existing builder configurations are unaffected,
as can be observed from the generated config files.

Test: tools/luci/config/main.star
Change-Id: Id21f2347e968c66214cf1650cbc4f52f1798b5c2