Autoupdate of 2 submodules
*frameworks/native updated to 0b3be83fda03783a9c6dee8515354784aa5f22f6
Merge "binder: ndk: add asBinderReference method to BpCInterface" into main
*system/tools/aidl updated to f6e1975214b9c065c4f7cd290e913e8f8734cfc1
eliminate ref-counting in AIDL generated client methods

`asBinder` returns a `SpAIBinder` by value. That type is ref counted and
so we end up with a `AIBinder_incStrong` and `AIBinder_decStrong` call
for each copy. It is called twice in each client method, so an 4 extra
function calls. Switching to the new `asBinderReference` function
eliminates those function calls and reduces the total binary size of a
sample AIDL .so file by 5%.

Test: m
Bug: 338458975
Change-Id: I830a459284c7bf6f205f3f5b92960935822a4af8