CommonClassItemTest: Make Kotlin test more consistent with Java
Previously, the kotlin variant of the "Test duplicate without type
substitutions" test used `@hide` to indicate that a public class was
not intended to part of an API. that was not. However, the java variant
used an inaccessible package private class. That difference will cause
differences in behavior in a follow-up change that defers creating
`ClassItem`s for inaccessible classes. So, this change makes the kotlin
test more consistent by making its class inaccessible by marking it as
`internal`.

Bug: 325341742
Test: ./gradlew
Change-Id: I49d2364eb833641295907ce56360ec1acee3f58f