Set MSCEn bit in HCRX_EL2 register by default
qemu 8.2 and the Linux kernel now support the
new optimized MOPS memory operations from v8.8
and v9.3. These trap to EL1. In unused-EL2 mode,
we need them to trap correctly to EL1 as the
correct exception kind, which requires that
the MSCEn bit be set in HCRX_EL2 by default.

Bug: 364748249
Change-Id: If6cfeb70fc1f51ce97f96e786a4cf18ff52df6cb