[scudo] Update secondary cache time-based release logic (#107507)
Secondary cache entries are now released to the OS from least recent to
most recent entries. This helps to avoid unnecessary scans of the cache
since entries ready to be released (specifically, entries that are
considered old relative to the configurable release interval) will
always be at the tail of the list of committed entries by the LRU
ordering. For this same reason, the `OldestTime` variable is no longer
needed to indicate when releases are necessary so it has been removed.

GitOrigin-RevId: e5271fef8fd8931370f04702ba2f9e8b2ab0e523
Change-Id: I4114a9d40eff6b14e6e059713c35f297305be5d8