Add test for using permission name as part of a word in javadoc
If `@RequiresPermission` is used without setting `conditional = true`
then text will be automatically added to the documentation explaining
that the permission is required. Metalava attempts to detect when the
documentation already mentions the permission and reports that as an
issue to avoid duplicating the documentation. However, its heuristic
for determining whether the permission is mentioned is simplistic and
only checks to see if the permission is mentioned in the text at all.
For most permissions that works because the permission name is very
specific and unlikely to occur anywhere else but some permissions, e.g.
`DUMP` could very easily be present in the documentation for some other
reason.

This change adds a test to show the issue. A follow-up change will fix
it. The test fails on Turbine because Turbine does not yet support
manipulating the javadoc like Psi does.

Bug: 367787970
Test: ./gradlew
Change-Id: Ib8d65dea8f8c7948e4278cf8c9bc102b2427dd92