Fix heuristic for detecting permission name in documentation
Previously, the heuristic would look for the permission's unqualified
name in the documentation. That meant if the permission was `DUMP` it
would think that `DUMP_FILE` or `VERBOSE_DUMP` was referencing the
permission when it clearly is not. This change fixes that by only
looking for the permission as a whole word.

Bug: 367787970
Test: ./gradlew
Change-Id: I6fbcc7e0c7175dfc5c56ecfe24167e273f797876