Autoupdate of 4 submodules
*external/flashrom updated to 8b48ebd018e890ff5deac72fd29a08fc45e90798
Move flashrom to the vendor partition

Test: m
Bug: b:366170141
Change-Id: Iea440736477799196496fc2cd8f069507ae1792b

*external/kmod updated to 1a5764621b950188c6fba2e3d8ea6ae502bed26d
Make libkmod available for vendor partition targets

libpci requires libkmod. libpci is required by flashrom, which has to be
moved to the vendor partition.

Test: m
Bug: b:366170141
Change-Id: If395083c85f645515794fec2ab313091a0dbc31e

*external/pciutils updated to 9517d971cd54bc2e3275fe58f1ee54fe8fe6a0fd
Make libpci available for vendor targets

libpci is dependency of flashrom, which has to be moved to the vendor
partition.

Test: m
Bug: b:366170141
Change-Id: I9d9488d3a41b2b021a056de37f0f12eaff7616a6

*external/vboot_reference updated to ef8f3c8635a6832738060a71e01c9ede579c6712
Move futility and cgpt to vendor partition

Test: m
Bub: b:366170141
Change-Id: I96b9b86868540d20e7d8488fbcb38e91b3d3b40e