Remove CodebaseSnapshotTaker.currentPackage
Previously, the `currentPackage` was set in `visitPackage(...)` and
cleared in `afterVisitPackage(...)` and used in `visitClass(...)`.
Keeping the state of this correct was difficult due to it being
reentered via `createClassFromUnderlyingModel(...)`. Future changes
would make that worse. This change removes it and creates the package
when requested for a class simplying the rest of the code.

Bug: 368021325
Test: ./gradlew
Change-Id: Ia7759f79d4e6690ecbf14f3c3fa91e8592805532