Remove CodebaseSnapshotTaker.currentClass
Previously, the `currentClass` represented the innermost of a stack of
nested classes (linked through the `containingClass()` method). A new
`ClassItem` was pushed in `visitClass(...)` and popped in
`afterVisitClass(...)`. That required that the classes were nested but
that meant it could not take a snapshot of a partial `Codebase`, e.g.
a system API `Codebase` where the outer class was part of the public
API and the nested class was part of the system API. In that case the
outer class would have `emit = false` and so the nested class would be
ignored.

This change removes the `currentClass` and adds an extension function
that allows the containing `ClassItem` to be retrieved as needed. In
the process it fixes the outstanding "TODO". Previously, creating
`java.util.Map.Entry` through `createClassFromUnderlyingModel(...)`
would not set its `containingClass()` to `java.util.Map` because
`currentClass` was not set correctly. However, this change fixes that.

Follow-up changes will continue the work to support partial
`Codebase`s.

Bug: 368021325
Test: ./gradlew
Change-Id: I4cbcfa52cd54a058f62a840f912925142e84bc49