Return value from CodebaseSnapshotTaker.inScope(...)
Previously, it did not return anything, i.e. was of type `Unit`. This
changes it to return the same type as returned by the `body` lambda.
This will make following changes simpler. The existing uses have been
updated to take advantage of this by moving some code outside the
lambda. That will be more consistent with following changes.

Bug: 368021325
Test: ./gradlew
Change-Id: Ic53d1e846cf9e322ce107717076dba6c65095acc