Add SnapshotTypeItemFactoryContext
Previously, the various `*TypeItem.snapshot()` methods and the
`List<ParameterItem>.snapshot(...)` method were defined in the
`CodebaseSnapshotTaker` and would access the `typeItemFactory` property
to retrieve the `SnapshotTypeItemFactory`. That separation of the
setting of `typeItemFactory` and methods that used it made it easy to
misuse.

This change adds inner class `SnapshotTypeItemFactoryContext` to
encapsulate the `typeItemFactory` and moves the methods that access it
into the class. That ensures that the callers of those methods must
have called `inScope {...}`. It is preparation for removing the
`CodebaseSnapshotTaker.typeItemFactory` which relies on preserving
class nesting which is in turn blocking taking a snapshot of a
partial `Codebase`.

Bug: 368021325
Test: ./gradlew
Change-Id: I6ded52c0779cf02128b4c5a8e1bac03f20d36e72