Enforce -Wmissing-prototypes
This lint check will raise an error if a function is implemented
without a corresponding prototype declaration included in scope.

The main use case will be to ensure mock functions do not become out of
sync causing the original function to be linked instead.

Because of the large amount of errors, this change suppresses
errors in all locations, to be cleaned up gradually.

Bug: 369381361
Test: m com.android.btservices
Flag: EXEMPT, lint fixes
Change-Id: I888b5e284e237d8a6e2864bdde55c0a7fc0c348d