ui: Fix UiMain lifecycle
Today UiMain has an onremove and a disposable stack, but
today (before this CL) those are pure placebo. In fact,
today UiMain is never destroyed.
This CL fixes that by wrapping it in a component that takes
care of creating a new UiMain instance for each new trace.
It does so by carefully linearizing with the point where
globals is initialized, so that everybody else sees globals'
entries in a consistent state.
This is required because we register Notes editor and Details
tabs in UiMain (which itself is okay). Until now we had only
one instance forever for those tabs, and it was fine because
their state was living in globals.state, which got recycled on
every trace. Now we are moving to a world where each manager
holds its own state, which means registering a tab only once
won't be enough.
In general feels nice to tie the lifecycle of the UiMain
component to a trace lifetime, as today it does register a
bunch of commands that need to be unregistered and re-registered
on each trace.
Also remove the overview track from globals and put it into
tickmark panel, which is the only user of it.

Change-Id: If8c9776aab5635b0a3d2ec917d2fe619e376ae9b