Autoupdate of 3 submodules
*frameworks/support updated to 4fdaf543b5a5709708b623d122712353ccc30707
Merge "Ink: Update from internal state" into androidx-main
*golden updated to b512ef722202ad6820fbf074ce928b77bb9f79e1
Update golden images (auto-generated from Scuba): invocation I45200010318509277

Note: The CLs in this topic include a regression seen in these screenshots where input points are interpreted incorrectly. Fixes for this are in progress, but in order to unblock API review, the aim is to land this topic as is, send out a CL for API+2 review, and then create a follow-up topic/CL with the fixes and updated screenshot goldens.

Change-Id: I1d9e0d7ed0ffe086b6685ffc4ba5e3780bfbb07b
Flag: TEST_ONLY
Test: atest MyRoboTests

*prebuilts/androidx/ink updated to 9387df44a94a221e52b354571fb4e54de7718178
Update Ink prebuilt .so

Change-Id: I0b7c5d4f43bb6c6807cbe5e69397816574ea8a8b