Autoupdate of 2 submodules
*hardware/generic/goldfish updated to c39ec5f8bda1a10be02587dbb370fb8c28f59a72
vsock: Enable adb through vsock proxy

This introduces a virtio-goldfish-hostfwd-socket that can be used to
proxy a host port to a guest port. This includes a port of the qemu
looper which will allow us to run the async sockets on the qemu threads.

This allows us to proxy port from the host to port 5555 inside the
guest, where the adbd lives.

Test: Can connect via adb shell.
Bug: 331947598
Change-Id: I048b41b10e7bbd37399506e38957defdb74abeff

*hardware/google/aemu updated to c6410b6d2ae6dcf93a383146a249466e79b0717a
Merge "Add support for sending in SimpleAsyncSocketAdapter" into main