Custommize mountdisks.BASE as well
Also use parted to get the offset and size of the partitions in bytes.
We need to know the size of the root partition (not just its offset)
because the end of the disk is gpt.

Bug: N/A
Test: cd package/modules/Virtualization/build/debian &&
sudo docker run -v $(pwd):/root/git --privileged -it ubuntu::22.04
bash /root/git/build.sh

Change-Id: Ic4367ff7fd8555ee89735e0cc1cd7f98ca5b41ff