Remove aidl "generator" modules
The generator modules were created during a defaultable hook, and
then later on they created the actual java/cpp modules. Creating modules
late in the build causes problems for mutators that expect to run
on all modules, as they won't be run on these modules. This affects
visibility enforcement of aidl modules, and blocks my cl for creating
a "base configuration" variant.

Instead, make it so that the defaultable hook creates the cpp/java
modules immediately. We then add the dependencies on the imported aidl
modules in a later mutator. Aidl rust modules already did something
similar to this, but with the rust.SourceProvider interface.

Bug: 368670280
Test: m nothing --no-skip-soong-tests
Change-Id: I5fcb39af4b3223f2dbb3ab20d6f50b34adab5d49