Switch crosvm base config from qemu-x86 to qemu-x86_64
This enables the X86_RUN_64 option, which makes the bootloader pass from
16-bit to 32-bit to 64-bit, rather than 16-bit to 32-bit. In 64-bit
mode, the bootloader can run 64-bit EFI applications.

Only enabled for crosvm because it gets stuck on QEMU at an earlier
stage, I think the transition into 64-bit mode. At crosvm it makes it to
a shell, but the MTRRs are in a state incompatible with Linux. This
change adds a command that initializes the MTRR state when the MTRRs are
accessible. In practice this only amoutns to `crosvm` because QEMU fails
on the `mtrr` command, and `mtrr` isn't available outside of x86.

Bug: b/365955006
Change-Id: I796faa2f42e09a1f82b55be6e4fdf7a70a8a88bf
Test: ../tools/bazel run //u-boot:crosvm_x86_dist --notoolchain_from_sources
Test: ../tools/bazel run //u-boot:crosvm_x86_64_dist --notoolchain_from_sources
Test: cvd create --bootloader=path/to/u-boot.rom
Test: cvd create --bootloader=path/to/u-boot.rom --vm_manager=qemu_cli