Decrease size of arena structure.
Remove as many of the stats as possible from the bin and arena. Also,
remove other parts of the structure that are not used by Android.

Remove tcache related data structures from the arena structure if the
tcache is disabled.

Original sizeof of arena_t in 32 bit: 15088
After all changes using low memory config: 11344

Test: bionic unit tests pass.
Change-Id: I68dd9290d72f30b4770557863933da7d7f6f9273