Remove usage of base::SetCurrentDirectory from avb tests
As part of the effort to uprev libchrome in Android, we are
removing Android dependencies to libchrome utility functions
that are not strictly required.

This change substitutes base::SetCurrentDirectory with its
implementation. It also removes an unused base::FilePath
from the same function.

Bug: 360917504
Test: atest libavb_host_unittest
Change-Id: I8497057ae58cd3114465cfbc3ca102427f9b9730
Merged-In: I8497057ae58cd3114465cfbc3ca102427f9b9730