ui: Improve usability of workspace API
Changes include:
- Removed `GroupNode`, the only type of node is now `TrackNode`.
  `TrackNode`s now contain group-like properties such as `children` and
  can be expanded/collapsed.
- Rename `displayName` to `title` everywhere.
- Normalize names of methods that add children to nodes, following the
  naming convention: `addChildXXX()`.

`TrackNode`s have a `key` and a `uri`. Since we use `TrackNode`s for
both tracks and groups, the `uri` must be optional, as most groups
don't have a track displayed on them. However, we still need a
consistent way to uniquely identify all nodes within the workspace for:

  - Locating standard groups so plugins can add tracks to these groups
    in a decentralized manner.
  - Scrolling to a track.

Take this `img` element as an analogy:

```html
<img id="foo" src="foo.png">
```

This element has an `id` which is used to identify this element within
a document, and a `src` attribute which identifies the file that
should be displayed. The element can still exist without a `src`
attribute, it just won't show an image.

The 'id' and 'src' of an `img` tag are analogous to 'key' and 'uri' of
a `TrackNode` respectively. TrackNodes have an id to uniquely identify
the node, and an optional uri to identify the track that should be
displayed on this node.

Note: For implementation reasons, a track's `key` is required and
immutable. If one is not supplied at construct time, a uuid is created
and used instead.

Change-Id: If166fbff94aba2f4a6b43fccc42408e947d2ba4c