Add separate visitor for handling references within a snapshot
The `CodebaseSnapshotTaker` relies on an `ItemVisitor` to determine
which `Item`s from the source `Codebase` need to be defined within the
snapshot. That will differ between signature `current.txt`,
`removed.txt` and stubs. However, `Item`s which are referenced from the
snapshot were hard coded to use the `NonFilteringApiVisitor`, which
allows references to any `Item` in the source, including hidden
classes. That is not suitable for stubs which may need to filter
inaccessible constructors from super classes before determining which
constructor to use.

This change adds support for the visitor to use for references to be
supplied to `takeSnapshot(...)` but keeps it as
`NonFilteringApiVisitor` for now.

Bug: 368021325
Test: ./gradlew
Change-Id: I44911a21e9ca999660bbcca6db8f9757a325fcf4