apksigner: support codename Baklava
As part of the signing operation, apksigner will extract the min SDK
value from the unsigned APK's <uses-sdk android:minSdkVersion="..."/>
manifest entry. If the min SDK value is a string codename, apksigner
uses a hard-coded mapping of first letter of the development codename to
corresponding numerical value. Wrapping the codenames back from V to B
confuses the logic.

Add an explicit check for the codename "Baklava".

Note: the codename mapping seems to have been left unattended from
around the Android O time and the implementation of
getMinSdkVersionForCodename should be updated or rewritten (in a
separate CL).

Test: m
Bug: 354023031
Flag: EXEMPT introducing new version codes can not be flagged
Change-Id: Icfedc4df24ab7945a02b7f6727fb2ef6a772b737