gbl: create libutils
We have a bunch of separate implementations of `aligned_subslice()`
scattered throughout the code, which have some slightly different
behaviors and are not tested.

This CL creates a libutils for common utility functions like this that
are needed from various GBL projects. Some utilities may be able to go
directly in other existing libraries, but for this function there didn't
seem to be anything at quite the right place in the dependency hierarchy
(depends on libsafemath and liberror, but not on libgbl or libstorage).

This will also simplify the future refactoring of moving Android boot
logic into GBL.

Bug: b/363074091
Change-Id: Ia9770d10658e4aa5a2110630359069ef1c7aea9f