ui: simplify handling of AreaSelection
Today Area selection contains a list of trackUris,
and then every selection aggregator ends up resolving them
by calling trackManager.getTrack().
Instead this CL does the resolution in selectionManager, before
setting the selection object. This allows getting rid to various
calls to globals.trackManager in the selection aggregators.
In turn this allows moving the wattson plugin from
core_plugins -> plugins, as now there is nothing special with
it anymore.
Also remove recording of action args in metatracing. Nobody is
using it (especially now that Actions is disappearing) and causes
noticeable jank with this change because I'm pushing the track
descriptors into AreaSelection, and in turn the PivotTableController
passes the area selection as an action argument. So we end up
calling metatracing.flattenArgs on all the resolved tracks for no
reason

Change-Id: I0c7f77d162c419aeb82fb69f4074a4e9f1d21613