fuchsia: update the prebuilts
Updates the Fuchsia IDK to a recent build. GBL will use a few of the
files contained here for ZBI structure definitions.

This code was last updated in 2019 and isn't included in the main AOSP
manifest, so presumably nobody cares about it anymore and we can do what
we want. On that assumption, this deletes all the old code and replaces
it with just the files we need in order to minimize the ~8GiB download
to a handful of MiB.

If needs come up for other files, this could be expanded to include more
or all of the IDK.

New build: https://chrome-infra-packages.appspot.com/p/fuchsia/sdk/core/linux-amd64/+/f22

NO_IFTTT=clobbering all preexisting files completely

Bug: b/369600577
Change-Id: Id6812e8fb6be3d93b6e10d8aa8e2b87636119cb4