Autoupdate of 2 submodules
*system/apex updated to 2d533c2ccfd5861e0dca82a38c809f1cec39a3a8
apexd: Use a new interface for VintfObject

For a group install (aka multi-package install), checkvintf should be
done with mounting all of staged apexes in a session. This change is a
preparation for upcomping change to run checkvintf with all staged
apexes.

To support that, PathReplacingFileSystem's interface is extended to
support multiple replacements.

Bug: 364999002
Test: ApexTestCases
Test: VendorApexTestCases
Change-Id: Ie70d18d7413779e19d496d0553a096734c8c3870

*system/libvintf updated to b4c708886574ef4a94b0819c062f3fcb3c2487a4
PathReplacingFileSystem with multiple replacements

This is to support multiple temp-mounts of staged apexes when running
checkvintf with staged apexes.

Bug: 364999002
Test: libvintf_test (tests added)
Change-Id: Ia3904997e54794f00da768ecd70af2f9059723f7