Replace actualItem with actualItemToSnapshot
The `actualItem` extension property was added for use by other parts of
Metalava to handle reverting changes. However, that was never going to
work long term as it is limited in which API changes it can handle and
is very fragile as it required modifying every place that used the part
of the API that was being reverted. Longer term the plan is to replace
those adhoc uses with the snapshot mechanism that will combine the
previously released API and the latest API into a single `Codebase`.

The snapshot mechanism took a copy of `actualItem` for its own use,
called `actualItemToSnapshot` and rather than keep both of them this
change replaces `actualItem` with `actualItemToSnapshot`. Once the
snapshot mechanism is in place those uses will be removed.

Bug: 370045964
Test: ./gradlew
Change-Id: I4d7efb24caa77c96e69f7c4e6ae527f562543171