Restrict use of framework-annotations
This filegroup target should only be used when the raw srcs are needed.
Module targets can use `framework-annotations-lib` as a libs reference,
if necessary. Otherwise, the annotations are bundled directly with the
framework.

This refactor is in anticipation of bundling keep rules with the
framework-annotations-lib target, instead of defining global keep rules.

Bug: 275175207
Test: m
Flag: EXEMPT refactor
Change-Id: I266501244ca372de4357999d485e8d1cfe0a2d9b