Remove usages of AddReverseDependency()
AddReverseDependency() doesn't work well with transition mutators.
(transitionmutators cannot be reversed) Most of the time the only
reason AddReverseDependency() works is because it will fall back
to adding a dependency on the only variant if there's only 1 variant
of the module, even if the two module's variations don't match.
To support incremental soong and changing global configurations,
remove usages of AddReverseDependency().

In this case, it was replaced with a SingletonModule that finds all
the aidl modules after the fact. This is also not good, finding all
modules in the tree has lots of problems, but it's at least a step
in the right direction. Removing the SingletonModule will likely require
behavior changes to the aidl modules.

Bug: 369916167
Test: Diff'd out/soong/.intermediates/system/tools/aidl/build/aidl_metadata_json/aidl_metadata.json before/after this cl and saw no diffs
Change-Id: Idb3fcba9475befa79bad60824688223bf977c2f9