Remove usages of AddReverseDependency()
AddReverseDependency() doesn't work well with transition mutators.
(transitionmutators cannot be reversed) Most of the time the only
reason AddReverseDependency() works is because it will fall back
to adding a dependency on the only variant if there's only 1 variant
of the module, even if the two module's variations don't match.
To support incremental soong and changing global configurations,
remove usages of AddReverseDependency().

In this case, it was replaced with a SingletonModule that finds all
the hidl modules after the fact. This is also not good, finding all
modules in the tree has lots of problems, but it's at least a step
in the right direction. Removing the SingletonModule will likely require
behavior changes to the hidl modules.

Bug: 369916167
Test: m nothing --no-skip-soong-tests
Change-Id: I5f485bddb506f698e6d82e1b87c41dd376ea50aa