Do not use static storage for translation primitives
If a guest program exits with multple threads running, static storage
objects may be destroyed before all pthread_key destructors are
finished. Since the latter are running the guest code we need to
guarantee that translation primitives are still alive. So we
allocate them on heap instead of static storage and never free.

Bug: 369726631
Test: tree-hugger, tests from the bug
Change-Id: I63a2ab26ab3eb4052ab209c95dbb9a02f27c7726