Create integration tests to compare atest behavior on prod and dev binaries
Add integration tests to compare the atest's behavior on its prod and dev binaries to make us aware of potential changes of behavior before a release is made. This CL only includes a default list of commands copied from the command verification tests to test. The next step is to use the input param to verify a much larger number of frequent user commands.

Test: atest atest_dry_run_diff_tests
Change-Id: Icf3460039850ce6517abdb79bdfb71e46fcb5ec0