Add TestFixture to encapsulate annotationManager
Some tests currently expect to use the `noOpAnnotationManager` and so
switching the tests to use `DefaultAnnotationManager` will break those
tests unless they can explicitly use `noOpAnnotationManager`. This
change adds `TestFixture` class to encapsulate the `AnnotationManager`
to use for the test and supply a default instance. It adds an optional
parameter to the `run*CodebaseTest` method to allow tests to override
the default.

A new `TestFixture` class is added instead of just passing an
`AnnotationManager` directly to make it easy to add other objects that
affect the behavior of the test, e.g. The work to move `PackageFilter`
into `metalava-model` so it can be used when parsing sources will need
to be able to supply a `PackageFilter` instance in order to test it
properly.

Bug: 370045964
Bug: 370463331
Test: ./gradlew
Change-Id: I10cf1733466dd47f9389421e842e7fc95e65513e