Resolve classes when checking imports should be included
Previously, the implementations of `getImports(Predicate<Item>)` used
`findClass(...)` methods to get an imported `ClassItem` and if one
could not be found then the import was ignored. That meant the imports
were dependent on whether a specific class had been previously
resolved which made the returned value fragile. This change switches
the implementations to use `resolveClass(String)` or
`findOrCreateClass(PsiClass)` to avoid that fragility.

Bug: 371955471
Test: ./gradlew
Change-Id: I1649dbda1ddc44b3a5208dcda43d69b361d1b501