Add test for filtering classes using a package filter
Previously, the API package filtering was only tested in the `metalava`
project, by checking the outputs, e.g. signature tests. This change
adds a test to show the effect of package filtering on the `Codebase`
and `ClassItem`. Basically, if a class is filtered out by the package
filtering then it behaves as if it was not specified on the command
line but was instead supplied on the source path. This tweaks the
existing source path test to make the two tests consistent.

Bug: 370463331
Test: ./gradlew
Change-Id: I2e264b53662fee896e053fa195fed80cbe2102d0