Separate NoOpAnnotationInfo from AnnotationInfo
Previously, AnnotationInfo was an open class that was used both as the
base class for `LazyAnnotationInfo` and also as the implementation used
by `NoOpAnnotationManager`. That slightly complicated the behavior of
each implementation. This change creates a `NoOpAnnotationInfo` for the
latter and changes `AnnotationInfo`into an interface. That necessitated
extracting the `computeTypeNullability(...)` method for use by both
implementations of `AnnotationInfo`.

Bug: 370698104
Test: ./gradlew
Change-Id: Iea4079d3fa6c3f7d1bc848f5855d6749b06cec0f