Delegate AnnotationItem.targets to AnnotationInfo
Previously, each `AnnotationItem` computed and stored its own `targets`
even though it was the same for all annotations of a specific class.
This change moves the computation and storage into `AnnotationInfo`.
The `AnnotationManager.computeTargets(...)` method was only needed on
`DefaultAnnotationManager` and so it was pushed down to there and
removed from `NoOpAnnotationManager`.

This will reduce the cost of computing and storing the `targets`.

Bug: 370698104
Test: ./gradlew
Change-Id: I82351fb710d625386ec7700fdd8b9fdd8d8291fd