Add a new known hash to drm
The types were split into another package so some newer devices are
using this new package.

The topic was cherry-picked but the branch configs made a mess. So we
are only cherry-picking this new hash for tests.

The new hash is added first, because the last hash in the file is
expected to the be current hash.

Bug: 293945485
Bug: 364817993
Test: build cuttlefish on this branch

Merged-In: I9e90ada03144b49245dc6b55a0fb2fa2eb6f3893
Change-Id: I6d3ef12301532d8b3fb8729abbbe6bd2dde86c91