Autoupdate of 2 submodules
*hardware/generic/goldfish updated to 6def97ba4eb78a74dcca0135cae64fb3c610f68c
Fix AsyncSocket concurrency issues

This fixes a series of concurrency issues for the AsyncSocket
implementation and it adds a series of unit tests to validate that it
behaves properly.

The main issue is that not every looper is thread safe. I.e.
registration of FdWatchers can only happen on the looper thread.

Bug: 368117031
Change-Id: I61cd9b4079d44e11713d644909171613b487d60d

*hardware/google/aemu updated to 60f73da9b07937eab0e3c34915f03a2b9b3f1b39
Merge "Fix aysnc socket concurrency issues." into main