Combine sourceFile() and getSourceFile()
Previously, these were defined separately which lead to bugs as their
behavior was slightly different. This change combines them by moving
the `ClassItem` implementations of `getSourceFile()` into
`sourceFile()` and delegating to the containing class for nested
classes.

This fixes a bug in the snapshot code due to `CodebaseSnapshotTaker`
calling `sourceFile()` to get the `SourceFile` but that method was not
overridden in `FilteringApiVisitor` which meant that the snapshot used
an unfiltered `SourceFile` which lead to invalid imports being listed
in stubs files.

Bug: 371164069
Test: ./gradlew
Change-Id: I8e96671b1607d4b2bbdb09e132821ff85c9f23ca