yukawa: Add --ramdisk_offset to BOARD_MKBOOTIMG_ARGS
When building boot.img using mkbootimg, we use the default ramdisk
offset of 0x1000000.

The boot.img information has as following:
0x11000000      ramdisk start
0x11080000      kernel start

Since the ramdisk is bigger than 512 KiB (0x00080000), there
is clearly an overlap between the kernel and the ramdisk.

This happened to work all right because U-Boot was ignoring the
ramdisk offset from the boot.img header.
However, a recent patch fixed this in U-Boot [1]. That fix was
incomplete for boot image v2 and has been completed with [2].

Because of the 2 fixes, boot is now broken on Yukawa because the ramdisk
gets overwritten by the kernel.

Change the ramdisk location by providing a --ramdisk_offset to mkbootimg.

[1] https://lore.kernel.org/r/all/20240519130951.2558467-1-r.stratiienko@gmail.com/
[2] https://lore.kernel.org/r/all/20241003-android-fix-boot-v2-v1-1-13e8e44af4b7@baylibre.com/
Change-Id: I59369dc2cea1e9e7a0227fa94841b806fb332b34
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>