Support fully qualifying snapshot of PsiItemDocumentation
Previously, calling `snapshot(Item)` on `PsiItemDocumentation` would
create a `DefaultItemDocumentation` containing the same text. The
problem with that is `DefaultItemDocumentation` does not support fully
qualifying the documentation which causes differences in behavior when
generating stubs from a snapshot.

This change implements `PsiItemDocumentation.snapshot(Item)` to return
`this` so that it supports fully qualifying. It is a temporary measure
until `DefaultItemDocumentation` supports it properly. That caused some
testsuite tests to fail in `metalava-model-snapshot-testing` that
called `duplicate(Item)` on the snapshot but did not pass it a
`PsiItem` so the cast failed. That was fixed in `duplicate(Item)`
by check the `Item` and returning `DefaultItemDocumentation` if it was
not a `PsiItem`.

Bug: 353737744
Test: ./gradlew
Change-Id: Ie502a4bf063571e3c6cc5a3c3c54cdf8f6871d5c