Add test to show that @deprecatedSince is treated as @deprecated
Currently, the check to see whether an item is deprecated using a
javadoc '@deprecated' tag just looks to see if the comment contains the
text `@deprecated`. However, that will match `@deprecatedSince` which
should not be treated as deprecating the item. This change adds tests
to show the behavior when `@deprecatedSince` is used, with and without
`@deprecated`. In the latter case the behavior is incorrect. A
follow-up change will correct it.

Bug: 368021325
Test: ./gradlew
Change-Id: I733c2d04cb660e67a05f99c20de815f7e8c12f66