Prevent @deprecatedSince from causing item to be deprecated
Previously, the check to see whether an item was deprecated using a
Javadoc `@deprecated` tag just looked for that text in the
documentation which meant it matched `@deprecatedSince` too. This
change adds `ItemDocumentation.hasTagSection(...)` which will check
to make sure that `@deprecated` string is either at the end of the
documentation or followed by a white space. It uses that to check for
`@deprecated` and fixes the previously incorrect test.

It switches the order of checks in `DefaultItem`'s init block to put
the `!modifiers.isDeprecated()` check first to avoid calling
`hasTagSection(...)` unnecessarily.

Bug: 368021325
Test: ./gradlew
Change-Id: I7036fb9c13a197c118c8e82e03a80031015650b0