Enforce total ordering when picking best stub constructor
Previously, the order of the list of `ConstructorItem`s passed to
`pickBest(...)` affected the result because the comparator that was
used did not define a total ordering so when two different
`ConstructorItem`'s compared equally it fell back to picking the
first.

This change adds the length of the parameter's erased types to the
criteria by which the `ConstructorItem`s are chosen. That was chosen
as the erased type is often used in the `super(...)` call and so the
smaller they are the smaller the stubs will be. If that does not
differentiate between `ConstructorItem`s then it will fall back to the
`CallableItem.comparator` which defines a total ordering.

That fixed the previously broken tests and will ensure that generating
stubs from snaphots will not result in different `super(...)` calls
than when generating directly from sources.

Bug: 368021325
Test: ./gradlew
Change-Id: I1105ba382a8484f26cef781f5a4526ed34dc0055