Autoupdate of 2 submodules
*build/blueprint updated to 32f934eb8046717612a4e440a0efd14d33ec256b
Remove the 1-variant fallback from reverse dependencies

The 1-variant fallback causes issues for the "base configuration"
feature and incremental soong.

This is a resubmission of aosp/3288584 with fixes to missing
dependencies for reverse dependencies. We had issues with missing
dependencies on mac os.

Bug: 369916167
Test: m nothing --no-skip-soong-tests
Change-Id: I817cbf8e95b3f65348b7471272200058b46faa33

*build/soong updated to b040e73cb5d54bdc3d1bd2655a8f4ed1376bdeaf
Merge "Use AddReverseVariationDependency to avoid 1-variant fallback" into main