Fix `VerificationResults` use in `OatWriter`.
Since
    https://android-review.googlesource.com/2236075
the `OatWriter` has been using a null `VerificationResults`
because `CreateOatWriters()` is called before initializing
the `verification_results_` in `Dex2Oat`. As the `OatWriter`
is using it only to select between two `ClassStatus` values
that currently lead to the very same behavior in the
`ClassLinker::VerifyClassUsingOatFile()` (but with different
comments and therefore prone to diverging), this did not
yet lead to any observable difference in behavior.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing.
Change-Id: I3f0b67497358ac98b4980691f80044a5b0f0fc17