Remove unused workaround controlled by PREPEND_LOCAL_CLASS
Among other changes, change https://r.android.com/775206 added a
workaround to deal with some issues doclava had with references to
members of the referencing class. That workaround was controlled by
`PREPEND_LOCAL_CLASS` which has been set to `false` since it was
added. As that workaround clearly has not been needed this change
removes the unused code.

Bug: 371997321
Test: ./gradlew
Change-Id: I800ed447f53e062e504208b839e8118ced090f48