Coalesce compatible mutators
Coalesce compatible mutators into a mutator group that as a single
visitor on each module.  This relaxes the mutator ordering to no
longer require that a mutator be finished on all modules before
the next mutator is run on any module.  A mutator on a module is
still guaranteed not to start before it is finished on all
dependencies.

Reduces analysis time from 2:07 to 1:50, with room for future
improvment as the rest of the mutators are fixed to be well behaved.

Bug: 372076859
Test: all soong tests pass with race detector enabled, no change to build.ninja
Flag: EXEMPT refactor
Change-Id: I03750e89bd66c62f85488a6c2f5939e1329532d4