Update directDeps immediately when adding new dependencies
Updating the list of dependencies was delegated to the outer goroutine
and handled at the end of the mutator in order to avoid interactions
between adding dependencies and propagating newly created variations.
Creating variations is now localized to TransitionMutators, and never
mixed with adding dependencies.

Add newly added dependencies to the directDeps list, as well as to the
forwardDeps list that is used to wait for dependencies to be ready
when running parallel mutators.  reverseDeps of the newly added
dependency doesn't need to be updated, it is only used in bottom up
mutators to propagate finished dependencies, and newly added
dependencies are guarnateed to be finished by the time the dependency
is added.  Similarly, waitingCount on the current module does not
need to be updated.

Bug: 372076859
Test: soong tests pass with race detector enabled
Flag: EXEMPT refactor
Change-Id: Ib2b4f4f7e4d9c539fc3b4bed0fb0e61fef9d1f3a