Load app image without holding the mutator lock.
Open app image files and decompress the contents without
holding the mutator lock. Note that we still acquire the
mutator lock for some parts of the app image loading even
though we're touching only the app image being loaded (not
yet part of the heap) and reading immutable data from boot
images that the app image depends on.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Bug: 330444460
Change-Id: Icb93c7cc6a6ee45d8184fc4c786e555b90f5ce4e