Don't keep storing NUD failures indefinitely.
The current code that stores NUD failures in the memory store
will stop storing NUD failures in the memory store if it already
contains 10 failures in the past 6 hours. This is intended to
stop the code from constantly storing failures and filling up the
database.

However, if there have not been 10 failures in the past 6 hours,
then the code will never stop writing to the database until
IpClient is restarted and queries the memory store again. Prevent
this by also storing the number of failures seen since IpClient
was last started, and applying the threshold to those failures as
well.

Bug: 329010113
Test: new unit tests
Change-Id: I78c0bb3ad10886bff2d83e8ae4be3b2ce618c8ad