Add DesktopModeSessionTaskUpdate TaskEvent for initialising state.
Statsd needs an initial value for the state field of this atom,
otherwise metrics that slice by the state will be bucketed in an
UNKNOWN bucket.

To not intefere with any other uses of this atom, we define a new
task_event just to mark the initialisation.

Bug: 370478775
Flag: EXEMPT logging only
Test: Manual + atest DesktopModeLoggerTransitionObserverTest
Change-Id: I23f9ad22486fa0c52f18a6c7622d7cdf4030c166
Merged-In: I23f9ad22486fa0c52f18a6c7622d7cdf4030c166