Autoupdate of 2 submodules
*build/blueprint updated to b1bb3f6c896fe14ced235ec987d342a33ef04922
Annotate mutators that use methods that prevent mutator coalescing

Mutator coalescing reduces the overhead of visiting every module
for every mutator by calling a series of mutator methods on a
a single module in a row.  This is only valid for well-behaved
mutators.  Add methods on MutatorHandle that allow annotating
mutators that are not well behaved, and use that to prevent
coalescing mutators.

Bug: 372076859
Test: all soong tests pass with race detector on
Flag: EXEMPT refactor
Change-Id: I44e28edb939451a983b3c300e08bd3855dfc6004

*build/soong updated to 7cf18df25318fcade9487db328e182f222eb567a
Merge "Annotate mutators that use methods that prevent mutator coalescing" into main