lib: trusty: trusty_app: Fix prg_hdr->p_memsz usage
Page align prg_hdr->p_memsz to find the first page of the next segment
instead of adding PAGE_SIZE. Adding PAGE_SIZE only works if p_memsz is
not page aligned. If the last segment is only one page large, and the
second last segment's size was page aligned, the check to see if there
is a mapping directly after the second last segment would fail. Then the
last segment would fail to map since it hit the guard page of the
previous segment.

Bug: 354981705
Change-Id: Ia5d2ef10d70bc6a885d63aa0fe1bdeb8db609f5d