Improve error message when failing to snapshot a class
A snapshot codebase should be a completely self-contained set of
classes such that every class referenced in the codebase is defined in
the codebase. If correct that means that resolving a class reference
(e.g. from an annotation or type) will always return a class. However,
just in case there is an error in processing this change replaces the
`!!` operator with logic to produce a more meaningful error message.

Bug: 357833584
Test: ./gradlew
Change-Id: I976e4946419a4a59250c671fcd02c84625c3aa5d