Add a package private constructor for file facade class
Previously, a file facade stub class would not have a package private
constructor, despite it not having an implicit constructor and should
not be instantiated. That was because the logic would only create the
constructor if the class had some constructors but they were all
filtered out or was pre filtered. That worked for most cases except
this one and it was hard to understand why it mostly worked.

The code tried to do two things:
1. Select a constructor to use, creating one if needed.

2. Optimize the memory footprint by using a shared `EMPTY` object to
   represent a class whose super class has an accessible no-args
   constructor and provides a no-args constructor for its subclasses.

Unfortunately, the optimization ended up preventing the file facade
stub class from having a default package private constructor.

This change restructures the code as follows to make it easier to
understand and fix this issue:
1. If the source class has a no-args constructor in the
   `ClassItem`'s list of constructors that ends up in the stub class
   after filtering then its subclasses do not need to generate a
   `super(...)` call so it is represented as `null` to make the
   existing optimization work.

2. Otherwise, it will always create a default private constructor
   to prevent an implicit constructor being added. That is what fixes
   the fiole facade stub class issue.

Bug: 372637655
Test: ./gradlew
Change-Id: I62e869c164f486f36274b8130eba57cca11973a1