MediaPreparer places a completion sentinel
MediaPreparer used "is the directory present" to decide if assets had
been delivered to the device; this was subject to corruption if
interrupted in mid-download. New scheme places a sentinel file in the
appropriate directory AFTER all assets have been pushed. This is robus
in the face of aborted runs while trying to install the test assets.

Update unit test to use our new signal for on-device media presence.

Bug: 372516779
Test: atest CtsMediaMuxerTest (for full content push)
Test: atest CtsMediaStressTest (for selective content push)
Change-Id: I3b2db62ee31368e7316af1cbd8fa26ff4d365b87