More minor optimizations to updateDependencies
Include duplicates in forwardDeps.  There is no need to avoid
duplicates, as forwardDeps is only used to propagate ready state
up and down the graph.  As long as each entry in forwardDeps has
a corresponding entry in the reverseDeps of the dependency then
the propagation and waiting counts will line up.

Also preallocate module.forwardDeps if it is too small.

Test: all soong tests pass
Flag: EXEMPT refactor
Change-Id: I260c823bc9ad3b9d511f942f9c859e37eabbb4fc