Fix incorrect behaviour for UrlRequestBuilder wrapper
This CL fixes two different methods that are both buggy which are:

`setCacheDisabled`: Cronet Engine API does not support taking a
parameter for this method, while HttpEngine supports that. In order to
correctly match the logic, the value of the param is stored and only
used when `build` is called.

`setDirectExecutorAllowed`: This is similar to `setCacheDisabled` and
has been fixed in the same way by caching the value and only calling the
appropriate API during `build`.

Bug: 372852416
Bug: 372653292
Change-Id: I2a02995342a7026ef51ee0b75ddb6035d2931e4d