verifier: Cache types by type index.
Speed up `RegType` lookup in `RegTypeCache` by adding a
cache based on `dex::TypeIndex` for types referenced by the
current dex file. This slows down the construction of the
`RegTypeCache` but the benefits of faster lookup outweigh
the slower construction now that we're constructing the
`RegTypeCache` only once per class and not once per method.

Remove the `MethodParamListDescriptorIterator` as it was
providing only descriptors from the resolved method's dex
file and we can get the same descriptors from the current
dex file based on the invoke's `dex::ProtoIndex`. This also
allows using the new `TypeIndex`-based cache for this work.

Clean up the `RegTypeCache` interface by inlining functions
`FindClass()` and `InsertClass()` to the only remaining user
`FromClass()` and update tests to use `FromClass()` instead.
Move the descriptor retrieval from callers to `FromClass()`.

Avoid some unnecessary string copy with `AddString()` in
`RegTypeCache::FromClass()`. Add a TODO comment to avoid the
string copy in `RegTypeCache::From()`.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Bug: 181943478
Change-Id: Iea429a33983104165c2ad2cf108d2b7d88f8b9e7