Do not disable adb_keys module type
The module can be added as a dependency, in which the reverse dependency
will fail with "depending on a disabled module" error. Given that the
module is setting `Disabled()` in `GenerateBuildActions()`, the reverse
dependencies does not have the information about whether the module will
be disabled or not in DepsMutator.

Test: m nothing
Bug: 371107666
Change-Id: I6a0a03a1e777bfb2655a55671696a54453b3475e