Autoupdate of 4 submodules
*frameworks/base updated to 4b757dafd40e7df6db322cf1bc2bd20590636ce7
Merge "Remove inconsistent @Deprecated behavior in *removed.txt files" into main
*packages/modules/Wifi updated to d13df95e26d90b92ee3acc5c196a7f900e0eac00
Merge "Remove @Deprecated inconsistency in *removed.txt files" into main
*packages/services/Car updated to 9a064c83601d80edc96ce73116adcdebce44e4b5
Remove @Deprecated inconsistency in *removed.txt files

Previously, `@Deprecated` was handled differently in `*removed.txt`
files compared to `*current.txt`. In the former `@Deprecated` on a
class was not propagated to its class members or nested classes but in
the latter it was. This change removes that inconsistency by making it
behave the same way in the `*removed.txt` files as it already does in
the `*current.txt` files.

Bug: 373584764
Test: m checkapi
Change-Id: I73f7d56dbfffa0f64801c8c9acdf50a1040f0d9c

*tools/metalava updated to 1cf269f3c2f96e6045805735805145b1e485dbbe
Merge remote-tracking branch 'aosp/metalava-main' into 'aosp/main'

Merge performed by:
  scripts/merge-from-metalava-main.sh

Changes includes in this merge (from newest to oldest):
* Propagate @Deprecated to class members in removed files
  - https://r.android.com/3308881

This merge includes a number of changes so this contains a list of all
the affected bugs in order from oldest to newest.

Bug: 373584764
Test: m checkapi
Change-Id: I118a24453a05de4e600aca976050346b67feb9d8