Autoupdate of 2 submodules
*frameworks/base updated to 4b757dafd40e7df6db322cf1bc2bd20590636ce7
Merge "Remove inconsistent @Deprecated behavior in *removed.txt files" into main
*packages/services/Car updated to 9a064c83601d80edc96ce73116adcdebce44e4b5
Remove @Deprecated inconsistency in *removed.txt files

Previously, `@Deprecated` was handled differently in `*removed.txt`
files compared to `*current.txt`. In the former `@Deprecated` on a
class was not propagated to its class members or nested classes but in
the latter it was. This change removes that inconsistency by making it
behave the same way in the `*removed.txt` files as it already does in
the `*current.txt` files.

Bug: 373584764
Test: m checkapi
Change-Id: I73f7d56dbfffa0f64801c8c9acdf50a1040f0d9c