Support testing the internal state of metalava main command
Previously, the `DriverTest.check(...)` method would only allow testing
of the outputs of the metalava main command, i.e. the output file
contents and stdout/stderr. This allows a lambda to be provided which
is called with some internal state of the metalava main command.
Initially, that is limited to the `Codebase` and `Options` but may be
extended in future.

This is needed to test some of the early changes in the API surfaces
work.

Bug: 374096206
Test: ./gradlew
Change-Id: Id4f3fdf4678edf24792aac138812be2355f3eb9c