Add ApiSurfaces and configure in Options
This adds a new `ApiSurfaces` class which tracks which API surfaces are
being managed and configures it in `Options` from the existing options.
This is needed to allow the `Item`s from the previously released API to
be associated with the correct API surface.

The main issue is determining whether a base API surface is required,
as that is not something that can be explicitly specified in the
command line and so has to be inferred. Longer term the whole structure
of the API surfaces WILL be explicitly specified in a configuration
file to allow the users of Metalava more control over how it behaves.

Bug: 374096206
Test: ./gradlew
Change-Id: Ib49e79894de8bc606c5b753d54b622f4cc285489