Set the default value for accept_ra_min_lft to 900s to watches
Assuming a single RA source on the network, this will result in RA
triggered wakeups to be at least 150s apart on WearOS devices (up from
30s).

This change also inlines the default constant value to make the code
more readable. Contrary to its annotation, it is not actually used in
tests.

Test: TH
Change-Id: I99647c9e333b4033c76766d36e3ca6f95ba80ce2