Fix read barrier assertions for app image loading.
Transitions between Runnable and suspended state can run
arbitrary checkpoints that can use read barriers, so we
need to restrict the `ScopedDebugDisallowReadBarriers` to
those parts of the code that contain no state transitions.

Note that these state transitions are actually unnecessary
as we're actually touching only the app image being loaded
and immutable class data in boot images that the app image
depends on. Once we remove these unnecessary transitions,
we shall move the `ScopedDebugDisallowReadBarriers` back
to the enclosing scope.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Bug: 330444460
Change-Id: I632de6a861817658e14c39ea34806561d4fab37b