Add API lint to flag value class definitions
Value classes can't really be used from Java. This lint can be selectively disabled in AndroidX for libraries that only target Kotlin clients.

This sets the issue to HIDDEN to avoid breaking androidx_with_metalava. Once the metalava version in AndroidX is updated to one which defines the issue, it will be possible to suppress it in AndroidX and it can be set to ERROR.

Bug: 292585793
Test: KotlinInteropChecksTest
Change-Id: I7a16ab90db52819b35f611d8e657090c8a0fda1d