Add --name option to the run command
It is used to optionally set the name of the VM. It can be useful when
there are multiple crosvm processes running and we want to distinguish
them. This is especially a must in Android which uses process name for
the resource attribution purpose.

BUG=b/375094126
TEST=tools/dev_container tools/presubmit
Test: ditto

Change-Id: Id4af3b8f2b216ad3239a4467a0c30a4630d44a73
Reviewed-on:
https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5955889
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Jiyong Park <jiyong@google.com>