Add Codebase.apiSurfaces
Although generally each `Codebase` created within a single invocation
of Metalava will track the same set of `ApiSurface` they need their own
reference to it, to avoid depending on global state.

This change adds `apiSurfaces` to `Codebase` and also adds it to
`TestFixture` so that it can be configured by test suite tests. It adds
a couple of tests that verify that the `Codebase.apiSurfaces` is set
correctly.

Follow-on changes will make more use of this.

Bug: 374096206
Test: ./gradlew
Change-Id: I8b37d2ad7ebbe9d81a1f6220d03f52b99030451f