Wrap AnnotationManager in Codebase.Config
Makes it easier to pass additional, configuration derived objects,
through to `Codebase`. This is largely a mechanical change that
replaces `AnnotationManager` paremeter with `Codebase.Config` which
contains an `AnnotationManager`.

Preparation for passing `ApiSurfaces` through to `Codebase`.

This was largely a mechanical change. The `DefaultCodebase` constructor
parameter of type `AnnotationManager` was replaced with one of type
`Codebase.Config` and then additional changes were needed to make the
whole of Metalava compile and run. That resulted in similar changes
propagating out through the call stack.

Bug: 375208805
Test: ./gradlew
Change-Id: I1b8f4cc2769af07b6a63605b0c64b05afdf21ba0