Remove special support for single SignatureFile
Historically, signature files were parsed separately but recent changes
have switched most code to using multiple signature files to handle
loading full APIs from multiple separate signature files, e.g. the
system API needs to load the public API from one signature file and the
additional system APIs from another signature file that just contains a
delta on the public API.

As a result of that transition there was a mixture of methods for
handling a single `SignatureFile` and a list of them. This change
removes/replaces the methods for handling a single `SignatureFile` with
the methods for handling a list. That simplifies follow-up changes that
need to change the parameters of those methods.

Bug: 374096206
Test: ./gradlew
Change-Id: I51b3671b832a2b46343dfe25195656668ff8ddb4