Avoid creating SignatureFile directly
Previously, a couple of places could not use the
`SignatureFile.fromFiles(...)` methods as they needed to control the
setting of `forMainApiSurface` so created them directly. This change
improves the `fromFiles(List<File>)` method to take a predicate and
stores its return value in `forMainApiSurface`.

This will make a follow-up change that will switch `SignatureFile`
from a concrete class to something more abstract.

Bug: 375160645
Test: ./gradlew
Change-Id: Idbd32fad3da1a446826923f74f0088c2e77541eb