Remove interpret_one_instruction from switch interpreter
This was used when mterp bailed out on executing an instruction. The
idea was to fallback to switch interpreter for that instruction and
continue with mterp. We no longer have mterp and removed the support for
it (aosp/1763167). We no longer use this capability and is always set to
false. So clean it up by removing this.

Test: art/test.py
Change-Id: Iaabf5d5afb240d7775b23d9b3fa8aee00a0c5660