Use `ArenaAllocator` in `RegTypeCache`.
Instead of `ScopedArenaAllocator`.

`RegTypeCache` is not using any additional functionality
provided by `ScopedArenaAllocator` (i.e. nested allocators).
However, `ScopedArenaAllocator` does not guarantee the
allocated memory to be zeroed (even though it actually is
if we don't use nested allocators), so we must clear it
explicitly which comes with a performance penalty.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Bug: 181943478
Change-Id: I0b7b7ff14c2c8fb203bc6f2a29095e17e3c65f00