Autoupdate of 2 submodules
*external/avb updated to c32891f55ee7e8d74bd9cad291ca671b28761d44
Remove dependencies on the 1-variant fallback

When adding a dependencies, if the variants don't match, but the
dependency only has 1 variant anyways, soong will always use that
variant. This makes it hard to add new variants to soong, because the
1-variant fallback stops being used and you start getting missing
variant errors. Make changes to bp files such that all dependencies
correctly specify the variant to use.

Bug: 372091092
Flag: EXEMPT refactor
Test: m nothing
Change-Id: I439f072a25ffa1b0b69a9174f003e48f823f3910

*packages/modules/Virtualization updated to 2b94409adc74d01ff74561c7723ab1231346d7ad
Remove dependencies on the 1-variant fallback

When adding a dependencies, if the variants don't match, but the
dependency only has 1 variant anyways, soong will always use that
variant. This makes it hard to add new variants to soong, because the
1-variant fallback stops being used and you start getting missing
variant errors. Make changes to bp files such that all dependencies
correctly specify the variant to use.

Bug: 372091092
Flag: EXEMPT refactor
Test: m nothing
Change-Id: Ief9e1962fb08349fa1c73d3f86db22a29d62ddf5