Remove Android.bp.
This crate does not appear to be used anywhere, at least by Soong.

It would be nice to delete it because it hasn't been updated upstream in
a long time, and is problematic because it includes undeclared GPL code
(which we have deleted).

Removing Android.bp is a first step to getting rid of the crate
entirely.

Bug: http://b/339424309
Test: treehugger
Change-Id: I4020e10b1b189ee06f15ff46c9cca7f7908a8fae