ui: Fix trace off-by-one in UiMain
Before this CL, this.trace was set only in oncreate.
But that is too late, because by that point, mithril
already called view() once. Creating view(), in turn
causes the instantiation of the sub-components.
As a result we ended up instantiating components with
an empty trace attrs on the first call (after the second
call, the non-null trace got passed correctly).
This creates surprising behaviour if the sub-component end
up binding the trace object passed oncreate. It doesn't happen
today but I hit this while writing a CL.
This CL fixes it by moving most of the logic in the constructor.

Change-Id: I7d45318ca4790d889dd815faadb8307a38e10b57