lib:storage: default pointer returns for error paths
Some of the functions take a pointer to be filled with a handle to an
opened file or directory state, or file size on success.

On error these pointers can also be filled to invalid values to reduce
the risk of uninitialised values being further used by (erroneous) calling code.

Bug: 375454649
Test: build.py
Change-Id: I1acce4cef79be8dcf7e6511a4775b35dd18d1679