Autoupdate of 2 submodules
*external/crosvm updated to 4bb58649839f21d089485996e51fd5cbf2c0a539
Add --name option to the run command

It is used to optionally set the name of the VM. It can be useful when
there are multiple crosvm processes running and we want to distinguish
them. This is especially a must in Android which uses process name for
the resource attribution purpose.

BUG=b/375094126
TEST=tools/dev_container tools/presubmit
Test: ditto

Change-Id: Id4af3b8f2b216ad3239a4467a0c30a4630d44a73
Reviewed-on:
https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5955889
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Jiyong Park <jiyong@google.com>

*packages/modules/Virtualization updated to 10ca6f99fe2d490213e4b523173c8745bd3e6a68
Merge "Give crosvm process a unique name" into main