Copy selectedApiVariants into snapshot
Copy the `selectedApiVariants` from the original `SelectedItem`s to
their snapshots. This switches `selectedApiVariants` from a `val` to a
`var` to allow the `ApiVariantSet` to be reused. This was chosen rather
than passing it in as a constructor parameter because this was simpler
and caused less changes and the constructor parameter would only be
used by snapshot.

Bug: 374096206
Test: ./gradlew
Change-Id: I018b68d9117d09122a5ec949464db3a10feeaf1c