Always provide Reporter to Codebase
Previously, providing the `Codebase` with a `Reporter` was optional
although it would fail if the `Reporter` was actually used. That was
a legacy of the `Reporter` being added recently and not providing a
way to tunnel it through to `Codebase`s created from signature files by
`ApiFile.parseApi(...)`. The latter was addressed by the changes in
b/375208805 so it no longer needs to be optional.

This change stops it being optional and uses `BasicReporter.ERR` as the
default value if none is provided.

Bug: 375208805
Test: ./gradlew
Change-Id: If57c905818e14e826d8ff060348acf99a87f9f0e