Avoid visiting ParameterItems unnecessarily in BaseItemVisitor
Currently, `BaseItemVisitor` will visit all `Item`s, which includes
`ParameterItem`s even though a lot of the visitors do nothing with the
`ParameterItem`s. This change adds support for controlling whether to
visit `ParameterItem` (enabled by default) and updates usages that do
not need to visit them to disable it.

Bug: 375592373
Test: ./gradlew
Change-Id: Ieae69d0683b09e2edb81bf5fdc25d11b64661a06