Update metalava to lint version 31.8.0-alpha08
Previously with K2, the `UExpression` for the default parameter value `Bitmap.Config.ARGB_8888` evaluated to null, so metalava used `CodePrinter.toSourceString` on the expression to turn it into a string. With updated K2 the expression evaluates to a non-null value, and metalava used `CodePrinter.constantToSource` on the computed value to turn it into an incorrect string.

This changes the condition determining wither to use `toSourceString` or `constantToSource` -- it previously only didn't use `constantToSource` on `Pair`s, because that is what K1 evaluated the expression to in this case. This updates it to only use `constantToSource` on primitives and strings instead.

Bug: 375651649
Test: ApiFileTest.`Default Values and Names in Kotlin` passes with this update, which it doesn't on the integration branch
Change-Id: I0ee836e240f8ce406566ccf0cd0ac961fbc1baf0