Remove checking for AnnotationItem in ComparisonVisitor
Previously, the `ComparisonVisitor` would check to see if an `Item`
was an `AnnotationItem` but that is actually impossible as neither
`AnnotationItem` nor any of its implementations implement `Item`.
It was some legacy code from a time when `AnnotationItem` did extend
`Item`.

This change removes that and cleans up a minor issue in the KDoc
reported by the IDE.

Bug: 375739126
Test: ./gradlew
Change-Id: I069ea5ee816737d19846966e3ea061d94cb06d3d