Switch ItemTree to store SelectableItem instead of Item
Previously, `ItemTree` stored `Item` because it needed to be able to
store `ParameterItem`s. However, that is no longer true so this
switches it to use `SelectableItem`. It also tweaks the
`addedItem(...)` and `removedItem(...)` methods to use `SelectableItem`
to make it clear that they will never be called with `ParameterItem`s.

Bug: 375739126
Test: ./gradlew
Change-Id: I1f045d65617790d492c267a0ac1c989ca918b3cf