Disambiguate ComparisonVisitor added/removed/compare method names
Previously, `ComparisonVisitor` `added(...)`, `removed(...)` and
`compare(...)` method names only differed by type. That made it
difficult to tell which one was being called and made it very easy to
accidentally select the generic `Item` one instead of the type specific
one. That was particularly true for `compare(...)` and `removed(...)`
methods which required both types were correct in order to call the
type specific method.

This change renames the methods to disambiguate them by name.

Bug: 375739126
Test: ./gradlew
Change-Id: I997eb2add5f1596e52e3eb9650c0ff1005f0dfe5