Correct removed...Item(...) "from" parameter types for non-packages
Previously, all of the `removed...Item(...)` methods for the
non-package `Item`s had a `from` parameter type that was nullable.
However, each of those types is guaranteed to have a parent from which
it was removed. e.g. `ClassItem` has a containing `ClassItem` or a
containing `PackageItem`, member items have a containing `ClassItem`.

This change stops them from being optional and removes any unnecessary
null checks in the overriding methods. It also narrows the type for the
`removeClassItem(...)` `from` parameter from `Item?` to
`SelectableItem`.

Bug: 375739126
Test: ./gradlew
Change-Id: I3d08e6c1ed9b9d5715340dceba900f82c8520131