Correct removedPackageItem(...) "from" parameter type
Previously, it was `Item?` but a `PackageItem` can only be removed from
another `PackageItem`. It could theoretically have no containing
`PackageItem` if it is the root `PackageItem` although there is
probably no way to do that in practice. This changes the type from
`Item?` to `PackageItem?` keeping it nullable just to be on the safe
side.

Bug: 375739126
Test: ./gradlew
Change-Id: If8aff96c838d716beeff3ce547849c691fb04df5