Property actualItemToSnapshot only makes sense on SelectableItem
Previously, `actualItemToSnapshot` was an extension property on
`Item` used to support reverting changes back to the previously
released API. However, it does not make sense on `ParameterItem` as
that can only be reverted if and only if the containing `CallableItem`
is reverted. This change rectifies that by moving it to
`SelectableItem` which does not include `ParameterItem`.

Bug: 375592373
Test: ./gradlew
Change-Id: I2f8d35d79717abd592320b950aa774f757b7ae82