Stop marking parameters as being part of the API
Parameters are only part of an API if and only if their containing
callable, i.e. method or constructor, is part of the API so there is no
point in tracking them separately. This is needed because
`markForMainApiSurface()` sets `emit` but a follow-up change will move
`emit` from `Item` to `SelectableItem`.

Bug: 375592373
Test: ./gradlew
Change-Id: I312034bcac32bed9caa0dd80c2747f94a5918a9d