Support host-cross variant on golang modules
This doesn't actually add support for cross-compiling go modules,
even when built on the cross-compile variant, the results will still
just be built for the current host.

However, it does make the variants of golang modules match up better
with other host tools that are being built for a cross-compile variant.

Bug: 369916167
Test: m nothing on aosp_cf_arm64_only_phone with aosp/3318965
Change-Id: I2245d09d2cfc9bcfb819d678cdae541dd87a441d