Revert^4 "IsMemorySharedMethod fix for intrinsics"
This reverts commit a5557d9da3e603a0531e566abf4d08579b9b8141.

Reason for revert: Fixed adding SetHotCounter in SetIntrinsic. We are
expecting both intrinsics and memory shared methods to have that set
as 0 i.e. hot. Intrinsics are always considered hot, and memory shared
methods use another counter. This is now DCHECKed in `Jit::AddSamples`.

Changes can be seen in patchsets 1..5. I also noticed that there were
comments in x86(_64) `s` files that needed to be moved to the right
place.

Test: art/test/testrunner/testrunner.py --host --64 -b --jit
Test: No timeout in presubmit-cts-deviceside-cloud (see b/78151261)
Test: No timeout in nexus_unit_test_multi_device_platform
Bug: 78151261
Bug: 162110941
Change-Id: Iee99f0d42fcdb3f5ab2b20b4b7a8126c7c03b358