ui: Close all related visualised args tracks on close click
This CL simply restored the previous behaviour of visualised args tracks
where clicking the close button on a single arg track closed all arg
tracks of the same name. This was broken a while ago and never fixed
as at the time it was not possible given the plugin architecture.

Change-Id: I66c0a10f24ae91ade0bb84c70a1bfced07cd6feb