Autoupdate of 2 submodules
*packages/modules/Virtualization updated to b007a822ed179a047f186f541f0a8a8c38447853
Merge changes from topic "switch-to-os-arg" into main

* changes:
  Add microdroid_16k kernel
  Update kernel to builds 12570979
  vm tool: rename --gki arg to more generic --os one
  Generalize our parameterized tests

*tools/tradefederation/core updated to 2933449e22aab4d0a853ca45a700ace066572dab
MicrodroidBuilder: deprecate gki() and introduce os()

The os() is a more generic concept that allows us to extend testing to
microdroid_16k kernels as well. Other patch in the topic renames the
--gki arg of the /apex/com.android.virt/bin/vm cli to the --os one.

We will keep the gki() one in case there are tests on Android V that are
actually using it (although it is highly unlikely).

Bug: 333730505
Test: presubmit
Change-Id: I604965a5b646e36099bbab5f15fec5f4d320b798