Remove stripJavaLangPrefix(...) method
Previously, it was used to support the `StripJavaLangPrefix.LEGACY`
behavior. This change moves that support into `appendTypeString(...)`
and removes the method.

The `stripJavaLangPrefix(String)` method would not strip the
`java.lang.` prefix if, after the prefix, it found a `.` before it
found a `<` or reached the end of the string. The purpose of that check
was to not remove the prefix from something like
`java.lang.reflect.Method` but it would also not remove it from
`java.lang.String...` or `java.lang.Thread.UncaughtExceptionHandler`.
It would however remove it from `java.lang.Comparable<String>...` as it
found the `<` before the `.` of the varargs ellipsis.

Replicating that was tricky because the logic to remove the prefix is
implemented for `ClassTypeItem` outermost class  but it does not know
whether it is part of a varargs type or part of a nested type.
Conversely, the `ArrayTypeItem` which does know whether it is part of a
varargs type and `ClassTypeItem` nested class do not know how to remove
the prefix.

That was fixed by having `ArrayTypeItem` replace `LEGACY` with a
special `VARARGS` value in the `TypeStringConfiguration` that it passes
down to the `ClassTypeItem`. `ClassTypeItem` then knows it is
processing a varargs type and can make sure to strip the prefix if and
only if it is at the start and is a generic type.

Similarly, a `ClassTypeItem` nested class will replace `LEGACY` with
`NEVER` in the `TypeStringConfiguration` that it passes down to the
`ClassTypeItem` for processing the outermost type. That will prevent it
from stripping the prefix for any nested class type.

Bug: 321061298
Test: ./gradlew
Change-Id: Ica266f8f81fd047ad1af958a636872c8cdc95932