Expand CommonTypeStringTest to test toCanonicalType()
Previously, it only tested `toTypeString(...)` with a variety of
configurations. This change extends it, to support specifying the
test method to use and then adds some `ConfigurationTestCase`s to the
existing test cases to show the behavior of `toCanonicalType()`. There
is one test case where `toCanonicalType()` does the wrong thing but
that will be fixed in a follow-up.

Bug: 321061298
Test: ./gradlew
Change-Id: Ibb7de611d881c19a97dc52e3c79449dee6483334