Add tests to show java.lang. prefix stripping edge cases
The `stripJavaLangPrefix(String)` method will not strip the
`java.lang.` prefix if, after the prefix, it finds a `.` before it
finds a `<` or reaches the end of the string. The purpose of that check
was to not remove the prefix from something like
`java.lang.reflect.Method` but it will also not remove it from
`java.lang.String...`.

It will however remove it from `java.lang.Comparable<String>...` as it
will find the `<` before the `.` of the varargs ellipsis.

It will not strip it from `java.lang.Thread.UncaughtExceptionHandler`
as the `.` after `Thread` will be treated as a different package even
though it is a nested class.

This change adds tests to verify that behavior is maintained as it is
relied upon downstream.

Bug: 321061298
Test: ./gradlew
Change-Id: Ifdefd00d0ef9efaf1366b24679a4bff5a1458f45