Improve Gpt buffer ownership flexibility
Currently, in order to create a GptCache, users have to allocate a
buffer, maintains it separately and pass the borrowed slice `&mut [u8]`
to `GptCache::from_uninit()`. This often results in mulitples data
structure needed. This CL changes the buffer type into a generic type that
implements DerefMut<Target = [u8]>, so that it can either be a borrowed
buffer, or vector, or other custom buffer object.

GptCache is renamed to Gpt. Also removes the parameter of `max_entries`
and instead let it be drived from the provided buffer size.

Bug: 376693338
Change-Id: If8ee0cc8e14bfd4441d9cef260ce03c7c82edce0