Move filter from TypeStringConfiguration to TypeStringParameters
Previously, `TypeStringConfiguration` in `CommonTypeStringTest` had a
`filter` property because it was a copy of `TypeStringConfiguration` in
`TypeItem` which also had it. However, since then `filter` was removed
from the latter by https://r.android.com/3149628 so they are not quite
the same anymore. This change removes `filter` from the test copy in
preparation to dedup-ing the two classes.

Bug: 376974661
Test: ./gradlew
Change-Id: I7aed905d0910c3f066053845eb6531aa5f82e26e