Make toCanonicalType use toTypeString(stripJavaLangPrefix=NEVER)
Previously, it would call `toTypeString()` and replace "java.lang."
with "". That was incorrect as it would remove the prefix from
something like `java.lang.annotation.Annotation` which results in an
invalid type. This change switches it to use
`stripJavaLangPrefix=NEVER` which handles that case correctly.

The slight behavioral difference should not cause any problems because
it is only used in two places:

1. To compare types by comparing their `toCannicalType()` result so
   two types that were previously different would still be treated as
   different and two types that were previous the same would still be
   treated the same for the most part. The exception would be if there
   were two types where one was simply the second with `java.lang.`
   prefixed, e.g. `annotation.Annotation` and
   `java.lang.annotation.Annotation`. That would be incredibly unlikely
   and would have caused other problems to be reported, e.g. removed
   methods.

2. In `ProguardWriter.getCleanTypeName(...) to get the type of a
   non-array, non-class type, e.g. primitive types or type variables,
   neither of which will have a `java.lang.` prefix.

The difference in behavior did break a test added in the previous
change so that was fixed.

Bug: 321061298
Test: ./gradlew
Change-Id: Ib56237905172164de0c5470285ce2a293547c037