Reuse TypeStringConfiguration instances
Previously, production code used the multiple parameter method
`TypeItem.toTypeString(...)` which created a `TypeStringConfiguration`
on every call. However, each use site used the same settings for
multiple calls meaning that it was unnecessary to create a new instance
on every call. This change refactors each use site to create and reuse
a single instance. That includes the `toTypeString()` call which uses
the `DEFAULT` instance.

Note: Both the multiple parameter `toTypeString(...)` and the single
parameter `toTypeString(TypeStringConfiguration)` provide defaults for
all their parameters so they both match `toTypeString()`. However,
instead of reporting an ambiguity as was expected it appears to be
resolved to use the `toTypeString(TypeStringConfiguration)`.

After this change the multiple parameter method is only used by tests
so it will be moved out of `TypeItem` into a test fixture extension
function. It will also be renamed to avoid any ambiguity.

Bug: 376974661
Test: ./gradlew
Change-Id: I12e1fe296b127d7175ffecc4fbbcb88bde5ec20f