Move multiple parameter TypeItem.toTypeString(...) to test fixture
This method is only used by testing so this change moves it to be an
extension function in the test fixture. It renames it to
`testTypeString(...)` to make it clearer that it is not using the
`TypeItem.toTypeString(TypeStringConfiguration)` method. It also
removes any parameters that are unused by callers.

Bug: 376974661
Test: ./gradlew
Change-Id: I27904f54cc5bfc83e3c36752c58e46272c7e26ab