Add TypeStringConfiguration.eraseGenerics
Adds support to `toTypeString(...)` to erase generics from the string
representation and uses that in `toErasedTypeString()` and removes its
specialized implementation. This will make it moderately slower as
the `toTypeString(...)` has more complicated code paths but the result
is still cached so it should not have a measurable impact on
performance.

Bug: 321061298
Test: ./gradlew
Change-Id: Ifc08c418068d3e4e4b207bdc4f4f07b13f343d16