Reduce cost of maintaining TypeStringConfiguration.isDefault
Previously, the `TypeStringConfiguration.isDefault` property was
computed by testing each property to see if it matched the default and
if they all did then it would be `true`, otherwise it would be `false`.
The problem with that is adding new properties required changing the
expression and each expression effectively duplicated the default
values for the property.

This change avoids that by using the fact that
`TypeStringConfiguration` is a `data class` which has a generated
`equals(Object)` that compares all the properties and comparing it
against a known `DEFAULT`. The value is computed lazily to avoid the
comparison against `DEFAULT` being done while creating the instance to
assign to `DEFAULT` at which point `DEFAULT` would be `null`.

Bug: 376974661
Test: ./gradlew
Change-Id: I6e219e69d01fe199e28e691bea256f72b7e836ef