Wait and retry when class-pointer is null during marking
Wait upto 1 second retrying to load class-pointer of the object being scanned
during CMC's marking-phase. In case of CC it was found to workaround similar
issue observed in there.

For now, we are going to crash even if we eventually find a valid class
pointer for debugging the root cause.

Bug: 373609505
Bug: 375544858
Test: art/test/testrunner/testrunner.py --host
Test: monkey test
Change-Id: Ideeec6b4c72d1be87bec14b434a9777b08d3b253