Disallow checking compatibility against .jar file
Historically, compatibility was checked against previously released
`.jar` files but that is no longer true so this change disallows
providing one. It fixes up the tests, removes `JarBasedApi` and removes
the `jarLoader` parameter from `PreviouslyReleasedApi.load(...)` as it
is no longer used.

Now that previously released APIs can never use `.jar` files it is no
longer necessary to highlight that they could not be used with
`--revert-annotation`.

Bug: 377445833
Test: ./gradlew
Change-Id: If91589f4828b253414eb37725666a7fedde652d4