Check compatibility against signature files not .jar files
Previously, the `CompatibilityCheckAndroidApisTest` checked against the
previously released `.jar` files. However, that is not a realistic test
as Metalava is never used that way (it used to be but not for a while
now). Instead it always checks compatibility against signature files.
This change switches the test to do that in preparation for removing
support to being able to check compatibility against `.jar` files
altogether.

This has the added advantages of making the tests substantially faster
and stops them failing on Turbine.

Bug: 377445833
Test: ./gradlew
Change-Id: Ifa3c40a4921117bdb2d1ce1ec626d9fb409e094e