ANDROID: Rename libbit_field to prevent conflict
As ag/26373528 (commit ac9261cc7b17552ff974957f4c6007bc7021b14c) added
the crate to AOSP at external/rust/crates/bit_field, for which
cargo_embargo will generate the default Soong module name libbit_field,
rename this fork local to crosvm to avoid conflicting with that name.

Ideally (and if it makes sense with the current build setup), this
commit will be followed by a clean-up making crosvm depend on that newly
introduced AOSP fork of bit_field.

Re-generate the blueprints with

  cargo_embargo generate cargo_embargo.json

and add the result to this commit.

Test: cargo_embargo generate cargo_embargo.json
Change-Id: I6642498050ba68da74bf1bd1d5b74b91e10c12dc