Remove trendy_team_desktop_hwsec from trusty/user/desktop/.
trusty/user/desktop is not present in partner Android builds, which prevents this team from being used from external/gsc-utils. After this is merged, I will send a second CL to add the team definition to external/gsc-utils, which we currently believe will make it visible everywhere it needs to be visible.

Bug: 359274525
Test: `mm .`
Change-Id: Ie0948fe387aa9ad8ff4f93922fd0808f29785c72