Move fastboot variable getters into GblFastboot
Moves fastboot variables getters into GblFastboot as private methods.
This gets rid of the macro `fb_vars_get_body`/'fb_vars_get_all_body`
which have not been intuitive to understand and also avoids dozens of
repetition of type GblFastboot in function signatures which have to be
updated every time there is a change in the definition.

Change-Id: Icb757a7bb9025f1b23456609480b6e7a9d42c651