kleaf: Move _trim_attrs() out of kernel_config_settings.
Now that the value of trimming is not explicitly mentioned in
kernel_config_settings(), it does not belong in
kernel_config_settings.bzl. Move it out to individual rules.

This change also removes trim_nonlisted_kmi attribute from kernel_env
and modules_prepare since they are no longer used. The attribute was set
in kernel_env and modules_prepare only to calculate the progress_message
properly. Now that it is already included in post_defconfig_fragments,
the attributes are no longer necessary.

Bug: 298268266
Change-Id: I4fb841eae203dbdc1fc1e34c4b1ee97100dec3f5