Fix use-after-free in dex2oat mini-debug-info compression.
This might happen if the device runs out of disk space,
which causes the WriteOutputFiles to return early and
dex2oat starts to shutdown and free memory before exit.

However, the mini-debug-info compression thread still
keeps running, so there is short window of time where
it might access freed memory before dex2oat exits.

This CL uses RAII to ensure that the compression task
finishes before return from the frame that started it.

Bug: 375314886
Test: test.py --host
Change-Id: I69b679551477de5e3df9c0867daefd8e22c52833