create security context for aconfigd-system rust binary and aconfigd_system socket.
aconfigd-system is the rust replacement for cc_binary aconfigd. And
aconfigd_system socket is the replacement for aconfigd socket. The
transition will be flag guarded, only one daemon will be turn on based
on the flag. The binaries and sockets should have identical policy.

The rust binary will implement the main function logic including socket
messaging marshaling. The detailed logic to handle a request is still in
a cc_library for now. Will conevert it to a rust_library later.

Bug: b/327170179
Test: m
Change-Id: I70975833d229bf04bcb8e2cc0970e69a63b1c366