Add selectedApiVariants test for a system API delta
Adds a test to show what happens when loading a `Codebase` from a set
of signature files that includes a base signature file (corresponding
to the public API) and a main signature file (corresponding to the
system API delta).

The base test fails in `metalava-model-snapshot-testing` because it
only takes a snapshot of items with `emit=true` but items loaded from
the base API surface have `emit=false`. That will be fixed in a
follow-up change.

Bug: 374096206
Test: ./gradlew
Change-Id: I9a14bc3d1bfb50427acc48c86e7ae69ec2ae3b23