Take a snapshot of every item in metalava-model-snapshot-testing
Previously, it would only take a snapshot of items with `emit=true`.
That was not an issue before as there were no tests where items with
`emit=true` needed to be copied. However, a preceding change added a
test that checks the `selectedApiVariants` state of all items in a
`Codebase` loaded from a base and a main file. Items loaded from the
base file will have `emit=false`.

This change stops the `SnapshotCodebaseTransformer` from filtering
out items with `emit=false` which fixes the test.

Bug: 374096206
Test: ./gradlew
Change-Id: I52e821da58196eeeb9d32daab861b2603873d3c8