Call C++ aconfigd implementation from Rust binary
To begin migrating the system aconfigd to Rust, make a Rust binary that
reads and writes socket messages, and defers the rest of the work to the
C++ aconfigd implementation.

The Rust binary is not used anywhere, so this CL introduces no behavioral
change yet. Follow-up flag-guarded CLs will start the binary from init.rc.

Test: manually confirmed that platform and storage files are created
Test: manually confirmed that socket can send and receive messages
Bug: 370499640
Change-Id: I41590115fda4cad69a9c0dd60253fce460b75c05