Pin to C17.
The vmlinux.h files contain #defines of bool/false/true which are all keywords in C23, and it's non-trivial to update those.

Even if/when that happens, it would probably be worth trying to find out what C version upstream expects to build against, and permanently pin to that. Some adjacent code appears to target C89 (!) but I couldn't find anything for this specific set of files.

Change-Id: Id1e7a47395ee0a842b70b60331e175c31f276dc7