Autoupdate of 2 submodules
*build/make updated to bfa0a82d3da7d4e4bcfbfe2b438ee42a67962bc7
Merge "refactor" into main
*packages/modules/ConfigInfrastructure updated to 6a252fc018378a63a01735d79f9054be76241ad1
Add AconfigPackageLocal

This change add AconfigPackageLocal. This class is intended only to be
used by generated code, and should be only used by aconfig team internal
to read the flag locally.

The class doesn't use the same implementation as AconfigPackage to
reduce the call stack, so that the execution is more efficient

Test: atest AconfigPackageTests
Bug: 367765164
Flag: NONE not api
Change-Id: I2876263630901dfe3113892467369e96a95692a6