Remove obsolete libc++ patches
Some of these patches (e.g.
avoid-triggering-fdsan-in-filebuf-test.patch) are very old, and while
they try to make the libc++ tests pass, the tests were marked with
"XFAIL: LIBCXX-ANDROID-FIXME" when Android support was officially added,
so they may unexpectedly pass (which would be a failure). The libc++
tests from the llvm_android build aren't automatically run on Android,
so these failures wouldn't be noticed.

hide-locale-lit-features-for-bionic-v2.patch does not apply cleanly to
ToT LLVM, but I think it's unnecessary for the older LLVM versions too.

Bug: http://b/378560392
Test: build.py, run tests manually on Android
Change-Id: I91a5bd7a6cf8b84a54d03cdb368aa46f0edaf344