verifier: Rewrite representation of constants.
There is no reason to keep precise constant values for any
verification checks. Instead of the precise and imprecise
constant types, define concrete register types for constant
ranges that we actually care about. Note that we currently
do not guard against narrowing conversions, so even the
information we keep is effectively unused for now, but we
shall need this to fix bug 270660613.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Bug: 181943478
Bug: 270660613
Change-Id: Ib6b4d077cd179580eee9ec7e938eb4318e4e55f1