Use null as the default value for --current-version
Previously, the default value for `--current-version` was -1. That made
it easy to accidentally use an unspecified value. This change switches
it to use `null` as the default value and fixes up the code to make
sure it is either using a specified value, or is purposely using an
uninitialized default value.

Bug: 378479241
Test: ./gradlew
Change-Id: I39bc148de6be97588609a5a84ab884479b7efab8