Clean up handling of jar patterns when generating XML API levels
Previously, the `findAndroidJars(...)` would take an args array from
which it would extract the `ARG_ANDROID_JAR_PATTERN` arguments.
However, that had already been done to populate the
`androidJarPatterns`. The only difference is the latter would have some
defaults appended that did not come from `--android-jar-pattern`. This
change uses the `androidJarPatterns` list and explains that the last
two entries are defaults.

This change is necessary because once `--android-jar-pattern` has been
converted to a Clikt property the `--android-jar-pattern` options will
no longer be in the `args` passed to `Options`.

Also, cleans up some stale comments and a stale warning suppression.

Bug: 378479241
Test: ./gradlew
Change-Id: I2c650a4dfb5d4a3f5c71fc544aa174e798e342c8