Add general mechanism for printing API levels and use for XML
Previously, `createApiLevelsXml(...)` would create an XML
representation of an `Api` instance. This change refactors that to
use a generic `ApiPrinter` and passes in an `ApiXmlPrinter` to write it
as XML.

Follow-up changes will move the XML printing code embedded within `Api`
and `ApiElement` into `ApiXmlPrinter` and refactor the Json printing to
use the same mechanism.

Bug: 378479241
Test: ./gradlew
Change-Id: Id807efde79498f9796ecd32c69361452d4c04445