Remove Api.mMin and pass it in to Api.print(...) method
Previously, `Api.mMin` was initialized in the constructor to the first
API level included in the `Api` but it was only used in the
`Api.print(...)` method. This change removes the property and instead
passes it to `Api.print(...)` method. That necessitated adding a
constructor property `ApiXmlPrinter.firstApiLevel` and passing that
through to `Api.print(...)`.

Bug: 378479241
Test: ./gradlew
Change-Id: If07caa73172b5afd4e87660ebc5fa704d7711cb6