runtime: arm64: Fix inline assembly
The runtime library on sdk_phone64_arm64_riscv64 was not building due to
errors in the inline assembly related to stack pointer arithmetic.  The
add and sub instructions were missing the rs1 operand.  The "J"
constraint used for the callee-saved stack size has a limited range into
which kCalleeSavedFrameSize does not fit.

Bug: 346603097
Test: sdk_phone64_arm64_riscv64 builds
Change-Id: Ifedcd93f1e9d03607d57d332cf83ca23546bd444