Initialize SensorDevice struct members to prevent test flakiness
This change initializes the members of the Sensor and SensorDevice
structs to their default values. Previously, these members were left
uninitialized, which could lead to test flakiness in certain build
configurations due to the use of uninitialized variables. This commit
ensures that all struct members have predictable initial values,
improving the reliability and stability of the tests.

In addition, this commit removes the unused int2hex function, further
cleaning up the codebase.

Change-Id: I083eed9a5e4298e3015619eaa9d61e1d27d259f1