Generalize using a file or string contents in DriverTest.check(...)
Previously, `DriverTest.check(...)` would allow signature files to be
provided as either string contents or the path to a file. This change
extracts a `useExistingFileOrCreateNewFile(...)` method that can be
used for any file type. It moves the signature specific behavior of
trimming and adding a header into a `transformer` lambda.

A follow-up change will use the new method to allow the
`applyApiLevelsXml` to be provided as a file.

Bug: 378479241
Test: ./gradlew
Change-Id: I42199a3ec28eea7a787b7e84914e12e8441c5245