Fix applying api-versions.xml file that has just been generated
Recent changes to move the `--generate-api-levels` from `Options` to
`ApiLevelsGenerationOptions` broke the logic that allowed the
`api-versions.xml` file that was just generated from being applied to
the documentation stubs. That logic relied on the
`--generate-api-levels` being in the `args` given to
`Options.parse(...)` but moving it meant that was not the case. That
was not detected because there were no tests that checked that
behavior.

This change modifies an existing `ApiGeneratorTest` test method to
apply the `api-versions.xml` file that has just been generated. That
broke with `.../api-versions.xml is not a file` because
`--apply-api-levels` was expecting an existing file. This change then
fixed the logic.

Bug: 378479241
Test: ./gradlew
Change-Id: I118bbe1423d1d68a63445592249ef27e9dc82755