Call findAndroidJars(...) on demand
Previously, `findAndroidJars(...)` was called by `Options.parse(...)`
if and only if the `generateApiLevelXml` property was not-null and its
return value was stored in the `apiLevelJars`. That meant the latter
was always non-null if `generateApiLevelXml` was non-null. However,
`apiLevelJars` was nullable which meant that any code that used it had
to check it was not null even if it had checked `generateApiLevelXml`
was not null.

This change only calls `findAndroidJars(...)` when `apiLevelJars` is
accessed, changes it to be non-nullable and simplifies the code that
accesses it.

Yet more preparation for moving `findAndroidJars(...)` into
`ApiLevelsGenerationOptions`.

Bug: 378479241
Test: ./gradlew
Change-Id: I404072c0c12ee1c8ae5f42f701c9178e7455d32d