Don't use echo -e
A rule that uses echo -e may execute in a dash shell instead of bash
when running on the build servers.  The dash echo builtin is posix
compliant, and does not treat -e as an option[1].  Replace all uses
of echo -e with echo if it does not need escape sequences, or printf
if it does.

[1] https://wiki.ubuntu.com/DashAsBinSh#echo

Bug: 378931009
Test: manual
Change-Id: I2974f59ea10a613bc3bd36e3591fd81f38a1769a