UPSTREAM: virtcpufreq_v2: Use a shared perf for vCPUs in the same freq_domain
Returning the current frequency of the physical cpu is going to inflate
sizes of tasks inside the guest when the vCPU's performance is at a
point below pCPU@FMin. Using a simplified example, if the FMin of the
pCPU is 1000Mhz, a task that took 5ms to run @ 400Mhz will instead be
accounted as if it ran at 5ms @ 1000Mhz, resulting in a much larger
and incorrect utilization value according to the guest scheduler.

In the shared freq_domain scenario, linux will select one policy_cpu to
write when selecting performance. This performance state needs to be
shared between the vCPUs in the same domain to report the correct
performance to the other vCPUs that are frequently reading its own
performance state. So to ensure correct utilization tracking of tasks
inside the guest for all vCPUs, return the current domain's requested
performance when perf dips below pCPU@FMin.

Bug=b:284367589
Test: tools/presubmit
Change-Id: Ie4cb65426dd0a1555f8413190daf6027da4a34e5
Signed-off-by: David Dai <davidai@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/6020998
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>