Preallocate array and map stores during parsing.
This preallocation was removed in aosp/2494034 because fuzzing
found that malformed CBOR could cause it to attempt to preallocate
too much, causing a crash.  Not preallocating doesn't eliminate
the crash if the CBOR actually contains more data than can be
stored in an allocation, but defers it.  But that's a reason not
to preallocate an unbounded amount, not a reason to avoid the
optimization entirely.  The argument that Android use cases of
libcppbor are not performance-sensitive is weakened by the fact
that libcppbor is now being used in google3.

This CL restores the preallocation but limits the maximum size.

Bug: 273450455
Test: atest cppbor_host_test

Change-Id: I41d5531e7fd4589199dcb6abb263308361f4f9bd