Add base listener to delegate test events to custom listeners.
Introduce a new TestListener class that acts as a proxy for delegating
test events to custom listeners. This removes the need to register each
listener with the test runner. This also bypass the JUnit 4 limitation
of supporting only hardcoded listeners in the test manifest file.

Rename IgnoredAnnotaionListener to IgnoredTestsReport and move it to a
new listeners package consisting of additional custom listeners.

BUG: b/377927896
Change-Id: I1c93538b95d8689b572cfce485e561a7a33f2391