Move contents of ModuleDefaults.bp to Android.bp
There's only 1 file in the `build` variable of the Android.bp file,
and no other modules in the Android.bp file, so it doesn't really make
sense to be a separate file.

Also having a `package` module in a non-Android.bp module is relatively
complex and may have suprising side effects, such as applying to
other .bp files in the same folder processed after that one
but not to the files processed before.

Bug: 336820042
Test: Presubmits
Change-Id: I5b06766c24cc160b53b15ee6efeb39e591f9dcc8