Autoupdate of 3 submodules
*frameworks/native updated to b9bbaadf952047242ddd34619806a408b0add9d3
Merge "gpuWork: remove unused reference to mock_bpf" into main
*system/bpf updated to b01b5c8a926d82d8d8b42075e20707dd6030a788
MOCK_BPF: Remove bpf mocking library

Removing the bpf mocking system because we feel that it is not
useful in properly testing bpf programs. There is currently only one
real user of it (time_in_state_test.cpp), which has significant issues
in that the tests are written in such a way that each one relies on the
state of the previous test. Also, the per CPU mocking capability is
limited in its capabilities.

Removal simplifies the transition to libbpf, and reduces legacy tech
debt.

Bug: 359646531
Test: TH
Change-Id: I402a6dec8e8b750161bf36fea2340cf54a1b7433
Signed-off-by: Neill Kapron <nkapron@google.com>

*system/bpfprogs updated to 7ae341829284928137f93ce1b7639653ba674d52
MOCK_BPF: Remove bpf mocking library

Removing the bpf mocking system because we feel that it is not useful
in properly testing bpf programs. There is currently only one real user
of it (time_in_state_test.cpp), which has significant issues in that the
tests are written in such a way that each one relies on the state of the
previous test. Also, the per CPU mocking capability is limited in its
capabilities.

Removal simplifies the transition to libbpf and reduces legacy tech
debt.

Bug: 359646531
Test: TH
Change-Id: Ic2e63ab33266e98a296ed33bddba9f3bb0029fc2
Signed-off-by: Neill Kapron <nkapron@google.com>