Don't reuse the old snapshot after a long running bg dexopt to do the file cleanup.
In particular this avoids deleting the dexopt artifacts for any new
apps installed since the background dexopt started. It may still
incorrectly delete files for new installs during the cleanup run, but
that time window is much shorter (on the order of 5-10 secs, vs 5-20
mins for a whole background dexopt run).

The dexopt and the cleanup are two independent operations, so they
don't need to use the same snapshot.

Test: atest ArtServiceTests
Bug: 378773852
Change-Id: I75c8f5b4934fb86facdd8d31ae18aab63426a346