Simplify ApiClass.alwaysHidden
Previously, it was an `Int` that use a negative value to indicate that
it was hidden at the API level `abs(alwaysHidden)` and positive numbers
to indicate that it was public at the API level `alwaysHidden`.
However, the code only ever cared whether it was negative and did not
use the API level that was encoded into it. So, this change replaces it
with a `Boolean` to simplify follow-up work that will replace the
`Int` with a proper `ApiVersion` class.

The only reason that hidden classes need to be tracked is because
legacy `android.jar` files could contain non-public classes. In future,
Android will use historical signature files that do not have that
problem. At which point `alwaysHidden` can probably be removed.

Bug: 379646963
Test: ./gradlew
Change-Id: I2810c487d5d220476c8e5cead1050a8191c1475a