remove bpf_native_test VTS test.
I'm not really seeing much utility in this test (without bpf working
the device shouldn't even be able to boot, since networking startup
code is heavily dependent on the presense of a bunch of pinned eBPF
maps), and the way it includes headers files from Connectivity module
(ie. tethering apex) and how it relies on root being able to ignore
sepolicy (to create pin/unpin maps, load bpf programs) seems outright
wrong.  It's also not clear how to fix any of this in the new libbpf
world...

(also the test fails on RiscV, since it is not flagged as 64-bit)

Basically we have 3 tests:
  BpfTest:bpfMapPinTest - outright not useful, since pinning is extensively tested by tethering apex startup code
  BpfRaceTest:testRaceWithBarrier - slightly flaky
  BpfRaceTest:DISABLED_testRaceWithoutBarrier - already disabled due to excessive flakiness

So the only semi-useful test in here is 'testRaceWithBarrier',
but even that seems to provide extremely limited value.

Test: TreeHugger
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: I02328c147f3ddeeee68a4510d76f163601e37115