Remove secondary constructors from ApiElement
Previously, `ApiElement` had a default primary constructor and two
secondary constructors. The latter was confusing and made it slightly
harder to reason about an `ApiElement`'s state because some properties
were initialized to a default in their definitions and then immediately
changed in a secondary constructor.

This change removes the secondary constructors and moves the parameters
to the primary constructor. The logic in the secondary constructor is
inlined into the property definitions.

Bug: 379646963
Test: ./gradlew
Change-Id: I92fb5eab9cd0a7c383157af621db6c00d84abc37