Use SdkVersion major/minor/patch in generateJson(...)
Previously, `generateJson(...)` would just use an `SdkVersion`'s
`level` property and then when it needed to access the real version
in `ApiJsonPrinter` would use the `level` property as an index into
`apiVersionNames`.

This change avoids the need to do that by using the `SdkVersion`'s
new ability to store `major`, `minor` and `patch` properties. It adds
a new `VersionedSignatureApi` class to encapsulate the `SdkVersion`
and a signature `File` and creates a list of those by combining the
`pastApiVersions` and `apiVersionNames`.

The `ApiJsonPrinter` then just uses the string representation of the
`SdkVersion` instead of performing the lookup in `apiVersionNames`.

Bug: 379646963
Test: ./gradlew
Change-Id: I0547e5a2b38267bd10b0f2ca1782fb87ce647a5b