Remove log start delay
The original purpose was to deal with potential flake in adb and/or the underlying transport layer, but we are long past that with many improvements done to adb. In places like virtual devices, this has already been overridden to 0.

Not starting log collection right after device is online may lead to losing critical logs for debugging especially if device disappears shortly after.

Bug: 373908707
Change-Id: I1a0f947451290b5cf4208877c6065e7ac0344707
Test: N/A