Add major/minor/patch/preReleaseQuality properties to SdkVersion
Previously, `SdkVersion` only supported a single `level` property.
This change adds `major`, `minor`, `patch` and `preReleaseQuality`
properties to match the structure of versions passed by AndroidX via
the `--api-version-names` option.

It temporarily keeps the `level` property as an alias for `major` to
avoid churn elsewhere but it will be removed once the code that uses it
has been migrated to use these new properties.

Bug: 379646963
Test: ./gradlew
Change-Id: I779374435d443eeed8265218d273399e2533a5c4