Fix the 918-fields test
The test relied on a synthetic field that the newer Java compiler
was not generating since the field was not being used. In order
to cause the synthetic generation of the field, which allows an
inner class to refer to its parent, the inner class has been
changed to have a reference to an outer field.

Bug: 378676805
Test: atest CtsJvmtiRunTest918HostTestCases:android.jvmti.cts.JvmtiHostTest918#testJvmti
Change-Id: I499154372fb4123c6ccd13e479cbeeb2a2ec1203