Autoupdate of 2 submodules
*build/release updated to 4abb1db4ad14d97f9c147074933b208aff0b7439
Add build feature to guard enable speed-profile compilation for Wifi.

Bug: 241823638
Change-Id: I68361239602e9e8fd4635649a28b92afc3ec5648
Test: `m`
Test: with flag on: go/art-benchmark?p=BootImageProfileId:41874 (code size reduction by -10.82MB)
Test: with flag off: go/art-benchmark?p=BootImageProfileId:41898 (no code change or metric changes)

*packages/modules/Wifi updated to 5621b644a2d03a59ac9693e6033ab27a4d906597
Add ART profile to wifi service.

This CL adds an ART profile to the wifi service. This profile will be used to improve the performance of the wifi service. The profile was generated by profiling the wifi service on Droidfood population.

The ART profile is expected to reduce the size of the Wifi service generated code size by 10.82MB! (15.8%)

The new behavior is guarded by a new build feature `RELEASE_SERVICE_WIFI_SPEED_PROFILE_ART_COMPILATION`

Bug: 241823638
Change-Id: I0298cf932bcd39e3b33cbc45df5d8556ba417001
Test: `m`
Test: with flag on: go/art-benchmark?p=BootImageProfileId:41874 (code size reduction by -10.82MB)
Test: with flag off: go/art-benchmark?p=BootImageProfileId:41898 (no code change or metric changes)