Autoupdate of 2 submodules
*hardware/generic/goldfish updated to e0cd27c1f4ca25605440218a350b75781789a63e
Merge "Fix: Prevent nullptr in AsyncSocket callbacks" into emu-dev
*hardware/google/aemu updated to 28aa9416bf71aef1507841202698e5d5a3e62b30
Fix: Prevent nullptr exception in AsyncSocket callbacks

The callbacks in AsyncSocket were being called from the looper
even after the socket was destroyed, leading to a nullptr exception.

This commit fixes the issue by ensuring that callbacks are not
scheduled if the socket is closing.

Bug: 380110981
Change-Id: Ic3077a4e4a8d6f92c7a2f2a06f99c89d3dd465a4