Remove magic ApiElement.NEVER value
Previously, `ApiElement.sinceExtension` used `ApiElement.NEVER` (which
was `ExtVersion.HIGHEST`) as a magic value to indicate that it had not
been set. The value of that was chosen to make it convenient to select
the earliest extension version in which an `ApiElement` first appeared,
if any. However, using a magic value did make it difficult to know
whether a specific `ExtVersion` variable could use that magic value or
not.

This change switches it to be of type `ExtVersion?` and use `null` to
indicate that it had not been set. That did require slightly more
complicated code to select the earliest version as it had to handle
`null` but did make other uses explicit whether they were optional
(of type `ExtVersion?`) or not (of type `ExtVersion`).

Bug: 380412656
Test: ./gradlew
Change-Id: I9758717898e6f18525ef2488dcda8f8c59cc372f