hwbcc: Add unified request struct
This is a first step in refactoring the hwbcc rust client to support
writing a hwbcc rust service. Writing a service is simplified if there
is a single message type for all requests.

BccMsg replaces the previous logic which sent a conditional Message
type based on the request (i.e. only a header or a SignDataMsg that
also included a header). Serialized requests to a hwbcc server
remain unchanged.

Bug: 346283699
Test: build.py --jobs=$(nproc) qemu-generic-arm64-test-debug --test="boot-test:com.android.trusty.rust.hwbcc.test"
Change-Id: Ibedd56c7667684db8ce34fbc819803901efccce1