x86/64: remove Multiboot header and entry point
The x86-64 kernel expects to be started in 64-bit long mode as of commit
7ed9ea8f858a ("arch: x86: Switch to 64-bit entry"), which works when the
kernel is loaded as ELF by crosvm.

However, the 64-bit kernel still contains a Multiboot header, and a
Multiboot-compliant loader would start the kernel in 32-bit protected
mode, which won't work since the 32-to-64 transition code was removed as
part of the previously mentioned commit. This doesn't matter in practice
currently for Trusty, but in order to avoid confusion and simplify the
entry point code, this change removes the Multiboot header and related
code, leaving only the 64-bit entry point at _start.

This means that the guest kernel currently has no way to determine the
memory map or total memory size, but this is no different than the
previous status quo, since crosvm did not provide Multiboot info to the
loaded guest kernel.

Bug: 372755978
Test: Build x86-64 lk.elf and execute it with crosvm
Change-Id: I893ac4aa766f21e8d99baa94f5223254530bed7b