Add ParentApiElement for use in ApiXmlPrinter
The `ApiXmlPrinter` only adds `since`, sdks`, `deprecated` and
`removed` attributes for a class or class member if they differ from
the parent element. It did that by passing down a parent `ApiElement`.
That necessitates `Api` having to extend `ApiElement` but `Api` is
special and behaves slightly differently to an `ApiElement`, e.g. it
does not actually change `since`, or many of the other properties.

However, `ApiXmlPrinter` does not need a whole `ApiElement` only a
subset of its capabilities so this change adds an interface to
represent the minimal information that the parent element has to
provide.

A follow-up change will make `Api` implement `ParentApiElement`
instead of `ApiElement`.

Bug: 379646963
Test: ./gradlew
Change-Id: I79a0144bb6e0b9bcd683c1fba166bbc781cc42ac