verifier: Stronger uninitialized `this` access checks.
Check uninitialized `this` field access early and make the
check more strict than before.

Ideally, two cases should be rejected now and the run-test
064-field-access is updated with the corresponding tests.
First, we no longer allow `iget` on uninitialized `this`.
Second, we do not allow iput on uninitialized `this` to
specify a field by referencing a subclass even if the
field id actually resolves to the class being verified.
However, the first case is actually something that we see
in some existing apps - they get the zero-initialized field
values from the object before the superclass constructor
call. To avoid breaking these apps, we continue to allow
such access if the class is resolved; when it's unresolved
for AOT, we reject the class and it shall be reverified at
runtime. Note that the RI would reject such access.

Note that doing this check early can also result in
reporting bad uninitialized `this` access instead of
another hard failure when both apply. It also suppresses
soft failures that would have been reported previously.

Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Change-Id: Ieb41115e6aaf618bc283f3aebf0739b606caa3b3