Autoupdate of 2 submodules
*build/make updated to 4e8e7eb815aa502350cb43efa73f626158501633
Merge "Salt init_boot and vendor_boot using the build number" into main
*build/soong updated to c117f933b15d15b1afe86dee9dfe37ee5685f9fe
Salt init_boot and vendor_boot using the build number

If you don't provide a salt, avbtool uses random bytes as the salt.
The randomness is bad for build caching and being able to diff
partitions to ensure code changes aren't breaking anything.

It's not exactly clear if the salt is necessary at all or not, but
as an intermediate step, salt the images using the build number instead
of purely random bytes.

I've only done the init_boot and vendor_boot partitions in this cl,
but I plan to expand it to all the other partitions. There are other
diffs besides the avb hash in those partitions right now though so it's
harder to test.

Bug: 380508890
Test: m soong_generated_init_boot_filesystem_test
Change-Id: I87afb2ef254fcdc981c72f107e94da34d1f07fb8