Make --generate-api-version-history format dependent on extension
Previously, `--generate-api-version-history` would always generate a
JSON file. This change makes the format dependent on the file
extension, which must be either `xml` or `json`.

Currently, this is purely for testing as it is easier to write
self-contained tests using signature files as the input instead of
JAR files. It will be used to add tests that test some edge cases that
are not currently handled correctly but which are not visible in the
JSON output, e.g. related to `removed`.

Longer term, when all API version history files are created from
signature files this will be used by Android, AndroidX, etc. to
select the correct format.

Bug: 380412656
Test: ./gradlew
Change-Id: Ia444ed68df93eb15956fe0ad3db8a4c1993c3ea1