Refactor whether to include current Codebase in generateXml(...)
Previously, the version to use for the current `Codebase` (i.e.
`sdkVersion`) was computed after deciding that the current `Codebase`
needed to be included. That meant that it was not available for use
elsewhere. This change determines which version, if any, to use for
the current `Codebase` first and only adds the current `Codebase` if
that is not `null`. That makes the `codebaseSdkVersion` available for
use in a follow-up change.

Bug: 380412656
Test: ./gradlew
Change-Id: I9197c52d52b0b975c057cd25dce9be4ac79858e5