Avoid having to quote $ in expected XML contents
The `api-versions.xml` file can include `$` characters in the `<sdk>`
element's `reference` attribute. That has to be quoted using `${'$'}`
which makes updating the contents and reading the contents a little
tricky. This change avoids that by adding constants such that
`$CONSTANT` is replaced with `$CONSTANT`.

This also corrects the indent of the affected expected XML contents.

Bug: 381187230
Test: ./gradlew
Change-Id: Idc8c7be3add70171fcc48e1a17f0ad742bb640aa