Fix test failure of CallLogManagerTest#testLogCorrectPhoneNumber
In CallLogManagerTest#testLogCorrectPhoneNumber, the expected number
is formatted to "US" convention, while the logged number is formatted
with current country iso. So the logged number doesn't match the
expected if running test when current country iso is not "US", for
example, when the system language is not English(United State).

The change is to format the expected number with current country iso.

Test: atest CallLogManagerTest
Bug: 384837722
Flag: NONE test fix.
Change-Id: Icfced42c646ff850a112d581a6ad08699b1ae73a