Make ExtractSystemApiLevelsTest use system jars
Previously, it would use public jars because it was originally just
checking to make sure that it would work when `--first-version` was set
to something greater than 1 and "did not care about the contents of the
jars". However, that last statement is no longer true as proven by the
fact that switching to system jars broke one of the assertions in the
test due to a class being added to system first and then moved to
public.

This change switches to use system jars to make the test more realistic
and fixes the broken assertion.

Bug: 381187230
Test: ./gradlew
Change-Id: Iec530c28a8bb1bbd0a909e58ef853594a8585bc7