Start reporting again the ArtDeviceDatumReported atom
Since the introduction of the ArtDeviceStatus pulled atom to replace the
ArtDeviceDatumReported pushed atom to report information about the Boot
Image status we have been seeing different distributions of data that
should have been more similar. This CL re-introduces the code reporting
the ArtDeviceDatumReported pushed atom to be able to compare the data
from the field and check if we see the same divergence in data.

Bug: 257028435
Test: statsd_testdrive 10205 550
Change-Id: Ia645819f05f40cbc253adb97ba04a14be17a736d