Remove comments on some fields in the ASN.1 schema.
The AIDL interface spec is the source of truth. We should avoid
documenting individual fields in two places since it's easy for
them to go out of sync (e.g. the comments on the version fields
were updated in the AIDL interface to reflect the new KeyMint
version (400) but the comments here weren't updated). Arguably
even the un-commented schemas in this file could just be deleted
but they're still helpful for cross-referencing with the actual
implementation (e.g. to see the types and encoding expectations).

Bug: 381383980
Test: n/a, comment update
Change-Id: I6a1b79bcb009d8edab33c7ab6113687a24f52e93