Use maps to store super classes and interfaces in ApiClass
Previously, `ApiClass` would store super classes and interfaces in a
list and then search that list by name. That required special code
to find and update the list. This switches them to use a mutable map
and use the standard map lookup and the `addToMap(...)` method.

This did originally replace `ConcurrentHashMap` with `mutableMap` but
that changed the order of the output for `ApiJsonPrinter` which does
not sort the elements. That will be fixed in a follow-up change.

Bug: 379646963
Test: ./gradlew
Change-Id: I509ddf5c12ca24f2a6e1b75c2abcd602dd02715a