Autoupdate of 2 submodules
*external/cronet updated to 1f2ef2ab439366817e3e18f21c413955a9eaf8c1
Add preload API method to httpEngine

This new system API is meant to be only called from
the zygote to preload HttpEngine code into the zygote. The
current implementation preloads a single ICronetEngineBuilder class
which is used to create an underlying CronetEngine that acts as
a backend to HttpEngine.

Added a CTS test to ensure that calling the new API does not crash. It
is generally hard to create a test that ensures that the classes are
preloaded as that would require creating a custom classloader and
running the test in its own isolated environment.

Change-Id: Icb8415feb15569e199ce0cb3d8879700596e7431

*packages/modules/Connectivity updated to c3a2764012603f8124713d1562436afa83d7bd05
Merge "Add a new module-only API to preload HttpEngine" into main