fdevent: remove obsolete personal copyright line
The original fdevent library was a tiny personal project of mine that I used when abd was originally written.  It's long-since been completely rewritten and leaving my personal copyright notice on this code (which is no longer even my code) is just confusing for people.  Let's fix it.

Flag: EXEMPT comment only change
Test: N/A comment only change
Change-Id: Iaff88dcfe2fbaed550fadc7ac27bea8fd7ebeafe