Remove prebuilts of Perfetto client libraries.
Build them from source instead on master-art (https://r.android.com/3313791).

libperfetto_client_experimental and perfetto_trace_protos are static
libraries which implements serialised protos in a shared memory area to
send trace data to traced.

- The communication with traced uses RPCs over a unix socket through a
  published ABI to set up the shared memory and control the tracing:
  https://perfetto.dev/docs/design-docs/api-and-abi#tracing-protocol-abi

- The serialised protos aren't interpreted by the system daemons -
  they're just forwarded to off-device Perfetto tools. Hence the
  serialisation code isn't a platform (or Mainline module) API, which
  means we don't need prebuilts to verify API compatibility (on
  master-art).

heapprofd_client_api is dropped as well. That's a stubbed library that
we'd ideally keep the prebuilt for (to catch e.g. header file
incompatibilities). However, given that the sources for all three
libraries are in the same repository that are getting added as source,
there's no feasible way to keep using the prebuilt for
heapprofd_client_api while switching the other two to sources.

Test: m (will test more with a new prebuilt drop in a follow-up CL)
Bug: 381897796
Change-Id: I1097845e92f5588c5d92937a3342c44e15359cfa