Autoupdate of 2 submodules
*packages/apps/Nfc updated to c96118c0116736b23f9dcfffd60254dae1d4e63d
Merge "Added call to NFA_SetNfcSecure()" into main
*system/nfc updated to 14fd8aaa87d5ed01b8230843588e8aa5133af229
Added NFA_SetNfcSecure() API

This is needed when setNfcSecure() is called:
only power state screen_on_unlocked is allowed in this case, but in
addition all other power states should be blocked. If not, the NFCC
will have no matching entries for the non supported power states, and
it will be up to NFCC to decide how to route. It might route the CE
to an host that will answer.

Bug: 378474016
Test: manual
Change-Id: Ibd4226af9625fa662908e16fa8e54187d9e8e9f8