Stop passing firstApiLevel into ApiXmlPrinter
Previously, `firstApiLevel` was passed separately to `allVersions` in
the `ApiXmlPrinter` constructor. However, the `firstApiLevel` should
simply be the first version in `allVersions`. Passing them in
separately can lead to inconsistencies where the `firstApiLevel` was
not actually the first version.

This change removes `firstApiLevel` and uses the first item in
`allVersions`. That did reveal a few inconsistencies in the tests which
were fixed as part of this change.

This does not affect AndroidX as it does not use `ApiXmlPrinter` and it
does not affect Android as the `firstApiLevel` was always used as the
first version in `allVersions` because the latter is created from the
former.

Bug: 378479241
Test: ./gradlew
Change-Id: I6009e74b2574ba08383f59c1a9fc09a7ae38c7a6