Cleanup some nastiness over the handling of lastApiVersion
Previously, `lastApiVersion` was set to the
following `ApiVersion.fromLevel(apiLevels.size - 1)`. That would fail
if `apiLevels` was empty because `ApiVersion.fromLevel(-1)` would fail
as the method makes sure that the level is greater than 0. However, it
worked, even when no historical versions were provided because:
1. It was only a few tests (specifically `FlaggedApiTest`s) that did
   not provide the historical versions.
2. The tests specified `--first-version 30` and `apiLevels` is padded
   from 0 to `firstApiLevel` (exclusive) that meant `apiLevels.size`
   was 30 not 0.
3. That ended up with `lastApiVersion` basically equal to 29.
4. The only place `lastApiVersion` was used was in the comparison
   `lastApiVersion < currentSdkVersion` and 29 was always less.

This change switches it to use `null` when no historical versions were
provided and updates the condition to check for `null` too. This makes
the follow-up change simpler.

Bug: 378479241
Test: ./gradlew
Change-Id: Idedf0a7fc86cdde514343062b7143fe495932757