Pass CodebaseFragment provider when create GenerateXmlConfig
Previously, `GenerateXmlConfig` was created in a custom getter of the
`generateXmlConfig` property. This change replaces that with a function
that takes a single `codebaseFragmentProvider` parameter. It uses that
to create a `List<VersionedApi>` containing an optional
`VersionedSourceApi` to encapsulate the `CodebaseFragment`. That is
stored in a new `GenerateXmlConfig.versionedApis` property and merged
into the `versionedApis` list created in `generateXml(...)`.

Allowed the removal of the `GeneralXmlConfig.codebaseSdkVersion`
property.

This is preparation for moving the whole `versionedApis` list creation
from `generateXml(...)` to `ApiLevelsGenerationOptions`.

Bug: 378479241
Test: ./gradlew
Change-Id: Ie8394255e8ece728f4e8ec60e356f9c93331311f