ui: Allow typed omnibox prompt choices
This CL improves the omnibox interface so that prompt choices can be
either a simple list of strings, or a list of objects with a getName()
mapping function to obtain the string to be shown in the dropdown list
for each item.

The latter allows non-string choices to be used in a type safe way,
without the caller having to do the mapping manually.

Change-Id: I615990733cf928051ebf9c0654712a7c07d5db62