arch: arm64: Make sure that SP1 is set on correct spsel
This change adds a single msr instruction that spsel to 1 before
modifying stack pointer. Prior to this change, start.S assumed that spsel is
set to 1. This may not be true in case that lk is not started directly
(eg. with pvmfw). This patch fixes undefined behaviour by eliminating
this assumption.

Bug: 375584439
Test: start trusty in a pVM with pvmfw
Change-Id: I13dc5b73b9860bac06c78649e11fe77c6638ebcd
Signed-off-by: Bartłomiej Grzesik <bgrzesik@google.com>