Add test for changing the severity of a category of issues
While working on b/379918901 `--hide-category compatibility` was passed
to Metalava and it did not work. Investigating why it was not working
revealed that there was no test for it so this change adds some tests
for a valid category (starts with an upper case letter) and an invalid
one.

A follow-up change will improve the error message to include the valid
categories.

Bug: 379918901
Test: ./gradlew
Change-Id: I93fdb9612000b2b92473111bcca07bca833bdc65