standalone: add additional warnings to Mac Clang build
Compiler warnings thrown by -Wall, -Wextra, and -Wpedantic can
differ significantly based on the compiler version. This is
particularly important to note since Apple Clang tends to lag
the official Clang by a few major versions.

This can cause issues Where CLs may experience Perfetto CI failures
since Clang 19.0 is used by the Perfetto CI, but Clang 16.0 is used
as the base for the latest version of Apple Clang that apple ships
with xcode.

To make the builds with Apple Clang behave more closely to those in
CI, throw extra warnings for the local Mac builds that are already
being thrown by the Perfetto CI. The intent of this change is to
minimize the number of CLs submitted from Mac development machines
that fail to pass the Perfetto CI once uploaded.

Change-Id: Iddbb4743040a1bc06d5c2710bc9e3fa5804ed77c