Replace generateFromVersionedApis(...) with generateApiHistory(...)
Previously, there were two methods to generate API history files, one
for generating XML (used by Android) and one for generating JSON (used
by AndroidX). This change is the culmination of a series of changes to
merge the functionality together.

The `generateApiHistory(...)` provides a super set of the functionality
provided by `generateFromVersionedApis(...)` so this change replaces
the call to the latter with a call to the former and removes the latter
along with the associate config class.

It is safe for the call to be replaced as the extra functionality of
the `generateApiHistory(...)` method has no effect as the caller either
configures it to have no effect (in the case of the SDK extensions and
missing class action) or (in the case of patching `SdkExtensions`
history) only has an effect when the `Api` contains the
`android.os.ext.SdkExtensions` class.

Bug: 378479241
Test: ./gradlew
Change-Id: I817ac98ccdc69343a55bb8b011a05dd3f05c4522