Autoupdate of 3 submodules
*external/avb updated to ee959a0ab688848c29e388d9de6c5338487246bd
Keep using old version of zerocopy for now.

This is part of a staged update of zerocopy.

Bug: 372549215
Test: Treehugger
Change-Id: Id0101962cb772a19f9fca93e824333cfe935a811
Merged-In: Id0101962cb772a19f9fca93e824333cfe935a811

*system/libfmq updated to 9eb2d04a2fe139786ff2288420bde6a9b3e05c81
Keep using old version of zerocopy for now.

This is part of a staged update of zerocopy.

Bug: 372549215
Test: Treehugger
Change-Id: I93dc16a5095abb9473b29dd6b68319586de885aa
Merged-In: I93dc16a5095abb9473b29dd6b68319586de885aa

*system/tools/mkbootimg updated to 9c1ea3aa70b261f6b77c6b52d62d80aa8feab2a4
Keep using old version of zerocopy for now.

This is part of a staged update of zerocopy.

Bug: 372549215
Test: Treehugger
Change-Id: I0d90f5752ef69fb0d094260c0486670d5c671035
Merged-In: I0d90f5752ef69fb0d094260c0486670d5c671035